% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gApplicationRunSimple}
\alias{gApplicationRunSimple}
\title{Run a GApplication (simple wrapper without command-line args)}
\usage{
gApplicationRunSimple(app)
}
\arguments{
\item{app}{GApplication pointer}
}
\value{
exit code (integer)
}
\description{
Run a GApplication (simple wrapper without command-line args)
}
