% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gdk_autogen.R
\name{gdkDisplayGetDefault}
\alias{gdkDisplayGetDefault}
\alias{gdkDisplayOpen}
\alias{gdkDisplayBeep}
\alias{gdkDisplayClose}
\alias{gdkDisplayCreateGlContext}
\alias{gdkDisplayDeviceIsGrabbed}
\alias{gdkDisplayFlush}
\alias{gdkDisplayGetAppLaunchContext}
\alias{gdkDisplayGetClipboard}
\alias{gdkDisplayGetDefaultSeat}
\alias{gdkDisplayGetDmabufFormats}
\alias{gdkDisplayGetMonitorAtSurface}
\alias{gdkDisplayGetMonitors}
\alias{gdkDisplayGetName}
\alias{gdkDisplayGetPrimaryClipboard}
\alias{gdkDisplayGetSetting}
\alias{gdkDisplayGetStartupNotificationId}
\alias{gdkDisplayIsClosed}
\alias{gdkDisplayIsComposited}
\alias{gdkDisplayIsRgba}
\alias{gdkDisplayListSeats}
\alias{gdkDisplayMapKeycode}
\alias{gdkDisplayMapKeyval}
\alias{gdkDisplayNotifyStartupComplete}
\alias{gdkDisplayPrepareGl}
\alias{gdkDisplayPutEvent}
\alias{gdkDisplaySupportsInputShapes}
\alias{gdkDisplaySupportsShadowWidth}
\alias{gdkDisplaySync}
\alias{gdkDisplayTranslateKey}
\alias{gdkDisplayManagerGet}
\alias{gdkDisplayManagerGetDefaultDisplay}
\alias{gdkDisplayManagerListDisplays}
\alias{gdkDisplayManagerOpenDisplay}
\alias{gdkDisplayManagerSetDefaultDisplay}
\alias{gdkMonitorGetConnector}
\alias{gdkMonitorGetDescription}
\alias{gdkMonitorGetDisplay}
\alias{gdkMonitorGetGeometry}
\alias{gdkMonitorGetHeightMm}
\alias{gdkMonitorGetManufacturer}
\alias{gdkMonitorGetModel}
\alias{gdkMonitorGetRefreshRate}
\alias{gdkMonitorGetScale}
\alias{gdkMonitorGetScaleFactor}
\alias{gdkMonitorGetSubpixelLayout}
\alias{gdkMonitorGetWidthMm}
\alias{gdkMonitorIsValid}
\alias{gdkSurfaceNewPopup}
\alias{gdkSurfaceNewToplevel}
\alias{gdkSurfaceBeep}
\alias{gdkSurfaceCreateCairoContext}
\alias{gdkSurfaceCreateGlContext}
\alias{gdkSurfaceCreateSimilarSurface}
\alias{gdkSurfaceCreateVulkanContext}
\alias{gdkSurfaceDestroy}
\alias{gdkSurfaceGetCursor}
\alias{gdkSurfaceGetDeviceCursor}
\alias{gdkSurfaceGetDevicePosition}
\alias{gdkSurfaceGetDisplay}
\alias{gdkSurfaceGetFrameClock}
\alias{gdkSurfaceGetHeight}
\alias{gdkSurfaceGetMapped}
\alias{gdkSurfaceGetScale}
\alias{gdkSurfaceGetScaleFactor}
\alias{gdkSurfaceGetWidth}
\alias{gdkSurfaceHide}
\alias{gdkSurfaceIsDestroyed}
\alias{gdkSurfaceQueueRender}
\alias{gdkSurfaceRequestLayout}
\alias{gdkSurfaceSetCursor}
\alias{gdkSurfaceSetDeviceCursor}
\alias{gdkSurfaceSetInputRegion}
\alias{gdkSurfaceSetOpaqueRegion}
\alias{gdkSurfaceTranslateCoordinates}
\title{gdkDisplayGetDefault}
\usage{
gdkDisplayGetDefault()

gdkDisplayOpen(display_name)

gdkDisplayBeep(display)

gdkDisplayClose(display)

gdkDisplayCreateGlContext(self)

gdkDisplayDeviceIsGrabbed(display, device)

gdkDisplayFlush(display)

gdkDisplayGetAppLaunchContext(display)

gdkDisplayGetClipboard(display)

gdkDisplayGetDefaultSeat(display)

gdkDisplayGetDmabufFormats(display)

gdkDisplayGetMonitorAtSurface(display, surface)

gdkDisplayGetMonitors(self)

gdkDisplayGetName(display)

gdkDisplayGetPrimaryClipboard(display)

gdkDisplayGetSetting(display, name, value)

gdkDisplayGetStartupNotificationId(display)

gdkDisplayIsClosed(display)

gdkDisplayIsComposited(display)

gdkDisplayIsRgba(display)

gdkDisplayListSeats(display)

gdkDisplayMapKeycode(display, keycode)

gdkDisplayMapKeyval(display, keyval)

gdkDisplayNotifyStartupComplete(display, startup_id)

gdkDisplayPrepareGl(self)

gdkDisplayPutEvent(display, event)

gdkDisplaySupportsInputShapes(display)

gdkDisplaySupportsShadowWidth(display)

gdkDisplaySync(display)

gdkDisplayTranslateKey(display, keycode, state, group)

gdkDisplayManagerGet()

gdkDisplayManagerGetDefaultDisplay(manager)

gdkDisplayManagerListDisplays(manager)

gdkDisplayManagerOpenDisplay(manager, name)

gdkDisplayManagerSetDefaultDisplay(manager, display)

gdkMonitorGetConnector(monitor)

gdkMonitorGetDescription(monitor)

gdkMonitorGetDisplay(monitor)

gdkMonitorGetGeometry(monitor)

gdkMonitorGetHeightMm(monitor)

gdkMonitorGetManufacturer(monitor)

gdkMonitorGetModel(monitor)

gdkMonitorGetRefreshRate(monitor)

gdkMonitorGetScale(monitor)

gdkMonitorGetScaleFactor(monitor)

gdkMonitorGetSubpixelLayout(monitor)

gdkMonitorGetWidthMm(monitor)

gdkMonitorIsValid(monitor)

gdkSurfaceNewPopup(parent, autohide)

gdkSurfaceNewToplevel(display)

gdkSurfaceBeep(surface)

gdkSurfaceCreateCairoContext(surface)

gdkSurfaceCreateGlContext(surface)

gdkSurfaceCreateSimilarSurface(surface, content, width, height)

gdkSurfaceCreateVulkanContext(surface)

gdkSurfaceDestroy(surface)

gdkSurfaceGetCursor(surface)

gdkSurfaceGetDeviceCursor(surface, device)

gdkSurfaceGetDevicePosition(surface, device)

gdkSurfaceGetDisplay(surface)

gdkSurfaceGetFrameClock(surface)

gdkSurfaceGetHeight(surface)

gdkSurfaceGetMapped(surface)

gdkSurfaceGetScale(surface)

gdkSurfaceGetScaleFactor(surface)

gdkSurfaceGetWidth(surface)

gdkSurfaceHide(surface)

gdkSurfaceIsDestroyed(surface)

gdkSurfaceQueueRender(surface)

gdkSurfaceRequestLayout(surface)

gdkSurfaceSetCursor(surface, cursor)

gdkSurfaceSetDeviceCursor(surface, device, cursor)

gdkSurfaceSetInputRegion(surface, region)

gdkSurfaceSetOpaqueRegion(surface, region)

gdkSurfaceTranslateCoordinates(from, to)
}
\arguments{
\item{display_name}{utf8}

\item{display}{Display}

\item{self}{Display}

\item{device}{Device}

\item{surface}{Surface}

\item{name}{utf8}

\item{value}{GObject.Value}

\item{keycode}{guint}

\item{keyval}{guint}

\item{startup_id}{utf8}

\item{event}{Event}

\item{state}{ModifierType}

\item{group}{gint}

\item{manager}{DisplayManager}

\item{monitor}{Monitor}

\item{parent}{Surface}

\item{autohide}{gboolean}

\item{content}{cairo.Content}

\item{width}{gint}

\item{height}{gint}

\item{cursor}{Cursor}

\item{region}{cairo.Region}

\item{from}{Surface}

\item{to}{Surface}
}
\value{
Display

Display

Return value from C function

Return value from C function

GLContext

gboolean

Return value from C function

AppLaunchContext

Clipboard

Seat

DmabufFormats

Monitor

Gio.ListModel

utf8

Clipboard

gboolean

utf8

gboolean

gboolean

gboolean

GLib.List

gboolean

gboolean

Return value from C function

gboolean

Return value from C function

gboolean

gboolean

Return value from C function

gboolean

DisplayManager

Display

GLib.SList

Display

Return value from C function

utf8

utf8

Display

Return value from C function

gint

utf8

utf8

gint

gdouble

gint

SubpixelLayout

gint

gboolean

Surface

Surface

Return value from C function

CairoContext

GLContext

cairo.Surface

VulkanContext

Return value from C function

Cursor

Cursor

gboolean

Display

FrameClock

gint

gboolean

gdouble

gint

gint

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean
}
\description{
gdkDisplayGetDefault

gdkDisplayOpen

gdkDisplayBeep

gdkDisplayClose

gdkDisplayCreateGlContext

gdkDisplayDeviceIsGrabbed

gdkDisplayFlush

gdkDisplayGetAppLaunchContext

gdkDisplayGetClipboard

gdkDisplayGetDefaultSeat

gdkDisplayGetDmabufFormats

gdkDisplayGetMonitorAtSurface

gdkDisplayGetMonitors

gdkDisplayGetName

gdkDisplayGetPrimaryClipboard

gdkDisplayGetSetting

gdkDisplayGetStartupNotificationId

gdkDisplayIsClosed

gdkDisplayIsComposited

gdkDisplayIsRgba

gdkDisplayListSeats

gdkDisplayMapKeycode

gdkDisplayMapKeyval

gdkDisplayNotifyStartupComplete

gdkDisplayPrepareGl

gdkDisplayPutEvent

gdkDisplaySupportsInputShapes

gdkDisplaySupportsShadowWidth

gdkDisplaySync

gdkDisplayTranslateKey

gdkDisplayManagerGet

gdkDisplayManagerGetDefaultDisplay

gdkDisplayManagerListDisplays

gdkDisplayManagerOpenDisplay

gdkDisplayManagerSetDefaultDisplay

gdkMonitorGetConnector

gdkMonitorGetDescription

gdkMonitorGetDisplay

gdkMonitorGetGeometry

gdkMonitorGetHeightMm

gdkMonitorGetManufacturer

gdkMonitorGetModel

gdkMonitorGetRefreshRate

gdkMonitorGetScale

gdkMonitorGetScaleFactor

gdkMonitorGetSubpixelLayout

gdkMonitorGetWidthMm

gdkMonitorIsValid

gdkSurfaceNewPopup

gdkSurfaceNewToplevel

gdkSurfaceBeep

gdkSurfaceCreateCairoContext

gdkSurfaceCreateGlContext

gdkSurfaceCreateSimilarSurface

gdkSurfaceCreateVulkanContext

gdkSurfaceDestroy

gdkSurfaceGetCursor

gdkSurfaceGetDeviceCursor

gdkSurfaceGetDevicePosition

gdkSurfaceGetDisplay

gdkSurfaceGetFrameClock

gdkSurfaceGetHeight

gdkSurfaceGetMapped

gdkSurfaceGetScale

gdkSurfaceGetScaleFactor

gdkSurfaceGetWidth

gdkSurfaceHide

gdkSurfaceIsDestroyed

gdkSurfaceQueueRender

gdkSurfaceRequestLayout

gdkSurfaceSetCursor

gdkSurfaceSetDeviceCursor

gdkSurfaceSetInputRegion

gdkSurfaceSetOpaqueRegion

gdkSurfaceTranslateCoordinates
}
