% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gdk_autogen.R
\name{gdkDeviceGetCapsLockState}
\alias{gdkDeviceGetCapsLockState}
\alias{gdkDeviceGetDeviceTool}
\alias{gdkDeviceGetDirection}
\alias{gdkDeviceGetDisplay}
\alias{gdkDeviceGetHasCursor}
\alias{gdkDeviceGetModifierState}
\alias{gdkDeviceGetName}
\alias{gdkDeviceGetNumLockState}
\alias{gdkDeviceGetNumTouches}
\alias{gdkDeviceGetProductId}
\alias{gdkDeviceGetScrollLockState}
\alias{gdkDeviceGetSeat}
\alias{gdkDeviceGetSource}
\alias{gdkDeviceGetSurfaceAtPosition}
\alias{gdkDeviceGetTimestamp}
\alias{gdkDeviceGetVendorId}
\alias{gdkDeviceHasBidiLayouts}
\alias{gdkDevicePadGetFeatureGroup}
\alias{gdkDevicePadGetGroupNModes}
\alias{gdkDevicePadGetNFeatures}
\alias{gdkDevicePadGetNGroups}
\alias{gdkDeviceToolGetAxes}
\alias{gdkDeviceToolGetHardwareId}
\alias{gdkDeviceToolGetSerial}
\alias{gdkDeviceToolGetToolType}
\alias{gdkEventGetAxes}
\alias{gdkEventGetAxis}
\alias{gdkEventGetDevice}
\alias{gdkEventGetDeviceTool}
\alias{gdkEventGetDisplay}
\alias{gdkEventGetEventSequence}
\alias{gdkEventGetEventType}
\alias{gdkEventGetHistory}
\alias{gdkEventGetModifierState}
\alias{gdkEventGetPointerEmulated}
\alias{gdkEventGetPosition}
\alias{gdkEventGetSeat}
\alias{gdkEventGetSurface}
\alias{gdkEventGetTime}
\alias{gdkEventRef}
\alias{gdkEventTriggersContextMenu}
\alias{gdkEventUnref}
\title{gdkDeviceGetCapsLockState}
\usage{
gdkDeviceGetCapsLockState(device)

gdkDeviceGetDeviceTool(device)

gdkDeviceGetDirection(device)

gdkDeviceGetDisplay(device)

gdkDeviceGetHasCursor(device)

gdkDeviceGetModifierState(device)

gdkDeviceGetName(device)

gdkDeviceGetNumLockState(device)

gdkDeviceGetNumTouches(device)

gdkDeviceGetProductId(device)

gdkDeviceGetScrollLockState(device)

gdkDeviceGetSeat(device)

gdkDeviceGetSource(device)

gdkDeviceGetSurfaceAtPosition(device)

gdkDeviceGetTimestamp(device)

gdkDeviceGetVendorId(device)

gdkDeviceHasBidiLayouts(device)

gdkDevicePadGetFeatureGroup(pad, feature, feature_idx)

gdkDevicePadGetGroupNModes(pad, group_idx)

gdkDevicePadGetNFeatures(pad, feature)

gdkDevicePadGetNGroups(pad)

gdkDeviceToolGetAxes(tool)

gdkDeviceToolGetHardwareId(tool)

gdkDeviceToolGetSerial(tool)

gdkDeviceToolGetToolType(tool)

gdkEventGetAxes(event)

gdkEventGetAxis(event, axis_use)

gdkEventGetDevice(event)

gdkEventGetDeviceTool(event)

gdkEventGetDisplay(event)

gdkEventGetEventSequence(event)

gdkEventGetEventType(event)

gdkEventGetHistory(event)

gdkEventGetModifierState(event)

gdkEventGetPointerEmulated(event)

gdkEventGetPosition(event)

gdkEventGetSeat(event)

gdkEventGetSurface(event)

gdkEventGetTime(event)

gdkEventRef(event)

gdkEventTriggersContextMenu(event)

gdkEventUnref(event)
}
\arguments{
\item{device}{Device}

\item{pad}{DevicePad}

\item{feature}{DevicePadFeature}

\item{feature_idx}{gint}

\item{group_idx}{gint}

\item{tool}{DeviceTool}

\item{event}{Event}

\item{axis_use}{AxisUse}
}
\value{
gboolean

DeviceTool

Pango.Direction

Display

gboolean

ModifierType

utf8

gboolean

guint

utf8

gboolean

Seat

InputSource

Surface

guint32

utf8

gboolean

gint

gint

gint

gint

AxisFlags

guint64

guint64

DeviceToolType

gboolean

gboolean

Device

DeviceTool

Display

EventSequence

EventType

TimeCoord

ModifierType

gboolean

gboolean

Seat

Surface

guint32

Event

gboolean

Return value from C function
}
\description{
gdkDeviceGetCapsLockState

gdkDeviceGetDeviceTool

gdkDeviceGetDirection

gdkDeviceGetDisplay

gdkDeviceGetHasCursor

gdkDeviceGetModifierState

gdkDeviceGetName

gdkDeviceGetNumLockState

gdkDeviceGetNumTouches

gdkDeviceGetProductId

gdkDeviceGetScrollLockState

gdkDeviceGetSeat

gdkDeviceGetSource

gdkDeviceGetSurfaceAtPosition

gdkDeviceGetTimestamp

gdkDeviceGetVendorId

gdkDeviceHasBidiLayouts

gdkDevicePadGetFeatureGroup

gdkDevicePadGetGroupNModes

gdkDevicePadGetNFeatures

gdkDevicePadGetNGroups

gdkDeviceToolGetAxes

gdkDeviceToolGetHardwareId

gdkDeviceToolGetSerial

gdkDeviceToolGetToolType

gdkEventGetAxes

gdkEventGetAxis

gdkEventGetDevice

gdkEventGetDeviceTool

gdkEventGetDisplay

gdkEventGetEventSequence

gdkEventGetEventType

gdkEventGetHistory

gdkEventGetModifierState

gdkEventGetPointerEmulated

gdkEventGetPosition

gdkEventGetSeat

gdkEventGetSurface

gdkEventGetTime

gdkEventRef

gdkEventTriggersContextMenu

gdkEventUnref
}
