% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gdk_autogen.R
\name{gdkTextureNewForPixbuf}
\alias{gdkTextureNewForPixbuf}
\alias{gdkTextureNewFromBytes}
\alias{gdkTextureNewFromFile}
\alias{gdkTextureNewFromFilename}
\alias{gdkTextureNewFromResource}
\alias{gdkTextureDownload}
\alias{gdkTextureGetColorState}
\alias{gdkTextureGetFormat}
\alias{gdkTextureGetHeight}
\alias{gdkTextureGetWidth}
\alias{gdkTextureSaveToPng}
\alias{gdkTextureSaveToPngBytes}
\alias{gdkTextureSaveToTiff}
\alias{gdkTextureSaveToTiffBytes}
\alias{gdkTextureDownloaderNew}
\alias{gdkTextureDownloaderCopy}
\alias{gdkTextureDownloaderDownloadBytes}
\alias{gdkTextureDownloaderDownloadBytesWithPlanes}
\alias{gdkTextureDownloaderDownloadInto}
\alias{gdkTextureDownloaderFree}
\alias{gdkTextureDownloaderGetColorState}
\alias{gdkTextureDownloaderGetFormat}
\alias{gdkTextureDownloaderGetTexture}
\alias{gdkTextureDownloaderSetColorState}
\alias{gdkTextureDownloaderSetFormat}
\alias{gdkTextureDownloaderSetTexture}
\alias{gdkTextureErrorQuark}
\alias{gdkPixbufGetFromSurface}
\alias{gdkPixbufGetFromTexture}
\title{gdkTextureNewForPixbuf}
\usage{
gdkTextureNewForPixbuf(pixbuf)

gdkTextureNewFromBytes(bytes)

gdkTextureNewFromFile(file)

gdkTextureNewFromFilename(path)

gdkTextureNewFromResource(resource_path)

gdkTextureDownload(texture, data, stride)

gdkTextureGetColorState(self)

gdkTextureGetFormat(self)

gdkTextureGetHeight(texture)

gdkTextureGetWidth(texture)

gdkTextureSaveToPng(texture, filename)

gdkTextureSaveToPngBytes(texture)

gdkTextureSaveToTiff(texture, filename)

gdkTextureSaveToTiffBytes(texture)

gdkTextureDownloaderNew(texture)

gdkTextureDownloaderCopy(self)

gdkTextureDownloaderDownloadBytes(self)

gdkTextureDownloaderDownloadBytesWithPlanes(self)

gdkTextureDownloaderDownloadInto(self, data, stride)

gdkTextureDownloaderFree(self)

gdkTextureDownloaderGetColorState(self)

gdkTextureDownloaderGetFormat(self)

gdkTextureDownloaderGetTexture(self)

gdkTextureDownloaderSetColorState(self, color_state)

gdkTextureDownloaderSetFormat(self, format)

gdkTextureDownloaderSetTexture(self, texture)

gdkTextureErrorQuark()

gdkPixbufGetFromSurface(surface, src_x, src_y, width, height)

gdkPixbufGetFromTexture(texture)
}
\arguments{
\item{pixbuf}{GdkPixbuf.Pixbuf}

\item{bytes}{GLib.Bytes}

\item{file}{Gio.File}

\item{path}{filename}

\item{resource_path}{utf8}

\item{texture}{Texture}

\item{data}{guint8}

\item{stride}{gsize}

\item{self}{TextureDownloader}

\item{filename}{filename}

\item{color_state}{ColorState}

\item{format}{MemoryFormat}

\item{surface}{cairo.Surface}

\item{src_x}{gint}

\item{src_y}{gint}

\item{width}{gint}

\item{height}{gint}
}
\value{
Texture

Texture

Texture

Texture

Texture

Return value from C function

ColorState

MemoryFormat

gint

gint

gboolean

GLib.Bytes

gboolean

GLib.Bytes

TextureDownloader

TextureDownloader

GLib.Bytes

GLib.Bytes

Return value from C function

Return value from C function

ColorState

MemoryFormat

Texture

Return value from C function

Return value from C function

Return value from C function

GLib.Quark

GdkPixbuf.Pixbuf

GdkPixbuf.Pixbuf
}
\description{
gdkTextureNewForPixbuf

gdkTextureNewFromBytes

gdkTextureNewFromFile

gdkTextureNewFromFilename

gdkTextureNewFromResource

gdkTextureDownload

gdkTextureGetColorState

gdkTextureGetFormat

gdkTextureGetHeight

gdkTextureGetWidth

gdkTextureSaveToPng

gdkTextureSaveToPngBytes

gdkTextureSaveToTiff

gdkTextureSaveToTiffBytes

gdkTextureDownloaderNew

gdkTextureDownloaderCopy

gdkTextureDownloaderDownloadBytes

gdkTextureDownloaderDownloadBytesWithPlanes

gdkTextureDownloaderDownloadInto

gdkTextureDownloaderFree

gdkTextureDownloaderGetColorState

gdkTextureDownloaderGetFormat

gdkTextureDownloaderGetTexture

gdkTextureDownloaderSetColorState

gdkTextureDownloaderSetFormat

gdkTextureDownloaderSetTexture

gdkTextureErrorQuark

gdkPixbufGetFromSurface

gdkPixbufGetFromTexture
}
