% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gdk_autogen.R
\name{gdkAppLaunchContextGetDisplay}
\alias{gdkAppLaunchContextGetDisplay}
\alias{gdkAppLaunchContextSetDesktop}
\alias{gdkAppLaunchContextSetIcon}
\alias{gdkAppLaunchContextSetIconName}
\alias{gdkAppLaunchContextSetTimestamp}
\alias{gdkButtonEventGetButton}
\alias{gdkCairoContextCairoCreate}
\alias{gdkClipboardGetContent}
\alias{gdkClipboardGetDisplay}
\alias{gdkClipboardGetFormats}
\alias{gdkClipboardIsLocal}
\alias{gdkClipboardReadFinish}
\alias{gdkClipboardReadTextFinish}
\alias{gdkClipboardReadTextureFinish}
\alias{gdkClipboardReadValueFinish}
\alias{gdkClipboardSetContent}
\alias{gdkClipboardSetValue}
\alias{gdkClipboardStoreFinish}
\alias{gdkContentDeserializerGetCancellable}
\alias{gdkContentDeserializerGetGtype}
\alias{gdkContentDeserializerGetInputStream}
\alias{gdkContentDeserializerGetMimeType}
\alias{gdkContentDeserializerGetPriority}
\alias{gdkContentDeserializerGetTaskData}
\alias{gdkContentDeserializerGetUserData}
\alias{gdkContentDeserializerGetValue}
\alias{gdkContentDeserializerReturnError}
\alias{gdkContentDeserializerReturnSuccess}
\alias{gdkContentFormatsNew}
\alias{gdkContentFormatsNewForGtype}
\alias{gdkContentFormatsContainGtype}
\alias{gdkContentFormatsContainMimeType}
\alias{gdkContentFormatsGetGtypes}
\alias{gdkContentFormatsGetMimeTypes}
\alias{gdkContentFormatsMatch}
\alias{gdkContentFormatsMatchGtype}
\alias{gdkContentFormatsMatchMimeType}
\alias{gdkContentFormatsPrint}
\alias{gdkContentFormatsRef}
\alias{gdkContentFormatsToString}
\alias{gdkContentFormatsUnion}
\alias{gdkContentFormatsUnionDeserializeGtypes}
\alias{gdkContentFormatsUnionDeserializeMimeTypes}
\alias{gdkContentFormatsUnionSerializeGtypes}
\alias{gdkContentFormatsUnionSerializeMimeTypes}
\alias{gdkContentFormatsUnref}
\alias{gdkContentFormatsParse}
\alias{gdkContentFormatsBuilderNew}
\alias{gdkContentFormatsBuilderAddFormats}
\alias{gdkContentFormatsBuilderAddGtype}
\alias{gdkContentFormatsBuilderAddMimeType}
\alias{gdkContentFormatsBuilderRef}
\alias{gdkContentFormatsBuilderToFormats}
\alias{gdkContentFormatsBuilderUnref}
\alias{gdkContentProviderNewForBytes}
\alias{gdkContentProviderNewForValue}
\alias{gdkContentProviderNewUnion}
\alias{gdkContentProviderContentChanged}
\alias{gdkContentProviderGetValue}
\alias{gdkContentProviderRefFormats}
\alias{gdkContentProviderRefStorableFormats}
\alias{gdkContentProviderWriteMimeTypeFinish}
\alias{gdkContentSerializerGetCancellable}
\alias{gdkContentSerializerGetGtype}
\alias{gdkContentSerializerGetMimeType}
\alias{gdkContentSerializerGetOutputStream}
\alias{gdkContentSerializerGetPriority}
\alias{gdkContentSerializerGetTaskData}
\alias{gdkContentSerializerGetUserData}
\alias{gdkContentSerializerGetValue}
\alias{gdkContentSerializerReturnError}
\alias{gdkContentSerializerReturnSuccess}
\alias{gdkCrossingEventGetDetail}
\alias{gdkCrossingEventGetFocus}
\alias{gdkCrossingEventGetMode}
\alias{gdkCursorNewFromName}
\alias{gdkCursorNewFromTexture}
\alias{gdkCursorGetFallback}
\alias{gdkCursorGetHotspotX}
\alias{gdkCursorGetHotspotY}
\alias{gdkCursorGetName}
\alias{gdkCursorGetTexture}
\alias{gdkDndEventGetDrop}
\alias{gdkDmabufErrorQuark}
\alias{gdkDragBegin}
\alias{gdkDragDropDone}
\alias{gdkDragGetActions}
\alias{gdkDragGetContent}
\alias{gdkDragGetDevice}
\alias{gdkDragGetDisplay}
\alias{gdkDragGetDragSurface}
\alias{gdkDragGetFormats}
\alias{gdkDragGetSelectedAction}
\alias{gdkDragGetSurface}
\alias{gdkDragSetHotspot}
\alias{gdkDragActionIsUnique}
\alias{gdkDragSurfacePresent}
\alias{gdkDragSurfaceSizeSetSize}
\alias{gdkDrawContextBeginFrame}
\alias{gdkDrawContextEndFrame}
\alias{gdkDrawContextGetDisplay}
\alias{gdkDrawContextGetFrameRegion}
\alias{gdkDrawContextGetSurface}
\alias{gdkDrawContextIsInFrame}
\alias{gdkDropFinish}
\alias{gdkDropGetActions}
\alias{gdkDropGetDevice}
\alias{gdkDropGetDisplay}
\alias{gdkDropGetDrag}
\alias{gdkDropGetFormats}
\alias{gdkDropGetSurface}
\alias{gdkDropReadFinish}
\alias{gdkDropReadValueFinish}
\alias{gdkDropStatus}
\alias{gdkEventsGetAngle}
\alias{gdkEventsGetCenter}
\alias{gdkEventsGetDistance}
\alias{gdkFileListNewFromArray}
\alias{gdkFileListNewFromList}
\alias{gdkFileListGetFiles}
\alias{gdkFocusEventGetIn}
\alias{gdkFrameClockBeginUpdating}
\alias{gdkFrameClockEndUpdating}
\alias{gdkFrameClockGetCurrentTimings}
\alias{gdkFrameClockGetFps}
\alias{gdkFrameClockGetFrameCounter}
\alias{gdkFrameClockGetFrameTime}
\alias{gdkFrameClockGetHistoryStart}
\alias{gdkFrameClockGetRefreshInfo}
\alias{gdkFrameClockGetTimings}
\alias{gdkFrameClockRequestPhase}
\alias{gdkFrameTimingsGetComplete}
\alias{gdkFrameTimingsGetFrameCounter}
\alias{gdkFrameTimingsGetFrameTime}
\alias{gdkFrameTimingsGetPredictedPresentationTime}
\alias{gdkFrameTimingsGetPresentationTime}
\alias{gdkFrameTimingsGetRefreshInterval}
\alias{gdkFrameTimingsRef}
\alias{gdkFrameTimingsUnref}
\alias{gdkGlContextClearCurrent}
\alias{gdkGlContextGetCurrent}
\alias{gdkGlContextGetAllowedApis}
\alias{gdkGlContextGetApi}
\alias{gdkGlContextGetDebugEnabled}
\alias{gdkGlContextGetDisplay}
\alias{gdkGlContextGetForwardCompatible}
\alias{gdkGlContextGetRequiredVersion}
\alias{gdkGlContextGetSharedContext}
\alias{gdkGlContextGetSurface}
\alias{gdkGlContextGetUseEs}
\alias{gdkGlContextGetVersion}
\alias{gdkGlContextIsLegacy}
\alias{gdkGlContextIsShared}
\alias{gdkGlContextMakeCurrent}
\alias{gdkGlContextRealize}
\alias{gdkGlContextSetAllowedApis}
\alias{gdkGlContextSetDebugEnabled}
\alias{gdkGlContextSetForwardCompatible}
\alias{gdkGlContextSetRequiredVersion}
\alias{gdkGlContextSetUseEs}
\alias{gdkGlErrorQuark}
\alias{gdkGlTextureRelease}
\alias{gdkGlTextureBuilderNew}
\alias{gdkGlTextureBuilderGetContext}
\alias{gdkGlTextureBuilderGetFormat}
\alias{gdkGlTextureBuilderGetHasMipmap}
\alias{gdkGlTextureBuilderGetHeight}
\alias{gdkGlTextureBuilderGetId}
\alias{gdkGlTextureBuilderGetSync}
\alias{gdkGlTextureBuilderGetUpdateRegion}
\alias{gdkGlTextureBuilderGetUpdateTexture}
\alias{gdkGlTextureBuilderGetWidth}
\alias{gdkGlTextureBuilderSetContext}
\alias{gdkGlTextureBuilderSetFormat}
\alias{gdkGlTextureBuilderSetHasMipmap}
\alias{gdkGlTextureBuilderSetHeight}
\alias{gdkGlTextureBuilderSetId}
\alias{gdkGlTextureBuilderSetSync}
\alias{gdkGlTextureBuilderSetUpdateRegion}
\alias{gdkGlTextureBuilderSetUpdateTexture}
\alias{gdkGlTextureBuilderSetWidth}
\alias{gdkGrabBrokenEventGetGrabSurface}
\alias{gdkGrabBrokenEventGetImplicit}
\alias{gdkKeyEventGetConsumedModifiers}
\alias{gdkKeyEventGetKeycode}
\alias{gdkKeyEventGetKeyval}
\alias{gdkKeyEventGetLayout}
\alias{gdkKeyEventGetLevel}
\alias{gdkKeyEventGetMatch}
\alias{gdkKeyEventIsModifier}
\alias{gdkKeyEventMatches}
\alias{gdkMemoryTextureNew}
\alias{gdkPadEventGetAxisValue}
\alias{gdkPadEventGetButton}
\alias{gdkPadEventGetGroupMode}
\alias{gdkPaintableNewEmpty}
\alias{gdkPaintableComputeConcreteSize}
\alias{gdkPaintableGetCurrentImage}
\alias{gdkPaintableGetFlags}
\alias{gdkPaintableGetIntrinsicAspectRatio}
\alias{gdkPaintableGetIntrinsicHeight}
\alias{gdkPaintableGetIntrinsicWidth}
\alias{gdkPaintableInvalidateContents}
\alias{gdkPaintableInvalidateSize}
\alias{gdkPaintableSnapshot}
\alias{gdkPopupGetAutohide}
\alias{gdkPopupGetParent}
\alias{gdkPopupGetPositionX}
\alias{gdkPopupGetPositionY}
\alias{gdkPopupGetRectAnchor}
\alias{gdkPopupGetSurfaceAnchor}
\alias{gdkPopupPresent}
\alias{gdkPopupLayoutNew}
\alias{gdkPopupLayoutCopy}
\alias{gdkPopupLayoutEqual}
\alias{gdkPopupLayoutGetAnchorHints}
\alias{gdkPopupLayoutGetAnchorRect}
\alias{gdkPopupLayoutGetOffset}
\alias{gdkPopupLayoutGetRectAnchor}
\alias{gdkPopupLayoutGetShadowWidth}
\alias{gdkPopupLayoutGetSurfaceAnchor}
\alias{gdkPopupLayoutRef}
\alias{gdkPopupLayoutSetAnchorHints}
\alias{gdkPopupLayoutSetAnchorRect}
\alias{gdkPopupLayoutSetOffset}
\alias{gdkPopupLayoutSetRectAnchor}
\alias{gdkPopupLayoutSetShadowWidth}
\alias{gdkPopupLayoutSetSurfaceAnchor}
\alias{gdkPopupLayoutUnref}
\alias{gdkRgbaCopy}
\alias{gdkRgbaEqual}
\alias{gdkRgbaFree}
\alias{gdkRgbaHash}
\alias{gdkRgbaIsClear}
\alias{gdkRgbaIsOpaque}
\alias{gdkRgbaParse}
\alias{gdkRgbaToString}
\alias{gdkRectangleContainsPoint}
\alias{gdkRectangleEqual}
\alias{gdkRectangleIntersect}
\alias{gdkRectangleUnion}
\alias{gdkScrollEventGetDeltas}
\alias{gdkScrollEventGetDirection}
\alias{gdkScrollEventGetUnit}
\alias{gdkScrollEventIsStop}
\alias{gdkSeatGetCapabilities}
\alias{gdkSeatGetDevices}
\alias{gdkSeatGetDisplay}
\alias{gdkSeatGetKeyboard}
\alias{gdkSeatGetPointer}
\alias{gdkSeatGetTools}
\alias{gdkToplevelBeginMove}
\alias{gdkToplevelBeginResize}
\alias{gdkToplevelFocus}
\alias{gdkToplevelGetState}
\alias{gdkToplevelInhibitSystemShortcuts}
\alias{gdkToplevelLower}
\alias{gdkToplevelMinimize}
\alias{gdkToplevelPresent}
\alias{gdkToplevelRestoreSystemShortcuts}
\alias{gdkToplevelSetDecorated}
\alias{gdkToplevelSetDeletable}
\alias{gdkToplevelSetIconList}
\alias{gdkToplevelSetModal}
\alias{gdkToplevelSetStartupId}
\alias{gdkToplevelSetTitle}
\alias{gdkToplevelSetTransientFor}
\alias{gdkToplevelShowWindowMenu}
\alias{gdkToplevelSupportsEdgeConstraints}
\alias{gdkToplevelTitlebarGesture}
\alias{gdkToplevelLayoutNew}
\alias{gdkToplevelLayoutCopy}
\alias{gdkToplevelLayoutEqual}
\alias{gdkToplevelLayoutGetFullscreen}
\alias{gdkToplevelLayoutGetFullscreenMonitor}
\alias{gdkToplevelLayoutGetMaximized}
\alias{gdkToplevelLayoutGetResizable}
\alias{gdkToplevelLayoutRef}
\alias{gdkToplevelLayoutSetFullscreen}
\alias{gdkToplevelLayoutSetMaximized}
\alias{gdkToplevelLayoutSetResizable}
\alias{gdkToplevelLayoutUnref}
\alias{gdkToplevelSizeGetBounds}
\alias{gdkToplevelSizeSetMinSize}
\alias{gdkToplevelSizeSetShadowWidth}
\alias{gdkToplevelSizeSetSize}
\alias{gdkTouchEventGetEmulatingPointer}
\alias{gdkTouchpadEventGetDeltas}
\alias{gdkTouchpadEventGetGesturePhase}
\alias{gdkTouchpadEventGetNFingers}
\alias{gdkTouchpadEventGetPinchAngleDelta}
\alias{gdkTouchpadEventGetPinchScale}
\alias{gdkVulkanErrorQuark}
\alias{gdkCairoDrawFromGl}
\alias{gdkCairoRectangle}
\alias{gdkCairoRegion}
\alias{gdkCairoRegionCreateFromSurface}
\alias{gdkCairoSetSourcePixbuf}
\alias{gdkCairoSetSourceRgba}
\alias{gdkContentDeserializeFinish}
\alias{gdkContentSerializeFinish}
\alias{gdkInternMimeType}
\alias{gdkKeyvalConvertCase}
\alias{gdkKeyvalFromName}
\alias{gdkKeyvalIsLower}
\alias{gdkKeyvalIsUpper}
\alias{gdkKeyvalName}
\alias{gdkKeyvalToLower}
\alias{gdkKeyvalToUnicode}
\alias{gdkKeyvalToUpper}
\alias{gdkSetAllowedBackends}
\alias{gdkUnicodeToKeyval}
\title{gdkAppLaunchContextGetDisplay}
\usage{
gdkAppLaunchContextGetDisplay(context)

gdkAppLaunchContextSetDesktop(context, desktop)

gdkAppLaunchContextSetIcon(context, icon)

gdkAppLaunchContextSetIconName(context, icon_name)

gdkAppLaunchContextSetTimestamp(context, timestamp)

gdkButtonEventGetButton(event)

gdkCairoContextCairoCreate(self)

gdkClipboardGetContent(clipboard)

gdkClipboardGetDisplay(clipboard)

gdkClipboardGetFormats(clipboard)

gdkClipboardIsLocal(clipboard)

gdkClipboardReadFinish(clipboard, result)

gdkClipboardReadTextFinish(clipboard, result)

gdkClipboardReadTextureFinish(clipboard, result)

gdkClipboardReadValueFinish(clipboard, result)

gdkClipboardSetContent(clipboard, provider)

gdkClipboardSetValue(clipboard, value)

gdkClipboardStoreFinish(clipboard, result)

gdkContentDeserializerGetCancellable(deserializer)

gdkContentDeserializerGetGtype(deserializer)

gdkContentDeserializerGetInputStream(deserializer)

gdkContentDeserializerGetMimeType(deserializer)

gdkContentDeserializerGetPriority(deserializer)

gdkContentDeserializerGetTaskData(deserializer)

gdkContentDeserializerGetUserData(deserializer)

gdkContentDeserializerGetValue(deserializer)

gdkContentDeserializerReturnError(deserializer, error)

gdkContentDeserializerReturnSuccess(deserializer)

gdkContentFormatsNew(mime_types, n_mime_types)

gdkContentFormatsNewForGtype(type)

gdkContentFormatsContainGtype(formats, type)

gdkContentFormatsContainMimeType(formats, mime_type)

gdkContentFormatsGetGtypes(formats)

gdkContentFormatsGetMimeTypes(formats)

gdkContentFormatsMatch(first, second)

gdkContentFormatsMatchGtype(first, second)

gdkContentFormatsMatchMimeType(first, second)

gdkContentFormatsPrint(formats, string)

gdkContentFormatsRef(formats)

gdkContentFormatsToString(formats)

gdkContentFormatsUnion(first, second)

gdkContentFormatsUnionDeserializeGtypes(formats)

gdkContentFormatsUnionDeserializeMimeTypes(formats)

gdkContentFormatsUnionSerializeGtypes(formats)

gdkContentFormatsUnionSerializeMimeTypes(formats)

gdkContentFormatsUnref(formats)

gdkContentFormatsParse(string)

gdkContentFormatsBuilderNew()

gdkContentFormatsBuilderAddFormats(builder, formats)

gdkContentFormatsBuilderAddGtype(builder, type)

gdkContentFormatsBuilderAddMimeType(builder, mime_type)

gdkContentFormatsBuilderRef(builder)

gdkContentFormatsBuilderToFormats(builder)

gdkContentFormatsBuilderUnref(builder)

gdkContentProviderNewForBytes(mime_type, bytes)

gdkContentProviderNewForValue(value)

gdkContentProviderNewUnion(providers, n_providers)

gdkContentProviderContentChanged(provider)

gdkContentProviderGetValue(provider)

gdkContentProviderRefFormats(provider)

gdkContentProviderRefStorableFormats(provider)

gdkContentProviderWriteMimeTypeFinish(provider, result)

gdkContentSerializerGetCancellable(serializer)

gdkContentSerializerGetGtype(serializer)

gdkContentSerializerGetMimeType(serializer)

gdkContentSerializerGetOutputStream(serializer)

gdkContentSerializerGetPriority(serializer)

gdkContentSerializerGetTaskData(serializer)

gdkContentSerializerGetUserData(serializer)

gdkContentSerializerGetValue(serializer)

gdkContentSerializerReturnError(serializer, error)

gdkContentSerializerReturnSuccess(serializer)

gdkCrossingEventGetDetail(event)

gdkCrossingEventGetFocus(event)

gdkCrossingEventGetMode(event)

gdkCursorNewFromName(name, fallback)

gdkCursorNewFromTexture(texture, hotspot_x, hotspot_y, fallback)

gdkCursorGetFallback(cursor)

gdkCursorGetHotspotX(cursor)

gdkCursorGetHotspotY(cursor)

gdkCursorGetName(cursor)

gdkCursorGetTexture(cursor)

gdkDndEventGetDrop(event)

gdkDmabufErrorQuark()

gdkDragBegin(surface, device, content, actions, dx, dy)

gdkDragDropDone(drag, success)

gdkDragGetActions(drag)

gdkDragGetContent(drag)

gdkDragGetDevice(drag)

gdkDragGetDisplay(drag)

gdkDragGetDragSurface(drag)

gdkDragGetFormats(drag)

gdkDragGetSelectedAction(drag)

gdkDragGetSurface(drag)

gdkDragSetHotspot(drag, hot_x, hot_y)

gdkDragActionIsUnique(action)

gdkDragSurfacePresent(drag_surface, width, height)

gdkDragSurfaceSizeSetSize(size, width, height)

gdkDrawContextBeginFrame(context, region)

gdkDrawContextEndFrame(context)

gdkDrawContextGetDisplay(context)

gdkDrawContextGetFrameRegion(context)

gdkDrawContextGetSurface(context)

gdkDrawContextIsInFrame(context)

gdkDropFinish(self, action)

gdkDropGetActions(self)

gdkDropGetDevice(self)

gdkDropGetDisplay(self)

gdkDropGetDrag(self)

gdkDropGetFormats(self)

gdkDropGetSurface(self)

gdkDropReadFinish(self, result)

gdkDropReadValueFinish(self, result)

gdkDropStatus(self, actions, preferred)

gdkEventsGetAngle(event1, event2)

gdkEventsGetCenter(event1, event2)

gdkEventsGetDistance(event1, event2)

gdkFileListNewFromArray(files, n_files)

gdkFileListNewFromList(files)

gdkFileListGetFiles(file_list)

gdkFocusEventGetIn(event)

gdkFrameClockBeginUpdating(frame_clock)

gdkFrameClockEndUpdating(frame_clock)

gdkFrameClockGetCurrentTimings(frame_clock)

gdkFrameClockGetFps(frame_clock)

gdkFrameClockGetFrameCounter(frame_clock)

gdkFrameClockGetFrameTime(frame_clock)

gdkFrameClockGetHistoryStart(frame_clock)

gdkFrameClockGetRefreshInfo(frame_clock, base_time)

gdkFrameClockGetTimings(frame_clock, frame_counter)

gdkFrameClockRequestPhase(frame_clock, phase)

gdkFrameTimingsGetComplete(timings)

gdkFrameTimingsGetFrameCounter(timings)

gdkFrameTimingsGetFrameTime(timings)

gdkFrameTimingsGetPredictedPresentationTime(timings)

gdkFrameTimingsGetPresentationTime(timings)

gdkFrameTimingsGetRefreshInterval(timings)

gdkFrameTimingsRef(timings)

gdkFrameTimingsUnref(timings)

gdkGlContextClearCurrent()

gdkGlContextGetCurrent()

gdkGlContextGetAllowedApis(self)

gdkGlContextGetApi(self)

gdkGlContextGetDebugEnabled(context)

gdkGlContextGetDisplay(context)

gdkGlContextGetForwardCompatible(context)

gdkGlContextGetRequiredVersion(context)

gdkGlContextGetSharedContext(context)

gdkGlContextGetSurface(context)

gdkGlContextGetUseEs(context)

gdkGlContextGetVersion(context)

gdkGlContextIsLegacy(context)

gdkGlContextIsShared(self, other)

gdkGlContextMakeCurrent(context)

gdkGlContextRealize(context)

gdkGlContextSetAllowedApis(self, apis)

gdkGlContextSetDebugEnabled(context, enabled)

gdkGlContextSetForwardCompatible(context, compatible)

gdkGlContextSetRequiredVersion(context, major, minor)

gdkGlContextSetUseEs(context, use_es)

gdkGlErrorQuark()

gdkGlTextureRelease(self)

gdkGlTextureBuilderNew()

gdkGlTextureBuilderGetContext(self)

gdkGlTextureBuilderGetFormat(self)

gdkGlTextureBuilderGetHasMipmap(self)

gdkGlTextureBuilderGetHeight(self)

gdkGlTextureBuilderGetId(self)

gdkGlTextureBuilderGetSync(self)

gdkGlTextureBuilderGetUpdateRegion(self)

gdkGlTextureBuilderGetUpdateTexture(self)

gdkGlTextureBuilderGetWidth(self)

gdkGlTextureBuilderSetContext(self, context)

gdkGlTextureBuilderSetFormat(self, format)

gdkGlTextureBuilderSetHasMipmap(self, has_mipmap)

gdkGlTextureBuilderSetHeight(self, height)

gdkGlTextureBuilderSetId(self, id)

gdkGlTextureBuilderSetSync(self, sync)

gdkGlTextureBuilderSetUpdateRegion(self, region)

gdkGlTextureBuilderSetUpdateTexture(self, texture)

gdkGlTextureBuilderSetWidth(self, width)

gdkGrabBrokenEventGetGrabSurface(event)

gdkGrabBrokenEventGetImplicit(event)

gdkKeyEventGetConsumedModifiers(event)

gdkKeyEventGetKeycode(event)

gdkKeyEventGetKeyval(event)

gdkKeyEventGetLayout(event)

gdkKeyEventGetLevel(event)

gdkKeyEventGetMatch(event)

gdkKeyEventIsModifier(event)

gdkKeyEventMatches(event, keyval, modifiers)

gdkMemoryTextureNew(width, height, format, bytes, stride)

gdkPadEventGetAxisValue(event)

gdkPadEventGetButton(event)

gdkPadEventGetGroupMode(event)

gdkPaintableNewEmpty(intrinsic_width, intrinsic_height)

gdkPaintableComputeConcreteSize(
  paintable,
  specified_width,
  specified_height,
  default_width,
  default_height
)

gdkPaintableGetCurrentImage(paintable)

gdkPaintableGetFlags(paintable)

gdkPaintableGetIntrinsicAspectRatio(paintable)

gdkPaintableGetIntrinsicHeight(paintable)

gdkPaintableGetIntrinsicWidth(paintable)

gdkPaintableInvalidateContents(paintable)

gdkPaintableInvalidateSize(paintable)

gdkPaintableSnapshot(paintable, snapshot, width, height)

gdkPopupGetAutohide(popup)

gdkPopupGetParent(popup)

gdkPopupGetPositionX(popup)

gdkPopupGetPositionY(popup)

gdkPopupGetRectAnchor(popup)

gdkPopupGetSurfaceAnchor(popup)

gdkPopupPresent(popup, width, height, layout)

gdkPopupLayoutNew(anchor_rect, rect_anchor, surface_anchor)

gdkPopupLayoutCopy(layout)

gdkPopupLayoutEqual(layout, other)

gdkPopupLayoutGetAnchorHints(layout)

gdkPopupLayoutGetAnchorRect(layout)

gdkPopupLayoutGetOffset(layout)

gdkPopupLayoutGetRectAnchor(layout)

gdkPopupLayoutGetShadowWidth(layout)

gdkPopupLayoutGetSurfaceAnchor(layout)

gdkPopupLayoutRef(layout)

gdkPopupLayoutSetAnchorHints(layout, anchor_hints)

gdkPopupLayoutSetAnchorRect(layout, anchor_rect)

gdkPopupLayoutSetOffset(layout, dx, dy)

gdkPopupLayoutSetRectAnchor(layout, anchor)

gdkPopupLayoutSetShadowWidth(layout, left, right, top, bottom)

gdkPopupLayoutSetSurfaceAnchor(layout, anchor)

gdkPopupLayoutUnref(layout)

gdkRgbaCopy(rgba)

gdkRgbaEqual(p1, p2)

gdkRgbaFree(rgba)

gdkRgbaHash(p)

gdkRgbaIsClear(rgba)

gdkRgbaIsOpaque(rgba)

gdkRgbaParse(rgba, spec)

gdkRgbaToString(rgba)

gdkRectangleContainsPoint(rect, x, y)

gdkRectangleEqual(rect1, rect2)

gdkRectangleIntersect(src1, src2)

gdkRectangleUnion(src1, src2)

gdkScrollEventGetDeltas(event)

gdkScrollEventGetDirection(event)

gdkScrollEventGetUnit(event)

gdkScrollEventIsStop(event)

gdkSeatGetCapabilities(seat)

gdkSeatGetDevices(seat, capabilities)

gdkSeatGetDisplay(seat)

gdkSeatGetKeyboard(seat)

gdkSeatGetPointer(seat)

gdkSeatGetTools(seat)

gdkToplevelBeginMove(toplevel, device, button, x, y, timestamp)

gdkToplevelBeginResize(toplevel, edge, device, button, x, y, timestamp)

gdkToplevelFocus(toplevel, timestamp)

gdkToplevelGetState(toplevel)

gdkToplevelInhibitSystemShortcuts(toplevel, event)

gdkToplevelLower(toplevel)

gdkToplevelMinimize(toplevel)

gdkToplevelPresent(toplevel, layout)

gdkToplevelRestoreSystemShortcuts(toplevel)

gdkToplevelSetDecorated(toplevel, decorated)

gdkToplevelSetDeletable(toplevel, deletable)

gdkToplevelSetIconList(toplevel, surfaces)

gdkToplevelSetModal(toplevel, modal)

gdkToplevelSetStartupId(toplevel, startup_id)

gdkToplevelSetTitle(toplevel, title)

gdkToplevelSetTransientFor(toplevel, parent)

gdkToplevelShowWindowMenu(toplevel, event)

gdkToplevelSupportsEdgeConstraints(toplevel)

gdkToplevelTitlebarGesture(toplevel, gesture)

gdkToplevelLayoutNew()

gdkToplevelLayoutCopy(layout)

gdkToplevelLayoutEqual(layout, other)

gdkToplevelLayoutGetFullscreen(layout)

gdkToplevelLayoutGetFullscreenMonitor(layout)

gdkToplevelLayoutGetMaximized(layout)

gdkToplevelLayoutGetResizable(layout)

gdkToplevelLayoutRef(layout)

gdkToplevelLayoutSetFullscreen(layout, fullscreen, monitor)

gdkToplevelLayoutSetMaximized(layout, maximized)

gdkToplevelLayoutSetResizable(layout, resizable)

gdkToplevelLayoutUnref(layout)

gdkToplevelSizeGetBounds(size)

gdkToplevelSizeSetMinSize(size, min_width, min_height)

gdkToplevelSizeSetShadowWidth(size, left, right, top, bottom)

gdkToplevelSizeSetSize(size, width, height)

gdkTouchEventGetEmulatingPointer(event)

gdkTouchpadEventGetDeltas(event)

gdkTouchpadEventGetGesturePhase(event)

gdkTouchpadEventGetNFingers(event)

gdkTouchpadEventGetPinchAngleDelta(event)

gdkTouchpadEventGetPinchScale(event)

gdkVulkanErrorQuark()

gdkCairoDrawFromGl(
  cr,
  surface,
  source,
  source_type,
  buffer_scale,
  x,
  y,
  width,
  height
)

gdkCairoRectangle(cr, rectangle)

gdkCairoRegion(cr, region)

gdkCairoRegionCreateFromSurface(surface)

gdkCairoSetSourcePixbuf(cr, pixbuf, pixbuf_x, pixbuf_y)

gdkCairoSetSourceRgba(cr, rgba)

gdkContentDeserializeFinish(result)

gdkContentSerializeFinish(result)

gdkInternMimeType(string)

gdkKeyvalConvertCase(symbol)

gdkKeyvalFromName(keyval_name)

gdkKeyvalIsLower(keyval)

gdkKeyvalIsUpper(keyval)

gdkKeyvalName(keyval)

gdkKeyvalToLower(keyval)

gdkKeyvalToUnicode(keyval)

gdkKeyvalToUpper(keyval)

gdkSetAllowedBackends(backends)

gdkUnicodeToKeyval(wc)
}
\arguments{
\item{context}{GLContext}

\item{desktop}{gint}

\item{icon}{Gio.Icon}

\item{icon_name}{utf8}

\item{timestamp}{guint32}

\item{event}{TouchpadEvent}

\item{self}{GLTextureBuilder}

\item{clipboard}{Clipboard}

\item{result}{Gio.AsyncResult}

\item{provider}{ContentProvider}

\item{value}{GObject.Value}

\item{deserializer}{ContentDeserializer}

\item{error}{GLib.Error}

\item{mime_types}{utf8}

\item{n_mime_types}{guint}

\item{type}{GType}

\item{formats}{ContentFormats}

\item{mime_type}{utf8}

\item{first}{ContentFormats}

\item{second}{ContentFormats}

\item{string}{utf8}

\item{builder}{ContentFormatsBuilder}

\item{bytes}{GLib.Bytes}

\item{providers}{ContentProvider}

\item{n_providers}{gsize}

\item{serializer}{ContentSerializer}

\item{name}{utf8}

\item{fallback}{Cursor}

\item{texture}{Texture}

\item{hotspot_x}{gint}

\item{hotspot_y}{gint}

\item{cursor}{Cursor}

\item{surface}{cairo.Surface}

\item{device}{Device}

\item{content}{ContentProvider}

\item{actions}{DragAction}

\item{dx}{gint}

\item{dy}{gint}

\item{drag}{Drag}

\item{success}{gboolean}

\item{hot_x}{gint}

\item{hot_y}{gint}

\item{action}{DragAction}

\item{drag_surface}{DragSurface}

\item{width}{gint}

\item{height}{gint}

\item{size}{ToplevelSize}

\item{region}{cairo.Region}

\item{preferred}{DragAction}

\item{event1}{Event}

\item{event2}{Event}

\item{files}{GLib.SList}

\item{n_files}{gsize}

\item{file_list}{FileList}

\item{frame_clock}{FrameClock}

\item{base_time}{gint64}

\item{frame_counter}{gint64}

\item{phase}{FrameClockPhase}

\item{timings}{FrameTimings}

\item{other}{ToplevelLayout}

\item{apis}{GLAPI}

\item{enabled}{gboolean}

\item{compatible}{gboolean}

\item{major}{gint}

\item{minor}{gint}

\item{use_es}{gint}

\item{format}{MemoryFormat}

\item{has_mipmap}{gboolean}

\item{id}{guint}

\item{sync}{gpointer}

\item{keyval}{guint}

\item{modifiers}{ModifierType}

\item{stride}{gsize}

\item{intrinsic_width}{gint}

\item{intrinsic_height}{gint}

\item{paintable}{Paintable}

\item{specified_width}{gdouble}

\item{specified_height}{gdouble}

\item{default_width}{gdouble}

\item{default_height}{gdouble}

\item{snapshot}{Snapshot}

\item{popup}{Popup}

\item{layout}{ToplevelLayout}

\item{anchor_rect}{Rectangle}

\item{rect_anchor}{Gravity}

\item{surface_anchor}{Gravity}

\item{anchor_hints}{AnchorHints}

\item{anchor}{Gravity}

\item{left}{gint}

\item{right}{gint}

\item{top}{gint}

\item{bottom}{gint}

\item{rgba}{RGBA}

\item{p1}{RGBA}

\item{p2}{RGBA}

\item{p}{RGBA}

\item{spec}{utf8}

\item{rect}{Rectangle}

\item{x}{gint}

\item{y}{gint}

\item{rect1}{Rectangle}

\item{rect2}{Rectangle}

\item{src1}{Rectangle}

\item{src2}{Rectangle}

\item{seat}{Seat}

\item{capabilities}{SeatCapabilities}

\item{toplevel}{Toplevel}

\item{button}{gint}

\item{edge}{SurfaceEdge}

\item{decorated}{gboolean}

\item{deletable}{gboolean}

\item{surfaces}{GLib.List}

\item{modal}{gboolean}

\item{startup_id}{utf8}

\item{title}{utf8}

\item{parent}{Surface}

\item{gesture}{TitlebarGesture}

\item{fullscreen}{gboolean}

\item{monitor}{Monitor}

\item{maximized}{gboolean}

\item{resizable}{gboolean}

\item{min_width}{gint}

\item{min_height}{gint}

\item{cr}{cairo.Context}

\item{source}{gint}

\item{source_type}{gint}

\item{buffer_scale}{gint}

\item{rectangle}{Rectangle}

\item{pixbuf}{GdkPixbuf.Pixbuf}

\item{pixbuf_x}{gdouble}

\item{pixbuf_y}{gdouble}

\item{symbol}{guint}

\item{keyval_name}{utf8}

\item{backends}{utf8}

\item{wc}{guint32}
}
\value{
Display

Return value from C function

Return value from C function

Return value from C function

Return value from C function

guint

cairo.Context

ContentProvider

Display

ContentFormats

gboolean

Gio.InputStream

utf8

Texture

GObject.Value

gboolean

Return value from C function

gboolean

Gio.Cancellable

GType

Gio.InputStream

utf8

gint

gpointer

gpointer

GObject.Value

Return value from C function

Return value from C function

ContentFormats

ContentFormats

gboolean

gboolean

GType

utf8

gboolean

GType

utf8

Return value from C function

ContentFormats

utf8

ContentFormats

ContentFormats

ContentFormats

ContentFormats

ContentFormats

Return value from C function

ContentFormats

ContentFormatsBuilder

Return value from C function

Return value from C function

Return value from C function

ContentFormatsBuilder

ContentFormats

Return value from C function

ContentProvider

ContentProvider

ContentProvider

Return value from C function

gboolean

ContentFormats

ContentFormats

gboolean

Gio.Cancellable

GType

utf8

Gio.OutputStream

gint

gpointer

gpointer

GObject.Value

Return value from C function

Return value from C function

NotifyType

gboolean

CrossingMode

Cursor

Cursor

Cursor

gint

gint

utf8

Texture

Drop

GLib.Quark

Drag

Return value from C function

DragAction

ContentProvider

Device

Display

Surface

ContentFormats

DragAction

Surface

Return value from C function

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Display

cairo.Region

Surface

gboolean

Return value from C function

DragAction

Device

Display

Drag

ContentFormats

Surface

Gio.InputStream

GObject.Value

Return value from C function

gboolean

gboolean

gboolean

FileList

FileList

GLib.SList

gboolean

Return value from C function

Return value from C function

FrameTimings

gdouble

gint64

gint64

gint64

Return value from C function

FrameTimings

Return value from C function

gboolean

gint64

gint64

gint64

gint64

gint64

FrameTimings

Return value from C function

Return value from C function

GLContext

GLAPI

GLAPI

gboolean

Display

gboolean

Return value from C function

GLContext

Surface

gboolean

Return value from C function

gboolean

gboolean

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

GLib.Quark

Return value from C function

GLTextureBuilder

GLContext

MemoryFormat

gboolean

gint

guint

gpointer

cairo.Region

Texture

gint

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Surface

gboolean

ModifierType

guint

guint

guint

guint

gboolean

gboolean

KeyMatch

MemoryTexture

Return value from C function

guint

Return value from C function

Paintable

Return value from C function

Paintable

PaintableFlags

gdouble

gint

gint

Return value from C function

Return value from C function

Return value from C function

gboolean

Surface

gint

gint

Gravity

Gravity

gboolean

PopupLayout

PopupLayout

gboolean

AnchorHints

Rectangle

Return value from C function

Gravity

Return value from C function

Gravity

PopupLayout

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

RGBA

gboolean

Return value from C function

guint

gboolean

gboolean

gboolean

utf8

gboolean

gboolean

gboolean

Return value from C function

Return value from C function

ScrollDirection

ScrollUnit

gboolean

SeatCapabilities

GLib.List

Display

Device

Device

GLib.List

Return value from C function

Return value from C function

Return value from C function

ToplevelState

Return value from C function

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

gboolean

gboolean

ToplevelLayout

ToplevelLayout

gboolean

gboolean

Monitor

gboolean

gboolean

ToplevelLayout

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

TouchpadGesturePhase

guint

gdouble

gdouble

GLib.Quark

Return value from C function

Return value from C function

Return value from C function

cairo.Region

Return value from C function

Return value from C function

gboolean

gboolean

utf8

Return value from C function

guint

gboolean

gboolean

utf8

guint

guint32

guint

Return value from C function

guint
}
\description{
gdkAppLaunchContextGetDisplay

gdkAppLaunchContextSetDesktop

gdkAppLaunchContextSetIcon

gdkAppLaunchContextSetIconName

gdkAppLaunchContextSetTimestamp

gdkButtonEventGetButton

gdkCairoContextCairoCreate

gdkClipboardGetContent

gdkClipboardGetDisplay

gdkClipboardGetFormats

gdkClipboardIsLocal

gdkClipboardReadFinish

gdkClipboardReadTextFinish

gdkClipboardReadTextureFinish

gdkClipboardReadValueFinish

gdkClipboardSetContent

gdkClipboardSetValue

gdkClipboardStoreFinish

gdkContentDeserializerGetCancellable

gdkContentDeserializerGetGtype

gdkContentDeserializerGetInputStream

gdkContentDeserializerGetMimeType

gdkContentDeserializerGetPriority

gdkContentDeserializerGetTaskData

gdkContentDeserializerGetUserData

gdkContentDeserializerGetValue

gdkContentDeserializerReturnError

gdkContentDeserializerReturnSuccess

gdkContentFormatsNew

gdkContentFormatsNewForGtype

gdkContentFormatsContainGtype

gdkContentFormatsContainMimeType

gdkContentFormatsGetGtypes

gdkContentFormatsGetMimeTypes

gdkContentFormatsMatch

gdkContentFormatsMatchGtype

gdkContentFormatsMatchMimeType

gdkContentFormatsPrint

gdkContentFormatsRef

gdkContentFormatsToString

gdkContentFormatsUnion

gdkContentFormatsUnionDeserializeGtypes

gdkContentFormatsUnionDeserializeMimeTypes

gdkContentFormatsUnionSerializeGtypes

gdkContentFormatsUnionSerializeMimeTypes

gdkContentFormatsUnref

gdkContentFormatsParse

gdkContentFormatsBuilderNew

gdkContentFormatsBuilderAddFormats

gdkContentFormatsBuilderAddGtype

gdkContentFormatsBuilderAddMimeType

gdkContentFormatsBuilderRef

gdkContentFormatsBuilderToFormats

gdkContentFormatsBuilderUnref

gdkContentProviderNewForBytes

gdkContentProviderNewForValue

gdkContentProviderNewUnion

gdkContentProviderContentChanged

gdkContentProviderGetValue

gdkContentProviderRefFormats

gdkContentProviderRefStorableFormats

gdkContentProviderWriteMimeTypeFinish

gdkContentSerializerGetCancellable

gdkContentSerializerGetGtype

gdkContentSerializerGetMimeType

gdkContentSerializerGetOutputStream

gdkContentSerializerGetPriority

gdkContentSerializerGetTaskData

gdkContentSerializerGetUserData

gdkContentSerializerGetValue

gdkContentSerializerReturnError

gdkContentSerializerReturnSuccess

gdkCrossingEventGetDetail

gdkCrossingEventGetFocus

gdkCrossingEventGetMode

gdkCursorNewFromName

gdkCursorNewFromTexture

gdkCursorGetFallback

gdkCursorGetHotspotX

gdkCursorGetHotspotY

gdkCursorGetName

gdkCursorGetTexture

gdkDndEventGetDrop

gdkDmabufErrorQuark

gdkDragBegin

gdkDragDropDone

gdkDragGetActions

gdkDragGetContent

gdkDragGetDevice

gdkDragGetDisplay

gdkDragGetDragSurface

gdkDragGetFormats

gdkDragGetSelectedAction

gdkDragGetSurface

gdkDragSetHotspot

gdkDragActionIsUnique

gdkDragSurfacePresent

gdkDragSurfaceSizeSetSize

gdkDrawContextBeginFrame

gdkDrawContextEndFrame

gdkDrawContextGetDisplay

gdkDrawContextGetFrameRegion

gdkDrawContextGetSurface

gdkDrawContextIsInFrame

gdkDropFinish

gdkDropGetActions

gdkDropGetDevice

gdkDropGetDisplay

gdkDropGetDrag

gdkDropGetFormats

gdkDropGetSurface

gdkDropReadFinish

gdkDropReadValueFinish

gdkDropStatus

gdkEventsGetAngle

gdkEventsGetCenter

gdkEventsGetDistance

gdkFileListNewFromArray

gdkFileListNewFromList

gdkFileListGetFiles

gdkFocusEventGetIn

gdkFrameClockBeginUpdating

gdkFrameClockEndUpdating

gdkFrameClockGetCurrentTimings

gdkFrameClockGetFps

gdkFrameClockGetFrameCounter

gdkFrameClockGetFrameTime

gdkFrameClockGetHistoryStart

gdkFrameClockGetRefreshInfo

gdkFrameClockGetTimings

gdkFrameClockRequestPhase

gdkFrameTimingsGetComplete

gdkFrameTimingsGetFrameCounter

gdkFrameTimingsGetFrameTime

gdkFrameTimingsGetPredictedPresentationTime

gdkFrameTimingsGetPresentationTime

gdkFrameTimingsGetRefreshInterval

gdkFrameTimingsRef

gdkFrameTimingsUnref

gdkGlContextClearCurrent

gdkGlContextGetCurrent

gdkGlContextGetAllowedApis

gdkGlContextGetApi

gdkGlContextGetDebugEnabled

gdkGlContextGetDisplay

gdkGlContextGetForwardCompatible

gdkGlContextGetRequiredVersion

gdkGlContextGetSharedContext

gdkGlContextGetSurface

gdkGlContextGetUseEs

gdkGlContextGetVersion

gdkGlContextIsLegacy

gdkGlContextIsShared

gdkGlContextMakeCurrent

gdkGlContextRealize

gdkGlContextSetAllowedApis

gdkGlContextSetDebugEnabled

gdkGlContextSetForwardCompatible

gdkGlContextSetRequiredVersion

gdkGlContextSetUseEs

gdkGlErrorQuark

gdkGlTextureRelease

gdkGlTextureBuilderNew

gdkGlTextureBuilderGetContext

gdkGlTextureBuilderGetFormat

gdkGlTextureBuilderGetHasMipmap

gdkGlTextureBuilderGetHeight

gdkGlTextureBuilderGetId

gdkGlTextureBuilderGetSync

gdkGlTextureBuilderGetUpdateRegion

gdkGlTextureBuilderGetUpdateTexture

gdkGlTextureBuilderGetWidth

gdkGlTextureBuilderSetContext

gdkGlTextureBuilderSetFormat

gdkGlTextureBuilderSetHasMipmap

gdkGlTextureBuilderSetHeight

gdkGlTextureBuilderSetId

gdkGlTextureBuilderSetSync

gdkGlTextureBuilderSetUpdateRegion

gdkGlTextureBuilderSetUpdateTexture

gdkGlTextureBuilderSetWidth

gdkGrabBrokenEventGetGrabSurface

gdkGrabBrokenEventGetImplicit

gdkKeyEventGetConsumedModifiers

gdkKeyEventGetKeycode

gdkKeyEventGetKeyval

gdkKeyEventGetLayout

gdkKeyEventGetLevel

gdkKeyEventGetMatch

gdkKeyEventIsModifier

gdkKeyEventMatches

gdkMemoryTextureNew

gdkPadEventGetAxisValue

gdkPadEventGetButton

gdkPadEventGetGroupMode

gdkPaintableNewEmpty

gdkPaintableComputeConcreteSize

gdkPaintableGetCurrentImage

gdkPaintableGetFlags

gdkPaintableGetIntrinsicAspectRatio

gdkPaintableGetIntrinsicHeight

gdkPaintableGetIntrinsicWidth

gdkPaintableInvalidateContents

gdkPaintableInvalidateSize

gdkPaintableSnapshot

gdkPopupGetAutohide

gdkPopupGetParent

gdkPopupGetPositionX

gdkPopupGetPositionY

gdkPopupGetRectAnchor

gdkPopupGetSurfaceAnchor

gdkPopupPresent

gdkPopupLayoutNew

gdkPopupLayoutCopy

gdkPopupLayoutEqual

gdkPopupLayoutGetAnchorHints

gdkPopupLayoutGetAnchorRect

gdkPopupLayoutGetOffset

gdkPopupLayoutGetRectAnchor

gdkPopupLayoutGetShadowWidth

gdkPopupLayoutGetSurfaceAnchor

gdkPopupLayoutRef

gdkPopupLayoutSetAnchorHints

gdkPopupLayoutSetAnchorRect

gdkPopupLayoutSetOffset

gdkPopupLayoutSetRectAnchor

gdkPopupLayoutSetShadowWidth

gdkPopupLayoutSetSurfaceAnchor

gdkPopupLayoutUnref

gdkRgbaCopy

gdkRgbaEqual

gdkRgbaFree

gdkRgbaHash

gdkRgbaIsClear

gdkRgbaIsOpaque

gdkRgbaParse

gdkRgbaToString

gdkRectangleContainsPoint

gdkRectangleEqual

gdkRectangleIntersect

gdkRectangleUnion

gdkScrollEventGetDeltas

gdkScrollEventGetDirection

gdkScrollEventGetUnit

gdkScrollEventIsStop

gdkSeatGetCapabilities

gdkSeatGetDevices

gdkSeatGetDisplay

gdkSeatGetKeyboard

gdkSeatGetPointer

gdkSeatGetTools

gdkToplevelBeginMove

gdkToplevelBeginResize

gdkToplevelFocus

gdkToplevelGetState

gdkToplevelInhibitSystemShortcuts

gdkToplevelLower

gdkToplevelMinimize

gdkToplevelPresent

gdkToplevelRestoreSystemShortcuts

gdkToplevelSetDecorated

gdkToplevelSetDeletable

gdkToplevelSetIconList

gdkToplevelSetModal

gdkToplevelSetStartupId

gdkToplevelSetTitle

gdkToplevelSetTransientFor

gdkToplevelShowWindowMenu

gdkToplevelSupportsEdgeConstraints

gdkToplevelTitlebarGesture

gdkToplevelLayoutNew

gdkToplevelLayoutCopy

gdkToplevelLayoutEqual

gdkToplevelLayoutGetFullscreen

gdkToplevelLayoutGetFullscreenMonitor

gdkToplevelLayoutGetMaximized

gdkToplevelLayoutGetResizable

gdkToplevelLayoutRef

gdkToplevelLayoutSetFullscreen

gdkToplevelLayoutSetMaximized

gdkToplevelLayoutSetResizable

gdkToplevelLayoutUnref

gdkToplevelSizeGetBounds

gdkToplevelSizeSetMinSize

gdkToplevelSizeSetShadowWidth

gdkToplevelSizeSetSize

gdkTouchEventGetEmulatingPointer

gdkTouchpadEventGetDeltas

gdkTouchpadEventGetGesturePhase

gdkTouchpadEventGetNFingers

gdkTouchpadEventGetPinchAngleDelta

gdkTouchpadEventGetPinchScale

gdkVulkanErrorQuark

gdkCairoDrawFromGl

gdkCairoRectangle

gdkCairoRegion

gdkCairoRegionCreateFromSurface

gdkCairoSetSourcePixbuf

gdkCairoSetSourceRgba

gdkContentDeserializeFinish

gdkContentSerializeFinish

gdkInternMimeType

gdkKeyvalConvertCase

gdkKeyvalFromName

gdkKeyvalIsLower

gdkKeyvalIsUpper

gdkKeyvalName

gdkKeyvalToLower

gdkKeyvalToUnicode

gdkKeyvalToUpper

gdkSetAllowedBackends

gdkUnicodeToKeyval
}
