% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gio_autogen.R
\name{gAppInfoCreateFromCommandline}
\alias{gAppInfoCreateFromCommandline}
\alias{gAppInfoGetAll}
\alias{gAppInfoGetAllForType}
\alias{gAppInfoGetDefaultForType}
\alias{gAppInfoGetDefaultForUriScheme}
\alias{gAppInfoGetFallbackForType}
\alias{gAppInfoGetRecommendedForType}
\alias{gAppInfoLaunchDefaultForUri}
\alias{gAppInfoLaunchDefaultForUriAsync}
\alias{gAppInfoLaunchDefaultForUriFinish}
\alias{gAppInfoResetTypeAssociations}
\alias{gAppInfoAddSupportsType}
\alias{gAppInfoCanDelete}
\alias{gAppInfoCanRemoveSupportsType}
\alias{gAppInfoDelete}
\alias{gAppInfoDup}
\alias{gAppInfoEqual}
\alias{gAppInfoGetCommandline}
\alias{gAppInfoGetDescription}
\alias{gAppInfoGetDisplayName}
\alias{gAppInfoGetExecutable}
\alias{gAppInfoGetIcon}
\alias{gAppInfoGetId}
\alias{gAppInfoGetName}
\alias{gAppInfoGetSupportedTypes}
\alias{gAppInfoLaunch}
\alias{gAppInfoLaunchUris}
\alias{gAppInfoRemoveSupportsType}
\alias{gAppInfoSetAsDefaultForExtension}
\alias{gAppInfoSetAsDefaultForType}
\alias{gAppInfoSetAsLastUsedForType}
\alias{gAppInfoShouldShow}
\alias{gAppInfoSupportsFiles}
\alias{gAppInfoSupportsUris}
\alias{gAppInfoMonitorGet}
\alias{gApplicationNew}
\alias{gApplicationGetDefault}
\alias{gApplicationIdIsValid}
\alias{gApplicationActivate}
\alias{gApplicationAddMainOption}
\alias{gApplicationAddMainOptionEntries}
\alias{gApplicationAddOptionGroup}
\alias{gApplicationBindBusyProperty}
\alias{gApplicationGetApplicationId}
\alias{gApplicationGetDbusConnection}
\alias{gApplicationGetDbusObjectPath}
\alias{gApplicationGetFlags}
\alias{gApplicationGetInactivityTimeout}
\alias{gApplicationGetIsBusy}
\alias{gApplicationGetIsRegistered}
\alias{gApplicationGetIsRemote}
\alias{gApplicationGetResourceBasePath}
\alias{gApplicationHold}
\alias{gApplicationMarkBusy}
\alias{gApplicationOpen}
\alias{gApplicationQuit}
\alias{gApplicationRegister}
\alias{gApplicationRelease}
\alias{gApplicationRun}
\alias{gApplicationSendNotification}
\alias{gApplicationSetActionGroup}
\alias{gApplicationSetApplicationId}
\alias{gApplicationSetDefault}
\alias{gApplicationSetFlags}
\alias{gApplicationSetInactivityTimeout}
\alias{gApplicationSetOptionContextDescription}
\alias{gApplicationSetOptionContextParameterString}
\alias{gApplicationSetOptionContextSummary}
\alias{gApplicationSetResourceBasePath}
\alias{gApplicationUnbindBusyProperty}
\alias{gApplicationUnmarkBusy}
\alias{gApplicationWithdrawNotification}
\alias{gApplicationCommandLineCreateFileForArg}
\alias{gApplicationCommandLineGetArguments}
\alias{gApplicationCommandLineGetCwd}
\alias{gApplicationCommandLineGetEnviron}
\alias{gApplicationCommandLineGetExitStatus}
\alias{gApplicationCommandLineGetIsRemote}
\alias{gApplicationCommandLineGetOptionsDict}
\alias{gApplicationCommandLineGetPlatformData}
\alias{gApplicationCommandLineGetStdin}
\alias{gApplicationCommandLineGetenv}
\alias{gApplicationCommandLineSetExitStatus}
\title{gAppInfoCreateFromCommandline}
\usage{
gAppInfoCreateFromCommandline(commandline, application_name, flags)

gAppInfoGetAll()

gAppInfoGetAllForType(content_type)

gAppInfoGetDefaultForType(content_type, must_support_uris)

gAppInfoGetDefaultForUriScheme(uri_scheme)

gAppInfoGetFallbackForType(content_type)

gAppInfoGetRecommendedForType(content_type)

gAppInfoLaunchDefaultForUri(uri, context)

gAppInfoLaunchDefaultForUriAsync(uri, context, cancellable, callback)

gAppInfoLaunchDefaultForUriFinish(result)

gAppInfoResetTypeAssociations(content_type)

gAppInfoAddSupportsType(appinfo, content_type)

gAppInfoCanDelete(appinfo)

gAppInfoCanRemoveSupportsType(appinfo)

gAppInfoDelete(appinfo)

gAppInfoDup(appinfo)

gAppInfoEqual(appinfo1, appinfo2)

gAppInfoGetCommandline(appinfo)

gAppInfoGetDescription(appinfo)

gAppInfoGetDisplayName(appinfo)

gAppInfoGetExecutable(appinfo)

gAppInfoGetIcon(appinfo)

gAppInfoGetId(appinfo)

gAppInfoGetName(appinfo)

gAppInfoGetSupportedTypes(appinfo)

gAppInfoLaunch(appinfo, files, context)

gAppInfoLaunchUris(appinfo, uris, context)

gAppInfoRemoveSupportsType(appinfo, content_type)

gAppInfoSetAsDefaultForExtension(appinfo, extension)

gAppInfoSetAsDefaultForType(appinfo, content_type)

gAppInfoSetAsLastUsedForType(appinfo, content_type)

gAppInfoShouldShow(appinfo)

gAppInfoSupportsFiles(appinfo)

gAppInfoSupportsUris(appinfo)

gAppInfoMonitorGet()

gApplicationNew(application_id, flags)

gApplicationGetDefault()

gApplicationIdIsValid(application_id)

gApplicationActivate(application)

gApplicationAddMainOption(
  application,
  long_name,
  short_name,
  flags,
  arg,
  description,
  arg_description
)

gApplicationAddMainOptionEntries(application, entries)

gApplicationAddOptionGroup(application, group)

gApplicationBindBusyProperty(application, object, property)

gApplicationGetApplicationId(application)

gApplicationGetDbusConnection(application)

gApplicationGetDbusObjectPath(application)

gApplicationGetFlags(application)

gApplicationGetInactivityTimeout(application)

gApplicationGetIsBusy(application)

gApplicationGetIsRegistered(application)

gApplicationGetIsRemote(application)

gApplicationGetResourceBasePath(application)

gApplicationHold(application)

gApplicationMarkBusy(application)

gApplicationOpen(application, files, n_files, hint)

gApplicationQuit(application)

gApplicationRegister(application, cancellable)

gApplicationRelease(application)

gApplicationRun(application, argc, argv)

gApplicationSendNotification(application, id, notification)

gApplicationSetActionGroup(application, action_group)

gApplicationSetApplicationId(application, application_id)

gApplicationSetDefault(application)

gApplicationSetFlags(application, flags)

gApplicationSetInactivityTimeout(application, inactivity_timeout)

gApplicationSetOptionContextDescription(application, description)

gApplicationSetOptionContextParameterString(application, parameter_string)

gApplicationSetOptionContextSummary(application, summary)

gApplicationSetResourceBasePath(application, resource_path)

gApplicationUnbindBusyProperty(application, object, property)

gApplicationUnmarkBusy(application)

gApplicationWithdrawNotification(application, id)

gApplicationCommandLineCreateFileForArg(cmdline, arg)

gApplicationCommandLineGetArguments(cmdline)

gApplicationCommandLineGetCwd(cmdline)

gApplicationCommandLineGetEnviron(cmdline)

gApplicationCommandLineGetExitStatus(cmdline)

gApplicationCommandLineGetIsRemote(cmdline)

gApplicationCommandLineGetOptionsDict(cmdline)

gApplicationCommandLineGetPlatformData(cmdline)

gApplicationCommandLineGetStdin(cmdline)

gApplicationCommandLineGetenv(cmdline, name)

gApplicationCommandLineSetExitStatus(cmdline, exit_status)
}
\arguments{
\item{commandline}{filename}

\item{application_name}{utf8}

\item{flags}{ApplicationFlags}

\item{content_type}{utf8}

\item{must_support_uris}{gboolean}

\item{uri_scheme}{utf8}

\item{uri}{utf8}

\item{context}{AppLaunchContext}

\item{cancellable}{Cancellable}

\item{callback}{function — AsyncReadyCallback function(source_object, res)}

\item{result}{AsyncResult}

\item{appinfo}{AppInfo}

\item{appinfo1}{AppInfo}

\item{appinfo2}{AppInfo}

\item{files}{File}

\item{uris}{GLib.List}

\item{extension}{filename}

\item{application_id}{utf8}

\item{application}{Application}

\item{long_name}{utf8}

\item{short_name}{gchar}

\item{arg}{filename}

\item{description}{utf8}

\item{arg_description}{utf8}

\item{entries}{GLib.OptionEntry}

\item{group}{GLib.OptionGroup}

\item{object}{GObject.Object}

\item{property}{utf8}

\item{n_files}{gint}

\item{hint}{utf8}

\item{argc}{gint}

\item{argv}{filename}

\item{id}{utf8}

\item{notification}{Notification}

\item{action_group}{ActionGroup}

\item{inactivity_timeout}{guint}

\item{parameter_string}{utf8}

\item{summary}{utf8}

\item{resource_path}{utf8}

\item{cmdline}{ApplicationCommandLine}

\item{name}{filename}

\item{exit_status}{gint}
}
\value{
AppInfo

GLib.List

GLib.List

AppInfo

AppInfo

GLib.List

GLib.List

gboolean

Return value from C function

gboolean

Return value from C function

gboolean

gboolean

gboolean

gboolean

AppInfo

gboolean

filename

utf8

utf8

filename

Icon

utf8

utf8

utf8

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

AppInfoMonitor

Application

Application

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

utf8

DBusConnection

utf8

ApplicationFlags

guint

gboolean

gboolean

gboolean

utf8

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

gint

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

File

filename

filename

filename

gint

gboolean

GLib.VariantDict

GLib.Variant

InputStream

utf8

Return value from C function
}
\description{
gAppInfoCreateFromCommandline

gAppInfoGetAll

gAppInfoGetAllForType

gAppInfoGetDefaultForType

gAppInfoGetDefaultForUriScheme

gAppInfoGetFallbackForType

gAppInfoGetRecommendedForType

gAppInfoLaunchDefaultForUri

gAppInfoLaunchDefaultForUriAsync

gAppInfoLaunchDefaultForUriFinish

gAppInfoResetTypeAssociations

gAppInfoAddSupportsType

gAppInfoCanDelete

gAppInfoCanRemoveSupportsType

gAppInfoDelete

gAppInfoDup

gAppInfoEqual

gAppInfoGetCommandline

gAppInfoGetDescription

gAppInfoGetDisplayName

gAppInfoGetExecutable

gAppInfoGetIcon

gAppInfoGetId

gAppInfoGetName

gAppInfoGetSupportedTypes

gAppInfoLaunch

gAppInfoLaunchUris

gAppInfoRemoveSupportsType

gAppInfoSetAsDefaultForExtension

gAppInfoSetAsDefaultForType

gAppInfoSetAsLastUsedForType

gAppInfoShouldShow

gAppInfoSupportsFiles

gAppInfoSupportsUris

gAppInfoMonitorGet

gApplicationNew

gApplicationGetDefault

gApplicationIdIsValid

gApplicationActivate

gApplicationAddMainOption

gApplicationAddMainOptionEntries

gApplicationAddOptionGroup

gApplicationBindBusyProperty

gApplicationGetApplicationId

gApplicationGetDbusConnection

gApplicationGetDbusObjectPath

gApplicationGetFlags

gApplicationGetInactivityTimeout

gApplicationGetIsBusy

gApplicationGetIsRegistered

gApplicationGetIsRemote

gApplicationGetResourceBasePath

gApplicationHold

gApplicationMarkBusy

gApplicationOpen

gApplicationQuit

gApplicationRegister

gApplicationRelease

gApplicationRun

gApplicationSendNotification

gApplicationSetActionGroup

gApplicationSetApplicationId

gApplicationSetDefault

gApplicationSetFlags

gApplicationSetInactivityTimeout

gApplicationSetOptionContextDescription

gApplicationSetOptionContextParameterString

gApplicationSetOptionContextSummary

gApplicationSetResourceBasePath

gApplicationUnbindBusyProperty

gApplicationUnmarkBusy

gApplicationWithdrawNotification

gApplicationCommandLineCreateFileForArg

gApplicationCommandLineGetArguments

gApplicationCommandLineGetCwd

gApplicationCommandLineGetEnviron

gApplicationCommandLineGetExitStatus

gApplicationCommandLineGetIsRemote

gApplicationCommandLineGetOptionsDict

gApplicationCommandLineGetPlatformData

gApplicationCommandLineGetStdin

gApplicationCommandLineGetenv

gApplicationCommandLineSetExitStatus
}
