% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gio_autogen.R
\name{gDbusActionGroupGet}
\alias{gDbusActionGroupGet}
\alias{gDbusAnnotationInfoRef}
\alias{gDbusAnnotationInfoUnref}
\alias{gDbusAnnotationInfoLookup}
\alias{gDbusArgInfoRef}
\alias{gDbusArgInfoUnref}
\alias{gDbusAuthObserverNew}
\alias{gDbusAuthObserverAllowMechanism}
\alias{gDbusAuthObserverAuthorizeAuthenticatedPeer}
\alias{gDbusConnectionNewFinish}
\alias{gDbusConnectionNewForAddressFinish}
\alias{gDbusConnectionNewForAddressSync}
\alias{gDbusConnectionNewSync}
\alias{gDbusConnectionCallFinish}
\alias{gDbusConnectionCallSync}
\alias{gDbusConnectionCallWithUnixFdListFinish}
\alias{gDbusConnectionCallWithUnixFdListSync}
\alias{gDbusConnectionCloseFinish}
\alias{gDbusConnectionCloseSync}
\alias{gDbusConnectionEmitSignal}
\alias{gDbusConnectionExportActionGroup}
\alias{gDbusConnectionExportMenuModel}
\alias{gDbusConnectionFlushFinish}
\alias{gDbusConnectionFlushSync}
\alias{gDbusConnectionGetCapabilities}
\alias{gDbusConnectionGetExitOnClose}
\alias{gDbusConnectionGetFlags}
\alias{gDbusConnectionGetGuid}
\alias{gDbusConnectionGetLastSerial}
\alias{gDbusConnectionGetPeerCredentials}
\alias{gDbusConnectionGetStream}
\alias{gDbusConnectionGetUniqueName}
\alias{gDbusConnectionIsClosed}
\alias{gDbusConnectionRegisterObjectWithClosures}
\alias{gDbusConnectionRegisterObjectWithClosures2}
\alias{gDbusConnectionRemoveFilter}
\alias{gDbusConnectionSendMessage}
\alias{gDbusConnectionSendMessageWithReplyFinish}
\alias{gDbusConnectionSendMessageWithReplySync}
\alias{gDbusConnectionSetExitOnClose}
\alias{gDbusConnectionSignalUnsubscribe}
\alias{gDbusConnectionStartMessageProcessing}
\alias{gDbusConnectionUnexportActionGroup}
\alias{gDbusConnectionUnexportMenuModel}
\alias{gDbusConnectionUnregisterObject}
\alias{gDbusConnectionUnregisterSubtree}
\alias{gDbusErrorEncodeGerror}
\alias{gDbusErrorGetRemoteError}
\alias{gDbusErrorIsRemoteError}
\alias{gDbusErrorNewForDbusError}
\alias{gDbusErrorQuark}
\alias{gDbusErrorRegisterError}
\alias{gDbusErrorRegisterErrorDomain}
\alias{gDbusErrorStripRemoteError}
\alias{gDbusErrorUnregisterError}
\alias{gDbusInterfaceDupObject}
\alias{gDbusInterfaceGetInfo}
\alias{gDbusInterfaceSetObject}
\alias{gDbusInterfaceInfoCacheBuild}
\alias{gDbusInterfaceInfoCacheRelease}
\alias{gDbusInterfaceInfoGenerateXml}
\alias{gDbusInterfaceInfoLookupMethod}
\alias{gDbusInterfaceInfoLookupProperty}
\alias{gDbusInterfaceInfoLookupSignal}
\alias{gDbusInterfaceInfoRef}
\alias{gDbusInterfaceInfoUnref}
\alias{gDbusInterfaceSkeletonExport}
\alias{gDbusInterfaceSkeletonFlush}
\alias{gDbusInterfaceSkeletonGetConnection}
\alias{gDbusInterfaceSkeletonGetConnections}
\alias{gDbusInterfaceSkeletonGetFlags}
\alias{gDbusInterfaceSkeletonGetInfo}
\alias{gDbusInterfaceSkeletonGetObjectPath}
\alias{gDbusInterfaceSkeletonGetProperties}
\alias{gDbusInterfaceSkeletonGetVtable}
\alias{gDbusInterfaceSkeletonHasConnection}
\alias{gDbusInterfaceSkeletonSetFlags}
\alias{gDbusInterfaceSkeletonUnexport}
\alias{gDbusInterfaceSkeletonUnexportFromConnection}
\alias{gDbusMenuModelGet}
\alias{gDbusMessageNew}
\alias{gDbusMessageNewFromBlob}
\alias{gDbusMessageNewMethodCall}
\alias{gDbusMessageNewSignal}
\alias{gDbusMessageBytesNeeded}
\alias{gDbusMessageCopy}
\alias{gDbusMessageGetArg0}
\alias{gDbusMessageGetArg0Path}
\alias{gDbusMessageGetBody}
\alias{gDbusMessageGetByteOrder}
\alias{gDbusMessageGetDestination}
\alias{gDbusMessageGetErrorName}
\alias{gDbusMessageGetFlags}
\alias{gDbusMessageGetHeader}
\alias{gDbusMessageGetHeaderFields}
\alias{gDbusMessageGetInterface}
\alias{gDbusMessageGetLocked}
\alias{gDbusMessageGetMember}
\alias{gDbusMessageGetMessageType}
\alias{gDbusMessageGetNumUnixFds}
\alias{gDbusMessageGetPath}
\alias{gDbusMessageGetReplySerial}
\alias{gDbusMessageGetSender}
\alias{gDbusMessageGetSerial}
\alias{gDbusMessageGetSignature}
\alias{gDbusMessageGetUnixFdList}
\alias{gDbusMessageLock}
\alias{gDbusMessageNewMethodErrorLiteral}
\alias{gDbusMessageNewMethodReply}
\alias{gDbusMessagePrint}
\alias{gDbusMessageSetBody}
\alias{gDbusMessageSetByteOrder}
\alias{gDbusMessageSetDestination}
\alias{gDbusMessageSetErrorName}
\alias{gDbusMessageSetFlags}
\alias{gDbusMessageSetHeader}
\alias{gDbusMessageSetInterface}
\alias{gDbusMessageSetMember}
\alias{gDbusMessageSetMessageType}
\alias{gDbusMessageSetNumUnixFds}
\alias{gDbusMessageSetPath}
\alias{gDbusMessageSetReplySerial}
\alias{gDbusMessageSetSender}
\alias{gDbusMessageSetSerial}
\alias{gDbusMessageSetSignature}
\alias{gDbusMessageSetUnixFdList}
\alias{gDbusMessageToBlob}
\alias{gDbusMessageToGerror}
\alias{gDbusMethodInfoRef}
\alias{gDbusMethodInfoUnref}
\alias{gDbusMethodInvocationGetConnection}
\alias{gDbusMethodInvocationGetInterfaceName}
\alias{gDbusMethodInvocationGetMessage}
\alias{gDbusMethodInvocationGetMethodInfo}
\alias{gDbusMethodInvocationGetMethodName}
\alias{gDbusMethodInvocationGetObjectPath}
\alias{gDbusMethodInvocationGetParameters}
\alias{gDbusMethodInvocationGetPropertyInfo}
\alias{gDbusMethodInvocationGetSender}
\alias{gDbusMethodInvocationReturnDbusError}
\alias{gDbusMethodInvocationReturnErrorLiteral}
\alias{gDbusMethodInvocationReturnGerror}
\alias{gDbusMethodInvocationReturnValue}
\alias{gDbusMethodInvocationReturnValueWithUnixFdList}
\alias{gDbusNodeInfoNewForXml}
\alias{gDbusNodeInfoGenerateXml}
\alias{gDbusNodeInfoLookupInterface}
\alias{gDbusNodeInfoRef}
\alias{gDbusNodeInfoUnref}
\alias{gDbusObjectGetInterface}
\alias{gDbusObjectGetInterfaces}
\alias{gDbusObjectGetObjectPath}
\alias{gDbusObjectManagerGetInterface}
\alias{gDbusObjectManagerGetObject}
\alias{gDbusObjectManagerGetObjectPath}
\alias{gDbusObjectManagerGetObjects}
\alias{gDbusObjectManagerClientNewFinish}
\alias{gDbusObjectManagerClientNewForBusFinish}
\alias{gDbusObjectManagerClientGetConnection}
\alias{gDbusObjectManagerClientGetFlags}
\alias{gDbusObjectManagerClientGetName}
\alias{gDbusObjectManagerClientGetNameOwner}
\alias{gDbusObjectManagerServerNew}
\alias{gDbusObjectManagerServerExport}
\alias{gDbusObjectManagerServerExportUniquely}
\alias{gDbusObjectManagerServerGetConnection}
\alias{gDbusObjectManagerServerIsExported}
\alias{gDbusObjectManagerServerSetConnection}
\alias{gDbusObjectManagerServerUnexport}
\alias{gDbusObjectProxyNew}
\alias{gDbusObjectProxyGetConnection}
\alias{gDbusObjectSkeletonNew}
\alias{gDbusObjectSkeletonAddInterface}
\alias{gDbusObjectSkeletonFlush}
\alias{gDbusObjectSkeletonRemoveInterface}
\alias{gDbusObjectSkeletonRemoveInterfaceByName}
\alias{gDbusObjectSkeletonSetObjectPath}
\alias{gDbusPropertyInfoRef}
\alias{gDbusPropertyInfoUnref}
\alias{gDbusProxyNewFinish}
\alias{gDbusProxyNewForBusFinish}
\alias{gDbusProxyNewForBusSync}
\alias{gDbusProxyNewSync}
\alias{gDbusProxyCallFinish}
\alias{gDbusProxyCallSync}
\alias{gDbusProxyCallWithUnixFdListFinish}
\alias{gDbusProxyCallWithUnixFdListSync}
\alias{gDbusProxyGetCachedProperty}
\alias{gDbusProxyGetCachedPropertyNames}
\alias{gDbusProxyGetConnection}
\alias{gDbusProxyGetDefaultTimeout}
\alias{gDbusProxyGetFlags}
\alias{gDbusProxyGetInterfaceInfo}
\alias{gDbusProxyGetInterfaceName}
\alias{gDbusProxyGetName}
\alias{gDbusProxyGetNameOwner}
\alias{gDbusProxyGetObjectPath}
\alias{gDbusProxySetCachedProperty}
\alias{gDbusProxySetDefaultTimeout}
\alias{gDbusProxySetInterfaceInfo}
\alias{gDbusServerNewSync}
\alias{gDbusServerGetClientAddress}
\alias{gDbusServerGetFlags}
\alias{gDbusServerGetGuid}
\alias{gDbusServerIsActive}
\alias{gDbusServerStart}
\alias{gDbusServerStop}
\alias{gDbusSignalInfoRef}
\alias{gDbusSignalInfoUnref}
\alias{gDbusAddressEscapeValue}
\alias{gDbusAddressGetForBusSync}
\alias{gDbusAddressGetStreamFinish}
\alias{gDbusAddressGetStreamSync}
\alias{gDbusEscapeObjectPath}
\alias{gDbusEscapeObjectPathBytestring}
\alias{gDbusGenerateGuid}
\alias{gDbusGvalueToGvariant}
\alias{gDbusGvariantToGvalue}
\alias{gDbusIsAddress}
\alias{gDbusIsErrorName}
\alias{gDbusIsGuid}
\alias{gDbusIsInterfaceName}
\alias{gDbusIsMemberName}
\alias{gDbusIsName}
\alias{gDbusIsSupportedAddress}
\alias{gDbusIsUniqueName}
\alias{gDbusUnescapeObjectPath}
\title{gDbusActionGroupGet}
\usage{
gDbusActionGroupGet(connection, bus_name, object_path)

gDbusAnnotationInfoRef(info)

gDbusAnnotationInfoUnref(info)

gDbusAnnotationInfoLookup(annotations, name)

gDbusArgInfoRef(info)

gDbusArgInfoUnref(info)

gDbusAuthObserverNew()

gDbusAuthObserverAllowMechanism(observer, mechanism)

gDbusAuthObserverAuthorizeAuthenticatedPeer(observer, stream, credentials)

gDbusConnectionNewFinish(res)

gDbusConnectionNewForAddressFinish(res)

gDbusConnectionNewForAddressSync(address, flags, observer, cancellable)

gDbusConnectionNewSync(stream, guid, flags, observer, cancellable)

gDbusConnectionCallFinish(connection, res)

gDbusConnectionCallSync(
  connection,
  bus_name,
  object_path,
  interface_name,
  method_name,
  parameters,
  reply_type,
  flags,
  timeout_msec,
  cancellable
)

gDbusConnectionCallWithUnixFdListFinish(connection, res)

gDbusConnectionCallWithUnixFdListSync(
  connection,
  bus_name,
  object_path,
  interface_name,
  method_name,
  parameters,
  reply_type,
  flags,
  timeout_msec,
  fd_list,
  cancellable
)

gDbusConnectionCloseFinish(connection, res)

gDbusConnectionCloseSync(connection, cancellable)

gDbusConnectionEmitSignal(
  connection,
  destination_bus_name,
  object_path,
  interface_name,
  signal_name,
  parameters
)

gDbusConnectionExportActionGroup(connection, object_path, action_group)

gDbusConnectionExportMenuModel(connection, object_path, menu)

gDbusConnectionFlushFinish(connection, res)

gDbusConnectionFlushSync(connection, cancellable)

gDbusConnectionGetCapabilities(connection)

gDbusConnectionGetExitOnClose(connection)

gDbusConnectionGetFlags(connection)

gDbusConnectionGetGuid(connection)

gDbusConnectionGetLastSerial(connection)

gDbusConnectionGetPeerCredentials(connection)

gDbusConnectionGetStream(connection)

gDbusConnectionGetUniqueName(connection)

gDbusConnectionIsClosed(connection)

gDbusConnectionRegisterObjectWithClosures(
  connection,
  object_path,
  interface_info,
  method_call_closure,
  get_property_closure,
  set_property_closure
)

gDbusConnectionRegisterObjectWithClosures2(
  connection,
  object_path,
  interface_info,
  method_call_closure,
  get_property_closure,
  set_property_closure
)

gDbusConnectionRemoveFilter(connection, filter_id)

gDbusConnectionSendMessage(connection, message, flags)

gDbusConnectionSendMessageWithReplyFinish(connection, res)

gDbusConnectionSendMessageWithReplySync(
  connection,
  message,
  flags,
  timeout_msec,
  cancellable
)

gDbusConnectionSetExitOnClose(connection, exit_on_close)

gDbusConnectionSignalUnsubscribe(connection, subscription_id)

gDbusConnectionStartMessageProcessing(connection)

gDbusConnectionUnexportActionGroup(connection, export_id)

gDbusConnectionUnexportMenuModel(connection, export_id)

gDbusConnectionUnregisterObject(connection, registration_id)

gDbusConnectionUnregisterSubtree(connection, registration_id)

gDbusErrorEncodeGerror(error)

gDbusErrorGetRemoteError(error)

gDbusErrorIsRemoteError(error)

gDbusErrorNewForDbusError(dbus_error_name, dbus_error_message)

gDbusErrorQuark()

gDbusErrorRegisterError(error_domain, error_code, dbus_error_name)

gDbusErrorRegisterErrorDomain(error_domain_quark_name, entries, num_entries)

gDbusErrorStripRemoteError(error)

gDbusErrorUnregisterError(error_domain, error_code, dbus_error_name)

gDbusInterfaceDupObject(interface_)

gDbusInterfaceGetInfo(interface_)

gDbusInterfaceSetObject(interface_, object)

gDbusInterfaceInfoCacheBuild(info)

gDbusInterfaceInfoCacheRelease(info)

gDbusInterfaceInfoGenerateXml(info, indent, string_builder)

gDbusInterfaceInfoLookupMethod(info, name)

gDbusInterfaceInfoLookupProperty(info, name)

gDbusInterfaceInfoLookupSignal(info, name)

gDbusInterfaceInfoRef(info)

gDbusInterfaceInfoUnref(info)

gDbusInterfaceSkeletonExport(interface_, connection, object_path)

gDbusInterfaceSkeletonFlush(interface_)

gDbusInterfaceSkeletonGetConnection(interface_)

gDbusInterfaceSkeletonGetConnections(interface_)

gDbusInterfaceSkeletonGetFlags(interface_)

gDbusInterfaceSkeletonGetInfo(interface_)

gDbusInterfaceSkeletonGetObjectPath(interface_)

gDbusInterfaceSkeletonGetProperties(interface_)

gDbusInterfaceSkeletonGetVtable(interface_)

gDbusInterfaceSkeletonHasConnection(interface_, connection)

gDbusInterfaceSkeletonSetFlags(interface_, flags)

gDbusInterfaceSkeletonUnexport(interface_)

gDbusInterfaceSkeletonUnexportFromConnection(interface_, connection)

gDbusMenuModelGet(connection, bus_name, object_path)

gDbusMessageNew()

gDbusMessageNewFromBlob(blob, blob_len, capabilities)

gDbusMessageNewMethodCall(name, path, interface_, method)

gDbusMessageNewSignal(path, interface_, signal)

gDbusMessageBytesNeeded(blob, blob_len)

gDbusMessageCopy(message)

gDbusMessageGetArg0(message)

gDbusMessageGetArg0Path(message)

gDbusMessageGetBody(message)

gDbusMessageGetByteOrder(message)

gDbusMessageGetDestination(message)

gDbusMessageGetErrorName(message)

gDbusMessageGetFlags(message)

gDbusMessageGetHeader(message, header_field)

gDbusMessageGetHeaderFields(message)

gDbusMessageGetInterface(message)

gDbusMessageGetLocked(message)

gDbusMessageGetMember(message)

gDbusMessageGetMessageType(message)

gDbusMessageGetNumUnixFds(message)

gDbusMessageGetPath(message)

gDbusMessageGetReplySerial(message)

gDbusMessageGetSender(message)

gDbusMessageGetSerial(message)

gDbusMessageGetSignature(message)

gDbusMessageGetUnixFdList(message)

gDbusMessageLock(message)

gDbusMessageNewMethodErrorLiteral(
  method_call_message,
  error_name,
  error_message
)

gDbusMessageNewMethodReply(method_call_message)

gDbusMessagePrint(message, indent)

gDbusMessageSetBody(message, body)

gDbusMessageSetByteOrder(message, byte_order)

gDbusMessageSetDestination(message, value)

gDbusMessageSetErrorName(message, value)

gDbusMessageSetFlags(message, flags)

gDbusMessageSetHeader(message, header_field, value)

gDbusMessageSetInterface(message, value)

gDbusMessageSetMember(message, value)

gDbusMessageSetMessageType(message, type)

gDbusMessageSetNumUnixFds(message, value)

gDbusMessageSetPath(message, value)

gDbusMessageSetReplySerial(message, value)

gDbusMessageSetSender(message, value)

gDbusMessageSetSerial(message, serial)

gDbusMessageSetSignature(message, value)

gDbusMessageSetUnixFdList(message, fd_list)

gDbusMessageToBlob(message, capabilities)

gDbusMessageToGerror(message)

gDbusMethodInfoRef(info)

gDbusMethodInfoUnref(info)

gDbusMethodInvocationGetConnection(invocation)

gDbusMethodInvocationGetInterfaceName(invocation)

gDbusMethodInvocationGetMessage(invocation)

gDbusMethodInvocationGetMethodInfo(invocation)

gDbusMethodInvocationGetMethodName(invocation)

gDbusMethodInvocationGetObjectPath(invocation)

gDbusMethodInvocationGetParameters(invocation)

gDbusMethodInvocationGetPropertyInfo(invocation)

gDbusMethodInvocationGetSender(invocation)

gDbusMethodInvocationReturnDbusError(invocation, error_name, error_message)

gDbusMethodInvocationReturnErrorLiteral(invocation, domain, code, message)

gDbusMethodInvocationReturnGerror(invocation, error)

gDbusMethodInvocationReturnValue(invocation, parameters)

gDbusMethodInvocationReturnValueWithUnixFdList(invocation, parameters, fd_list)

gDbusNodeInfoNewForXml(xml_data)

gDbusNodeInfoGenerateXml(info, indent, string_builder)

gDbusNodeInfoLookupInterface(info, name)

gDbusNodeInfoRef(info)

gDbusNodeInfoUnref(info)

gDbusObjectGetInterface(object, interface_name)

gDbusObjectGetInterfaces(object)

gDbusObjectGetObjectPath(object)

gDbusObjectManagerGetInterface(manager, object_path, interface_name)

gDbusObjectManagerGetObject(manager, object_path)

gDbusObjectManagerGetObjectPath(manager)

gDbusObjectManagerGetObjects(manager)

gDbusObjectManagerClientNewFinish(res)

gDbusObjectManagerClientNewForBusFinish(res)

gDbusObjectManagerClientGetConnection(manager)

gDbusObjectManagerClientGetFlags(manager)

gDbusObjectManagerClientGetName(manager)

gDbusObjectManagerClientGetNameOwner(manager)

gDbusObjectManagerServerNew(object_path)

gDbusObjectManagerServerExport(manager, object)

gDbusObjectManagerServerExportUniquely(manager, object)

gDbusObjectManagerServerGetConnection(manager)

gDbusObjectManagerServerIsExported(manager, object)

gDbusObjectManagerServerSetConnection(manager, connection)

gDbusObjectManagerServerUnexport(manager, object_path)

gDbusObjectProxyNew(connection, object_path)

gDbusObjectProxyGetConnection(proxy)

gDbusObjectSkeletonNew(object_path)

gDbusObjectSkeletonAddInterface(object, interface_)

gDbusObjectSkeletonFlush(object)

gDbusObjectSkeletonRemoveInterface(object, interface_)

gDbusObjectSkeletonRemoveInterfaceByName(object, interface_name)

gDbusObjectSkeletonSetObjectPath(object, object_path)

gDbusPropertyInfoRef(info)

gDbusPropertyInfoUnref(info)

gDbusProxyNewFinish(res)

gDbusProxyNewForBusFinish(res)

gDbusProxyNewForBusSync(
  bus_type,
  flags,
  info,
  name,
  object_path,
  interface_name,
  cancellable
)

gDbusProxyNewSync(
  connection,
  flags,
  info,
  name,
  object_path,
  interface_name,
  cancellable
)

gDbusProxyCallFinish(proxy, res)

gDbusProxyCallSync(
  proxy,
  method_name,
  parameters,
  flags,
  timeout_msec,
  cancellable
)

gDbusProxyCallWithUnixFdListFinish(proxy, res)

gDbusProxyCallWithUnixFdListSync(
  proxy,
  method_name,
  parameters,
  flags,
  timeout_msec,
  fd_list,
  cancellable
)

gDbusProxyGetCachedProperty(proxy, property_name)

gDbusProxyGetCachedPropertyNames(proxy)

gDbusProxyGetConnection(proxy)

gDbusProxyGetDefaultTimeout(proxy)

gDbusProxyGetFlags(proxy)

gDbusProxyGetInterfaceInfo(proxy)

gDbusProxyGetInterfaceName(proxy)

gDbusProxyGetName(proxy)

gDbusProxyGetNameOwner(proxy)

gDbusProxyGetObjectPath(proxy)

gDbusProxySetCachedProperty(proxy, property_name, value)

gDbusProxySetDefaultTimeout(proxy, timeout_msec)

gDbusProxySetInterfaceInfo(proxy, info)

gDbusServerNewSync(address, flags, guid, observer, cancellable)

gDbusServerGetClientAddress(server)

gDbusServerGetFlags(server)

gDbusServerGetGuid(server)

gDbusServerIsActive(server)

gDbusServerStart(server)

gDbusServerStop(server)

gDbusSignalInfoRef(info)

gDbusSignalInfoUnref(info)

gDbusAddressEscapeValue(string)

gDbusAddressGetForBusSync(bus_type, cancellable)

gDbusAddressGetStreamFinish(res)

gDbusAddressGetStreamSync(address, cancellable)

gDbusEscapeObjectPath(s)

gDbusEscapeObjectPathBytestring(bytes)

gDbusGenerateGuid()

gDbusGvalueToGvariant(gvalue, type)

gDbusGvariantToGvalue(value)

gDbusIsAddress(string)

gDbusIsErrorName(string)

gDbusIsGuid(string)

gDbusIsInterfaceName(string)

gDbusIsMemberName(string)

gDbusIsName(string)

gDbusIsSupportedAddress(string)

gDbusIsUniqueName(string)

gDbusUnescapeObjectPath(s)
}
\arguments{
\item{connection}{DBusConnection}

\item{bus_name}{utf8}

\item{object_path}{utf8}

\item{info}{DBusSignalInfo}

\item{annotations}{DBusAnnotationInfo}

\item{name}{utf8}

\item{observer}{DBusAuthObserver}

\item{mechanism}{utf8}

\item{stream}{IOStream}

\item{credentials}{Credentials}

\item{res}{AsyncResult}

\item{address}{utf8}

\item{flags}{DBusServerFlags}

\item{cancellable}{Cancellable}

\item{guid}{utf8}

\item{interface_name}{utf8}

\item{method_name}{utf8}

\item{parameters}{GLib.Variant}

\item{reply_type}{GLib.VariantType}

\item{timeout_msec}{gint}

\item{fd_list}{UnixFDList}

\item{destination_bus_name}{utf8}

\item{signal_name}{utf8}

\item{action_group}{ActionGroup}

\item{menu}{MenuModel}

\item{interface_info}{DBusInterfaceInfo}

\item{method_call_closure}{GObject.Closure}

\item{get_property_closure}{GObject.Closure}

\item{set_property_closure}{GObject.Closure}

\item{filter_id}{guint}

\item{message}{utf8}

\item{exit_on_close}{gboolean}

\item{subscription_id}{guint}

\item{export_id}{guint}

\item{registration_id}{guint}

\item{error}{GLib.Error}

\item{dbus_error_name}{utf8}

\item{dbus_error_message}{utf8}

\item{error_domain}{GLib.Quark}

\item{error_code}{gint}

\item{error_domain_quark_name}{utf8}

\item{entries}{DBusErrorEntry}

\item{num_entries}{guint}

\item{interface_}{DBusInterfaceSkeleton}

\item{object}{DBusObjectSkeleton}

\item{indent}{guint}

\item{string_builder}{GLib.String}

\item{blob}{guint8}

\item{blob_len}{gsize}

\item{capabilities}{DBusCapabilityFlags}

\item{path}{utf8}

\item{method}{utf8}

\item{signal}{utf8}

\item{header_field}{DBusMessageHeaderField}

\item{method_call_message}{DBusMessage}

\item{error_name}{utf8}

\item{error_message}{utf8}

\item{body}{GLib.Variant}

\item{byte_order}{DBusMessageByteOrder}

\item{value}{GLib.Variant}

\item{type}{GLib.VariantType}

\item{serial}{guint32}

\item{invocation}{DBusMethodInvocation}

\item{domain}{GLib.Quark}

\item{code}{gint}

\item{xml_data}{utf8}

\item{manager}{DBusObjectManagerServer}

\item{proxy}{DBusProxy}

\item{bus_type}{BusType}

\item{property_name}{utf8}

\item{server}{DBusServer}

\item{string}{utf8}

\item{s}{utf8}

\item{bytes}{guint8}

\item{gvalue}{GObject.Value}
}
\value{
DBusActionGroup

DBusAnnotationInfo

Return value from C function

utf8

DBusArgInfo

Return value from C function

DBusAuthObserver

gboolean

gboolean

DBusConnection

DBusConnection

DBusConnection

DBusConnection

GLib.Variant

GLib.Variant

GLib.Variant

GLib.Variant

gboolean

gboolean

gboolean

guint

guint

gboolean

gboolean

DBusCapabilityFlags

gboolean

DBusConnectionFlags

utf8

guint32

Credentials

IOStream

utf8

gboolean

guint

guint

Return value from C function

gboolean

DBusMessage

DBusMessage

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

gboolean

utf8

utf8

gboolean

GLib.Error

GLib.Quark

gboolean

Return value from C function

gboolean

gboolean

DBusObject

DBusInterfaceInfo

Return value from C function

Return value from C function

Return value from C function

Return value from C function

DBusMethodInfo

DBusPropertyInfo

DBusSignalInfo

DBusInterfaceInfo

Return value from C function

gboolean

Return value from C function

DBusConnection

GLib.List

DBusInterfaceSkeletonFlags

DBusInterfaceInfo

utf8

GLib.Variant

DBusInterfaceVTable

gboolean

Return value from C function

Return value from C function

Return value from C function

DBusMenuModel

DBusMessage

DBusMessage

DBusMessage

DBusMessage

gssize

DBusMessage

utf8

utf8

GLib.Variant

DBusMessageByteOrder

utf8

utf8

DBusMessageFlags

GLib.Variant

guint8

utf8

gboolean

utf8

DBusMessageType

guint32

utf8

guint32

utf8

guint32

utf8

UnixFDList

Return value from C function

DBusMessage

DBusMessage

utf8

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

guint8

gboolean

DBusMethodInfo

Return value from C function

DBusConnection

utf8

DBusMessage

DBusMethodInfo

utf8

utf8

GLib.Variant

DBusPropertyInfo

utf8

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

DBusNodeInfo

Return value from C function

DBusInterfaceInfo

DBusNodeInfo

Return value from C function

DBusInterface

GLib.List

utf8

DBusInterface

DBusObject

utf8

GLib.List

DBusObjectManagerClient

DBusObjectManagerClient

DBusConnection

DBusObjectManagerClientFlags

utf8

utf8

DBusObjectManagerServer

Return value from C function

Return value from C function

DBusConnection

gboolean

Return value from C function

gboolean

DBusObjectProxy

DBusConnection

DBusObjectSkeleton

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

DBusPropertyInfo

Return value from C function

DBusProxy

DBusProxy

DBusProxy

DBusProxy

GLib.Variant

GLib.Variant

GLib.Variant

GLib.Variant

GLib.Variant

utf8

DBusConnection

gint

DBusProxyFlags

DBusInterfaceInfo

utf8

utf8

utf8

utf8

Return value from C function

Return value from C function

Return value from C function

DBusServer

utf8

DBusServerFlags

utf8

gboolean

Return value from C function

Return value from C function

DBusSignalInfo

Return value from C function

utf8

utf8

IOStream

IOStream

utf8

utf8

utf8

GLib.Variant

Return value from C function

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

guint8
}
\description{
gDbusActionGroupGet

gDbusAnnotationInfoRef

gDbusAnnotationInfoUnref

gDbusAnnotationInfoLookup

gDbusArgInfoRef

gDbusArgInfoUnref

gDbusAuthObserverNew

gDbusAuthObserverAllowMechanism

gDbusAuthObserverAuthorizeAuthenticatedPeer

gDbusConnectionNewFinish

gDbusConnectionNewForAddressFinish

gDbusConnectionNewForAddressSync

gDbusConnectionNewSync

gDbusConnectionCallFinish

gDbusConnectionCallSync

gDbusConnectionCallWithUnixFdListFinish

gDbusConnectionCallWithUnixFdListSync

gDbusConnectionCloseFinish

gDbusConnectionCloseSync

gDbusConnectionEmitSignal

gDbusConnectionExportActionGroup

gDbusConnectionExportMenuModel

gDbusConnectionFlushFinish

gDbusConnectionFlushSync

gDbusConnectionGetCapabilities

gDbusConnectionGetExitOnClose

gDbusConnectionGetFlags

gDbusConnectionGetGuid

gDbusConnectionGetLastSerial

gDbusConnectionGetPeerCredentials

gDbusConnectionGetStream

gDbusConnectionGetUniqueName

gDbusConnectionIsClosed

gDbusConnectionRegisterObjectWithClosures

gDbusConnectionRegisterObjectWithClosures2

gDbusConnectionRemoveFilter

gDbusConnectionSendMessage

gDbusConnectionSendMessageWithReplyFinish

gDbusConnectionSendMessageWithReplySync

gDbusConnectionSetExitOnClose

gDbusConnectionSignalUnsubscribe

gDbusConnectionStartMessageProcessing

gDbusConnectionUnexportActionGroup

gDbusConnectionUnexportMenuModel

gDbusConnectionUnregisterObject

gDbusConnectionUnregisterSubtree

gDbusErrorEncodeGerror

gDbusErrorGetRemoteError

gDbusErrorIsRemoteError

gDbusErrorNewForDbusError

gDbusErrorQuark

gDbusErrorRegisterError

gDbusErrorRegisterErrorDomain

gDbusErrorStripRemoteError

gDbusErrorUnregisterError

gDbusInterfaceDupObject

gDbusInterfaceGetInfo

gDbusInterfaceSetObject

gDbusInterfaceInfoCacheBuild

gDbusInterfaceInfoCacheRelease

gDbusInterfaceInfoGenerateXml

gDbusInterfaceInfoLookupMethod

gDbusInterfaceInfoLookupProperty

gDbusInterfaceInfoLookupSignal

gDbusInterfaceInfoRef

gDbusInterfaceInfoUnref

gDbusInterfaceSkeletonExport

gDbusInterfaceSkeletonFlush

gDbusInterfaceSkeletonGetConnection

gDbusInterfaceSkeletonGetConnections

gDbusInterfaceSkeletonGetFlags

gDbusInterfaceSkeletonGetInfo

gDbusInterfaceSkeletonGetObjectPath

gDbusInterfaceSkeletonGetProperties

gDbusInterfaceSkeletonGetVtable

gDbusInterfaceSkeletonHasConnection

gDbusInterfaceSkeletonSetFlags

gDbusInterfaceSkeletonUnexport

gDbusInterfaceSkeletonUnexportFromConnection

gDbusMenuModelGet

gDbusMessageNew

gDbusMessageNewFromBlob

gDbusMessageNewMethodCall

gDbusMessageNewSignal

gDbusMessageBytesNeeded

gDbusMessageCopy

gDbusMessageGetArg0

gDbusMessageGetArg0Path

gDbusMessageGetBody

gDbusMessageGetByteOrder

gDbusMessageGetDestination

gDbusMessageGetErrorName

gDbusMessageGetFlags

gDbusMessageGetHeader

gDbusMessageGetHeaderFields

gDbusMessageGetInterface

gDbusMessageGetLocked

gDbusMessageGetMember

gDbusMessageGetMessageType

gDbusMessageGetNumUnixFds

gDbusMessageGetPath

gDbusMessageGetReplySerial

gDbusMessageGetSender

gDbusMessageGetSerial

gDbusMessageGetSignature

gDbusMessageGetUnixFdList

gDbusMessageLock

gDbusMessageNewMethodErrorLiteral

gDbusMessageNewMethodReply

gDbusMessagePrint

gDbusMessageSetBody

gDbusMessageSetByteOrder

gDbusMessageSetDestination

gDbusMessageSetErrorName

gDbusMessageSetFlags

gDbusMessageSetHeader

gDbusMessageSetInterface

gDbusMessageSetMember

gDbusMessageSetMessageType

gDbusMessageSetNumUnixFds

gDbusMessageSetPath

gDbusMessageSetReplySerial

gDbusMessageSetSender

gDbusMessageSetSerial

gDbusMessageSetSignature

gDbusMessageSetUnixFdList

gDbusMessageToBlob

gDbusMessageToGerror

gDbusMethodInfoRef

gDbusMethodInfoUnref

gDbusMethodInvocationGetConnection

gDbusMethodInvocationGetInterfaceName

gDbusMethodInvocationGetMessage

gDbusMethodInvocationGetMethodInfo

gDbusMethodInvocationGetMethodName

gDbusMethodInvocationGetObjectPath

gDbusMethodInvocationGetParameters

gDbusMethodInvocationGetPropertyInfo

gDbusMethodInvocationGetSender

gDbusMethodInvocationReturnDbusError

gDbusMethodInvocationReturnErrorLiteral

gDbusMethodInvocationReturnGerror

gDbusMethodInvocationReturnValue

gDbusMethodInvocationReturnValueWithUnixFdList

gDbusNodeInfoNewForXml

gDbusNodeInfoGenerateXml

gDbusNodeInfoLookupInterface

gDbusNodeInfoRef

gDbusNodeInfoUnref

gDbusObjectGetInterface

gDbusObjectGetInterfaces

gDbusObjectGetObjectPath

gDbusObjectManagerGetInterface

gDbusObjectManagerGetObject

gDbusObjectManagerGetObjectPath

gDbusObjectManagerGetObjects

gDbusObjectManagerClientNewFinish

gDbusObjectManagerClientNewForBusFinish

gDbusObjectManagerClientGetConnection

gDbusObjectManagerClientGetFlags

gDbusObjectManagerClientGetName

gDbusObjectManagerClientGetNameOwner

gDbusObjectManagerServerNew

gDbusObjectManagerServerExport

gDbusObjectManagerServerExportUniquely

gDbusObjectManagerServerGetConnection

gDbusObjectManagerServerIsExported

gDbusObjectManagerServerSetConnection

gDbusObjectManagerServerUnexport

gDbusObjectProxyNew

gDbusObjectProxyGetConnection

gDbusObjectSkeletonNew

gDbusObjectSkeletonAddInterface

gDbusObjectSkeletonFlush

gDbusObjectSkeletonRemoveInterface

gDbusObjectSkeletonRemoveInterfaceByName

gDbusObjectSkeletonSetObjectPath

gDbusPropertyInfoRef

gDbusPropertyInfoUnref

gDbusProxyNewFinish

gDbusProxyNewForBusFinish

gDbusProxyNewForBusSync

gDbusProxyNewSync

gDbusProxyCallFinish

gDbusProxyCallSync

gDbusProxyCallWithUnixFdListFinish

gDbusProxyCallWithUnixFdListSync

gDbusProxyGetCachedProperty

gDbusProxyGetCachedPropertyNames

gDbusProxyGetConnection

gDbusProxyGetDefaultTimeout

gDbusProxyGetFlags

gDbusProxyGetInterfaceInfo

gDbusProxyGetInterfaceName

gDbusProxyGetName

gDbusProxyGetNameOwner

gDbusProxyGetObjectPath

gDbusProxySetCachedProperty

gDbusProxySetDefaultTimeout

gDbusProxySetInterfaceInfo

gDbusServerNewSync

gDbusServerGetClientAddress

gDbusServerGetFlags

gDbusServerGetGuid

gDbusServerIsActive

gDbusServerStart

gDbusServerStop

gDbusSignalInfoRef

gDbusSignalInfoUnref

gDbusAddressEscapeValue

gDbusAddressGetForBusSync

gDbusAddressGetStreamFinish

gDbusAddressGetStreamSync

gDbusEscapeObjectPath

gDbusEscapeObjectPathBytestring

gDbusGenerateGuid

gDbusGvalueToGvariant

gDbusGvariantToGvalue

gDbusIsAddress

gDbusIsErrorName

gDbusIsGuid

gDbusIsInterfaceName

gDbusIsMemberName

gDbusIsName

gDbusIsSupportedAddress

gDbusIsUniqueName

gDbusUnescapeObjectPath
}
