% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R, R/Gio_autogen.R
\name{gFileErrorFromErrno}
\alias{gFileErrorFromErrno}
\alias{gFileErrorQuark}
\alias{gFileGetContents}
\alias{gFileOpenTmp}
\alias{gFileReadLink}
\alias{gFileSetContents}
\alias{gFileTest}
\alias{gFileNewForCommandlineArg}
\alias{gFileNewForCommandlineArgAndCwd}
\alias{gFileNewForPath}
\alias{gFileNewForUri}
\alias{gFileNewTmp}
\alias{gFileParseName}
\alias{gFileAppendTo}
\alias{gFileAppendToFinish}
\alias{gFileCopyAttributes}
\alias{gFileCopyFinish}
\alias{gFileCreate}
\alias{gFileCreateFinish}
\alias{gFileCreateReadwrite}
\alias{gFileCreateReadwriteFinish}
\alias{gFileDelete}
\alias{gFileDeleteFinish}
\alias{gFileDup}
\alias{gFileEjectMountableFinish}
\alias{gFileEjectMountableWithOperationFinish}
\alias{gFileEnumerateChildren}
\alias{gFileEnumerateChildrenFinish}
\alias{gFileEqual}
\alias{gFileFindEnclosingMount}
\alias{gFileFindEnclosingMountFinish}
\alias{gFileGetBasename}
\alias{gFileGetChild}
\alias{gFileGetChildForDisplayName}
\alias{gFileGetParent}
\alias{gFileGetParseName}
\alias{gFileGetPath}
\alias{gFileGetRelativePath}
\alias{gFileGetUri}
\alias{gFileGetUriScheme}
\alias{gFileHasParent}
\alias{gFileHasPrefix}
\alias{gFileHasUriScheme}
\alias{gFileHash}
\alias{gFileIsNative}
\alias{gFileLoadBytes}
\alias{gFileLoadBytesFinish}
\alias{gFileLoadContents}
\alias{gFileLoadContentsFinish}
\alias{gFileLoadPartialContentsFinish}
\alias{gFileMakeDirectory}
\alias{gFileMakeDirectoryFinish}
\alias{gFileMakeDirectoryWithParents}
\alias{gFileMakeSymbolicLink}
\alias{gFileMeasureDiskUsageFinish}
\alias{gFileMonitor}
\alias{gFileMonitorDirectory}
\alias{gFileMonitorFile}
\alias{gFileMountEnclosingVolumeFinish}
\alias{gFileMountMountableFinish}
\alias{gFileOpenReadwrite}
\alias{gFileOpenReadwriteFinish}
\alias{gFilePeekPath}
\alias{gFilePollMountableFinish}
\alias{gFileQueryDefaultHandler}
\alias{gFileQueryExists}
\alias{gFileQueryFileType}
\alias{gFileQueryFilesystemInfo}
\alias{gFileQueryFilesystemInfoFinish}
\alias{gFileQueryInfo}
\alias{gFileQueryInfoFinish}
\alias{gFileQuerySettableAttributes}
\alias{gFileQueryWritableNamespaces}
\alias{gFileRead}
\alias{gFileReadFinish}
\alias{gFileReplace}
\alias{gFileReplaceContents}
\alias{gFileReplaceContentsFinish}
\alias{gFileReplaceFinish}
\alias{gFileReplaceReadwrite}
\alias{gFileReplaceReadwriteFinish}
\alias{gFileResolveRelativePath}
\alias{gFileSetAttribute}
\alias{gFileSetAttributeByteString}
\alias{gFileSetAttributeInt32}
\alias{gFileSetAttributeInt64}
\alias{gFileSetAttributeString}
\alias{gFileSetAttributeUint32}
\alias{gFileSetAttributeUint64}
\alias{gFileSetAttributesFinish}
\alias{gFileSetAttributesFromInfo}
\alias{gFileSetDisplayName}
\alias{gFileSetDisplayNameFinish}
\alias{gFileStartMountableFinish}
\alias{gFileStopMountableFinish}
\alias{gFileSupportsThreadContexts}
\alias{gFileTrash}
\alias{gFileTrashFinish}
\alias{gFileUnmountMountableFinish}
\alias{gFileUnmountMountableWithOperationFinish}
\alias{gFileAttributeInfoListNew}
\alias{gFileAttributeInfoListAdd}
\alias{gFileAttributeInfoListDup}
\alias{gFileAttributeInfoListLookup}
\alias{gFileAttributeInfoListRef}
\alias{gFileAttributeInfoListUnref}
\alias{gFileAttributeMatcherNew}
\alias{gFileAttributeMatcherEnumerateNamespace}
\alias{gFileAttributeMatcherEnumerateNext}
\alias{gFileAttributeMatcherMatches}
\alias{gFileAttributeMatcherMatchesOnly}
\alias{gFileAttributeMatcherRef}
\alias{gFileAttributeMatcherSubtract}
\alias{gFileAttributeMatcherToString}
\alias{gFileAttributeMatcherUnref}
\alias{gFileEnumeratorClose}
\alias{gFileEnumeratorCloseFinish}
\alias{gFileEnumeratorGetChild}
\alias{gFileEnumeratorGetContainer}
\alias{gFileEnumeratorHasPending}
\alias{gFileEnumeratorIsClosed}
\alias{gFileEnumeratorIterate}
\alias{gFileEnumeratorNextFile}
\alias{gFileEnumeratorNextFilesFinish}
\alias{gFileEnumeratorSetPending}
\alias{gFileIoStreamGetEtag}
\alias{gFileIoStreamQueryInfo}
\alias{gFileIoStreamQueryInfoFinish}
\alias{gFileIconNew}
\alias{gFileIconGetFile}
\alias{gFileInfoNew}
\alias{gFileInfoClearStatus}
\alias{gFileInfoCopyInto}
\alias{gFileInfoDup}
\alias{gFileInfoGetAttributeAsString}
\alias{gFileInfoGetAttributeBoolean}
\alias{gFileInfoGetAttributeByteString}
\alias{gFileInfoGetAttributeData}
\alias{gFileInfoGetAttributeInt32}
\alias{gFileInfoGetAttributeInt64}
\alias{gFileInfoGetAttributeObject}
\alias{gFileInfoGetAttributeStatus}
\alias{gFileInfoGetAttributeString}
\alias{gFileInfoGetAttributeStringv}
\alias{gFileInfoGetAttributeType}
\alias{gFileInfoGetAttributeUint32}
\alias{gFileInfoGetAttributeUint64}
\alias{gFileInfoGetContentType}
\alias{gFileInfoGetDeletionDate}
\alias{gFileInfoGetDisplayName}
\alias{gFileInfoGetEditName}
\alias{gFileInfoGetEtag}
\alias{gFileInfoGetFileType}
\alias{gFileInfoGetIcon}
\alias{gFileInfoGetIsBackup}
\alias{gFileInfoGetIsHidden}
\alias{gFileInfoGetIsSymlink}
\alias{gFileInfoGetModificationTime}
\alias{gFileInfoGetName}
\alias{gFileInfoGetSize}
\alias{gFileInfoGetSortOrder}
\alias{gFileInfoGetSymbolicIcon}
\alias{gFileInfoGetSymlinkTarget}
\alias{gFileInfoHasAttribute}
\alias{gFileInfoHasNamespace}
\alias{gFileInfoListAttributes}
\alias{gFileInfoRemoveAttribute}
\alias{gFileInfoSetAttribute}
\alias{gFileInfoSetAttributeBoolean}
\alias{gFileInfoSetAttributeByteString}
\alias{gFileInfoSetAttributeInt32}
\alias{gFileInfoSetAttributeInt64}
\alias{gFileInfoSetAttributeMask}
\alias{gFileInfoSetAttributeObject}
\alias{gFileInfoSetAttributeStatus}
\alias{gFileInfoSetAttributeString}
\alias{gFileInfoSetAttributeStringv}
\alias{gFileInfoSetAttributeUint32}
\alias{gFileInfoSetAttributeUint64}
\alias{gFileInfoSetContentType}
\alias{gFileInfoSetDisplayName}
\alias{gFileInfoSetEditName}
\alias{gFileInfoSetFileType}
\alias{gFileInfoSetIcon}
\alias{gFileInfoSetIsHidden}
\alias{gFileInfoSetIsSymlink}
\alias{gFileInfoSetModificationTime}
\alias{gFileInfoSetName}
\alias{gFileInfoSetSize}
\alias{gFileInfoSetSortOrder}
\alias{gFileInfoSetSymbolicIcon}
\alias{gFileInfoSetSymlinkTarget}
\alias{gFileInfoUnsetAttributeMask}
\alias{gFileInputStreamQueryInfo}
\alias{gFileInputStreamQueryInfoFinish}
\alias{gFileMonitorCancel}
\alias{gFileMonitorEmitEvent}
\alias{gFileMonitorIsCancelled}
\alias{gFileMonitorSetRateLimit}
\alias{gFileOutputStreamGetEtag}
\alias{gFileOutputStreamQueryInfo}
\alias{gFileOutputStreamQueryInfoFinish}
\title{gFileErrorFromErrno}
\usage{
gFileErrorFromErrno(err_no)

gFileErrorQuark()

gFileGetContents(filename)

gFileOpenTmp(tmpl)

gFileReadLink(filename)

gFileSetContents(filename, contents, length)

gFileTest(filename, test)

gFileNewForCommandlineArg(arg)

gFileNewForCommandlineArgAndCwd(arg, cwd)

gFileNewForPath(path)

gFileNewForUri(uri)

gFileNewTmp(tmpl)

gFileParseName(parse_name)

gFileAppendTo(file, flags, cancellable)

gFileAppendToFinish(file, res)

gFileCopyAttributes(source, destination, flags, cancellable)

gFileCopyFinish(file, res)

gFileCreate(file, flags, cancellable)

gFileCreateFinish(file, res)

gFileCreateReadwrite(file, flags, cancellable)

gFileCreateReadwriteFinish(file, res)

gFileDelete(file, cancellable)

gFileDeleteFinish(file, result)

gFileDup(file)

gFileEjectMountableFinish(file, result)

gFileEjectMountableWithOperationFinish(file, result)

gFileEnumerateChildren(file, attributes, flags, cancellable)

gFileEnumerateChildrenFinish(file, res)

gFileEqual(file1, file2)

gFileFindEnclosingMount(file, cancellable)

gFileFindEnclosingMountFinish(file, res)

gFileGetBasename(file)

gFileGetChild(file, name)

gFileGetChildForDisplayName(file, display_name)

gFileGetParent(file)

gFileGetParseName(file)

gFileGetPath(file)

gFileGetRelativePath(parent, descendant)

gFileGetUri(file)

gFileGetUriScheme(file)

gFileHasParent(file, parent)

gFileHasPrefix(file, prefix)

gFileHasUriScheme(file, uri_scheme)

gFileHash(file)

gFileIsNative(file)

gFileLoadBytes(file, cancellable)

gFileLoadBytesFinish(file, result)

gFileLoadContents(file, cancellable)

gFileLoadContentsFinish(file, res)

gFileLoadPartialContentsFinish(file, res)

gFileMakeDirectory(file, cancellable)

gFileMakeDirectoryFinish(file, result)

gFileMakeDirectoryWithParents(file, cancellable)

gFileMakeSymbolicLink(file, symlink_value, cancellable)

gFileMeasureDiskUsageFinish(file, result)

gFileMonitor(file, flags, cancellable)

gFileMonitorDirectory(file, flags, cancellable)

gFileMonitorFile(file, flags, cancellable)

gFileMountEnclosingVolumeFinish(location, result)

gFileMountMountableFinish(file, result)

gFileOpenReadwrite(file, cancellable)

gFileOpenReadwriteFinish(file, res)

gFilePeekPath(file)

gFilePollMountableFinish(file, result)

gFileQueryDefaultHandler(file, cancellable)

gFileQueryExists(file, cancellable)

gFileQueryFileType(file, flags, cancellable)

gFileQueryFilesystemInfo(file, attributes, cancellable)

gFileQueryFilesystemInfoFinish(file, res)

gFileQueryInfo(file, attributes, flags, cancellable)

gFileQueryInfoFinish(file, res)

gFileQuerySettableAttributes(file, cancellable)

gFileQueryWritableNamespaces(file, cancellable)

gFileRead(file, cancellable)

gFileReadFinish(file, res)

gFileReplace(file, etag, make_backup, flags, cancellable)

gFileReplaceContents(
  file,
  contents,
  length,
  etag,
  make_backup,
  flags,
  cancellable
)

gFileReplaceContentsFinish(file, res)

gFileReplaceFinish(file, res)

gFileReplaceReadwrite(file, etag, make_backup, flags, cancellable)

gFileReplaceReadwriteFinish(file, res)

gFileResolveRelativePath(file, relative_path)

gFileSetAttribute(file, attribute, type, value_p, flags, cancellable)

gFileSetAttributeByteString(file, attribute, value, flags, cancellable)

gFileSetAttributeInt32(file, attribute, value, flags, cancellable)

gFileSetAttributeInt64(file, attribute, value, flags, cancellable)

gFileSetAttributeString(file, attribute, value, flags, cancellable)

gFileSetAttributeUint32(file, attribute, value, flags, cancellable)

gFileSetAttributeUint64(file, attribute, value, flags, cancellable)

gFileSetAttributesFinish(file, result)

gFileSetAttributesFromInfo(file, info, flags, cancellable)

gFileSetDisplayName(file, display_name, cancellable)

gFileSetDisplayNameFinish(file, res)

gFileStartMountableFinish(file, result)

gFileStopMountableFinish(file, result)

gFileSupportsThreadContexts(file)

gFileTrash(file, cancellable)

gFileTrashFinish(file, result)

gFileUnmountMountableFinish(file, result)

gFileUnmountMountableWithOperationFinish(file, result)

gFileAttributeInfoListNew()

gFileAttributeInfoListAdd(list, name, type, flags)

gFileAttributeInfoListDup(list)

gFileAttributeInfoListLookup(list, name)

gFileAttributeInfoListRef(list)

gFileAttributeInfoListUnref(list)

gFileAttributeMatcherNew(attributes)

gFileAttributeMatcherEnumerateNamespace(matcher, ns)

gFileAttributeMatcherEnumerateNext(matcher)

gFileAttributeMatcherMatches(matcher, attribute)

gFileAttributeMatcherMatchesOnly(matcher, attribute)

gFileAttributeMatcherRef(matcher)

gFileAttributeMatcherSubtract(matcher, subtract)

gFileAttributeMatcherToString(matcher)

gFileAttributeMatcherUnref(matcher)

gFileEnumeratorClose(enumerator, cancellable)

gFileEnumeratorCloseFinish(enumerator, result)

gFileEnumeratorGetChild(enumerator, info)

gFileEnumeratorGetContainer(enumerator)

gFileEnumeratorHasPending(enumerator)

gFileEnumeratorIsClosed(enumerator)

gFileEnumeratorIterate(direnum, cancellable)

gFileEnumeratorNextFile(enumerator, cancellable)

gFileEnumeratorNextFilesFinish(enumerator, result)

gFileEnumeratorSetPending(enumerator, pending)

gFileIoStreamGetEtag(stream)

gFileIoStreamQueryInfo(stream, attributes, cancellable)

gFileIoStreamQueryInfoFinish(stream, result)

gFileIconNew(file)

gFileIconGetFile(icon)

gFileInfoNew()

gFileInfoClearStatus(info)

gFileInfoCopyInto(src_info, dest_info)

gFileInfoDup(other)

gFileInfoGetAttributeAsString(info, attribute)

gFileInfoGetAttributeBoolean(info, attribute)

gFileInfoGetAttributeByteString(info, attribute)

gFileInfoGetAttributeData(info, attribute)

gFileInfoGetAttributeInt32(info, attribute)

gFileInfoGetAttributeInt64(info, attribute)

gFileInfoGetAttributeObject(info, attribute)

gFileInfoGetAttributeStatus(info, attribute)

gFileInfoGetAttributeString(info, attribute)

gFileInfoGetAttributeStringv(info, attribute)

gFileInfoGetAttributeType(info, attribute)

gFileInfoGetAttributeUint32(info, attribute)

gFileInfoGetAttributeUint64(info, attribute)

gFileInfoGetContentType(info)

gFileInfoGetDeletionDate(info)

gFileInfoGetDisplayName(info)

gFileInfoGetEditName(info)

gFileInfoGetEtag(info)

gFileInfoGetFileType(info)

gFileInfoGetIcon(info)

gFileInfoGetIsBackup(info)

gFileInfoGetIsHidden(info)

gFileInfoGetIsSymlink(info)

gFileInfoGetModificationTime(info)

gFileInfoGetName(info)

gFileInfoGetSize(info)

gFileInfoGetSortOrder(info)

gFileInfoGetSymbolicIcon(info)

gFileInfoGetSymlinkTarget(info)

gFileInfoHasAttribute(info, attribute)

gFileInfoHasNamespace(info, name_space)

gFileInfoListAttributes(info, name_space)

gFileInfoRemoveAttribute(info, attribute)

gFileInfoSetAttribute(info, attribute, type, value_p)

gFileInfoSetAttributeBoolean(info, attribute, attr_value)

gFileInfoSetAttributeByteString(info, attribute, attr_value)

gFileInfoSetAttributeInt32(info, attribute, attr_value)

gFileInfoSetAttributeInt64(info, attribute, attr_value)

gFileInfoSetAttributeMask(info, mask)

gFileInfoSetAttributeObject(info, attribute, attr_value)

gFileInfoSetAttributeStatus(info, attribute, status)

gFileInfoSetAttributeString(info, attribute, attr_value)

gFileInfoSetAttributeStringv(info, attribute, attr_value)

gFileInfoSetAttributeUint32(info, attribute, attr_value)

gFileInfoSetAttributeUint64(info, attribute, attr_value)

gFileInfoSetContentType(info, content_type)

gFileInfoSetDisplayName(info, display_name)

gFileInfoSetEditName(info, edit_name)

gFileInfoSetFileType(info, type)

gFileInfoSetIcon(info, icon)

gFileInfoSetIsHidden(info, is_hidden)

gFileInfoSetIsSymlink(info, is_symlink)

gFileInfoSetModificationTime(info, mtime)

gFileInfoSetName(info, name)

gFileInfoSetSize(info, size)

gFileInfoSetSortOrder(info, sort_order)

gFileInfoSetSymbolicIcon(info, icon)

gFileInfoSetSymlinkTarget(info, symlink_target)

gFileInfoUnsetAttributeMask(info)

gFileInputStreamQueryInfo(stream, attributes, cancellable)

gFileInputStreamQueryInfoFinish(stream, result)

gFileMonitorCancel(monitor)

gFileMonitorEmitEvent(monitor, child, other_file, event_type)

gFileMonitorIsCancelled(monitor)

gFileMonitorSetRateLimit(monitor, limit_msecs)

gFileOutputStreamGetEtag(stream)

gFileOutputStreamQueryInfo(stream, attributes, cancellable)

gFileOutputStreamQueryInfoFinish(stream, result)
}
\arguments{
\item{err_no}{gint}

\item{filename}{filename}

\item{tmpl}{filename}

\item{contents}{guint8}

\item{length}{gsize}

\item{test}{FileTest}

\item{arg}{filename}

\item{cwd}{filename}

\item{path}{filename}

\item{uri}{utf8}

\item{parse_name}{utf8}

\item{file}{File}

\item{flags}{FileAttributeInfoFlags}

\item{cancellable}{Cancellable}

\item{res}{AsyncResult}

\item{source}{File}

\item{destination}{File}

\item{result}{AsyncResult}

\item{attributes}{utf8}

\item{file1}{File}

\item{file2}{File}

\item{name}{filename}

\item{display_name}{utf8}

\item{parent}{File}

\item{descendant}{File}

\item{prefix}{File}

\item{uri_scheme}{utf8}

\item{symlink_value}{filename}

\item{location}{File}

\item{etag}{utf8}

\item{make_backup}{gboolean}

\item{relative_path}{filename}

\item{attribute}{utf8}

\item{type}{FileType}

\item{value_p}{gpointer}

\item{value}{guint64}

\item{info}{FileInfo}

\item{list}{FileAttributeInfoList}

\item{matcher}{FileAttributeMatcher}

\item{ns}{utf8}

\item{subtract}{FileAttributeMatcher}

\item{enumerator}{FileEnumerator}

\item{direnum}{FileEnumerator}

\item{pending}{gboolean}

\item{stream}{FileOutputStream}

\item{icon}{Icon}

\item{src_info}{FileInfo}

\item{dest_info}{FileInfo}

\item{other}{FileInfo}

\item{name_space}{utf8}

\item{attr_value}{guint64}

\item{mask}{FileAttributeMatcher}

\item{status}{FileAttributeStatus}

\item{content_type}{utf8}

\item{edit_name}{utf8}

\item{is_hidden}{gboolean}

\item{is_symlink}{gboolean}

\item{mtime}{GLib.TimeVal}

\item{size}{gint64}

\item{sort_order}{gint32}

\item{symlink_target}{filename}

\item{monitor}{FileMonitor}

\item{child}{File}

\item{other_file}{File}

\item{event_type}{FileMonitorEvent}

\item{limit_msecs}{gint}
}
\value{
FileError

Quark

gboolean

gint

filename

gboolean

gboolean

File

File

File

File

File

File

FileOutputStream

FileOutputStream

gboolean

gboolean

FileOutputStream

FileOutputStream

FileIOStream

FileIOStream

gboolean

gboolean

File

gboolean

gboolean

FileEnumerator

FileEnumerator

gboolean

Mount

Mount

filename

File

File

File

utf8

filename

filename

utf8

utf8

gboolean

gboolean

gboolean

guint

gboolean

GLib.Bytes

GLib.Bytes

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

FileMonitor

FileMonitor

FileMonitor

gboolean

File

FileIOStream

FileIOStream

filename

gboolean

AppInfo

gboolean

FileType

FileInfo

FileInfo

FileInfo

FileInfo

FileAttributeInfoList

FileAttributeInfoList

FileInputStream

FileInputStream

FileOutputStream

gboolean

gboolean

FileOutputStream

FileIOStream

FileIOStream

File

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

File

File

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

FileAttributeInfoList

Return value from C function

FileAttributeInfoList

FileAttributeInfo

FileAttributeInfoList

Return value from C function

FileAttributeMatcher

gboolean

utf8

gboolean

gboolean

FileAttributeMatcher

FileAttributeMatcher

utf8

Return value from C function

gboolean

gboolean

File

File

gboolean

gboolean

gboolean

FileInfo

GLib.List

Return value from C function

utf8

FileInfo

FileInfo

FileIcon

File

FileInfo

Return value from C function

Return value from C function

FileInfo

utf8

gboolean

utf8

gboolean

gint32

gint64

GObject.Object

FileAttributeStatus

utf8

utf8

FileAttributeType

guint32

guint64

utf8

GLib.DateTime

utf8

utf8

utf8

FileType

Icon

gboolean

gboolean

gboolean

Return value from C function

filename

gint64

gint32

Icon

filename

gboolean

gboolean

utf8

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

FileInfo

FileInfo

gboolean

Return value from C function

gboolean

Return value from C function

utf8

FileInfo

FileInfo
}
\description{
gFileErrorFromErrno

gFileErrorQuark

gFileGetContents

gFileOpenTmp

gFileReadLink

gFileSetContents

gFileTest

gFileNewForCommandlineArg

gFileNewForCommandlineArgAndCwd

gFileNewForPath

gFileNewForUri

gFileNewTmp

gFileParseName

gFileAppendTo

gFileAppendToFinish

gFileCopyAttributes

gFileCopyFinish

gFileCreate

gFileCreateFinish

gFileCreateReadwrite

gFileCreateReadwriteFinish

gFileDelete

gFileDeleteFinish

gFileDup

gFileEjectMountableFinish

gFileEjectMountableWithOperationFinish

gFileEnumerateChildren

gFileEnumerateChildrenFinish

gFileEqual

gFileFindEnclosingMount

gFileFindEnclosingMountFinish

gFileGetBasename

gFileGetChild

gFileGetChildForDisplayName

gFileGetParent

gFileGetParseName

gFileGetPath

gFileGetRelativePath

gFileGetUri

gFileGetUriScheme

gFileHasParent

gFileHasPrefix

gFileHasUriScheme

gFileHash

gFileIsNative

gFileLoadBytes

gFileLoadBytesFinish

gFileLoadContents

gFileLoadContentsFinish

gFileLoadPartialContentsFinish

gFileMakeDirectory

gFileMakeDirectoryFinish

gFileMakeDirectoryWithParents

gFileMakeSymbolicLink

gFileMeasureDiskUsageFinish

gFileMonitor

gFileMonitorDirectory

gFileMonitorFile

gFileMountEnclosingVolumeFinish

gFileMountMountableFinish

gFileOpenReadwrite

gFileOpenReadwriteFinish

gFilePeekPath

gFilePollMountableFinish

gFileQueryDefaultHandler

gFileQueryExists

gFileQueryFileType

gFileQueryFilesystemInfo

gFileQueryFilesystemInfoFinish

gFileQueryInfo

gFileQueryInfoFinish

gFileQuerySettableAttributes

gFileQueryWritableNamespaces

gFileRead

gFileReadFinish

gFileReplace

gFileReplaceContents

gFileReplaceContentsFinish

gFileReplaceFinish

gFileReplaceReadwrite

gFileReplaceReadwriteFinish

gFileResolveRelativePath

gFileSetAttribute

gFileSetAttributeByteString

gFileSetAttributeInt32

gFileSetAttributeInt64

gFileSetAttributeString

gFileSetAttributeUint32

gFileSetAttributeUint64

gFileSetAttributesFinish

gFileSetAttributesFromInfo

gFileSetDisplayName

gFileSetDisplayNameFinish

gFileStartMountableFinish

gFileStopMountableFinish

gFileSupportsThreadContexts

gFileTrash

gFileTrashFinish

gFileUnmountMountableFinish

gFileUnmountMountableWithOperationFinish

gFileAttributeInfoListNew

gFileAttributeInfoListAdd

gFileAttributeInfoListDup

gFileAttributeInfoListLookup

gFileAttributeInfoListRef

gFileAttributeInfoListUnref

gFileAttributeMatcherNew

gFileAttributeMatcherEnumerateNamespace

gFileAttributeMatcherEnumerateNext

gFileAttributeMatcherMatches

gFileAttributeMatcherMatchesOnly

gFileAttributeMatcherRef

gFileAttributeMatcherSubtract

gFileAttributeMatcherToString

gFileAttributeMatcherUnref

gFileEnumeratorClose

gFileEnumeratorCloseFinish

gFileEnumeratorGetChild

gFileEnumeratorGetContainer

gFileEnumeratorHasPending

gFileEnumeratorIsClosed

gFileEnumeratorIterate

gFileEnumeratorNextFile

gFileEnumeratorNextFilesFinish

gFileEnumeratorSetPending

gFileIoStreamGetEtag

gFileIoStreamQueryInfo

gFileIoStreamQueryInfoFinish

gFileIconNew

gFileIconGetFile

gFileInfoNew

gFileInfoClearStatus

gFileInfoCopyInto

gFileInfoDup

gFileInfoGetAttributeAsString

gFileInfoGetAttributeBoolean

gFileInfoGetAttributeByteString

gFileInfoGetAttributeData

gFileInfoGetAttributeInt32

gFileInfoGetAttributeInt64

gFileInfoGetAttributeObject

gFileInfoGetAttributeStatus

gFileInfoGetAttributeString

gFileInfoGetAttributeStringv

gFileInfoGetAttributeType

gFileInfoGetAttributeUint32

gFileInfoGetAttributeUint64

gFileInfoGetContentType

gFileInfoGetDeletionDate

gFileInfoGetDisplayName

gFileInfoGetEditName

gFileInfoGetEtag

gFileInfoGetFileType

gFileInfoGetIcon

gFileInfoGetIsBackup

gFileInfoGetIsHidden

gFileInfoGetIsSymlink

gFileInfoGetModificationTime

gFileInfoGetName

gFileInfoGetSize

gFileInfoGetSortOrder

gFileInfoGetSymbolicIcon

gFileInfoGetSymlinkTarget

gFileInfoHasAttribute

gFileInfoHasNamespace

gFileInfoListAttributes

gFileInfoRemoveAttribute

gFileInfoSetAttribute

gFileInfoSetAttributeBoolean

gFileInfoSetAttributeByteString

gFileInfoSetAttributeInt32

gFileInfoSetAttributeInt64

gFileInfoSetAttributeMask

gFileInfoSetAttributeObject

gFileInfoSetAttributeStatus

gFileInfoSetAttributeString

gFileInfoSetAttributeStringv

gFileInfoSetAttributeUint32

gFileInfoSetAttributeUint64

gFileInfoSetContentType

gFileInfoSetDisplayName

gFileInfoSetEditName

gFileInfoSetFileType

gFileInfoSetIcon

gFileInfoSetIsHidden

gFileInfoSetIsSymlink

gFileInfoSetModificationTime

gFileInfoSetName

gFileInfoSetSize

gFileInfoSetSortOrder

gFileInfoSetSymbolicIcon

gFileInfoSetSymlinkTarget

gFileInfoUnsetAttributeMask

gFileInputStreamQueryInfo

gFileInputStreamQueryInfoFinish

gFileMonitorCancel

gFileMonitorEmitEvent

gFileMonitorIsCancelled

gFileMonitorSetRateLimit

gFileOutputStreamGetEtag

gFileOutputStreamQueryInfo

gFileOutputStreamQueryInfoFinish
}
