% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gio_autogen.R
\name{gSettingsNew}
\alias{gSettingsNew}
\alias{gSettingsNewFull}
\alias{gSettingsNewWithBackend}
\alias{gSettingsNewWithBackendAndPath}
\alias{gSettingsNewWithPath}
\alias{gSettingsListRelocatableSchemas}
\alias{gSettingsListSchemas}
\alias{gSettingsSync}
\alias{gSettingsUnbind}
\alias{gSettingsApply}
\alias{gSettingsBind}
\alias{gSettingsBindWritable}
\alias{gSettingsCreateAction}
\alias{gSettingsDelay}
\alias{gSettingsGetBoolean}
\alias{gSettingsGetChild}
\alias{gSettingsGetDefaultValue}
\alias{gSettingsGetDouble}
\alias{gSettingsGetEnum}
\alias{gSettingsGetFlags}
\alias{gSettingsGetHasUnapplied}
\alias{gSettingsGetInt}
\alias{gSettingsGetInt64}
\alias{gSettingsGetMapped}
\alias{gSettingsGetRange}
\alias{gSettingsGetString}
\alias{gSettingsGetStrv}
\alias{gSettingsGetUint}
\alias{gSettingsGetUint64}
\alias{gSettingsGetUserValue}
\alias{gSettingsGetValue}
\alias{gSettingsIsWritable}
\alias{gSettingsListChildren}
\alias{gSettingsListKeys}
\alias{gSettingsRangeCheck}
\alias{gSettingsReset}
\alias{gSettingsRevert}
\alias{gSettingsSetBoolean}
\alias{gSettingsSetDouble}
\alias{gSettingsSetEnum}
\alias{gSettingsSetFlags}
\alias{gSettingsSetInt}
\alias{gSettingsSetInt64}
\alias{gSettingsSetString}
\alias{gSettingsSetStrv}
\alias{gSettingsSetUint}
\alias{gSettingsSetUint64}
\alias{gSettingsSetValue}
\alias{gSettingsBackendFlattenTree}
\alias{gSettingsBackendGetDefault}
\alias{gSettingsBackendChanged}
\alias{gSettingsBackendChangedTree}
\alias{gSettingsBackendKeysChanged}
\alias{gSettingsBackendPathChanged}
\alias{gSettingsBackendPathWritableChanged}
\alias{gSettingsBackendWritableChanged}
\alias{gSettingsSchemaGetId}
\alias{gSettingsSchemaGetKey}
\alias{gSettingsSchemaGetPath}
\alias{gSettingsSchemaHasKey}
\alias{gSettingsSchemaListChildren}
\alias{gSettingsSchemaListKeys}
\alias{gSettingsSchemaRef}
\alias{gSettingsSchemaUnref}
\alias{gSettingsSchemaKeyGetDefaultValue}
\alias{gSettingsSchemaKeyGetDescription}
\alias{gSettingsSchemaKeyGetName}
\alias{gSettingsSchemaKeyGetRange}
\alias{gSettingsSchemaKeyGetSummary}
\alias{gSettingsSchemaKeyGetValueType}
\alias{gSettingsSchemaKeyRangeCheck}
\alias{gSettingsSchemaKeyRef}
\alias{gSettingsSchemaKeyUnref}
\alias{gSettingsSchemaSourceNewFromDirectory}
\alias{gSettingsSchemaSourceListSchemas}
\alias{gSettingsSchemaSourceLookup}
\alias{gSettingsSchemaSourceRef}
\alias{gSettingsSchemaSourceUnref}
\alias{gSettingsSchemaSourceGetDefault}
\title{gSettingsNew}
\usage{
gSettingsNew(schema_id)

gSettingsNewFull(schema, backend, path)

gSettingsNewWithBackend(schema_id, backend)

gSettingsNewWithBackendAndPath(schema_id, backend, path)

gSettingsNewWithPath(schema_id, path)

gSettingsListRelocatableSchemas()

gSettingsListSchemas()

gSettingsSync()

gSettingsUnbind(object, property)

gSettingsApply(settings)

gSettingsBind(settings, key, object, property, flags)

gSettingsBindWritable(settings, key, object, property, inverted)

gSettingsCreateAction(settings, key)

gSettingsDelay(settings)

gSettingsGetBoolean(settings, key)

gSettingsGetChild(settings, name)

gSettingsGetDefaultValue(settings, key)

gSettingsGetDouble(settings, key)

gSettingsGetEnum(settings, key)

gSettingsGetFlags(settings, key)

gSettingsGetHasUnapplied(settings)

gSettingsGetInt(settings, key)

gSettingsGetInt64(settings, key)

gSettingsGetMapped(settings, key, mapping)

gSettingsGetRange(settings, key)

gSettingsGetString(settings, key)

gSettingsGetStrv(settings, key)

gSettingsGetUint(settings, key)

gSettingsGetUint64(settings, key)

gSettingsGetUserValue(settings, key)

gSettingsGetValue(settings, key)

gSettingsIsWritable(settings, name)

gSettingsListChildren(settings)

gSettingsListKeys(settings)

gSettingsRangeCheck(settings, key, value)

gSettingsReset(settings, key)

gSettingsRevert(settings)

gSettingsSetBoolean(settings, key, value)

gSettingsSetDouble(settings, key, value)

gSettingsSetEnum(settings, key, value)

gSettingsSetFlags(settings, key, value)

gSettingsSetInt(settings, key, value)

gSettingsSetInt64(settings, key, value)

gSettingsSetString(settings, key, value)

gSettingsSetStrv(settings, key, value)

gSettingsSetUint(settings, key, value)

gSettingsSetUint64(settings, key, value)

gSettingsSetValue(settings, key, value)

gSettingsBackendFlattenTree(tree)

gSettingsBackendGetDefault()

gSettingsBackendChanged(backend, key, origin_tag)

gSettingsBackendChangedTree(backend, tree, origin_tag)

gSettingsBackendKeysChanged(backend, path, items, origin_tag)

gSettingsBackendPathChanged(backend, path, origin_tag)

gSettingsBackendPathWritableChanged(backend, path)

gSettingsBackendWritableChanged(backend, key)

gSettingsSchemaGetId(schema)

gSettingsSchemaGetKey(schema, name)

gSettingsSchemaGetPath(schema)

gSettingsSchemaHasKey(schema, name)

gSettingsSchemaListChildren(schema)

gSettingsSchemaListKeys(schema)

gSettingsSchemaRef(schema)

gSettingsSchemaUnref(schema)

gSettingsSchemaKeyGetDefaultValue(key)

gSettingsSchemaKeyGetDescription(key)

gSettingsSchemaKeyGetName(key)

gSettingsSchemaKeyGetRange(key)

gSettingsSchemaKeyGetSummary(key)

gSettingsSchemaKeyGetValueType(key)

gSettingsSchemaKeyRangeCheck(key, value)

gSettingsSchemaKeyRef(key)

gSettingsSchemaKeyUnref(key)

gSettingsSchemaSourceNewFromDirectory(directory, parent, trusted)

gSettingsSchemaSourceListSchemas(source, recursive)

gSettingsSchemaSourceLookup(source, schema_id, recursive)

gSettingsSchemaSourceRef(source)

gSettingsSchemaSourceUnref(source)

gSettingsSchemaSourceGetDefault()
}
\arguments{
\item{schema_id}{utf8}

\item{schema}{SettingsSchema}

\item{backend}{SettingsBackend}

\item{path}{utf8}

\item{object}{GObject.Object}

\item{property}{utf8}

\item{settings}{Settings}

\item{key}{SettingsSchemaKey}

\item{flags}{SettingsBindFlags}

\item{inverted}{gboolean}

\item{name}{utf8}

\item{mapping}{function — SettingsGetMapping callback}

\item{value}{GLib.Variant}

\item{tree}{GLib.Tree}

\item{origin_tag}{gpointer}

\item{items}{utf8}

\item{directory}{filename}

\item{parent}{SettingsSchemaSource}

\item{trusted}{gboolean}

\item{source}{SettingsSchemaSource}

\item{recursive}{gboolean}
}
\value{
Settings

Settings

Settings

Settings

Settings

utf8

utf8

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Action

Return value from C function

gboolean

Settings

GLib.Variant

gdouble

gint

guint

gboolean

gint

gint64

gpointer

GLib.Variant

utf8

utf8

guint

guint64

GLib.Variant

GLib.Variant

gboolean

utf8

utf8

gboolean

Return value from C function

Return value from C function

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

Return value from C function

SettingsBackend

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

utf8

SettingsSchemaKey

utf8

gboolean

utf8

utf8

SettingsSchema

Return value from C function

GLib.Variant

utf8

utf8

GLib.Variant

utf8

GLib.VariantType

gboolean

SettingsSchemaKey

Return value from C function

SettingsSchemaSource

Return value from C function

SettingsSchema

SettingsSchemaSource

Return value from C function

SettingsSchemaSource
}
\description{
gSettingsNew

gSettingsNewFull

gSettingsNewWithBackend

gSettingsNewWithBackendAndPath

gSettingsNewWithPath

gSettingsListRelocatableSchemas

gSettingsListSchemas

gSettingsSync

gSettingsUnbind

gSettingsApply

gSettingsBind

gSettingsBindWritable

gSettingsCreateAction

gSettingsDelay

gSettingsGetBoolean

gSettingsGetChild

gSettingsGetDefaultValue

gSettingsGetDouble

gSettingsGetEnum

gSettingsGetFlags

gSettingsGetHasUnapplied

gSettingsGetInt

gSettingsGetInt64

gSettingsGetMapped

gSettingsGetRange

gSettingsGetString

gSettingsGetStrv

gSettingsGetUint

gSettingsGetUint64

gSettingsGetUserValue

gSettingsGetValue

gSettingsIsWritable

gSettingsListChildren

gSettingsListKeys

gSettingsRangeCheck

gSettingsReset

gSettingsRevert

gSettingsSetBoolean

gSettingsSetDouble

gSettingsSetEnum

gSettingsSetFlags

gSettingsSetInt

gSettingsSetInt64

gSettingsSetString

gSettingsSetStrv

gSettingsSetUint

gSettingsSetUint64

gSettingsSetValue

gSettingsBackendFlattenTree

gSettingsBackendGetDefault

gSettingsBackendChanged

gSettingsBackendChangedTree

gSettingsBackendKeysChanged

gSettingsBackendPathChanged

gSettingsBackendPathWritableChanged

gSettingsBackendWritableChanged

gSettingsSchemaGetId

gSettingsSchemaGetKey

gSettingsSchemaGetPath

gSettingsSchemaHasKey

gSettingsSchemaListChildren

gSettingsSchemaListKeys

gSettingsSchemaRef

gSettingsSchemaUnref

gSettingsSchemaKeyGetDefaultValue

gSettingsSchemaKeyGetDescription

gSettingsSchemaKeyGetName

gSettingsSchemaKeyGetRange

gSettingsSchemaKeyGetSummary

gSettingsSchemaKeyGetValueType

gSettingsSchemaKeyRangeCheck

gSettingsSchemaKeyRef

gSettingsSchemaKeyUnref

gSettingsSchemaSourceNewFromDirectory

gSettingsSchemaSourceListSchemas

gSettingsSchemaSourceLookup

gSettingsSchemaSourceRef

gSettingsSchemaSourceUnref

gSettingsSchemaSourceGetDefault
}
