% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R, R/Gio_autogen.R
\name{gIoChannelNewFile}
\alias{gIoChannelNewFile}
\alias{gIoChannelUnixNew}
\alias{gIoChannelClose}
\alias{gIoChannelFlush}
\alias{gIoChannelGetBufferCondition}
\alias{gIoChannelGetBufferSize}
\alias{gIoChannelGetBuffered}
\alias{gIoChannelGetCloseOnUnref}
\alias{gIoChannelGetEncoding}
\alias{gIoChannelGetFlags}
\alias{gIoChannelGetLineTerm}
\alias{gIoChannelInit}
\alias{gIoChannelRead}
\alias{gIoChannelReadChars}
\alias{gIoChannelReadLine}
\alias{gIoChannelReadLineString}
\alias{gIoChannelReadToEnd}
\alias{gIoChannelReadUnichar}
\alias{gIoChannelRef}
\alias{gIoChannelSeek}
\alias{gIoChannelSeekPosition}
\alias{gIoChannelSetBufferSize}
\alias{gIoChannelSetBuffered}
\alias{gIoChannelSetCloseOnUnref}
\alias{gIoChannelSetEncoding}
\alias{gIoChannelSetFlags}
\alias{gIoChannelSetLineTerm}
\alias{gIoChannelShutdown}
\alias{gIoChannelUnixGetFd}
\alias{gIoChannelUnref}
\alias{gIoChannelWrite}
\alias{gIoChannelWriteChars}
\alias{gIoChannelWriteUnichar}
\alias{gIoChannelErrorFromErrno}
\alias{gIoChannelErrorQuark}
\alias{gIoAddWatchFull}
\alias{gIoCreateWatch}
\alias{gIoExtensionGetName}
\alias{gIoExtensionGetPriority}
\alias{gIoExtensionGetType}
\alias{gIoExtensionPointGetExtensionByName}
\alias{gIoExtensionPointGetExtensions}
\alias{gIoExtensionPointGetRequiredType}
\alias{gIoExtensionPointSetRequiredType}
\alias{gIoExtensionPointImplement}
\alias{gIoExtensionPointLookup}
\alias{gIoExtensionPointRegister}
\alias{gIoSchedulerJobSendToMainloop}
\alias{gIoSchedulerJobSendToMainloopAsync}
\alias{gIoStreamSpliceFinish}
\alias{gIoStreamClearPending}
\alias{gIoStreamClose}
\alias{gIoStreamCloseAsync}
\alias{gIoStreamCloseFinish}
\alias{gIoStreamGetInputStream}
\alias{gIoStreamGetOutputStream}
\alias{gIoStreamHasPending}
\alias{gIoStreamIsClosed}
\alias{gIoStreamSetPending}
\alias{gIoStreamSpliceAsync}
\alias{gInputStreamClearPending}
\alias{gInputStreamClose}
\alias{gInputStreamCloseAsync}
\alias{gInputStreamCloseFinish}
\alias{gInputStreamHasPending}
\alias{gInputStreamIsClosed}
\alias{gInputStreamRead}
\alias{gInputStreamReadAll}
\alias{gInputStreamReadAllAsync}
\alias{gInputStreamReadAllFinish}
\alias{gInputStreamReadAsync}
\alias{gInputStreamReadBytes}
\alias{gInputStreamReadBytesAsync}
\alias{gInputStreamReadBytesFinish}
\alias{gInputStreamReadFinish}
\alias{gInputStreamSetPending}
\alias{gInputStreamSkip}
\alias{gInputStreamSkipAsync}
\alias{gInputStreamSkipFinish}
\alias{gOutputStreamClearPending}
\alias{gOutputStreamClose}
\alias{gOutputStreamCloseAsync}
\alias{gOutputStreamCloseFinish}
\alias{gOutputStreamFlush}
\alias{gOutputStreamFlushAsync}
\alias{gOutputStreamFlushFinish}
\alias{gOutputStreamHasPending}
\alias{gOutputStreamIsClosed}
\alias{gOutputStreamIsClosing}
\alias{gOutputStreamSetPending}
\alias{gOutputStreamSplice}
\alias{gOutputStreamSpliceAsync}
\alias{gOutputStreamSpliceFinish}
\alias{gOutputStreamWrite}
\alias{gOutputStreamWriteAll}
\alias{gOutputStreamWriteAllAsync}
\alias{gOutputStreamWriteAllFinish}
\alias{gOutputStreamWriteAsync}
\alias{gOutputStreamWriteBytes}
\alias{gOutputStreamWriteBytesAsync}
\alias{gOutputStreamWriteBytesFinish}
\alias{gOutputStreamWriteFinish}
\alias{gIoErrorFromErrno}
\alias{gIoErrorQuark}
\alias{gIoModulesLoadAllInDirectory}
\alias{gIoModulesLoadAllInDirectoryWithScope}
\alias{gIoModulesScanAllInDirectory}
\alias{gIoModulesScanAllInDirectoryWithScope}
\alias{gIoSchedulerCancelAllJobs}
\alias{gIoSchedulerPushJob}
\title{gIoChannelNewFile}
\usage{
gIoChannelNewFile(filename, mode)

gIoChannelUnixNew(fd)

gIoChannelClose(channel)

gIoChannelFlush(channel)

gIoChannelGetBufferCondition(channel)

gIoChannelGetBufferSize(channel)

gIoChannelGetBuffered(channel)

gIoChannelGetCloseOnUnref(channel)

gIoChannelGetEncoding(channel)

gIoChannelGetFlags(channel)

gIoChannelGetLineTerm(channel)

gIoChannelInit(channel)

gIoChannelRead(channel, buf, count, bytes_read)

gIoChannelReadChars(channel, count)

gIoChannelReadLine(channel)

gIoChannelReadLineString(channel, buffer, terminator_pos)

gIoChannelReadToEnd(channel)

gIoChannelReadUnichar(channel)

gIoChannelRef(channel)

gIoChannelSeek(channel, offset, type)

gIoChannelSeekPosition(channel, offset, type)

gIoChannelSetBufferSize(channel, size)

gIoChannelSetBuffered(channel, buffered)

gIoChannelSetCloseOnUnref(channel, do_close)

gIoChannelSetEncoding(channel, encoding)

gIoChannelSetFlags(channel, flags)

gIoChannelSetLineTerm(channel, line_term, length)

gIoChannelShutdown(channel, flush)

gIoChannelUnixGetFd(channel)

gIoChannelUnref(channel)

gIoChannelWrite(channel, buf, count, bytes_written)

gIoChannelWriteChars(channel, buf, count)

gIoChannelWriteUnichar(channel, thechar)

gIoChannelErrorFromErrno(en)

gIoChannelErrorQuark()

gIoAddWatchFull(channel, priority, condition, func)

gIoCreateWatch(channel, condition)

gIoExtensionGetName(extension)

gIoExtensionGetPriority(extension)

gIoExtensionGetType(extension)

gIoExtensionPointGetExtensionByName(extension_point, name)

gIoExtensionPointGetExtensions(extension_point)

gIoExtensionPointGetRequiredType(extension_point)

gIoExtensionPointSetRequiredType(extension_point, type)

gIoExtensionPointImplement(
  extension_point_name,
  type,
  extension_name,
  priority
)

gIoExtensionPointLookup(name)

gIoExtensionPointRegister(name)

gIoSchedulerJobSendToMainloop(job, func)

gIoSchedulerJobSendToMainloopAsync(job, func)

gIoStreamSpliceFinish(result)

gIoStreamClearPending(stream)

gIoStreamClose(stream, cancellable)

gIoStreamCloseAsync(stream, io_priority, cancellable, callback)

gIoStreamCloseFinish(stream, result)

gIoStreamGetInputStream(stream)

gIoStreamGetOutputStream(stream)

gIoStreamHasPending(stream)

gIoStreamIsClosed(stream)

gIoStreamSetPending(stream)

gIoStreamSpliceAsync(
  stream1,
  stream2,
  flags,
  io_priority,
  cancellable,
  callback
)

gInputStreamClearPending(stream)

gInputStreamClose(stream, cancellable)

gInputStreamCloseAsync(stream, io_priority, cancellable, callback)

gInputStreamCloseFinish(stream, result)

gInputStreamHasPending(stream)

gInputStreamIsClosed(stream)

gInputStreamRead(stream, count, cancellable)

gInputStreamReadAll(stream, count, cancellable)

gInputStreamReadAllAsync(stream, count, io_priority, cancellable, callback)

gInputStreamReadAllFinish(stream, result)

gInputStreamReadAsync(stream, count, io_priority, cancellable, callback)

gInputStreamReadBytes(stream, count, cancellable)

gInputStreamReadBytesAsync(stream, count, io_priority, cancellable, callback)

gInputStreamReadBytesFinish(stream, result)

gInputStreamReadFinish(stream, result)

gInputStreamSetPending(stream)

gInputStreamSkip(stream, count, cancellable)

gInputStreamSkipAsync(stream, count, io_priority, cancellable, callback)

gInputStreamSkipFinish(stream, result)

gOutputStreamClearPending(stream)

gOutputStreamClose(stream, cancellable)

gOutputStreamCloseAsync(stream, io_priority, cancellable, callback)

gOutputStreamCloseFinish(stream, result)

gOutputStreamFlush(stream, cancellable)

gOutputStreamFlushAsync(stream, io_priority, cancellable, callback)

gOutputStreamFlushFinish(stream, result)

gOutputStreamHasPending(stream)

gOutputStreamIsClosed(stream)

gOutputStreamIsClosing(stream)

gOutputStreamSetPending(stream)

gOutputStreamSplice(stream, source, flags, cancellable)

gOutputStreamSpliceAsync(
  stream,
  source,
  flags,
  io_priority,
  cancellable,
  callback
)

gOutputStreamSpliceFinish(stream, result)

gOutputStreamWrite(stream, buffer, count, cancellable)

gOutputStreamWriteAll(stream, buffer, count, cancellable)

gOutputStreamWriteAllAsync(
  stream,
  buffer,
  count,
  io_priority,
  cancellable,
  callback
)

gOutputStreamWriteAllFinish(stream, result)

gOutputStreamWriteAsync(
  stream,
  buffer,
  count,
  io_priority,
  cancellable,
  callback
)

gOutputStreamWriteBytes(stream, bytes, cancellable)

gOutputStreamWriteBytesAsync(stream, bytes, io_priority, cancellable, callback)

gOutputStreamWriteBytesFinish(stream, result)

gOutputStreamWriteFinish(stream, result)

gIoErrorFromErrno(err_no)

gIoErrorQuark()

gIoModulesLoadAllInDirectory(dirname)

gIoModulesLoadAllInDirectoryWithScope(dirname, scope)

gIoModulesScanAllInDirectory(dirname)

gIoModulesScanAllInDirectoryWithScope(dirname, scope)

gIoSchedulerCancelAllJobs()

gIoSchedulerPushJob(job_func, io_priority, cancellable)
}
\arguments{
\item{filename}{filename}

\item{mode}{utf8}

\item{fd}{gint}

\item{channel}{IOChannel}

\item{buf}{guint8}

\item{count}{gsize}

\item{bytes_read}{gsize}

\item{buffer}{guint8}

\item{terminator_pos}{gsize}

\item{offset}{gint64}

\item{type}{GType}

\item{size}{gsize}

\item{buffered}{gboolean}

\item{do_close}{gboolean}

\item{encoding}{utf8}

\item{flags}{OutputStreamSpliceFlags}

\item{line_term}{utf8}

\item{length}{gint}

\item{flush}{gboolean}

\item{bytes_written}{gsize}

\item{thechar}{gunichar}

\item{en}{gint}

\item{priority}{gint}

\item{condition}{IOCondition}

\item{func}{function — GLib.SourceFunc callback}

\item{extension}{IOExtension}

\item{extension_point}{IOExtensionPoint}

\item{name}{utf8}

\item{extension_point_name}{utf8}

\item{extension_name}{utf8}

\item{job}{IOSchedulerJob}

\item{result}{AsyncResult}

\item{stream}{OutputStream}

\item{cancellable}{Cancellable}

\item{io_priority}{gint}

\item{callback}{function — AsyncReadyCallback callback}

\item{stream1}{IOStream}

\item{stream2}{IOStream}

\item{source}{InputStream}

\item{bytes}{GLib.Bytes}

\item{err_no}{gint}

\item{dirname}{filename}

\item{scope}{IOModuleScope}

\item{job_func}{function — IOSchedulerJobFunc callback}
}
\value{
IOChannel

IOChannel

Return value from C function

IOStatus

IOCondition

gsize

gboolean

gboolean

utf8

IOFlags

utf8

Return value from C function

IOError

IOStatus

IOStatus

IOStatus

IOStatus

IOStatus

IOChannel

IOError

IOStatus

Return value from C function

Return value from C function

Return value from C function

IOStatus

IOStatus

Return value from C function

IOStatus

gint

Return value from C function

IOError

IOStatus

IOStatus

IOChannelError

Quark

guint

Source

utf8

gint

GType

IOExtension

GLib.List

GType

Return value from C function

IOExtension

IOExtensionPoint

IOExtensionPoint

gboolean

Return value from C function

gboolean

Return value from C function

gboolean

Return value from C function

gboolean

InputStream

OutputStream

gboolean

gboolean

gboolean

Return value from C function

Return value from C function

gboolean

Return value from C function

gboolean

gboolean

gboolean

gssize

gboolean

Return value from C function

gboolean

Return value from C function

GLib.Bytes

Return value from C function

GLib.Bytes

gssize

gboolean

gssize

Return value from C function

gssize

Return value from C function

gboolean

Return value from C function

gboolean

gboolean

Return value from C function

gboolean

gboolean

gboolean

gboolean

gboolean

gssize

Return value from C function

gssize

gssize

gboolean

Return value from C function

gboolean

Return value from C function

gssize

Return value from C function

gssize

gssize

IOErrorEnum

GLib.Quark

GLib.List

GLib.List

Return value from C function

Return value from C function

Return value from C function

Return value from C function
}
\description{
gIoChannelNewFile

gIoChannelUnixNew

gIoChannelClose

gIoChannelFlush

gIoChannelGetBufferCondition

gIoChannelGetBufferSize

gIoChannelGetBuffered

gIoChannelGetCloseOnUnref

gIoChannelGetEncoding

gIoChannelGetFlags

gIoChannelGetLineTerm

gIoChannelInit

gIoChannelRead

gIoChannelReadChars

gIoChannelReadLine

gIoChannelReadLineString

gIoChannelReadToEnd

gIoChannelReadUnichar

gIoChannelRef

gIoChannelSeek

gIoChannelSeekPosition

gIoChannelSetBufferSize

gIoChannelSetBuffered

gIoChannelSetCloseOnUnref

gIoChannelSetEncoding

gIoChannelSetFlags

gIoChannelSetLineTerm

gIoChannelShutdown

gIoChannelUnixGetFd

gIoChannelUnref

gIoChannelWrite

gIoChannelWriteChars

gIoChannelWriteUnichar

gIoChannelErrorFromErrno

gIoChannelErrorQuark

gIoAddWatchFull

gIoCreateWatch

gIoExtensionGetName

gIoExtensionGetPriority

gIoExtensionGetType

gIoExtensionPointGetExtensionByName

gIoExtensionPointGetExtensions

gIoExtensionPointGetRequiredType

gIoExtensionPointSetRequiredType

gIoExtensionPointImplement

gIoExtensionPointLookup

gIoExtensionPointRegister

gIoSchedulerJobSendToMainloop

gIoSchedulerJobSendToMainloopAsync

gIoStreamSpliceFinish

gIoStreamClearPending

gIoStreamClose

gIoStreamCloseAsync

gIoStreamCloseFinish

gIoStreamGetInputStream

gIoStreamGetOutputStream

gIoStreamHasPending

gIoStreamIsClosed

gIoStreamSetPending

gIoStreamSpliceAsync

gInputStreamClearPending

gInputStreamClose

gInputStreamCloseAsync

gInputStreamCloseFinish

gInputStreamHasPending

gInputStreamIsClosed

gInputStreamRead

gInputStreamReadAll

gInputStreamReadAllAsync

gInputStreamReadAllFinish

gInputStreamReadAsync

gInputStreamReadBytes

gInputStreamReadBytesAsync

gInputStreamReadBytesFinish

gInputStreamReadFinish

gInputStreamSetPending

gInputStreamSkip

gInputStreamSkipAsync

gInputStreamSkipFinish

gOutputStreamClearPending

gOutputStreamClose

gOutputStreamCloseAsync

gOutputStreamCloseFinish

gOutputStreamFlush

gOutputStreamFlushAsync

gOutputStreamFlushFinish

gOutputStreamHasPending

gOutputStreamIsClosed

gOutputStreamIsClosing

gOutputStreamSetPending

gOutputStreamSplice

gOutputStreamSpliceAsync

gOutputStreamSpliceFinish

gOutputStreamWrite

gOutputStreamWriteAll

gOutputStreamWriteAllAsync

gOutputStreamWriteAllFinish

gOutputStreamWriteAsync

gOutputStreamWriteBytes

gOutputStreamWriteBytesAsync

gOutputStreamWriteBytesFinish

gOutputStreamWriteFinish

gIoErrorFromErrno

gIoErrorQuark

gIoModulesLoadAllInDirectory

gIoModulesLoadAllInDirectoryWithScope

gIoModulesScanAllInDirectory

gIoModulesScanAllInDirectoryWithScope

gIoSchedulerCancelAllJobs

gIoSchedulerPushJob
}
