% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R, R/Gio_autogen.R
\name{gByteArrayAppend}
\alias{gByteArrayAppend}
\alias{gByteArrayFree}
\alias{gByteArrayFreeToBytes}
\alias{gByteArrayNew}
\alias{gByteArrayNewTake}
\alias{gByteArrayPrepend}
\alias{gByteArrayRef}
\alias{gByteArrayRemoveIndex}
\alias{gByteArrayRemoveIndexFast}
\alias{gByteArrayRemoveRange}
\alias{gByteArraySetSize}
\alias{gByteArraySizedNew}
\alias{gByteArraySort}
\alias{gByteArraySortWithData}
\alias{gByteArrayUnref}
\alias{gBytesNew}
\alias{gBytesNewFromBytes}
\alias{gBytesNewTake}
\alias{gBytesCompare}
\alias{gBytesEqual}
\alias{gBytesGetData}
\alias{gBytesGetSize}
\alias{gBytesHash}
\alias{gBytesRef}
\alias{gBytesUnref}
\alias{gBytesUnrefToArray}
\alias{gBytesUnrefToData}
\alias{gBytesIconNew}
\alias{gBytesIconGetBytes}
\title{gByteArrayAppend}
\usage{
gByteArrayAppend(array, data, len)

gByteArrayFree(array, free_segment)

gByteArrayFreeToBytes(array)

gByteArrayNew()

gByteArrayNewTake(data, len)

gByteArrayPrepend(array, data, len)

gByteArrayRef(array)

gByteArrayRemoveIndex(array, index_)

gByteArrayRemoveIndexFast(array, index_)

gByteArrayRemoveRange(array, index_, length)

gByteArraySetSize(array, length)

gByteArraySizedNew(reserved_size)

gByteArraySort(array, compare_func)

gByteArraySortWithData(array, compare_func)

gByteArrayUnref(array)

gBytesNew(data, size)

gBytesNewFromBytes(bytes, offset, length)

gBytesNewTake(data, size)

gBytesCompare(bytes1, bytes2)

gBytesEqual(bytes1, bytes2)

gBytesGetData(bytes)

gBytesGetSize(bytes)

gBytesHash(bytes)

gBytesRef(bytes)

gBytesUnref(bytes)

gBytesUnrefToArray(bytes)

gBytesUnrefToData(bytes)

gBytesIconNew(bytes)

gBytesIconGetBytes(icon)
}
\arguments{
\item{array}{guint8}

\item{data}{guint8}

\item{len}{guint}

\item{free_segment}{gboolean}

\item{index_}{guint}

\item{length}{gsize}

\item{reserved_size}{guint}

\item{compare_func}{function — CompareDataFunc function(a, b)}

\item{size}{gsize}

\item{bytes}{GLib.Bytes}

\item{offset}{gsize}

\item{bytes1}{Bytes}

\item{bytes2}{Bytes}

\item{icon}{BytesIcon}
}
\value{
guint8

guint8

Bytes

guint8

guint8

guint8

guint8

guint8

guint8

guint8

guint8

guint8

Return value from C function

Return value from C function

Return value from C function

Bytes

Bytes

Bytes

gint

gboolean

guint8

gsize

guint

Bytes

Return value from C function

guint8

guint8

BytesIcon

GLib.Bytes
}
\description{
gByteArrayAppend

gByteArrayFree

gByteArrayFreeToBytes

gByteArrayNew

gByteArrayNewTake

gByteArrayPrepend

gByteArrayRef

gByteArrayRemoveIndex

gByteArrayRemoveIndexFast

gByteArrayRemoveRange

gByteArraySetSize

gByteArraySizedNew

gByteArraySort

gByteArraySortWithData

gByteArrayUnref

gBytesNew

gBytesNewFromBytes

gBytesNewTake

gBytesCompare

gBytesEqual

gBytesGetData

gBytesGetSize

gBytesHash

gBytesRef

gBytesUnref

gBytesUnrefToArray

gBytesUnrefToData

gBytesIconNew

gBytesIconGetBytes
}
