% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R
\name{gDateTimeNew}
\alias{gDateTimeNew}
\alias{gDateTimeNewFromIso8601}
\alias{gDateTimeNewFromTimevalLocal}
\alias{gDateTimeNewFromTimevalUtc}
\alias{gDateTimeNewFromUnixLocal}
\alias{gDateTimeNewFromUnixLocalUsec}
\alias{gDateTimeNewFromUnixUtc}
\alias{gDateTimeNewFromUnixUtcUsec}
\alias{gDateTimeNewLocal}
\alias{gDateTimeNewNow}
\alias{gDateTimeNewNowLocal}
\alias{gDateTimeNewNowUtc}
\alias{gDateTimeNewUtc}
\alias{gDateTimeAdd}
\alias{gDateTimeAddDays}
\alias{gDateTimeAddFull}
\alias{gDateTimeAddHours}
\alias{gDateTimeAddMinutes}
\alias{gDateTimeAddMonths}
\alias{gDateTimeAddSeconds}
\alias{gDateTimeAddWeeks}
\alias{gDateTimeAddYears}
\alias{gDateTimeCompare}
\alias{gDateTimeDifference}
\alias{gDateTimeEqual}
\alias{gDateTimeFormat}
\alias{gDateTimeFormatIso8601}
\alias{gDateTimeGetDayOfMonth}
\alias{gDateTimeGetDayOfWeek}
\alias{gDateTimeGetDayOfYear}
\alias{gDateTimeGetHour}
\alias{gDateTimeGetMicrosecond}
\alias{gDateTimeGetMinute}
\alias{gDateTimeGetMonth}
\alias{gDateTimeGetSecond}
\alias{gDateTimeGetSeconds}
\alias{gDateTimeGetTimezone}
\alias{gDateTimeGetTimezoneAbbreviation}
\alias{gDateTimeGetUtcOffset}
\alias{gDateTimeGetWeekNumberingYear}
\alias{gDateTimeGetWeekOfYear}
\alias{gDateTimeGetYear}
\alias{gDateTimeGetYmd}
\alias{gDateTimeHash}
\alias{gDateTimeIsDaylightSavings}
\alias{gDateTimeRef}
\alias{gDateTimeToLocal}
\alias{gDateTimeToTimeval}
\alias{gDateTimeToTimezone}
\alias{gDateTimeToUnix}
\alias{gDateTimeToUnixUsec}
\alias{gDateTimeToUtc}
\alias{gDateTimeUnref}
\alias{gTimeValAdd}
\alias{gTimeValToIso8601}
\alias{gTimeValFromIso8601}
\alias{gTimeZoneNew}
\alias{gTimeZoneNewIdentifier}
\alias{gTimeZoneNewLocal}
\alias{gTimeZoneNewOffset}
\alias{gTimeZoneNewUtc}
\alias{gTimeZoneAdjustTime}
\alias{gTimeZoneFindInterval}
\alias{gTimeZoneGetAbbreviation}
\alias{gTimeZoneGetIdentifier}
\alias{gTimeZoneGetOffset}
\alias{gTimeZoneIsDst}
\alias{gTimeZoneRef}
\alias{gTimeZoneUnref}
\title{gDateTimeNew}
\usage{
gDateTimeNew(tz, year, month, day, hour, minute, seconds)

gDateTimeNewFromIso8601(text, default_tz)

gDateTimeNewFromTimevalLocal(tv)

gDateTimeNewFromTimevalUtc(tv)

gDateTimeNewFromUnixLocal(t)

gDateTimeNewFromUnixLocalUsec(usecs)

gDateTimeNewFromUnixUtc(t)

gDateTimeNewFromUnixUtcUsec(usecs)

gDateTimeNewLocal(year, month, day, hour, minute, seconds)

gDateTimeNewNow(tz)

gDateTimeNewNowLocal()

gDateTimeNewNowUtc()

gDateTimeNewUtc(year, month, day, hour, minute, seconds)

gDateTimeAdd(datetime, timespan)

gDateTimeAddDays(datetime, days)

gDateTimeAddFull(datetime, years, months, days, hours, minutes, seconds)

gDateTimeAddHours(datetime, hours)

gDateTimeAddMinutes(datetime, minutes)

gDateTimeAddMonths(datetime, months)

gDateTimeAddSeconds(datetime, seconds)

gDateTimeAddWeeks(datetime, weeks)

gDateTimeAddYears(datetime, years)

gDateTimeCompare(dt1, dt2)

gDateTimeDifference(end, begin)

gDateTimeEqual(dt1, dt2)

gDateTimeFormat(datetime, format)

gDateTimeFormatIso8601(datetime)

gDateTimeGetDayOfMonth(datetime)

gDateTimeGetDayOfWeek(datetime)

gDateTimeGetDayOfYear(datetime)

gDateTimeGetHour(datetime)

gDateTimeGetMicrosecond(datetime)

gDateTimeGetMinute(datetime)

gDateTimeGetMonth(datetime)

gDateTimeGetSecond(datetime)

gDateTimeGetSeconds(datetime)

gDateTimeGetTimezone(datetime)

gDateTimeGetTimezoneAbbreviation(datetime)

gDateTimeGetUtcOffset(datetime)

gDateTimeGetWeekNumberingYear(datetime)

gDateTimeGetWeekOfYear(datetime)

gDateTimeGetYear(datetime)

gDateTimeGetYmd(datetime)

gDateTimeHash(datetime)

gDateTimeIsDaylightSavings(datetime)

gDateTimeRef(datetime)

gDateTimeToLocal(datetime)

gDateTimeToTimeval(datetime, tv)

gDateTimeToTimezone(datetime, tz)

gDateTimeToUnix(datetime)

gDateTimeToUnixUsec(datetime)

gDateTimeToUtc(datetime)

gDateTimeUnref(datetime)

gTimeValAdd(time_, microseconds)

gTimeValToIso8601(time_)

gTimeValFromIso8601(iso_date)

gTimeZoneNew(identifier)

gTimeZoneNewIdentifier(identifier)

gTimeZoneNewLocal()

gTimeZoneNewOffset(seconds)

gTimeZoneNewUtc()

gTimeZoneAdjustTime(tz, type)

gTimeZoneFindInterval(tz, type, time_)

gTimeZoneGetAbbreviation(tz, interval)

gTimeZoneGetIdentifier(tz)

gTimeZoneGetOffset(tz, interval)

gTimeZoneIsDst(tz, interval)

gTimeZoneRef(tz)

gTimeZoneUnref(tz)
}
\arguments{
\item{tz}{TimeZone}

\item{year}{gint}

\item{month}{gint}

\item{day}{gint}

\item{hour}{gint}

\item{minute}{gint}

\item{seconds}{gint32}

\item{text}{utf8}

\item{default_tz}{TimeZone}

\item{tv}{TimeVal}

\item{t}{gint64}

\item{usecs}{gint64}

\item{datetime}{DateTime}

\item{timespan}{TimeSpan}

\item{days}{gint}

\item{years}{gint}

\item{months}{gint}

\item{hours}{gint}

\item{minutes}{gint}

\item{weeks}{gint}

\item{dt1}{DateTime}

\item{dt2}{DateTime}

\item{end}{DateTime}

\item{begin}{DateTime}

\item{format}{utf8}

\item{time_}{gint64}

\item{microseconds}{glong}

\item{iso_date}{utf8}

\item{identifier}{utf8}

\item{type}{TimeType}

\item{interval}{gint}
}
\value{
DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

DateTime

gint

TimeSpan

gboolean

utf8

utf8

gint

gint

gint

gint

gint

gint

gint

gint

gdouble

TimeZone

utf8

TimeSpan

gint

gint

gint

Return value from C function

guint

gboolean

DateTime

DateTime

gboolean

DateTime

gint64

gint64

DateTime

Return value from C function

Return value from C function

utf8

gboolean

TimeZone

TimeZone

TimeZone

TimeZone

TimeZone

gint

gint

utf8

utf8

gint32

gboolean

TimeZone

Return value from C function
}
\description{
gDateTimeNew

gDateTimeNewFromIso8601

gDateTimeNewFromTimevalLocal

gDateTimeNewFromTimevalUtc

gDateTimeNewFromUnixLocal

gDateTimeNewFromUnixLocalUsec

gDateTimeNewFromUnixUtc

gDateTimeNewFromUnixUtcUsec

gDateTimeNewLocal

gDateTimeNewNow

gDateTimeNewNowLocal

gDateTimeNewNowUtc

gDateTimeNewUtc

gDateTimeAdd

gDateTimeAddDays

gDateTimeAddFull

gDateTimeAddHours

gDateTimeAddMinutes

gDateTimeAddMonths

gDateTimeAddSeconds

gDateTimeAddWeeks

gDateTimeAddYears

gDateTimeCompare

gDateTimeDifference

gDateTimeEqual

gDateTimeFormat

gDateTimeFormatIso8601

gDateTimeGetDayOfMonth

gDateTimeGetDayOfWeek

gDateTimeGetDayOfYear

gDateTimeGetHour

gDateTimeGetMicrosecond

gDateTimeGetMinute

gDateTimeGetMonth

gDateTimeGetSecond

gDateTimeGetSeconds

gDateTimeGetTimezone

gDateTimeGetTimezoneAbbreviation

gDateTimeGetUtcOffset

gDateTimeGetWeekNumberingYear

gDateTimeGetWeekOfYear

gDateTimeGetYear

gDateTimeGetYmd

gDateTimeHash

gDateTimeIsDaylightSavings

gDateTimeRef

gDateTimeToLocal

gDateTimeToTimeval

gDateTimeToTimezone

gDateTimeToUnix

gDateTimeToUnixUsec

gDateTimeToUtc

gDateTimeUnref

gTimeValAdd

gTimeValToIso8601

gTimeValFromIso8601

gTimeZoneNew

gTimeZoneNewIdentifier

gTimeZoneNewLocal

gTimeZoneNewOffset

gTimeZoneNewUtc

gTimeZoneAdjustTime

gTimeZoneFindInterval

gTimeZoneGetAbbreviation

gTimeZoneGetIdentifier

gTimeZoneGetOffset

gTimeZoneIsDst

gTimeZoneRef

gTimeZoneUnref
}
