% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R
\name{gErrorNewLiteral}
\alias{gErrorNewLiteral}
\alias{gErrorCopy}
\alias{gErrorFree}
\alias{gErrorMatches}
\alias{gLogDefaultHandler}
\alias{gLogRemoveHandler}
\alias{gLogSetAlwaysFatal}
\alias{gLogSetFatalMask}
\alias{gLogSetHandlerFull}
\alias{gLogSetWriterFunc}
\alias{gLogStructuredArray}
\alias{gLogVariant}
\alias{gLogWriterDefault}
\alias{gLogWriterFormatFields}
\alias{gLogWriterIsJournald}
\alias{gLogWriterJournald}
\alias{gLogWriterStandardStreams}
\alias{gLogWriterSupportsColor}
\title{gErrorNewLiteral}
\usage{
gErrorNewLiteral(domain, code, message)

gErrorCopy(error)

gErrorFree(error)

gErrorMatches(error, domain, code)

gLogDefaultHandler(log_domain, log_level, message, unused_data)

gLogRemoveHandler(log_domain, handler_id)

gLogSetAlwaysFatal(fatal_mask)

gLogSetFatalMask(log_domain, fatal_mask)

gLogSetHandlerFull(log_domain, log_levels, log_func)

gLogSetWriterFunc(func)

gLogStructuredArray(log_level, fields, n_fields)

gLogVariant(log_domain, log_level, fields)

gLogWriterDefault(log_level, fields, n_fields, user_data)

gLogWriterFormatFields(log_level, fields, n_fields, use_color)

gLogWriterIsJournald(output_fd)

gLogWriterJournald(log_level, fields, n_fields, user_data)

gLogWriterStandardStreams(log_level, fields, n_fields, user_data)

gLogWriterSupportsColor(output_fd)
}
\arguments{
\item{domain}{Quark}

\item{code}{gint}

\item{message}{utf8}

\item{error}{Error}

\item{log_domain}{utf8}

\item{log_level}{LogLevelFlags}

\item{unused_data}{gpointer}

\item{handler_id}{guint}

\item{fatal_mask}{LogLevelFlags}

\item{log_levels}{LogLevelFlags}

\item{log_func}{function — LogFunc callback}

\item{func}{function — LogWriterFunc callback}

\item{fields}{LogField}

\item{n_fields}{gsize}

\item{user_data}{gpointer}

\item{use_color}{gboolean}

\item{output_fd}{gint}
}
\value{
Error

Error

Return value from C function

gboolean

Return value from C function

Return value from C function

LogLevelFlags

LogLevelFlags

guint

Return value from C function

Return value from C function

Return value from C function

LogWriterOutput

utf8

gboolean

LogWriterOutput

LogWriterOutput

gboolean
}
\description{
gErrorNewLiteral

gErrorCopy

gErrorFree

gErrorMatches

gLogDefaultHandler

gLogRemoveHandler

gLogSetAlwaysFatal

gLogSetFatalMask

gLogSetHandlerFull

gLogSetWriterFunc

gLogStructuredArray

gLogVariant

gLogWriterDefault

gLogWriterFormatFields

gLogWriterIsJournald

gLogWriterJournald

gLogWriterStandardStreams

gLogWriterSupportsColor
}
