% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R, R/Gio_autogen.R
\name{gListPopAllocator}
\alias{gListPopAllocator}
\alias{gListPushAllocator}
\alias{gQueueClear}
\alias{gQueueFree}
\alias{gQueueGetLength}
\alias{gQueueIndex}
\alias{gQueueInit}
\alias{gQueueIsEmpty}
\alias{gQueuePeekHead}
\alias{gQueuePeekNth}
\alias{gQueuePeekTail}
\alias{gQueuePopHead}
\alias{gQueuePopNth}
\alias{gQueuePopTail}
\alias{gQueuePushHead}
\alias{gQueuePushNth}
\alias{gQueuePushTail}
\alias{gQueueRemove}
\alias{gQueueRemoveAll}
\alias{gQueueReverse}
\alias{gSlistPopAllocator}
\alias{gSlistPushAllocator}
\alias{gListModelGetItemType}
\alias{gListModelGetNItems}
\alias{gListModelGetObject}
\alias{gListModelItemsChanged}
\alias{gListStoreNew}
\alias{gListStoreAppend}
\alias{gListStoreFind}
\alias{gListStoreFindWithEqualFuncFull}
\alias{gListStoreInsert}
\alias{gListStoreRemove}
\alias{gListStoreRemoveAll}
\alias{gListStoreSplice}
\title{gListPopAllocator}
\usage{
gListPopAllocator()

gListPushAllocator(allocator)

gQueueClear(queue)

gQueueFree(queue)

gQueueGetLength(queue)

gQueueIndex(queue, data)

gQueueInit(queue)

gQueueIsEmpty(queue)

gQueuePeekHead(queue)

gQueuePeekNth(queue, n)

gQueuePeekTail(queue)

gQueuePopHead(queue)

gQueuePopNth(queue, n)

gQueuePopTail(queue)

gQueuePushHead(queue, data)

gQueuePushNth(queue, data, n)

gQueuePushTail(queue, data)

gQueueRemove(queue, data)

gQueueRemoveAll(queue, data)

gQueueReverse(queue)

gSlistPopAllocator()

gSlistPushAllocator(allocator)

gListModelGetItemType(list)

gListModelGetNItems(list)

gListModelGetObject(list, position)

gListModelItemsChanged(list, position, removed, added)

gListStoreNew(item_type)

gListStoreAppend(store, item)

gListStoreFind(store, item)

gListStoreFindWithEqualFuncFull(store, item, equal_func, user_data)

gListStoreInsert(store, position, item)

gListStoreRemove(store, position)

gListStoreRemoveAll(store)

gListStoreSplice(store, position, n_removals, additions, n_additions)
}
\arguments{
\item{allocator}{Allocator}

\item{queue}{Queue}

\item{data}{gpointer}

\item{n}{gint}

\item{list}{ListModel}

\item{position}{guint}

\item{removed}{guint}

\item{added}{guint}

\item{item_type}{GType}

\item{store}{ListStore}

\item{item}{GObject.Object}

\item{equal_func}{GLib.EqualFuncFull}

\item{user_data}{gpointer}

\item{n_removals}{guint}

\item{additions}{GObject.Object}

\item{n_additions}{guint}
}
\value{
Return value from C function

Return value from C function

Return value from C function

Return value from C function

guint

gint

Return value from C function

gboolean

gpointer

gpointer

gpointer

gpointer

gpointer

gpointer

Return value from C function

Return value from C function

Return value from C function

gboolean

guint

Return value from C function

Return value from C function

Return value from C function

GType

guint

GObject.Object

Return value from C function

ListStore

Return value from C function

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function
}
\description{
gListPopAllocator

gListPushAllocator

gQueueClear

gQueueFree

gQueueGetLength

gQueueIndex

gQueueInit

gQueueIsEmpty

gQueuePeekHead

gQueuePeekNth

gQueuePeekTail

gQueuePopHead

gQueuePopNth

gQueuePopTail

gQueuePushHead

gQueuePushNth

gQueuePushTail

gQueueRemove

gQueueRemoveAll

gQueueReverse

gSlistPopAllocator

gSlistPushAllocator

gListModelGetItemType

gListModelGetNItems

gListModelGetObject

gListModelItemsChanged

gListStoreNew

gListStoreAppend

gListStoreFind

gListStoreFindWithEqualFuncFull

gListStoreInsert

gListStoreRemove

gListStoreRemoveAll

gListStoreSplice
}
