% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R, R/GObject_autogen.R
\name{gMainContextNew}
\alias{gMainContextNew}
\alias{gMainContextAcquire}
\alias{gMainContextAddPoll}
\alias{gMainContextCheck}
\alias{gMainContextDispatch}
\alias{gMainContextFindSourceByFuncsUserData}
\alias{gMainContextFindSourceById}
\alias{gMainContextFindSourceByUserData}
\alias{gMainContextIsOwner}
\alias{gMainContextIteration}
\alias{gMainContextPending}
\alias{gMainContextPopThreadDefault}
\alias{gMainContextPrepare}
\alias{gMainContextPushThreadDefault}
\alias{gMainContextQuery}
\alias{gMainContextRef}
\alias{gMainContextRelease}
\alias{gMainContextRemovePoll}
\alias{gMainContextUnref}
\alias{gMainContextWait}
\alias{gMainContextWakeup}
\alias{gMainContextDefault}
\alias{gMainContextGetThreadDefault}
\alias{gMainContextRefThreadDefault}
\alias{gMainLoopNew}
\alias{gMainLoopGetContext}
\alias{gMainLoopIsRunning}
\alias{gMainLoopQuit}
\alias{gMainLoopRef}
\alias{gMainLoopRun}
\alias{gMainLoopUnref}
\alias{gSourceNew}
\alias{gSourceAddChildSource}
\alias{gSourceAddPoll}
\alias{gSourceAttach}
\alias{gSourceDestroy}
\alias{gSourceGetCanRecurse}
\alias{gSourceGetContext}
\alias{gSourceGetCurrentTime}
\alias{gSourceGetId}
\alias{gSourceGetName}
\alias{gSourceGetPriority}
\alias{gSourceGetReadyTime}
\alias{gSourceGetTime}
\alias{gSourceIsDestroyed}
\alias{gSourceRef}
\alias{gSourceRemoveChildSource}
\alias{gSourceRemovePoll}
\alias{gSourceSetCallbackIndirect}
\alias{gSourceSetCanRecurse}
\alias{gSourceSetFuncs}
\alias{gSourceSetName}
\alias{gSourceSetPriority}
\alias{gSourceSetReadyTime}
\alias{gSourceUnref}
\alias{gSourceRemove}
\alias{gSourceRemoveByFuncsUserData}
\alias{gSourceRemoveByUserData}
\alias{gSourceSetNameById}
\alias{gIdleRemoveByData}
\alias{gIdleSourceNew}
\alias{gMainCurrentSource}
\alias{gMainDepth}
\alias{gTimeoutSourceNew}
\alias{gTimeoutSourceNewSeconds}
\alias{gSourceSetClosure}
\alias{gSourceSetDummyCallback}
\title{gMainContextNew}
\usage{
gMainContextNew()

gMainContextAcquire(context)

gMainContextAddPoll(context, fd, priority)

gMainContextCheck(context, max_priority, fds, n_fds)

gMainContextDispatch(context)

gMainContextFindSourceByFuncsUserData(context, funcs, user_data)

gMainContextFindSourceById(context, source_id)

gMainContextFindSourceByUserData(context, user_data)

gMainContextIsOwner(context)

gMainContextIteration(context, may_block)

gMainContextPending(context)

gMainContextPopThreadDefault(context)

gMainContextPrepare(context)

gMainContextPushThreadDefault(context)

gMainContextQuery(context, max_priority, n_fds)

gMainContextRef(context)

gMainContextRelease(context)

gMainContextRemovePoll(context, fd)

gMainContextUnref(context)

gMainContextWait(context, cond, mutex)

gMainContextWakeup(context)

gMainContextDefault()

gMainContextGetThreadDefault()

gMainContextRefThreadDefault()

gMainLoopNew(context, is_running)

gMainLoopGetContext(loop)

gMainLoopIsRunning(loop)

gMainLoopQuit(loop)

gMainLoopRef(loop)

gMainLoopRun(loop)

gMainLoopUnref(loop)

gSourceNew(source_funcs, struct_size)

gSourceAddChildSource(source, child_source)

gSourceAddPoll(source, fd)

gSourceAttach(source, context)

gSourceDestroy(source)

gSourceGetCanRecurse(source)

gSourceGetContext(source)

gSourceGetCurrentTime(source, timeval)

gSourceGetId(source)

gSourceGetName(source)

gSourceGetPriority(source)

gSourceGetReadyTime(source)

gSourceGetTime(source)

gSourceIsDestroyed(source)

gSourceRef(source)

gSourceRemoveChildSource(source, child_source)

gSourceRemovePoll(source, fd)

gSourceSetCallbackIndirect(source, callback_data, callback_funcs)

gSourceSetCanRecurse(source, can_recurse)

gSourceSetFuncs(source, funcs)

gSourceSetName(source, name)

gSourceSetPriority(source, priority)

gSourceSetReadyTime(source, ready_time)

gSourceUnref(source)

gSourceRemove(tag)

gSourceRemoveByFuncsUserData(funcs, user_data)

gSourceRemoveByUserData(user_data)

gSourceSetNameById(tag, name)

gIdleRemoveByData(data)

gIdleSourceNew()

gMainCurrentSource()

gMainDepth()

gTimeoutSourceNew(interval)

gTimeoutSourceNewSeconds(interval)

gSourceSetClosure(source, closure)

gSourceSetDummyCallback(source)
}
\arguments{
\item{context}{MainContext}

\item{fd}{PollFD}

\item{priority}{gint}

\item{max_priority}{gint}

\item{fds}{PollFD}

\item{n_fds}{gint}

\item{funcs}{SourceFuncs}

\item{user_data}{gpointer}

\item{source_id}{guint}

\item{may_block}{gboolean}

\item{cond}{Cond}

\item{mutex}{Mutex}

\item{is_running}{gboolean}

\item{loop}{MainLoop}

\item{source_funcs}{SourceFuncs}

\item{struct_size}{guint}

\item{source}{GLib.Source}

\item{child_source}{Source}

\item{timeval}{TimeVal}

\item{callback_data}{gpointer}

\item{callback_funcs}{SourceCallbackFuncs}

\item{can_recurse}{gboolean}

\item{name}{utf8}

\item{ready_time}{gint64}

\item{tag}{guint}

\item{data}{gpointer}

\item{interval}{guint}

\item{closure}{Closure}
}
\value{
MainContext

gboolean

Return value from C function

gboolean

Return value from C function

Source

Source

Source

gboolean

gboolean

gboolean

Return value from C function

gboolean

Return value from C function

gint

MainContext

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

MainContext

MainContext

MainContext

MainLoop

MainContext

gboolean

Return value from C function

MainLoop

Return value from C function

Return value from C function

Source

Return value from C function

Return value from C function

guint

Return value from C function

gboolean

MainContext

Return value from C function

guint

utf8

gint

gint64

gint64

gboolean

Source

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

gboolean

gboolean

Return value from C function

gboolean

Source

Source

gint

Source

Source

Return value from C function

Return value from C function
}
\description{
gMainContextNew

gMainContextAcquire

gMainContextAddPoll

gMainContextCheck

gMainContextDispatch

gMainContextFindSourceByFuncsUserData

gMainContextFindSourceById

gMainContextFindSourceByUserData

gMainContextIsOwner

gMainContextIteration

gMainContextPending

gMainContextPopThreadDefault

gMainContextPrepare

gMainContextPushThreadDefault

gMainContextQuery

gMainContextRef

gMainContextRelease

gMainContextRemovePoll

gMainContextUnref

gMainContextWait

gMainContextWakeup

gMainContextDefault

gMainContextGetThreadDefault

gMainContextRefThreadDefault

gMainLoopNew

gMainLoopGetContext

gMainLoopIsRunning

gMainLoopQuit

gMainLoopRef

gMainLoopRun

gMainLoopUnref

gSourceNew

gSourceAddChildSource

gSourceAddPoll

gSourceAttach

gSourceDestroy

gSourceGetCanRecurse

gSourceGetContext

gSourceGetCurrentTime

gSourceGetId

gSourceGetName

gSourceGetPriority

gSourceGetReadyTime

gSourceGetTime

gSourceIsDestroyed

gSourceRef

gSourceRemoveChildSource

gSourceRemovePoll

gSourceSetCallbackIndirect

gSourceSetCanRecurse

gSourceSetFuncs

gSourceSetName

gSourceSetPriority

gSourceSetReadyTime

gSourceUnref

gSourceRemove

gSourceRemoveByFuncsUserData

gSourceRemoveByUserData

gSourceSetNameById

gIdleRemoveByData

gIdleSourceNew

gMainCurrentSource

gMainDepth

gTimeoutSourceNew

gTimeoutSourceNewSeconds

gSourceSetClosure

gSourceSetDummyCallback
}
