% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R, R/GObject_autogen.R
\name{gStringNew}
\alias{gStringNew}
\alias{gStringNewLen}
\alias{gStringNewTake}
\alias{gStringSizedNew}
\alias{gStringAppend}
\alias{gStringAppendC}
\alias{gStringAppendLen}
\alias{gStringAppendUnichar}
\alias{gStringAppendUriEscaped}
\alias{gStringAsciiDown}
\alias{gStringAsciiUp}
\alias{gStringAssign}
\alias{gStringCopy}
\alias{gStringDown}
\alias{gStringEqual}
\alias{gStringErase}
\alias{gStringFree}
\alias{gStringFreeAndSteal}
\alias{gStringFreeToBytes}
\alias{gStringHash}
\alias{gStringInsert}
\alias{gStringInsertC}
\alias{gStringInsertLen}
\alias{gStringInsertUnichar}
\alias{gStringOverwrite}
\alias{gStringOverwriteLen}
\alias{gStringPrepend}
\alias{gStringPrependC}
\alias{gStringPrependLen}
\alias{gStringPrependUnichar}
\alias{gStringReplace}
\alias{gStringSetSize}
\alias{gStringTruncate}
\alias{gStringUp}
\alias{gStringChunkClear}
\alias{gStringChunkFree}
\alias{gStringChunkInsert}
\alias{gStringChunkInsertConst}
\alias{gStringChunkInsertLen}
\alias{gStrvBuilderNew}
\alias{gStrvBuilderAdd}
\alias{gStrvBuilderAddv}
\alias{gStrvBuilderEnd}
\alias{gStrvBuilderRef}
\alias{gStrvBuilderTake}
\alias{gStrvBuilderUnref}
\alias{gStrvBuilderUnrefToStrv}
\alias{gStrEqual}
\alias{gStrHasPrefix}
\alias{gStrHasSuffix}
\alias{gStrHash}
\alias{gStrIsAscii}
\alias{gStrMatchString}
\alias{gStrToAscii}
\alias{gStrTokenizeAndFold}
\alias{gStripContext}
\alias{gStrrstrLen}
\alias{gStrsplitSet}
\alias{gStrstrLen}
\alias{gStrvContains}
\alias{gStrvEqual}
\alias{gStrvGetType}
\alias{gStrvLength}
\alias{gStrdupValueContents}
\title{gStringNew}
\usage{
gStringNew(init)

gStringNewLen(init, len)

gStringNewTake(init)

gStringSizedNew(dfl_size)

gStringAppend(string, val)

gStringAppendC(string, c)

gStringAppendLen(string, val, len)

gStringAppendUnichar(string, wc)

gStringAppendUriEscaped(string, unescaped, reserved_chars_allowed, allow_utf8)

gStringAsciiDown(string)

gStringAsciiUp(string)

gStringAssign(string, rval)

gStringCopy(string)

gStringDown(string)

gStringEqual(v, v2)

gStringErase(string, pos, len)

gStringFree(string, free_segment)

gStringFreeAndSteal(string)

gStringFreeToBytes(string)

gStringHash(str)

gStringInsert(string, pos, val)

gStringInsertC(string, pos, c)

gStringInsertLen(string, pos, val, len)

gStringInsertUnichar(string, pos, wc)

gStringOverwrite(string, pos, val)

gStringOverwriteLen(string, pos, val, len)

gStringPrepend(string, val)

gStringPrependC(string, c)

gStringPrependLen(string, val, len)

gStringPrependUnichar(string, wc)

gStringReplace(string, find, replace, limit)

gStringSetSize(string, len)

gStringTruncate(string, len)

gStringUp(string)

gStringChunkClear(chunk)

gStringChunkFree(chunk)

gStringChunkInsert(chunk, string)

gStringChunkInsertConst(chunk, string)

gStringChunkInsertLen(chunk, string, len)

gStrvBuilderNew()

gStrvBuilderAdd(builder, value)

gStrvBuilderAddv(builder, value)

gStrvBuilderEnd(builder)

gStrvBuilderRef(builder)

gStrvBuilderTake(builder, value)

gStrvBuilderUnref(builder)

gStrvBuilderUnrefToStrv(builder)

gStrEqual(v1, v2)

gStrHasPrefix(str, prefix)

gStrHasSuffix(str, suffix)

gStrHash(v)

gStrIsAscii(str)

gStrMatchString(search_term, potential_hit, accept_alternates)

gStrToAscii(str, from_locale)

gStrTokenizeAndFold(string, translit_locale)

gStripContext(msgid, msgval)

gStrrstrLen(haystack, haystack_len, needle)

gStrsplitSet(string, delimiters, max_tokens)

gStrstrLen(haystack, haystack_len, needle)

gStrvContains(strv, str)

gStrvEqual(strv1, strv2)

gStrvGetType()

gStrvLength(str_array)

gStrdupValueContents(value)
}
\arguments{
\item{init}{utf8}

\item{len}{gssize}

\item{dfl_size}{gsize}

\item{string}{utf8}

\item{val}{utf8}

\item{c}{gchar}

\item{wc}{gunichar}

\item{unescaped}{utf8}

\item{reserved_chars_allowed}{utf8}

\item{allow_utf8}{gboolean}

\item{rval}{utf8}

\item{v}{gpointer}

\item{v2}{gpointer}

\item{pos}{gsize}

\item{free_segment}{gboolean}

\item{str}{utf8}

\item{find}{utf8}

\item{replace}{utf8}

\item{limit}{guint}

\item{chunk}{StringChunk}

\item{builder}{StrvBuilder}

\item{value}{Value}

\item{v1}{gpointer}

\item{prefix}{utf8}

\item{suffix}{utf8}

\item{search_term}{utf8}

\item{potential_hit}{utf8}

\item{accept_alternates}{gboolean}

\item{from_locale}{utf8}

\item{translit_locale}{utf8}

\item{msgid}{utf8}

\item{msgval}{utf8}

\item{haystack}{utf8}

\item{haystack_len}{gssize}

\item{needle}{utf8}

\item{delimiters}{guint8}

\item{max_tokens}{gint}

\item{strv}{utf8}

\item{strv1}{utf8}

\item{strv2}{utf8}

\item{str_array}{utf8}
}
\value{
String

String

String

String

String

String

String

String

String

String

String

String

String

String

gboolean

String

utf8

utf8

Bytes

guint

String

String

String

String

String

String

String

String

String

String

guint

String

String

String

Return value from C function

Return value from C function

utf8

utf8

utf8

StrvBuilder

Return value from C function

Return value from C function

utf8

StrvBuilder

Return value from C function

Return value from C function

utf8

gboolean

gboolean

gboolean

guint

gboolean

gboolean

utf8

utf8

utf8

utf8

utf8

utf8

gboolean

gboolean

GType

guint

utf8
}
\description{
gStringNew

gStringNewLen

gStringNewTake

gStringSizedNew

gStringAppend

gStringAppendC

gStringAppendLen

gStringAppendUnichar

gStringAppendUriEscaped

gStringAsciiDown

gStringAsciiUp

gStringAssign

gStringCopy

gStringDown

gStringEqual

gStringErase

gStringFree

gStringFreeAndSteal

gStringFreeToBytes

gStringHash

gStringInsert

gStringInsertC

gStringInsertLen

gStringInsertUnichar

gStringOverwrite

gStringOverwriteLen

gStringPrepend

gStringPrependC

gStringPrependLen

gStringPrependUnichar

gStringReplace

gStringSetSize

gStringTruncate

gStringUp

gStringChunkClear

gStringChunkFree

gStringChunkInsert

gStringChunkInsertConst

gStringChunkInsertLen

gStrvBuilderNew

gStrvBuilderAdd

gStrvBuilderAddv

gStrvBuilderEnd

gStrvBuilderRef

gStrvBuilderTake

gStrvBuilderUnref

gStrvBuilderUnrefToStrv

gStrEqual

gStrHasPrefix

gStrHasSuffix

gStrHash

gStrIsAscii

gStrMatchString

gStrToAscii

gStrTokenizeAndFold

gStripContext

gStrrstrLen

gStrsplitSet

gStrstrLen

gStrvContains

gStrvEqual

gStrvGetType

gStrvLength

gStrdupValueContents
}
