% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R
\name{gVariantNewArray}
\alias{gVariantNewArray}
\alias{gVariantNewBoolean}
\alias{gVariantNewByte}
\alias{gVariantNewBytestring}
\alias{gVariantNewBytestringArray}
\alias{gVariantNewDictEntry}
\alias{gVariantNewDouble}
\alias{gVariantNewFixedArray}
\alias{gVariantNewFromBytes}
\alias{gVariantNewFromData}
\alias{gVariantNewHandle}
\alias{gVariantNewInt16}
\alias{gVariantNewInt32}
\alias{gVariantNewInt64}
\alias{gVariantNewMaybe}
\alias{gVariantNewObjectPath}
\alias{gVariantNewObjv}
\alias{gVariantNewSignature}
\alias{gVariantNewString}
\alias{gVariantNewStrv}
\alias{gVariantNewTuple}
\alias{gVariantNewUint16}
\alias{gVariantNewUint32}
\alias{gVariantNewUint64}
\alias{gVariantNewVariant}
\alias{gVariantByteswap}
\alias{gVariantCheckFormatString}
\alias{gVariantClassify}
\alias{gVariantCompare}
\alias{gVariantDupBytestring}
\alias{gVariantDupBytestringArray}
\alias{gVariantDupObjv}
\alias{gVariantDupString}
\alias{gVariantDupStrv}
\alias{gVariantEqual}
\alias{gVariantGetBoolean}
\alias{gVariantGetByte}
\alias{gVariantGetBytestring}
\alias{gVariantGetBytestringArray}
\alias{gVariantGetChildValue}
\alias{gVariantGetData}
\alias{gVariantGetDataAsBytes}
\alias{gVariantGetDouble}
\alias{gVariantGetHandle}
\alias{gVariantGetInt16}
\alias{gVariantGetInt32}
\alias{gVariantGetInt64}
\alias{gVariantGetMaybe}
\alias{gVariantGetNormalForm}
\alias{gVariantGetObjv}
\alias{gVariantGetSize}
\alias{gVariantGetString}
\alias{gVariantGetStrv}
\alias{gVariantGetType}
\alias{gVariantGetTypeString}
\alias{gVariantGetUint16}
\alias{gVariantGetUint32}
\alias{gVariantGetUint64}
\alias{gVariantGetVariant}
\alias{gVariantHash}
\alias{gVariantIsContainer}
\alias{gVariantIsFloating}
\alias{gVariantIsNormalForm}
\alias{gVariantIsOfType}
\alias{gVariantLookupValue}
\alias{gVariantNChildren}
\alias{gVariantPrint}
\alias{gVariantRef}
\alias{gVariantRefSink}
\alias{gVariantStore}
\alias{gVariantTakeRef}
\alias{gVariantUnref}
\alias{gVariantIsObjectPath}
\alias{gVariantIsSignature}
\alias{gVariantParse}
\alias{gVariantParseErrorPrintContext}
\alias{gVariantParseErrorQuark}
\alias{gVariantParserGetErrorQuark}
\alias{gVariantBuilderNew}
\alias{gVariantBuilderAddValue}
\alias{gVariantBuilderClose}
\alias{gVariantBuilderEnd}
\alias{gVariantBuilderOpen}
\alias{gVariantBuilderRef}
\alias{gVariantBuilderUnref}
\alias{gVariantDictNew}
\alias{gVariantDictClear}
\alias{gVariantDictContains}
\alias{gVariantDictEnd}
\alias{gVariantDictInsertValue}
\alias{gVariantDictLookupValue}
\alias{gVariantDictRef}
\alias{gVariantDictRemove}
\alias{gVariantDictUnref}
\alias{gVariantIterFree}
\alias{gVariantIterNChildren}
\alias{gVariantIterNextValue}
\alias{gVariantTypeNew}
\alias{gVariantTypeNewArray}
\alias{gVariantTypeNewDictEntry}
\alias{gVariantTypeNewMaybe}
\alias{gVariantTypeNewTuple}
\alias{gVariantTypeCopy}
\alias{gVariantTypeDupString}
\alias{gVariantTypeElement}
\alias{gVariantTypeEqual}
\alias{gVariantTypeFirst}
\alias{gVariantTypeFree}
\alias{gVariantTypeGetStringLength}
\alias{gVariantTypeHash}
\alias{gVariantTypeIsArray}
\alias{gVariantTypeIsBasic}
\alias{gVariantTypeIsContainer}
\alias{gVariantTypeIsDefinite}
\alias{gVariantTypeIsDictEntry}
\alias{gVariantTypeIsMaybe}
\alias{gVariantTypeIsSubtypeOf}
\alias{gVariantTypeIsTuple}
\alias{gVariantTypeIsVariant}
\alias{gVariantTypeKey}
\alias{gVariantTypeNItems}
\alias{gVariantTypeNext}
\alias{gVariantTypeValue}
\alias{gVariantTypeChecked}
\alias{gVariantTypeStringGetDepth}
\alias{gVariantTypeStringIsValid}
\alias{gVariantTypeStringScan}
\alias{gVariantGetGtype}
\title{gVariantNewArray}
\usage{
gVariantNewArray(child_type, children, n_children)

gVariantNewBoolean(value)

gVariantNewByte(value)

gVariantNewBytestring(string)

gVariantNewBytestringArray(strv, length)

gVariantNewDictEntry(key, value)

gVariantNewDouble(value)

gVariantNewFixedArray(element_type, elements, n_elements, element_size)

gVariantNewFromBytes(type, bytes, trusted)

gVariantNewFromData(type, data, size, trusted, notify, user_data)

gVariantNewHandle(value)

gVariantNewInt16(value)

gVariantNewInt32(value)

gVariantNewInt64(value)

gVariantNewMaybe(child_type, child)

gVariantNewObjectPath(object_path)

gVariantNewObjv(strv, length)

gVariantNewSignature(signature)

gVariantNewString(string)

gVariantNewStrv(strv, length)

gVariantNewTuple(children, n_children)

gVariantNewUint16(value)

gVariantNewUint32(value)

gVariantNewUint64(value)

gVariantNewVariant(value)

gVariantByteswap(value)

gVariantCheckFormatString(value, format_string, copy_only)

gVariantClassify(value)

gVariantCompare(one, two)

gVariantDupBytestring(value)

gVariantDupBytestringArray(value)

gVariantDupObjv(value)

gVariantDupString(value)

gVariantDupStrv(value)

gVariantEqual(one, two)

gVariantGetBoolean(value)

gVariantGetByte(value)

gVariantGetBytestring(value)

gVariantGetBytestringArray(value)

gVariantGetChildValue(value, index_)

gVariantGetData(value)

gVariantGetDataAsBytes(value)

gVariantGetDouble(value)

gVariantGetHandle(value)

gVariantGetInt16(value)

gVariantGetInt32(value)

gVariantGetInt64(value)

gVariantGetMaybe(value)

gVariantGetNormalForm(value)

gVariantGetObjv(value)

gVariantGetSize(value)

gVariantGetString(value)

gVariantGetStrv(value)

gVariantGetType(value)

gVariantGetTypeString(value)

gVariantGetUint16(value)

gVariantGetUint32(value)

gVariantGetUint64(value)

gVariantGetVariant(value)

gVariantHash(value)

gVariantIsContainer(value)

gVariantIsFloating(value)

gVariantIsNormalForm(value)

gVariantIsOfType(value, type)

gVariantLookupValue(dictionary, key, expected_type)

gVariantNChildren(value)

gVariantPrint(value, type_annotate)

gVariantRef(value)

gVariantRefSink(value)

gVariantStore(value, data)

gVariantTakeRef(value)

gVariantUnref(value)

gVariantIsObjectPath(string)

gVariantIsSignature(string)

gVariantParse(type, text, limit, endptr)

gVariantParseErrorPrintContext(error, source_str)

gVariantParseErrorQuark()

gVariantParserGetErrorQuark()

gVariantBuilderNew(type)

gVariantBuilderAddValue(builder, value)

gVariantBuilderClose(builder)

gVariantBuilderEnd(builder)

gVariantBuilderOpen(builder, type)

gVariantBuilderRef(builder)

gVariantBuilderUnref(builder)

gVariantDictNew(from_asv)

gVariantDictClear(dict)

gVariantDictContains(dict, key)

gVariantDictEnd(dict)

gVariantDictInsertValue(dict, key, value)

gVariantDictLookupValue(dict, key, expected_type)

gVariantDictRef(dict)

gVariantDictRemove(dict, key)

gVariantDictUnref(dict)

gVariantIterFree(iter)

gVariantIterNChildren(iter)

gVariantIterNextValue(iter)

gVariantTypeNew(type_string)

gVariantTypeNewArray(element)

gVariantTypeNewDictEntry(key, value)

gVariantTypeNewMaybe(element)

gVariantTypeNewTuple(items, length)

gVariantTypeCopy(type)

gVariantTypeDupString(type)

gVariantTypeElement(type)

gVariantTypeEqual(type1, type2)

gVariantTypeFirst(type)

gVariantTypeFree(type)

gVariantTypeGetStringLength(type)

gVariantTypeHash(type)

gVariantTypeIsArray(type)

gVariantTypeIsBasic(type)

gVariantTypeIsContainer(type)

gVariantTypeIsDefinite(type)

gVariantTypeIsDictEntry(type)

gVariantTypeIsMaybe(type)

gVariantTypeIsSubtypeOf(type, supertype)

gVariantTypeIsTuple(type)

gVariantTypeIsVariant(type)

gVariantTypeKey(type)

gVariantTypeNItems(type)

gVariantTypeNext(type)

gVariantTypeValue(type)

gVariantTypeChecked(type_string)

gVariantTypeStringGetDepth(type_string)

gVariantTypeStringIsValid(type_string)

gVariantTypeStringScan(string, limit)

gVariantGetGtype()
}
\arguments{
\item{child_type}{VariantType}

\item{children}{Variant}

\item{n_children}{gsize}

\item{value}{VariantType}

\item{string}{utf8}

\item{strv}{utf8}

\item{length}{gint}

\item{key}{VariantType}

\item{element_type}{VariantType}

\item{elements}{gpointer}

\item{n_elements}{gsize}

\item{element_size}{gsize}

\item{type}{VariantType}

\item{bytes}{Bytes}

\item{trusted}{gboolean}

\item{data}{gpointer}

\item{size}{gsize}

\item{notify}{DestroyNotify}

\item{user_data}{gpointer}

\item{child}{Variant}

\item{object_path}{utf8}

\item{signature}{utf8}

\item{format_string}{utf8}

\item{copy_only}{gboolean}

\item{one}{Variant}

\item{two}{Variant}

\item{index_}{gsize}

\item{dictionary}{Variant}

\item{expected_type}{VariantType}

\item{type_annotate}{gboolean}

\item{text}{utf8}

\item{limit}{utf8}

\item{endptr}{utf8}

\item{error}{Error}

\item{source_str}{utf8}

\item{builder}{VariantBuilder}

\item{from_asv}{Variant}

\item{dict}{VariantDict}

\item{iter}{VariantIter}

\item{type_string}{utf8}

\item{element}{VariantType}

\item{items}{VariantType}

\item{type1}{VariantType}

\item{type2}{VariantType}

\item{supertype}{VariantType}
}
\value{
Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

Variant

gboolean

VariantClass

gint

guint8

utf8

utf8

utf8

utf8

gboolean

gboolean

guint8

guint8

utf8

Variant

gpointer

Bytes

gdouble

gint32

gint16

gint32

gint64

Variant

Variant

utf8

gsize

utf8

utf8

VariantType

utf8

guint16

guint32

guint64

Variant

guint

gboolean

gboolean

gboolean

gboolean

Variant

gsize

utf8

Variant

Variant

Return value from C function

Variant

Return value from C function

gboolean

gboolean

Variant

utf8

Quark

Quark

VariantBuilder

Return value from C function

Return value from C function

Variant

Return value from C function

VariantBuilder

Return value from C function

VariantDict

Return value from C function

gboolean

Variant

Return value from C function

Variant

VariantDict

gboolean

Return value from C function

Return value from C function

gsize

Variant

VariantType

VariantType

VariantType

VariantType

VariantType

VariantType

utf8

VariantType

gboolean

VariantType

Return value from C function

gsize

guint

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

VariantType

gsize

VariantType

VariantType

VariantType

gsize

gboolean

gboolean

GType
}
\description{
gVariantNewArray

gVariantNewBoolean

gVariantNewByte

gVariantNewBytestring

gVariantNewBytestringArray

gVariantNewDictEntry

gVariantNewDouble

gVariantNewFixedArray

gVariantNewFromBytes

gVariantNewFromData

gVariantNewHandle

gVariantNewInt16

gVariantNewInt32

gVariantNewInt64

gVariantNewMaybe

gVariantNewObjectPath

gVariantNewObjv

gVariantNewSignature

gVariantNewString

gVariantNewStrv

gVariantNewTuple

gVariantNewUint16

gVariantNewUint32

gVariantNewUint64

gVariantNewVariant

gVariantByteswap

gVariantCheckFormatString

gVariantClassify

gVariantCompare

gVariantDupBytestring

gVariantDupBytestringArray

gVariantDupObjv

gVariantDupString

gVariantDupStrv

gVariantEqual

gVariantGetBoolean

gVariantGetByte

gVariantGetBytestring

gVariantGetBytestringArray

gVariantGetChildValue

gVariantGetData

gVariantGetDataAsBytes

gVariantGetDouble

gVariantGetHandle

gVariantGetInt16

gVariantGetInt32

gVariantGetInt64

gVariantGetMaybe

gVariantGetNormalForm

gVariantGetObjv

gVariantGetSize

gVariantGetString

gVariantGetStrv

gVariantGetType

gVariantGetTypeString

gVariantGetUint16

gVariantGetUint32

gVariantGetUint64

gVariantGetVariant

gVariantHash

gVariantIsContainer

gVariantIsFloating

gVariantIsNormalForm

gVariantIsOfType

gVariantLookupValue

gVariantNChildren

gVariantPrint

gVariantRef

gVariantRefSink

gVariantStore

gVariantTakeRef

gVariantUnref

gVariantIsObjectPath

gVariantIsSignature

gVariantParse

gVariantParseErrorPrintContext

gVariantParseErrorQuark

gVariantParserGetErrorQuark

gVariantBuilderNew

gVariantBuilderAddValue

gVariantBuilderClose

gVariantBuilderEnd

gVariantBuilderOpen

gVariantBuilderRef

gVariantBuilderUnref

gVariantDictNew

gVariantDictClear

gVariantDictContains

gVariantDictEnd

gVariantDictInsertValue

gVariantDictLookupValue

gVariantDictRef

gVariantDictRemove

gVariantDictUnref

gVariantIterFree

gVariantIterNChildren

gVariantIterNextValue

gVariantTypeNew

gVariantTypeNewArray

gVariantTypeNewDictEntry

gVariantTypeNewMaybe

gVariantTypeNewTuple

gVariantTypeCopy

gVariantTypeDupString

gVariantTypeElement

gVariantTypeEqual

gVariantTypeFirst

gVariantTypeFree

gVariantTypeGetStringLength

gVariantTypeHash

gVariantTypeIsArray

gVariantTypeIsBasic

gVariantTypeIsContainer

gVariantTypeIsDefinite

gVariantTypeIsDictEntry

gVariantTypeIsMaybe

gVariantTypeIsSubtypeOf

gVariantTypeIsTuple

gVariantTypeIsVariant

gVariantTypeKey

gVariantTypeNItems

gVariantTypeNext

gVariantTypeValue

gVariantTypeChecked

gVariantTypeStringGetDepth

gVariantTypeStringIsValid

gVariantTypeStringScan

gVariantGetGtype
}
