% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R
\name{glibCheckVersion}
\alias{glibCheckVersion}
\title{glibCheckVersion}
\usage{
glibCheckVersion(required_major, required_minor, required_micro)
}
\arguments{
\item{required_major}{guint}

\item{required_minor}{guint}

\item{required_micro}{guint}
}
\value{
utf8
}
\description{
glibCheckVersion
}
