% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GObject_autogen.R
\name{gObjectNewv}
\alias{gObjectNewv}
\alias{gObjectCompatControl}
\alias{gObjectInterfaceFindProperty}
\alias{gObjectInterfaceInstallProperty}
\alias{gObjectInterfaceListProperties}
\alias{gObjectBindProperty}
\alias{gObjectBindPropertyWithClosures}
\alias{gObjectForceFloating}
\alias{gObjectFreezeNotify}
\alias{gObjectGetData}
\alias{gObjectGetProperty}
\alias{gObjectGetQdata}
\alias{gObjectGetv}
\alias{gObjectIsFloating}
\alias{gObjectNotify}
\alias{gObjectNotifyByPspec}
\alias{gObjectRef}
\alias{gObjectRefSink}
\alias{gObjectRunDispose}
\alias{gObjectSetData}
\alias{gObjectSetProperty}
\alias{gObjectStealData}
\alias{gObjectStealQdata}
\alias{gObjectThawNotify}
\alias{gObjectUnref}
\alias{gObjectWatchClosure}
\alias{gObjectClassFindProperty}
\alias{gObjectClassInstallProperties}
\alias{gObjectClassInstallProperty}
\alias{gObjectClassListProperties}
\alias{gObjectClassOverrideProperty}
\alias{gSignalAccumulatorFirstWins}
\alias{gSignalAccumulatorTrueHandled}
\alias{gSignalChainFromOverridden}
\alias{gSignalConnectClosure}
\alias{gSignalConnectClosureById}
\alias{gSignalEmitv}
\alias{gSignalGetInvocationHint}
\alias{gSignalHandlerBlock}
\alias{gSignalHandlerDisconnect}
\alias{gSignalHandlerFind}
\alias{gSignalHandlerIsConnected}
\alias{gSignalHandlerUnblock}
\alias{gSignalHandlersBlockMatched}
\alias{gSignalHandlersDestroy}
\alias{gSignalHandlersDisconnectMatched}
\alias{gSignalHandlersUnblockMatched}
\alias{gSignalHasHandlerPending}
\alias{gSignalListIds}
\alias{gSignalLookup}
\alias{gSignalName}
\alias{gSignalOverrideClassClosure}
\alias{gSignalParseName}
\alias{gSignalQuery}
\alias{gSignalRemoveEmissionHook}
\alias{gSignalStopEmission}
\alias{gSignalStopEmissionByName}
\alias{gSignalTypeCclosureNew}
\title{gObjectNewv}
\usage{
gObjectNewv(object_type, n_parameters, parameters)

gObjectCompatControl(what, data)

gObjectInterfaceFindProperty(g_iface, property_name)

gObjectInterfaceInstallProperty(g_iface, pspec)

gObjectInterfaceListProperties(g_iface)

gObjectBindProperty(source, source_property, target, target_property, flags)

gObjectBindPropertyWithClosures(
  source,
  source_property,
  target,
  target_property,
  flags,
  transform_to,
  transform_from
)

gObjectForceFloating(object)

gObjectFreezeNotify(object)

gObjectGetData(object, key)

gObjectGetProperty(object, property_name, value)

gObjectGetQdata(object, quark)

gObjectGetv(object, n_properties, names, values)

gObjectIsFloating(object)

gObjectNotify(object, property_name)

gObjectNotifyByPspec(object, pspec)

gObjectRef(object)

gObjectRefSink(object)

gObjectRunDispose(object)

gObjectSetData(object, key, data)

gObjectSetProperty(object, property_name, value)

gObjectStealData(object, key)

gObjectStealQdata(object, quark)

gObjectThawNotify(object)

gObjectUnref(object)

gObjectWatchClosure(object, closure)

gObjectClassFindProperty(oclass, property_name)

gObjectClassInstallProperties(oclass, n_pspecs, pspecs)

gObjectClassInstallProperty(oclass, property_id, pspec)

gObjectClassListProperties(oclass)

gObjectClassOverrideProperty(oclass, property_id, name)

gSignalAccumulatorFirstWins(ihint, return_accu, handler_return, dummy)

gSignalAccumulatorTrueHandled(ihint, return_accu, handler_return, dummy)

gSignalChainFromOverridden(instance_and_params, return_value)

gSignalConnectClosure(instance, detailed_signal, closure, after)

gSignalConnectClosureById(instance, signal_id, detail, closure, after)

gSignalEmitv(instance_and_params, signal_id, detail)

gSignalGetInvocationHint(instance)

gSignalHandlerBlock(instance, handler_id)

gSignalHandlerDisconnect(instance, handler_id)

gSignalHandlerFind(instance, mask, signal_id, detail, closure, func, data)

gSignalHandlerIsConnected(instance, handler_id)

gSignalHandlerUnblock(instance, handler_id)

gSignalHandlersBlockMatched(
  instance,
  mask,
  signal_id,
  detail,
  closure,
  func,
  data
)

gSignalHandlersDestroy(instance)

gSignalHandlersDisconnectMatched(
  instance,
  mask,
  signal_id,
  detail,
  closure,
  func,
  data
)

gSignalHandlersUnblockMatched(
  instance,
  mask,
  signal_id,
  detail,
  closure,
  func,
  data
)

gSignalHasHandlerPending(instance, signal_id, detail, may_be_blocked)

gSignalListIds(itype)

gSignalLookup(name, itype)

gSignalName(signal_id)

gSignalOverrideClassClosure(signal_id, instance_type, class_closure)

gSignalParseName(detailed_signal, itype, force_detail_quark)

gSignalQuery(signal_id)

gSignalRemoveEmissionHook(signal_id, hook_id)

gSignalStopEmission(instance, signal_id, detail)

gSignalStopEmissionByName(instance, detailed_signal)

gSignalTypeCclosureNew(itype, struct_offset)
}
\arguments{
\item{object_type}{GType}

\item{n_parameters}{guint}

\item{parameters}{Parameter}

\item{what}{gsize}

\item{data}{gpointer}

\item{g_iface}{TypeInterface}

\item{property_name}{utf8}

\item{pspec}{ParamSpec}

\item{source}{Object}

\item{source_property}{utf8}

\item{target}{Object}

\item{target_property}{utf8}

\item{flags}{BindingFlags}

\item{transform_to}{Closure}

\item{transform_from}{Closure}

\item{object}{Object}

\item{key}{utf8}

\item{value}{Value}

\item{quark}{GLib.Quark}

\item{n_properties}{guint}

\item{names}{utf8}

\item{values}{Value}

\item{closure}{Closure}

\item{oclass}{ObjectClass}

\item{n_pspecs}{guint}

\item{pspecs}{ParamSpec}

\item{property_id}{guint}

\item{name}{utf8}

\item{ihint}{SignalInvocationHint}

\item{return_accu}{Value}

\item{handler_return}{Value}

\item{dummy}{gpointer}

\item{instance_and_params}{Value}

\item{return_value}{Value}

\item{instance}{Object}

\item{detailed_signal}{utf8}

\item{after}{gboolean}

\item{signal_id}{guint}

\item{detail}{GLib.Quark}

\item{handler_id}{gulong}

\item{mask}{SignalMatchType}

\item{func}{gpointer}

\item{may_be_blocked}{gboolean}

\item{itype}{GType}

\item{instance_type}{GType}

\item{class_closure}{Closure}

\item{force_detail_quark}{gboolean}

\item{hook_id}{gulong}

\item{struct_offset}{guint}
}
\value{
Object

gsize

ParamSpec

Return value from C function

ParamSpec

Binding

Binding

Return value from C function

Return value from C function

gpointer

Return value from C function

gpointer

Return value from C function

gboolean

Return value from C function

Return value from C function

Object

Object

Return value from C function

Return value from C function

Return value from C function

gpointer

gpointer

Return value from C function

Return value from C function

Return value from C function

ParamSpec

Return value from C function

Return value from C function

ParamSpec

Return value from C function

gboolean

gboolean

Return value from C function

gulong

gulong

Return value from C function

SignalInvocationHint

Return value from C function

Return value from C function

gulong

gboolean

Return value from C function

guint

Return value from C function

guint

guint

gboolean

guint

guint

utf8

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Closure
}
\description{
gObjectNewv

gObjectCompatControl

gObjectInterfaceFindProperty

gObjectInterfaceInstallProperty

gObjectInterfaceListProperties

gObjectBindProperty

gObjectBindPropertyWithClosures

gObjectForceFloating

gObjectFreezeNotify

gObjectGetData

gObjectGetProperty

gObjectGetQdata

gObjectGetv

gObjectIsFloating

gObjectNotify

gObjectNotifyByPspec

gObjectRef

gObjectRefSink

gObjectRunDispose

gObjectSetData

gObjectSetProperty

gObjectStealData

gObjectStealQdata

gObjectThawNotify

gObjectUnref

gObjectWatchClosure

gObjectClassFindProperty

gObjectClassInstallProperties

gObjectClassInstallProperty

gObjectClassListProperties

gObjectClassOverrideProperty

gSignalAccumulatorFirstWins

gSignalAccumulatorTrueHandled

gSignalChainFromOverridden

gSignalConnectClosure

gSignalConnectClosureById

gSignalEmitv

gSignalGetInvocationHint

gSignalHandlerBlock

gSignalHandlerDisconnect

gSignalHandlerFind

gSignalHandlerIsConnected

gSignalHandlerUnblock

gSignalHandlersBlockMatched

gSignalHandlersDestroy

gSignalHandlersDisconnectMatched

gSignalHandlersUnblockMatched

gSignalHasHandlerPending

gSignalListIds

gSignalLookup

gSignalName

gSignalOverrideClassClosure

gSignalParseName

gSignalQuery

gSignalRemoveEmissionHook

gSignalStopEmission

gSignalStopEmissionByName

gSignalTypeCclosureNew
}
