% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gObjectSetString}
\alias{gObjectSetString}
\title{Set a string property on a GObject}
\usage{
gObjectSetString(object, property, value)
}
\arguments{
\item{object}{GObject external pointer}

\item{property}{Property name (character)}

\item{value}{String value to set}
}
\description{
Convenience function for setting properties that accept string values
like colors, fonts, CSS classes, etc.
}
