% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gObjectSetters}
\alias{gObjectSetters}
\alias{gObjectSetString}
\alias{gObjectSetBoolean}
\alias{gObjectSetInt}
\alias{gObjectSetDouble}
\title{GObject Property Setters}
\usage{
gObjectSetString(object, property, value)

gObjectSetBoolean(object, property, value)

gObjectSetInt(object, property, value)

gObjectSetDouble(object, property, value)
}
\arguments{
\item{object}{GObject external pointer}

\item{property}{Property name (character)}

\item{value}{The value to set (type-specific)}
}
\description{
These convenience functions allow you to set properties on a GObject by
explicitly defining the data type. This ensures the underlying C logic
correctly casts the R values to the appropriate GTK types.
}
