% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphene_autogen.R
\name{grapheneBoxAlloc}
\alias{grapheneBoxAlloc}
\alias{grapheneBoxContainsBox}
\alias{grapheneBoxContainsPoint}
\alias{grapheneBoxEqual}
\alias{grapheneBoxExpand}
\alias{grapheneBoxExpandScalar}
\alias{grapheneBoxExpandVec3}
\alias{grapheneBoxFree}
\alias{grapheneBoxGetBoundingSphere}
\alias{grapheneBoxGetCenter}
\alias{grapheneBoxGetDepth}
\alias{grapheneBoxGetHeight}
\alias{grapheneBoxGetMax}
\alias{grapheneBoxGetMin}
\alias{grapheneBoxGetSize}
\alias{grapheneBoxGetVertices}
\alias{grapheneBoxGetWidth}
\alias{grapheneBoxInit}
\alias{grapheneBoxInitFromBox}
\alias{grapheneBoxInitFromPoints}
\alias{grapheneBoxInitFromVec3}
\alias{grapheneBoxInitFromVectors}
\alias{grapheneBoxIntersection}
\alias{grapheneBoxUnion}
\alias{grapheneBoxEmpty}
\alias{grapheneBoxInfinite}
\alias{grapheneBoxMinusOne}
\alias{grapheneBoxOne}
\alias{grapheneBoxOneMinusOne}
\alias{grapheneBoxZero}
\alias{grapheneEulerAlloc}
\alias{grapheneEulerEqual}
\alias{grapheneEulerFree}
\alias{grapheneEulerGetAlpha}
\alias{grapheneEulerGetBeta}
\alias{grapheneEulerGetGamma}
\alias{grapheneEulerGetOrder}
\alias{grapheneEulerGetX}
\alias{grapheneEulerGetY}
\alias{grapheneEulerGetZ}
\alias{grapheneEulerInit}
\alias{grapheneEulerInitFromEuler}
\alias{grapheneEulerInitFromMatrix}
\alias{grapheneEulerInitFromQuaternion}
\alias{grapheneEulerInitFromRadians}
\alias{grapheneEulerInitFromVec3}
\alias{grapheneEulerInitWithOrder}
\alias{grapheneEulerReorder}
\alias{grapheneEulerToMatrix}
\alias{grapheneEulerToQuaternion}
\alias{grapheneEulerToVec3}
\alias{grapheneFrustumAlloc}
\alias{grapheneFrustumContainsPoint}
\alias{grapheneFrustumEqual}
\alias{grapheneFrustumFree}
\alias{grapheneFrustumGetPlanes}
\alias{grapheneFrustumInit}
\alias{grapheneFrustumInitFromFrustum}
\alias{grapheneFrustumInitFromMatrix}
\alias{grapheneFrustumIntersectsBox}
\alias{grapheneFrustumIntersectsSphere}
\alias{grapheneMatrixAlloc}
\alias{grapheneMatrixDecompose}
\alias{grapheneMatrixDeterminant}
\alias{grapheneMatrixEqual}
\alias{grapheneMatrixEqualFast}
\alias{grapheneMatrixFree}
\alias{grapheneMatrixGetRow}
\alias{grapheneMatrixGetValue}
\alias{grapheneMatrixGetXScale}
\alias{grapheneMatrixGetXTranslation}
\alias{grapheneMatrixGetYScale}
\alias{grapheneMatrixGetYTranslation}
\alias{grapheneMatrixGetZScale}
\alias{grapheneMatrixGetZTranslation}
\alias{grapheneMatrixInitFrom2d}
\alias{grapheneMatrixInitFromFloat}
\alias{grapheneMatrixInitFromMatrix}
\alias{grapheneMatrixInitFromVec4}
\alias{grapheneMatrixInitFrustum}
\alias{grapheneMatrixInitIdentity}
\alias{grapheneMatrixInitLookAt}
\alias{grapheneMatrixInitOrtho}
\alias{grapheneMatrixInitPerspective}
\alias{grapheneMatrixInitRotate}
\alias{grapheneMatrixInitScale}
\alias{grapheneMatrixInitSkew}
\alias{grapheneMatrixInitTranslate}
\alias{grapheneMatrixInterpolate}
\alias{grapheneMatrixInverse}
\alias{grapheneMatrixIs2d}
\alias{grapheneMatrixIsBackfaceVisible}
\alias{grapheneMatrixIsIdentity}
\alias{grapheneMatrixIsSingular}
\alias{grapheneMatrixMultiply}
\alias{grapheneMatrixNear}
\alias{grapheneMatrixNormalize}
\alias{grapheneMatrixPerspective}
\alias{grapheneMatrixPrint}
\alias{grapheneMatrixProjectPoint}
\alias{grapheneMatrixProjectRect}
\alias{grapheneMatrixProjectRectBounds}
\alias{grapheneMatrixRotate}
\alias{grapheneMatrixRotateEuler}
\alias{grapheneMatrixRotateQuaternion}
\alias{grapheneMatrixRotateX}
\alias{grapheneMatrixRotateY}
\alias{grapheneMatrixRotateZ}
\alias{grapheneMatrixScale}
\alias{grapheneMatrixSkewXy}
\alias{grapheneMatrixSkewXz}
\alias{grapheneMatrixSkewYz}
\alias{grapheneMatrixTo2d}
\alias{grapheneMatrixToFloat}
\alias{grapheneMatrixTransformBounds}
\alias{grapheneMatrixTransformBox}
\alias{grapheneMatrixTransformPoint}
\alias{grapheneMatrixTransformPoint3d}
\alias{grapheneMatrixTransformRay}
\alias{grapheneMatrixTransformRect}
\alias{grapheneMatrixTransformSphere}
\alias{grapheneMatrixTransformVec3}
\alias{grapheneMatrixTransformVec4}
\alias{grapheneMatrixTranslate}
\alias{grapheneMatrixTranspose}
\alias{grapheneMatrixUnprojectPoint3d}
\alias{grapheneMatrixUntransformBounds}
\alias{grapheneMatrixUntransformPoint}
\alias{graphenePlaneAlloc}
\alias{graphenePlaneDistance}
\alias{graphenePlaneEqual}
\alias{graphenePlaneFree}
\alias{graphenePlaneGetConstant}
\alias{graphenePlaneGetNormal}
\alias{graphenePlaneInit}
\alias{graphenePlaneInitFromPlane}
\alias{graphenePlaneInitFromPoint}
\alias{graphenePlaneInitFromPoints}
\alias{graphenePlaneInitFromVec4}
\alias{graphenePlaneNegate}
\alias{graphenePlaneNormalize}
\alias{graphenePlaneTransform}
\alias{graphenePointAlloc}
\alias{graphenePointDistance}
\alias{graphenePointEqual}
\alias{graphenePointFree}
\alias{graphenePointInit}
\alias{graphenePointInitFromPoint}
\alias{graphenePointInitFromVec2}
\alias{graphenePointInterpolate}
\alias{graphenePointNear}
\alias{graphenePointToVec2}
\alias{graphenePointZero}
\alias{graphenePoint3dAlloc}
\alias{graphenePoint3dCross}
\alias{graphenePoint3dDistance}
\alias{graphenePoint3dDot}
\alias{graphenePoint3dEqual}
\alias{graphenePoint3dFree}
\alias{graphenePoint3dInit}
\alias{graphenePoint3dInitFromPoint}
\alias{graphenePoint3dInitFromVec3}
\alias{graphenePoint3dInterpolate}
\alias{graphenePoint3dLength}
\alias{graphenePoint3dNear}
\alias{graphenePoint3dNormalize}
\alias{graphenePoint3dNormalizeViewport}
\alias{graphenePoint3dScale}
\alias{graphenePoint3dToVec3}
\alias{graphenePoint3dZero}
\alias{grapheneQuadAlloc}
\alias{grapheneQuadBounds}
\alias{grapheneQuadContains}
\alias{grapheneQuadFree}
\alias{grapheneQuadGetPoint}
\alias{grapheneQuadInit}
\alias{grapheneQuadInitFromPoints}
\alias{grapheneQuadInitFromRect}
\alias{grapheneQuaternionAlloc}
\alias{grapheneQuaternionAdd}
\alias{grapheneQuaternionDot}
\alias{grapheneQuaternionEqual}
\alias{grapheneQuaternionFree}
\alias{grapheneQuaternionInit}
\alias{grapheneQuaternionInitFromAngleVec3}
\alias{grapheneQuaternionInitFromAngles}
\alias{grapheneQuaternionInitFromEuler}
\alias{grapheneQuaternionInitFromMatrix}
\alias{grapheneQuaternionInitFromQuaternion}
\alias{grapheneQuaternionInitFromRadians}
\alias{grapheneQuaternionInitFromVec4}
\alias{grapheneQuaternionInitIdentity}
\alias{grapheneQuaternionInvert}
\alias{grapheneQuaternionMultiply}
\alias{grapheneQuaternionNormalize}
\alias{grapheneQuaternionScale}
\alias{grapheneQuaternionSlerp}
\alias{grapheneQuaternionToAngleVec3}
\alias{grapheneQuaternionToAngles}
\alias{grapheneQuaternionToMatrix}
\alias{grapheneQuaternionToRadians}
\alias{grapheneQuaternionToVec4}
\alias{grapheneRayAlloc}
\alias{grapheneRayEqual}
\alias{grapheneRayFree}
\alias{grapheneRayGetClosestPointToPoint}
\alias{grapheneRayGetDirection}
\alias{grapheneRayGetDistanceToPlane}
\alias{grapheneRayGetDistanceToPoint}
\alias{grapheneRayGetOrigin}
\alias{grapheneRayGetPositionAt}
\alias{grapheneRayInit}
\alias{grapheneRayInitFromRay}
\alias{grapheneRayInitFromVec3}
\alias{grapheneRayIntersectBox}
\alias{grapheneRayIntersectSphere}
\alias{grapheneRayIntersectTriangle}
\alias{grapheneRayIntersectsBox}
\alias{grapheneRayIntersectsSphere}
\alias{grapheneRayIntersectsTriangle}
\alias{grapheneRectContainsPoint}
\alias{grapheneRectContainsRect}
\alias{grapheneRectEqual}
\alias{grapheneRectExpand}
\alias{grapheneRectFree}
\alias{grapheneRectGetArea}
\alias{grapheneRectGetBottomLeft}
\alias{grapheneRectGetBottomRight}
\alias{grapheneRectGetCenter}
\alias{grapheneRectGetHeight}
\alias{grapheneRectGetTopLeft}
\alias{grapheneRectGetTopRight}
\alias{grapheneRectGetVertices}
\alias{grapheneRectGetWidth}
\alias{grapheneRectGetX}
\alias{grapheneRectGetY}
\alias{grapheneRectInit}
\alias{grapheneRectInitFromRect}
\alias{grapheneRectInset}
\alias{grapheneRectInsetR}
\alias{grapheneRectInterpolate}
\alias{grapheneRectIntersection}
\alias{grapheneRectNormalize}
\alias{grapheneRectNormalizeR}
\alias{grapheneRectOffset}
\alias{grapheneRectOffsetR}
\alias{grapheneRectRound}
\alias{grapheneRectRoundExtents}
\alias{grapheneRectRoundToPixel}
\alias{grapheneRectScale}
\alias{grapheneRectUnion}
\alias{grapheneRectAlloc}
\alias{grapheneRectZero}
\alias{grapheneSizeAlloc}
\alias{grapheneSizeEqual}
\alias{grapheneSizeFree}
\alias{grapheneSizeInit}
\alias{grapheneSizeInitFromSize}
\alias{grapheneSizeInterpolate}
\alias{grapheneSizeScale}
\alias{grapheneSizeZero}
\alias{grapheneSphereAlloc}
\alias{grapheneSphereContainsPoint}
\alias{grapheneSphereDistance}
\alias{grapheneSphereEqual}
\alias{grapheneSphereFree}
\alias{grapheneSphereGetBoundingBox}
\alias{grapheneSphereGetCenter}
\alias{grapheneSphereGetRadius}
\alias{grapheneSphereInit}
\alias{grapheneSphereInitFromPoints}
\alias{grapheneSphereInitFromVectors}
\alias{grapheneSphereIsEmpty}
\alias{grapheneSphereTranslate}
\alias{grapheneTriangleAlloc}
\alias{grapheneTriangleContainsPoint}
\alias{grapheneTriangleEqual}
\alias{grapheneTriangleFree}
\alias{grapheneTriangleGetArea}
\alias{grapheneTriangleGetBarycoords}
\alias{grapheneTriangleGetBoundingBox}
\alias{grapheneTriangleGetMidpoint}
\alias{grapheneTriangleGetNormal}
\alias{grapheneTriangleGetPlane}
\alias{grapheneTriangleGetPoints}
\alias{grapheneTriangleGetUv}
\alias{grapheneTriangleGetVertices}
\alias{grapheneTriangleInitFromFloat}
\alias{grapheneTriangleInitFromPoint3d}
\alias{grapheneTriangleInitFromVec3}
\alias{grapheneVec2Alloc}
\alias{grapheneVec2Add}
\alias{grapheneVec2Divide}
\alias{grapheneVec2Dot}
\alias{grapheneVec2Equal}
\alias{grapheneVec2Free}
\alias{grapheneVec2GetX}
\alias{grapheneVec2GetY}
\alias{grapheneVec2Init}
\alias{grapheneVec2InitFromFloat}
\alias{grapheneVec2InitFromVec2}
\alias{grapheneVec2Interpolate}
\alias{grapheneVec2Length}
\alias{grapheneVec2Max}
\alias{grapheneVec2Min}
\alias{grapheneVec2Multiply}
\alias{grapheneVec2Near}
\alias{grapheneVec2Negate}
\alias{grapheneVec2Normalize}
\alias{grapheneVec2Scale}
\alias{grapheneVec2Subtract}
\alias{grapheneVec2ToFloat}
\alias{grapheneVec2One}
\alias{grapheneVec2XAxis}
\alias{grapheneVec2YAxis}
\alias{grapheneVec2Zero}
\alias{grapheneVec3Alloc}
\alias{grapheneVec3Add}
\alias{grapheneVec3Cross}
\alias{grapheneVec3Divide}
\alias{grapheneVec3Dot}
\alias{grapheneVec3Equal}
\alias{grapheneVec3Free}
\alias{grapheneVec3GetX}
\alias{grapheneVec3GetXy}
\alias{grapheneVec3GetXy0}
\alias{grapheneVec3GetXyz0}
\alias{grapheneVec3GetXyz1}
\alias{grapheneVec3GetXyzw}
\alias{grapheneVec3GetY}
\alias{grapheneVec3GetZ}
\alias{grapheneVec3Init}
\alias{grapheneVec3InitFromFloat}
\alias{grapheneVec3InitFromVec3}
\alias{grapheneVec3Interpolate}
\alias{grapheneVec3Length}
\alias{grapheneVec3Max}
\alias{grapheneVec3Min}
\alias{grapheneVec3Multiply}
\alias{grapheneVec3Near}
\alias{grapheneVec3Negate}
\alias{grapheneVec3Normalize}
\alias{grapheneVec3Scale}
\alias{grapheneVec3Subtract}
\alias{grapheneVec3ToFloat}
\alias{grapheneVec3One}
\alias{grapheneVec3XAxis}
\alias{grapheneVec3YAxis}
\alias{grapheneVec3ZAxis}
\alias{grapheneVec3Zero}
\alias{grapheneVec4Alloc}
\alias{grapheneVec4Add}
\alias{grapheneVec4Divide}
\alias{grapheneVec4Dot}
\alias{grapheneVec4Equal}
\alias{grapheneVec4Free}
\alias{grapheneVec4GetW}
\alias{grapheneVec4GetX}
\alias{grapheneVec4GetXy}
\alias{grapheneVec4GetXyz}
\alias{grapheneVec4GetY}
\alias{grapheneVec4GetZ}
\alias{grapheneVec4Init}
\alias{grapheneVec4InitFromFloat}
\alias{grapheneVec4InitFromVec2}
\alias{grapheneVec4InitFromVec3}
\alias{grapheneVec4InitFromVec4}
\alias{grapheneVec4Interpolate}
\alias{grapheneVec4Length}
\alias{grapheneVec4Max}
\alias{grapheneVec4Min}
\alias{grapheneVec4Multiply}
\alias{grapheneVec4Near}
\alias{grapheneVec4Negate}
\alias{grapheneVec4Normalize}
\alias{grapheneVec4Scale}
\alias{grapheneVec4Subtract}
\alias{grapheneVec4ToFloat}
\alias{grapheneVec4One}
\alias{grapheneVec4WAxis}
\alias{grapheneVec4XAxis}
\alias{grapheneVec4YAxis}
\alias{grapheneVec4ZAxis}
\alias{grapheneVec4Zero}
\title{grapheneBoxAlloc}
\usage{
grapheneBoxAlloc()

grapheneBoxContainsBox(a, b)

grapheneBoxContainsPoint(box, point)

grapheneBoxEqual(a, b)

grapheneBoxExpand(box, point)

grapheneBoxExpandScalar(box, scalar)

grapheneBoxExpandVec3(box, vec)

grapheneBoxFree(box)

grapheneBoxGetBoundingSphere(box)

grapheneBoxGetCenter(box)

grapheneBoxGetDepth(box)

grapheneBoxGetHeight(box)

grapheneBoxGetMax(box)

grapheneBoxGetMin(box)

grapheneBoxGetSize(box)

grapheneBoxGetVertices(box)

grapheneBoxGetWidth(box)

grapheneBoxInit(box, min, max)

grapheneBoxInitFromBox(box, src)

grapheneBoxInitFromPoints(box, n_points, points)

grapheneBoxInitFromVec3(box, min, max)

grapheneBoxInitFromVectors(box, n_vectors, vectors)

grapheneBoxIntersection(a, b)

grapheneBoxUnion(a, b)

grapheneBoxEmpty()

grapheneBoxInfinite()

grapheneBoxMinusOne()

grapheneBoxOne()

grapheneBoxOneMinusOne()

grapheneBoxZero()

grapheneEulerAlloc()

grapheneEulerEqual(a, b)

grapheneEulerFree(e)

grapheneEulerGetAlpha(e)

grapheneEulerGetBeta(e)

grapheneEulerGetGamma(e)

grapheneEulerGetOrder(e)

grapheneEulerGetX(e)

grapheneEulerGetY(e)

grapheneEulerGetZ(e)

grapheneEulerInit(e, x, y, z)

grapheneEulerInitFromEuler(e, src)

grapheneEulerInitFromMatrix(e, m, order)

grapheneEulerInitFromQuaternion(e, q, order)

grapheneEulerInitFromRadians(e, x, y, z, order)

grapheneEulerInitFromVec3(e, v, order)

grapheneEulerInitWithOrder(e, x, y, z, order)

grapheneEulerReorder(e, order)

grapheneEulerToMatrix(e)

grapheneEulerToQuaternion(e)

grapheneEulerToVec3(e)

grapheneFrustumAlloc()

grapheneFrustumContainsPoint(f, point)

grapheneFrustumEqual(a, b)

grapheneFrustumFree(f)

grapheneFrustumGetPlanes(f)

grapheneFrustumInit(f, p0, p1, p2, p3, p4, p5)

grapheneFrustumInitFromFrustum(f, src)

grapheneFrustumInitFromMatrix(f, matrix)

grapheneFrustumIntersectsBox(f, box)

grapheneFrustumIntersectsSphere(f, sphere)

grapheneMatrixAlloc()

grapheneMatrixDecompose(m)

grapheneMatrixDeterminant(m)

grapheneMatrixEqual(a, b)

grapheneMatrixEqualFast(a, b)

grapheneMatrixFree(m)

grapheneMatrixGetRow(m, index_)

grapheneMatrixGetValue(m, row, col)

grapheneMatrixGetXScale(m)

grapheneMatrixGetXTranslation(m)

grapheneMatrixGetYScale(m)

grapheneMatrixGetYTranslation(m)

grapheneMatrixGetZScale(m)

grapheneMatrixGetZTranslation(m)

grapheneMatrixInitFrom2d(m, xx, yx, xy, yy, x_0, y_0)

grapheneMatrixInitFromFloat(m, v)

grapheneMatrixInitFromMatrix(m, src)

grapheneMatrixInitFromVec4(m, v0, v1, v2, v3)

grapheneMatrixInitFrustum(m, left, right, bottom, top, z_near, z_far)

grapheneMatrixInitIdentity(m)

grapheneMatrixInitLookAt(m, eye, center, up)

grapheneMatrixInitOrtho(m, left, right, top, bottom, z_near, z_far)

grapheneMatrixInitPerspective(m, fovy, aspect, z_near, z_far)

grapheneMatrixInitRotate(m, angle, axis)

grapheneMatrixInitScale(m, x, y, z)

grapheneMatrixInitSkew(m, x_skew, y_skew)

grapheneMatrixInitTranslate(m, p)

grapheneMatrixInterpolate(a, b, factor)

grapheneMatrixInverse(m)

grapheneMatrixIs2d(m)

grapheneMatrixIsBackfaceVisible(m)

grapheneMatrixIsIdentity(m)

grapheneMatrixIsSingular(m)

grapheneMatrixMultiply(a, b)

grapheneMatrixNear(a, b, epsilon)

grapheneMatrixNormalize(m)

grapheneMatrixPerspective(m, depth)

grapheneMatrixPrint(m)

grapheneMatrixProjectPoint(m, p)

grapheneMatrixProjectRect(m, r)

grapheneMatrixProjectRectBounds(m, r)

grapheneMatrixRotate(m, angle, axis)

grapheneMatrixRotateEuler(m, e)

grapheneMatrixRotateQuaternion(m, q)

grapheneMatrixRotateX(m, angle)

grapheneMatrixRotateY(m, angle)

grapheneMatrixRotateZ(m, angle)

grapheneMatrixScale(m, factor_x, factor_y, factor_z)

grapheneMatrixSkewXy(m, factor)

grapheneMatrixSkewXz(m, factor)

grapheneMatrixSkewYz(m, factor)

grapheneMatrixTo2d(m)

grapheneMatrixToFloat(m)

grapheneMatrixTransformBounds(m, r)

grapheneMatrixTransformBox(m, b)

grapheneMatrixTransformPoint(m, p)

grapheneMatrixTransformPoint3d(m, p)

grapheneMatrixTransformRay(m, r)

grapheneMatrixTransformRect(m, r)

grapheneMatrixTransformSphere(m, s)

grapheneMatrixTransformVec3(m, v)

grapheneMatrixTransformVec4(m, v)

grapheneMatrixTranslate(m, pos)

grapheneMatrixTranspose(m)

grapheneMatrixUnprojectPoint3d(projection, modelview, point)

grapheneMatrixUntransformBounds(m, r, bounds)

grapheneMatrixUntransformPoint(m, p, bounds)

graphenePlaneAlloc()

graphenePlaneDistance(p, point)

graphenePlaneEqual(a, b)

graphenePlaneFree(p)

graphenePlaneGetConstant(p)

graphenePlaneGetNormal(p)

graphenePlaneInit(p, normal, constant)

graphenePlaneInitFromPlane(p, src)

graphenePlaneInitFromPoint(p, normal, point)

graphenePlaneInitFromPoints(p, a, b, c)

graphenePlaneInitFromVec4(p, src)

graphenePlaneNegate(p)

graphenePlaneNormalize(p)

graphenePlaneTransform(p, matrix, normal_matrix)

graphenePointAlloc()

graphenePointDistance(a, b)

graphenePointEqual(a, b)

graphenePointFree(p)

graphenePointInit(p, x, y)

graphenePointInitFromPoint(p, src)

graphenePointInitFromVec2(p, src)

graphenePointInterpolate(a, b, factor)

graphenePointNear(a, b, epsilon)

graphenePointToVec2(p)

graphenePointZero()

graphenePoint3dAlloc()

graphenePoint3dCross(a, b)

graphenePoint3dDistance(a, b)

graphenePoint3dDot(a, b)

graphenePoint3dEqual(a, b)

graphenePoint3dFree(p)

graphenePoint3dInit(p, x, y, z)

graphenePoint3dInitFromPoint(p, src)

graphenePoint3dInitFromVec3(p, v)

graphenePoint3dInterpolate(a, b, factor)

graphenePoint3dLength(p)

graphenePoint3dNear(a, b, epsilon)

graphenePoint3dNormalize(p)

graphenePoint3dNormalizeViewport(p, viewport, z_near, z_far)

graphenePoint3dScale(p, factor)

graphenePoint3dToVec3(p)

graphenePoint3dZero()

grapheneQuadAlloc()

grapheneQuadBounds(q)

grapheneQuadContains(q, p)

grapheneQuadFree(q)

grapheneQuadGetPoint(q, index_)

grapheneQuadInit(q, p1, p2, p3, p4)

grapheneQuadInitFromPoints(q, points)

grapheneQuadInitFromRect(q, r)

grapheneQuaternionAlloc()

grapheneQuaternionAdd(a, b)

grapheneQuaternionDot(a, b)

grapheneQuaternionEqual(a, b)

grapheneQuaternionFree(q)

grapheneQuaternionInit(q, x, y, z, w)

grapheneQuaternionInitFromAngleVec3(q, angle, axis)

grapheneQuaternionInitFromAngles(q, deg_x, deg_y, deg_z)

grapheneQuaternionInitFromEuler(q, e)

grapheneQuaternionInitFromMatrix(q, m)

grapheneQuaternionInitFromQuaternion(q, src)

grapheneQuaternionInitFromRadians(q, rad_x, rad_y, rad_z)

grapheneQuaternionInitFromVec4(q, src)

grapheneQuaternionInitIdentity(q)

grapheneQuaternionInvert(q)

grapheneQuaternionMultiply(a, b)

grapheneQuaternionNormalize(q)

grapheneQuaternionScale(q, factor)

grapheneQuaternionSlerp(a, b, factor)

grapheneQuaternionToAngleVec3(q)

grapheneQuaternionToAngles(q)

grapheneQuaternionToMatrix(q)

grapheneQuaternionToRadians(q)

grapheneQuaternionToVec4(q)

grapheneRayAlloc()

grapheneRayEqual(a, b)

grapheneRayFree(r)

grapheneRayGetClosestPointToPoint(r, p)

grapheneRayGetDirection(r)

grapheneRayGetDistanceToPlane(r, p)

grapheneRayGetDistanceToPoint(r, p)

grapheneRayGetOrigin(r)

grapheneRayGetPositionAt(r, t)

grapheneRayInit(r, origin, direction)

grapheneRayInitFromRay(r, src)

grapheneRayInitFromVec3(r, origin, direction)

grapheneRayIntersectBox(r, b)

grapheneRayIntersectSphere(r, s)

grapheneRayIntersectTriangle(r, t)

grapheneRayIntersectsBox(r, b)

grapheneRayIntersectsSphere(r, s)

grapheneRayIntersectsTriangle(r, t)

grapheneRectContainsPoint(r, p)

grapheneRectContainsRect(a, b)

grapheneRectEqual(a, b)

grapheneRectExpand(r, p)

grapheneRectFree(r)

grapheneRectGetArea(r)

grapheneRectGetBottomLeft(r)

grapheneRectGetBottomRight(r)

grapheneRectGetCenter(r)

grapheneRectGetHeight(r)

grapheneRectGetTopLeft(r)

grapheneRectGetTopRight(r)

grapheneRectGetVertices(r)

grapheneRectGetWidth(r)

grapheneRectGetX(r)

grapheneRectGetY(r)

grapheneRectInit(r, x, y, width, height)

grapheneRectInitFromRect(r, src)

grapheneRectInset(r, d_x, d_y)

grapheneRectInsetR(r, d_x, d_y)

grapheneRectInterpolate(a, b, factor)

grapheneRectIntersection(a, b)

grapheneRectNormalize(r)

grapheneRectNormalizeR(r)

grapheneRectOffset(r, d_x, d_y)

grapheneRectOffsetR(r, d_x, d_y)

grapheneRectRound(r)

grapheneRectRoundExtents(r)

grapheneRectRoundToPixel(r)

grapheneRectScale(r, s_h, s_v)

grapheneRectUnion(a, b)

grapheneRectAlloc()

grapheneRectZero()

grapheneSizeAlloc()

grapheneSizeEqual(a, b)

grapheneSizeFree(s)

grapheneSizeInit(s, width, height)

grapheneSizeInitFromSize(s, src)

grapheneSizeInterpolate(a, b, factor)

grapheneSizeScale(s, factor)

grapheneSizeZero()

grapheneSphereAlloc()

grapheneSphereContainsPoint(s, point)

grapheneSphereDistance(s, point)

grapheneSphereEqual(a, b)

grapheneSphereFree(s)

grapheneSphereGetBoundingBox(s)

grapheneSphereGetCenter(s)

grapheneSphereGetRadius(s)

grapheneSphereInit(s, center, radius)

grapheneSphereInitFromPoints(s, n_points, points, center)

grapheneSphereInitFromVectors(s, n_vectors, vectors, center)

grapheneSphereIsEmpty(s)

grapheneSphereTranslate(s, point)

grapheneTriangleAlloc()

grapheneTriangleContainsPoint(t, p)

grapheneTriangleEqual(a, b)

grapheneTriangleFree(t)

grapheneTriangleGetArea(t)

grapheneTriangleGetBarycoords(t, p)

grapheneTriangleGetBoundingBox(t)

grapheneTriangleGetMidpoint(t)

grapheneTriangleGetNormal(t)

grapheneTriangleGetPlane(t)

grapheneTriangleGetPoints(t)

grapheneTriangleGetUv(t, p, uv_a, uv_b, uv_c)

grapheneTriangleGetVertices(t)

grapheneTriangleInitFromFloat(t, a, b, c)

grapheneTriangleInitFromPoint3d(t, a, b, c)

grapheneTriangleInitFromVec3(t, a, b, c)

grapheneVec2Alloc()

grapheneVec2Add(a, b)

grapheneVec2Divide(a, b)

grapheneVec2Dot(a, b)

grapheneVec2Equal(v1, v2)

grapheneVec2Free(v)

grapheneVec2GetX(v)

grapheneVec2GetY(v)

grapheneVec2Init(v, x, y)

grapheneVec2InitFromFloat(v, src)

grapheneVec2InitFromVec2(v, src)

grapheneVec2Interpolate(v1, v2, factor)

grapheneVec2Length(v)

grapheneVec2Max(a, b)

grapheneVec2Min(a, b)

grapheneVec2Multiply(a, b)

grapheneVec2Near(v1, v2, epsilon)

grapheneVec2Negate(v)

grapheneVec2Normalize(v)

grapheneVec2Scale(v, factor)

grapheneVec2Subtract(a, b)

grapheneVec2ToFloat(v)

grapheneVec2One()

grapheneVec2XAxis()

grapheneVec2YAxis()

grapheneVec2Zero()

grapheneVec3Alloc()

grapheneVec3Add(a, b)

grapheneVec3Cross(a, b)

grapheneVec3Divide(a, b)

grapheneVec3Dot(a, b)

grapheneVec3Equal(v1, v2)

grapheneVec3Free(v)

grapheneVec3GetX(v)

grapheneVec3GetXy(v)

grapheneVec3GetXy0(v)

grapheneVec3GetXyz0(v)

grapheneVec3GetXyz1(v)

grapheneVec3GetXyzw(v, w)

grapheneVec3GetY(v)

grapheneVec3GetZ(v)

grapheneVec3Init(v, x, y, z)

grapheneVec3InitFromFloat(v, src)

grapheneVec3InitFromVec3(v, src)

grapheneVec3Interpolate(v1, v2, factor)

grapheneVec3Length(v)

grapheneVec3Max(a, b)

grapheneVec3Min(a, b)

grapheneVec3Multiply(a, b)

grapheneVec3Near(v1, v2, epsilon)

grapheneVec3Negate(v)

grapheneVec3Normalize(v)

grapheneVec3Scale(v, factor)

grapheneVec3Subtract(a, b)

grapheneVec3ToFloat(v)

grapheneVec3One()

grapheneVec3XAxis()

grapheneVec3YAxis()

grapheneVec3ZAxis()

grapheneVec3Zero()

grapheneVec4Alloc()

grapheneVec4Add(a, b)

grapheneVec4Divide(a, b)

grapheneVec4Dot(a, b)

grapheneVec4Equal(v1, v2)

grapheneVec4Free(v)

grapheneVec4GetW(v)

grapheneVec4GetX(v)

grapheneVec4GetXy(v)

grapheneVec4GetXyz(v)

grapheneVec4GetY(v)

grapheneVec4GetZ(v)

grapheneVec4Init(v, x, y, z, w)

grapheneVec4InitFromFloat(v, src)

grapheneVec4InitFromVec2(v, src, z, w)

grapheneVec4InitFromVec3(v, src, w)

grapheneVec4InitFromVec4(v, src)

grapheneVec4Interpolate(v1, v2, factor)

grapheneVec4Length(v)

grapheneVec4Max(a, b)

grapheneVec4Min(a, b)

grapheneVec4Multiply(a, b)

grapheneVec4Near(v1, v2, epsilon)

grapheneVec4Negate(v)

grapheneVec4Normalize(v)

grapheneVec4Scale(v, factor)

grapheneVec4Subtract(a, b)

grapheneVec4ToFloat(v)

grapheneVec4One()

grapheneVec4WAxis()

grapheneVec4XAxis()

grapheneVec4YAxis()

grapheneVec4ZAxis()

grapheneVec4Zero()
}
\arguments{
\item{a}{Vec4}

\item{b}{Vec4}

\item{box}{Box}

\item{point}{Point3D}

\item{scalar}{gfloat}

\item{vec}{Vec3}

\item{min}{Vec3}

\item{max}{Vec3}

\item{src}{Vec4}

\item{n_points}{guint}

\item{points}{Point3D}

\item{n_vectors}{guint}

\item{vectors}{Vec3}

\item{e}{Euler}

\item{x}{gfloat}

\item{y}{gfloat}

\item{z}{gfloat}

\item{m}{Matrix}

\item{order}{EulerOrder}

\item{q}{Quaternion}

\item{v}{Vec4}

\item{f}{Frustum}

\item{p0}{Plane}

\item{p1}{Point}

\item{p2}{Point}

\item{p3}{Point}

\item{p4}{Point}

\item{p5}{Plane}

\item{matrix}{Matrix}

\item{sphere}{Sphere}

\item{index_}{guint}

\item{row}{guint}

\item{col}{guint}

\item{xx}{gdouble}

\item{yx}{gdouble}

\item{xy}{gdouble}

\item{yy}{gdouble}

\item{x_0}{gdouble}

\item{y_0}{gdouble}

\item{v0}{Vec4}

\item{v1}{Vec4}

\item{v2}{Vec4}

\item{v3}{Vec4}

\item{left}{gfloat}

\item{right}{gfloat}

\item{bottom}{gfloat}

\item{top}{gfloat}

\item{z_near}{gfloat}

\item{z_far}{gfloat}

\item{eye}{Vec3}

\item{center}{Point3D}

\item{up}{Vec3}

\item{fovy}{gfloat}

\item{aspect}{gfloat}

\item{angle}{gfloat}

\item{axis}{Vec3}

\item{x_skew}{gfloat}

\item{y_skew}{gfloat}

\item{p}{Point3D}

\item{factor}{gfloat}

\item{epsilon}{gfloat}

\item{depth}{gfloat}

\item{r}{Rect}

\item{factor_x}{gfloat}

\item{factor_y}{gfloat}

\item{factor_z}{gfloat}

\item{s}{Sphere}

\item{pos}{Point3D}

\item{projection}{Matrix}

\item{modelview}{Matrix}

\item{bounds}{Rect}

\item{normal}{Vec3}

\item{constant}{gfloat}

\item{c}{Vec3}

\item{normal_matrix}{Matrix}

\item{viewport}{Rect}

\item{w}{gfloat}

\item{deg_x}{gfloat}

\item{deg_y}{gfloat}

\item{deg_z}{gfloat}

\item{rad_x}{gfloat}

\item{rad_y}{gfloat}

\item{rad_z}{gfloat}

\item{t}{Triangle}

\item{origin}{Vec3}

\item{direction}{Vec3}

\item{width}{gfloat}

\item{height}{gfloat}

\item{d_x}{gfloat}

\item{d_y}{gfloat}

\item{s_h}{gfloat}

\item{s_v}{gfloat}

\item{radius}{gfloat}

\item{uv_a}{Vec2}

\item{uv_b}{Vec2}

\item{uv_c}{Vec2}
}
\value{
Box

gboolean

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gfloat

gfloat

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gfloat

Box

Box

Box

Box

Box

gboolean

Return value from C function

Box

Box

Box

Box

Box

Box

Euler

gboolean

Return value from C function

gfloat

gfloat

gfloat

EulerOrder

gfloat

gfloat

gfloat

Euler

Euler

Euler

Euler

Euler

Euler

Euler

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Frustum

gboolean

gboolean

Return value from C function

Return value from C function

Frustum

Frustum

Frustum

gboolean

gboolean

Matrix

gboolean

gfloat

gboolean

gboolean

Return value from C function

Return value from C function

gfloat

gfloat

gfloat

gfloat

gfloat

gfloat

gfloat

Matrix

Matrix

Matrix

Matrix

Matrix

Matrix

Matrix

Matrix

Matrix

Matrix

Matrix

Matrix

Matrix

Return value from C function

gboolean

gboolean

gboolean

gboolean

gboolean

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

Plane

gfloat

gboolean

Return value from C function

gfloat

Return value from C function

Plane

Plane

Plane

Plane

Plane

Return value from C function

Return value from C function

Return value from C function

Point

gfloat

gboolean

Return value from C function

Point

Point

Point

Return value from C function

gboolean

Return value from C function

Point

Point3D

Return value from C function

gfloat

gfloat

gboolean

Return value from C function

Point3D

Point3D

Point3D

Return value from C function

gfloat

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Point3D

Quad

Return value from C function

gboolean

Return value from C function

Point

Quad

Quad

Quad

Quaternion

Return value from C function

gfloat

gboolean

Return value from C function

Quaternion

Quaternion

Quaternion

Quaternion

Quaternion

Quaternion

Quaternion

Quaternion

Quaternion

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Ray

gboolean

Return value from C function

Return value from C function

Return value from C function

gfloat

gfloat

Return value from C function

Return value from C function

Ray

Ray

Ray

RayIntersectionKind

RayIntersectionKind

RayIntersectionKind

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

Return value from C function

Return value from C function

gfloat

Return value from C function

Return value from C function

Return value from C function

gfloat

Return value from C function

Return value from C function

Return value from C function

gfloat

gfloat

gfloat

Rect

Rect

Rect

Return value from C function

Return value from C function

gboolean

Rect

Return value from C function

Rect

Return value from C function

Return value from C function

Return value from C function

Rect

Return value from C function

Return value from C function

Rect

Rect

Size

gboolean

Return value from C function

Size

Size

Return value from C function

Return value from C function

Size

Sphere

gboolean

gfloat

gboolean

Return value from C function

Return value from C function

Return value from C function

gfloat

Sphere

Sphere

Sphere

gboolean

Return value from C function

Triangle

gboolean

gboolean

Return value from C function

gfloat

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Triangle

Triangle

Triangle

Vec2

Return value from C function

Return value from C function

gfloat

gboolean

Return value from C function

gfloat

gfloat

Vec2

Vec2

Vec2

Return value from C function

gfloat

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Vec2

Vec2

Vec2

Vec2

Vec3

Return value from C function

Return value from C function

Return value from C function

gfloat

gboolean

Return value from C function

gfloat

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gfloat

gfloat

Vec3

Vec3

Vec3

Return value from C function

gfloat

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Vec3

Vec3

Vec3

Vec3

Vec3

Vec4

Return value from C function

Return value from C function

gfloat

gboolean

Return value from C function

gfloat

gfloat

Return value from C function

Return value from C function

gfloat

gfloat

Vec4

Vec4

Vec4

Vec4

Vec4

Return value from C function

gfloat

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Vec4

Vec4

Vec4

Vec4

Vec4

Vec4
}
\description{
grapheneBoxAlloc

grapheneBoxContainsBox

grapheneBoxContainsPoint

grapheneBoxEqual

grapheneBoxExpand

grapheneBoxExpandScalar

grapheneBoxExpandVec3

grapheneBoxFree

grapheneBoxGetBoundingSphere

grapheneBoxGetCenter

grapheneBoxGetDepth

grapheneBoxGetHeight

grapheneBoxGetMax

grapheneBoxGetMin

grapheneBoxGetSize

grapheneBoxGetVertices

grapheneBoxGetWidth

grapheneBoxInit

grapheneBoxInitFromBox

grapheneBoxInitFromPoints

grapheneBoxInitFromVec3

grapheneBoxInitFromVectors

grapheneBoxIntersection

grapheneBoxUnion

grapheneBoxEmpty

grapheneBoxInfinite

grapheneBoxMinusOne

grapheneBoxOne

grapheneBoxOneMinusOne

grapheneBoxZero

grapheneEulerAlloc

grapheneEulerEqual

grapheneEulerFree

grapheneEulerGetAlpha

grapheneEulerGetBeta

grapheneEulerGetGamma

grapheneEulerGetOrder

grapheneEulerGetX

grapheneEulerGetY

grapheneEulerGetZ

grapheneEulerInit

grapheneEulerInitFromEuler

grapheneEulerInitFromMatrix

grapheneEulerInitFromQuaternion

grapheneEulerInitFromRadians

grapheneEulerInitFromVec3

grapheneEulerInitWithOrder

grapheneEulerReorder

grapheneEulerToMatrix

grapheneEulerToQuaternion

grapheneEulerToVec3

grapheneFrustumAlloc

grapheneFrustumContainsPoint

grapheneFrustumEqual

grapheneFrustumFree

grapheneFrustumGetPlanes

grapheneFrustumInit

grapheneFrustumInitFromFrustum

grapheneFrustumInitFromMatrix

grapheneFrustumIntersectsBox

grapheneFrustumIntersectsSphere

grapheneMatrixAlloc

grapheneMatrixDecompose

grapheneMatrixDeterminant

grapheneMatrixEqual

grapheneMatrixEqualFast

grapheneMatrixFree

grapheneMatrixGetRow

grapheneMatrixGetValue

grapheneMatrixGetXScale

grapheneMatrixGetXTranslation

grapheneMatrixGetYScale

grapheneMatrixGetYTranslation

grapheneMatrixGetZScale

grapheneMatrixGetZTranslation

grapheneMatrixInitFrom2d

grapheneMatrixInitFromFloat

grapheneMatrixInitFromMatrix

grapheneMatrixInitFromVec4

grapheneMatrixInitFrustum

grapheneMatrixInitIdentity

grapheneMatrixInitLookAt

grapheneMatrixInitOrtho

grapheneMatrixInitPerspective

grapheneMatrixInitRotate

grapheneMatrixInitScale

grapheneMatrixInitSkew

grapheneMatrixInitTranslate

grapheneMatrixInterpolate

grapheneMatrixInverse

grapheneMatrixIs2d

grapheneMatrixIsBackfaceVisible

grapheneMatrixIsIdentity

grapheneMatrixIsSingular

grapheneMatrixMultiply

grapheneMatrixNear

grapheneMatrixNormalize

grapheneMatrixPerspective

grapheneMatrixPrint

grapheneMatrixProjectPoint

grapheneMatrixProjectRect

grapheneMatrixProjectRectBounds

grapheneMatrixRotate

grapheneMatrixRotateEuler

grapheneMatrixRotateQuaternion

grapheneMatrixRotateX

grapheneMatrixRotateY

grapheneMatrixRotateZ

grapheneMatrixScale

grapheneMatrixSkewXy

grapheneMatrixSkewXz

grapheneMatrixSkewYz

grapheneMatrixTo2d

grapheneMatrixToFloat

grapheneMatrixTransformBounds

grapheneMatrixTransformBox

grapheneMatrixTransformPoint

grapheneMatrixTransformPoint3d

grapheneMatrixTransformRay

grapheneMatrixTransformRect

grapheneMatrixTransformSphere

grapheneMatrixTransformVec3

grapheneMatrixTransformVec4

grapheneMatrixTranslate

grapheneMatrixTranspose

grapheneMatrixUnprojectPoint3d

grapheneMatrixUntransformBounds

grapheneMatrixUntransformPoint

graphenePlaneAlloc

graphenePlaneDistance

graphenePlaneEqual

graphenePlaneFree

graphenePlaneGetConstant

graphenePlaneGetNormal

graphenePlaneInit

graphenePlaneInitFromPlane

graphenePlaneInitFromPoint

graphenePlaneInitFromPoints

graphenePlaneInitFromVec4

graphenePlaneNegate

graphenePlaneNormalize

graphenePlaneTransform

graphenePointAlloc

graphenePointDistance

graphenePointEqual

graphenePointFree

graphenePointInit

graphenePointInitFromPoint

graphenePointInitFromVec2

graphenePointInterpolate

graphenePointNear

graphenePointToVec2

graphenePointZero

graphenePoint3dAlloc

graphenePoint3dCross

graphenePoint3dDistance

graphenePoint3dDot

graphenePoint3dEqual

graphenePoint3dFree

graphenePoint3dInit

graphenePoint3dInitFromPoint

graphenePoint3dInitFromVec3

graphenePoint3dInterpolate

graphenePoint3dLength

graphenePoint3dNear

graphenePoint3dNormalize

graphenePoint3dNormalizeViewport

graphenePoint3dScale

graphenePoint3dToVec3

graphenePoint3dZero

grapheneQuadAlloc

grapheneQuadBounds

grapheneQuadContains

grapheneQuadFree

grapheneQuadGetPoint

grapheneQuadInit

grapheneQuadInitFromPoints

grapheneQuadInitFromRect

grapheneQuaternionAlloc

grapheneQuaternionAdd

grapheneQuaternionDot

grapheneQuaternionEqual

grapheneQuaternionFree

grapheneQuaternionInit

grapheneQuaternionInitFromAngleVec3

grapheneQuaternionInitFromAngles

grapheneQuaternionInitFromEuler

grapheneQuaternionInitFromMatrix

grapheneQuaternionInitFromQuaternion

grapheneQuaternionInitFromRadians

grapheneQuaternionInitFromVec4

grapheneQuaternionInitIdentity

grapheneQuaternionInvert

grapheneQuaternionMultiply

grapheneQuaternionNormalize

grapheneQuaternionScale

grapheneQuaternionSlerp

grapheneQuaternionToAngleVec3

grapheneQuaternionToAngles

grapheneQuaternionToMatrix

grapheneQuaternionToRadians

grapheneQuaternionToVec4

grapheneRayAlloc

grapheneRayEqual

grapheneRayFree

grapheneRayGetClosestPointToPoint

grapheneRayGetDirection

grapheneRayGetDistanceToPlane

grapheneRayGetDistanceToPoint

grapheneRayGetOrigin

grapheneRayGetPositionAt

grapheneRayInit

grapheneRayInitFromRay

grapheneRayInitFromVec3

grapheneRayIntersectBox

grapheneRayIntersectSphere

grapheneRayIntersectTriangle

grapheneRayIntersectsBox

grapheneRayIntersectsSphere

grapheneRayIntersectsTriangle

grapheneRectContainsPoint

grapheneRectContainsRect

grapheneRectEqual

grapheneRectExpand

grapheneRectFree

grapheneRectGetArea

grapheneRectGetBottomLeft

grapheneRectGetBottomRight

grapheneRectGetCenter

grapheneRectGetHeight

grapheneRectGetTopLeft

grapheneRectGetTopRight

grapheneRectGetVertices

grapheneRectGetWidth

grapheneRectGetX

grapheneRectGetY

grapheneRectInit

grapheneRectInitFromRect

grapheneRectInset

grapheneRectInsetR

grapheneRectInterpolate

grapheneRectIntersection

grapheneRectNormalize

grapheneRectNormalizeR

grapheneRectOffset

grapheneRectOffsetR

grapheneRectRound

grapheneRectRoundExtents

grapheneRectRoundToPixel

grapheneRectScale

grapheneRectUnion

grapheneRectAlloc

grapheneRectZero

grapheneSizeAlloc

grapheneSizeEqual

grapheneSizeFree

grapheneSizeInit

grapheneSizeInitFromSize

grapheneSizeInterpolate

grapheneSizeScale

grapheneSizeZero

grapheneSphereAlloc

grapheneSphereContainsPoint

grapheneSphereDistance

grapheneSphereEqual

grapheneSphereFree

grapheneSphereGetBoundingBox

grapheneSphereGetCenter

grapheneSphereGetRadius

grapheneSphereInit

grapheneSphereInitFromPoints

grapheneSphereInitFromVectors

grapheneSphereIsEmpty

grapheneSphereTranslate

grapheneTriangleAlloc

grapheneTriangleContainsPoint

grapheneTriangleEqual

grapheneTriangleFree

grapheneTriangleGetArea

grapheneTriangleGetBarycoords

grapheneTriangleGetBoundingBox

grapheneTriangleGetMidpoint

grapheneTriangleGetNormal

grapheneTriangleGetPlane

grapheneTriangleGetPoints

grapheneTriangleGetUv

grapheneTriangleGetVertices

grapheneTriangleInitFromFloat

grapheneTriangleInitFromPoint3d

grapheneTriangleInitFromVec3

grapheneVec2Alloc

grapheneVec2Add

grapheneVec2Divide

grapheneVec2Dot

grapheneVec2Equal

grapheneVec2Free

grapheneVec2GetX

grapheneVec2GetY

grapheneVec2Init

grapheneVec2InitFromFloat

grapheneVec2InitFromVec2

grapheneVec2Interpolate

grapheneVec2Length

grapheneVec2Max

grapheneVec2Min

grapheneVec2Multiply

grapheneVec2Near

grapheneVec2Negate

grapheneVec2Normalize

grapheneVec2Scale

grapheneVec2Subtract

grapheneVec2ToFloat

grapheneVec2One

grapheneVec2XAxis

grapheneVec2YAxis

grapheneVec2Zero

grapheneVec3Alloc

grapheneVec3Add

grapheneVec3Cross

grapheneVec3Divide

grapheneVec3Dot

grapheneVec3Equal

grapheneVec3Free

grapheneVec3GetX

grapheneVec3GetXy

grapheneVec3GetXy0

grapheneVec3GetXyz0

grapheneVec3GetXyz1

grapheneVec3GetXyzw

grapheneVec3GetY

grapheneVec3GetZ

grapheneVec3Init

grapheneVec3InitFromFloat

grapheneVec3InitFromVec3

grapheneVec3Interpolate

grapheneVec3Length

grapheneVec3Max

grapheneVec3Min

grapheneVec3Multiply

grapheneVec3Near

grapheneVec3Negate

grapheneVec3Normalize

grapheneVec3Scale

grapheneVec3Subtract

grapheneVec3ToFloat

grapheneVec3One

grapheneVec3XAxis

grapheneVec3YAxis

grapheneVec3ZAxis

grapheneVec3Zero

grapheneVec4Alloc

grapheneVec4Add

grapheneVec4Divide

grapheneVec4Dot

grapheneVec4Equal

grapheneVec4Free

grapheneVec4GetW

grapheneVec4GetX

grapheneVec4GetXy

grapheneVec4GetXyz

grapheneVec4GetY

grapheneVec4GetZ

grapheneVec4Init

grapheneVec4InitFromFloat

grapheneVec4InitFromVec2

grapheneVec4InitFromVec3

grapheneVec4InitFromVec4

grapheneVec4Interpolate

grapheneVec4Length

grapheneVec4Max

grapheneVec4Min

grapheneVec4Multiply

grapheneVec4Near

grapheneVec4Negate

grapheneVec4Normalize

grapheneVec4Scale

grapheneVec4Subtract

grapheneVec4ToFloat

grapheneVec4One

grapheneVec4WAxis

grapheneVec4XAxis

grapheneVec4YAxis

grapheneVec4ZAxis

grapheneVec4Zero
}
