% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gSignalConnectR}
\alias{gSignalConnectR}
\title{Connect an R function to a GObject signal}
\usage{
gSignalConnectR(obj, signal, fun)
}
\arguments{
\item{obj}{External pointer to a GObject}

\item{signal}{Signal name (character)}

\item{fun}{R function to invoke on signal. For signals that expect a boolean return
(like "close-request"), your function should return TRUE to stop propagation or
FALSE to allow default behavior.}
}
\value{
handler id (numeric) - invisibly
}
\description{
Connect an R function to a GObject signal
}
