% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gSignalConnectRBoolean}
\alias{gSignalConnectRBoolean}
\title{Connect an R function to a signal that returns boolean}
\usage{
gSignalConnectRBoolean(obj, signal, fun)
}
\arguments{
\item{obj}{External pointer to a GObject}

\item{signal}{Signal name (character)}

\item{fun}{R function to invoke on signal - must return TRUE or FALSE}
}
\value{
handler id (numeric)
}
\description{
Use this for signals like "close-request" that expect a boolean return value.
Your function should return TRUE to stop propagation, FALSE to allow default behavior.
}
