% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gsk_autogen.R
\name{gskBlendNodeNew}
\alias{gskBlendNodeNew}
\alias{gskBlendNodeGetBlendMode}
\alias{gskBlendNodeGetBottomChild}
\alias{gskBlendNodeGetTopChild}
\alias{gskBlurNodeNew}
\alias{gskBlurNodeGetChild}
\alias{gskBlurNodeGetRadius}
\alias{gskBorderNodeNew}
\alias{gskBorderNodeGetColors}
\alias{gskBorderNodeGetOutline}
\alias{gskBorderNodeGetWidths}
\alias{gskCairoNodeNew}
\alias{gskCairoNodeGetDrawContext}
\alias{gskCairoNodeGetSurface}
\alias{gskCairoRendererNew}
\alias{gskClipNodeNew}
\alias{gskClipNodeGetChild}
\alias{gskClipNodeGetClip}
\alias{gskColorMatrixNodeNew}
\alias{gskColorMatrixNodeGetChild}
\alias{gskColorMatrixNodeGetColorMatrix}
\alias{gskColorMatrixNodeGetColorOffset}
\alias{gskColorNodeNew}
\alias{gskColorNodeGetColor}
\alias{gskConicGradientNodeNew}
\alias{gskConicGradientNodeGetAngle}
\alias{gskConicGradientNodeGetCenter}
\alias{gskConicGradientNodeGetColorStops}
\alias{gskConicGradientNodeGetNColorStops}
\alias{gskConicGradientNodeGetRotation}
\alias{gskContainerNodeNew}
\alias{gskContainerNodeGetChild}
\alias{gskContainerNodeGetNChildren}
\alias{gskCrossFadeNodeNew}
\alias{gskCrossFadeNodeGetEndChild}
\alias{gskCrossFadeNodeGetProgress}
\alias{gskCrossFadeNodeGetStartChild}
\alias{gskDebugNodeNew}
\alias{gskDebugNodeGetChild}
\alias{gskDebugNodeGetMessage}
\alias{gskFillNodeNew}
\alias{gskFillNodeGetChild}
\alias{gskFillNodeGetFillRule}
\alias{gskFillNodeGetPath}
\alias{gskGlRendererNew}
\alias{gskGlShaderNewFromBytes}
\alias{gskGlShaderNewFromResource}
\alias{gskGlShaderCompile}
\alias{gskGlShaderFindUniformByName}
\alias{gskGlShaderGetArgBool}
\alias{gskGlShaderGetArgFloat}
\alias{gskGlShaderGetArgInt}
\alias{gskGlShaderGetArgUint}
\alias{gskGlShaderGetArgVec2}
\alias{gskGlShaderGetArgVec3}
\alias{gskGlShaderGetArgVec4}
\alias{gskGlShaderGetArgsSize}
\alias{gskGlShaderGetNTextures}
\alias{gskGlShaderGetNUniforms}
\alias{gskGlShaderGetResource}
\alias{gskGlShaderGetSource}
\alias{gskGlShaderGetUniformName}
\alias{gskGlShaderGetUniformOffset}
\alias{gskGlShaderGetUniformType}
\alias{gskGlShaderNodeNew}
\alias{gskGlShaderNodeGetArgs}
\alias{gskGlShaderNodeGetChild}
\alias{gskGlShaderNodeGetNChildren}
\alias{gskGlShaderNodeGetShader}
\alias{gskInsetShadowNodeNew}
\alias{gskInsetShadowNodeGetBlurRadius}
\alias{gskInsetShadowNodeGetColor}
\alias{gskInsetShadowNodeGetDx}
\alias{gskInsetShadowNodeGetDy}
\alias{gskInsetShadowNodeGetOutline}
\alias{gskInsetShadowNodeGetSpread}
\alias{gskLinearGradientNodeNew}
\alias{gskLinearGradientNodeGetColorStops}
\alias{gskLinearGradientNodeGetEnd}
\alias{gskLinearGradientNodeGetNColorStops}
\alias{gskLinearGradientNodeGetStart}
\alias{gskMaskNodeNew}
\alias{gskMaskNodeGetMask}
\alias{gskMaskNodeGetMaskMode}
\alias{gskMaskNodeGetSource}
\alias{gskNglRendererNew}
\alias{gskOpacityNodeNew}
\alias{gskOpacityNodeGetChild}
\alias{gskOpacityNodeGetOpacity}
\alias{gskOutsetShadowNodeNew}
\alias{gskOutsetShadowNodeGetBlurRadius}
\alias{gskOutsetShadowNodeGetColor}
\alias{gskOutsetShadowNodeGetDx}
\alias{gskOutsetShadowNodeGetDy}
\alias{gskOutsetShadowNodeGetOutline}
\alias{gskOutsetShadowNodeGetSpread}
\alias{gskRadialGradientNodeNew}
\alias{gskRadialGradientNodeGetCenter}
\alias{gskRadialGradientNodeGetColorStops}
\alias{gskRadialGradientNodeGetEnd}
\alias{gskRadialGradientNodeGetHradius}
\alias{gskRadialGradientNodeGetNColorStops}
\alias{gskRadialGradientNodeGetStart}
\alias{gskRadialGradientNodeGetVradius}
\alias{gskRenderNodeDeserialize}
\alias{gskRenderNodeDraw}
\alias{gskRenderNodeGetBounds}
\alias{gskRenderNodeGetNodeType}
\alias{gskRenderNodeRef}
\alias{gskRenderNodeSerialize}
\alias{gskRenderNodeUnref}
\alias{gskRenderNodeWriteToFile}
\alias{gskRendererNewForSurface}
\alias{gskRendererGetSurface}
\alias{gskRendererIsRealized}
\alias{gskRendererRealize}
\alias{gskRendererRealizeForDisplay}
\alias{gskRendererRender}
\alias{gskRendererRenderTexture}
\alias{gskRendererUnrealize}
\alias{gskRepeatNodeNew}
\alias{gskRepeatNodeGetChild}
\alias{gskRepeatNodeGetChildBounds}
\alias{gskRepeatingLinearGradientNodeNew}
\alias{gskRepeatingRadialGradientNodeNew}
\alias{gskRoundedClipNodeNew}
\alias{gskRoundedClipNodeGetChild}
\alias{gskRoundedClipNodeGetClip}
\alias{gskRoundedRectContainsPoint}
\alias{gskRoundedRectContainsRect}
\alias{gskRoundedRectInit}
\alias{gskRoundedRectInitCopy}
\alias{gskRoundedRectInitFromRect}
\alias{gskRoundedRectIntersectsRect}
\alias{gskRoundedRectIsRectilinear}
\alias{gskRoundedRectNormalize}
\alias{gskRoundedRectOffset}
\alias{gskRoundedRectShrink}
\alias{gskSerializationErrorQuark}
\alias{gskShaderArgsBuilderNew}
\alias{gskShaderArgsBuilderRef}
\alias{gskShaderArgsBuilderSetBool}
\alias{gskShaderArgsBuilderSetFloat}
\alias{gskShaderArgsBuilderSetInt}
\alias{gskShaderArgsBuilderSetUint}
\alias{gskShaderArgsBuilderSetVec2}
\alias{gskShaderArgsBuilderSetVec3}
\alias{gskShaderArgsBuilderSetVec4}
\alias{gskShaderArgsBuilderToArgs}
\alias{gskShaderArgsBuilderUnref}
\alias{gskShadowNodeNew}
\alias{gskShadowNodeGetChild}
\alias{gskShadowNodeGetNShadows}
\alias{gskShadowNodeGetShadow}
\alias{gskStrokeNew}
\alias{gskStrokeCopy}
\alias{gskStrokeFree}
\alias{gskStrokeGetDash}
\alias{gskStrokeGetDashOffset}
\alias{gskStrokeGetLineCap}
\alias{gskStrokeGetLineJoin}
\alias{gskStrokeGetLineWidth}
\alias{gskStrokeGetMiterLimit}
\alias{gskStrokeSetDash}
\alias{gskStrokeSetDashOffset}
\alias{gskStrokeSetLineCap}
\alias{gskStrokeSetLineJoin}
\alias{gskStrokeSetLineWidth}
\alias{gskStrokeSetMiterLimit}
\alias{gskStrokeToCairo}
\alias{gskStrokeEqual}
\alias{gskStrokeNodeNew}
\alias{gskStrokeNodeGetChild}
\alias{gskStrokeNodeGetPath}
\alias{gskStrokeNodeGetStroke}
\alias{gskSubsurfaceNodeGetChild}
\alias{gskTextNodeNew}
\alias{gskTextNodeGetColor}
\alias{gskTextNodeGetFont}
\alias{gskTextNodeGetGlyphs}
\alias{gskTextNodeGetNumGlyphs}
\alias{gskTextNodeGetOffset}
\alias{gskTextNodeHasColorGlyphs}
\alias{gskTextureNodeNew}
\alias{gskTextureNodeGetTexture}
\alias{gskTextureScaleNodeNew}
\alias{gskTextureScaleNodeGetFilter}
\alias{gskTextureScaleNodeGetTexture}
\alias{gskTransformNew}
\alias{gskTransformEqual}
\alias{gskTransformGetCategory}
\alias{gskTransformInvert}
\alias{gskTransformMatrix}
\alias{gskTransformPerspective}
\alias{gskTransformPrint}
\alias{gskTransformRef}
\alias{gskTransformRotate}
\alias{gskTransformRotate3d}
\alias{gskTransformScale}
\alias{gskTransformScale3d}
\alias{gskTransformSkew}
\alias{gskTransformTo2d}
\alias{gskTransformTo2dComponents}
\alias{gskTransformToAffine}
\alias{gskTransformToMatrix}
\alias{gskTransformToString}
\alias{gskTransformToTranslate}
\alias{gskTransformTransform}
\alias{gskTransformTransformBounds}
\alias{gskTransformTransformPoint}
\alias{gskTransformTranslate}
\alias{gskTransformTranslate3d}
\alias{gskTransformUnref}
\alias{gskTransformParse}
\alias{gskTransformNodeNew}
\alias{gskTransformNodeGetChild}
\alias{gskTransformNodeGetTransform}
\alias{gskVulkanRendererNew}
\alias{gskValueDupRenderNode}
\alias{gskValueGetRenderNode}
\alias{gskValueSetRenderNode}
\alias{gskValueTakeRenderNode}
\title{gskBlendNodeNew}
\usage{
gskBlendNodeNew(bottom, top, blend_mode)

gskBlendNodeGetBlendMode(node)

gskBlendNodeGetBottomChild(node)

gskBlendNodeGetTopChild(node)

gskBlurNodeNew(child, radius)

gskBlurNodeGetChild(node)

gskBlurNodeGetRadius(node)

gskBorderNodeNew(outline, border_width, border_color)

gskBorderNodeGetColors(node)

gskBorderNodeGetOutline(node)

gskBorderNodeGetWidths(node)

gskCairoNodeNew(bounds)

gskCairoNodeGetDrawContext(node)

gskCairoNodeGetSurface(node)

gskCairoRendererNew()

gskClipNodeNew(child, clip)

gskClipNodeGetChild(node)

gskClipNodeGetClip(node)

gskColorMatrixNodeNew(child, color_matrix, color_offset)

gskColorMatrixNodeGetChild(node)

gskColorMatrixNodeGetColorMatrix(node)

gskColorMatrixNodeGetColorOffset(node)

gskColorNodeNew(rgba, bounds)

gskColorNodeGetColor(node)

gskConicGradientNodeNew(bounds, center, rotation, color_stops, n_color_stops)

gskConicGradientNodeGetAngle(node)

gskConicGradientNodeGetCenter(node)

gskConicGradientNodeGetColorStops(node)

gskConicGradientNodeGetNColorStops(node)

gskConicGradientNodeGetRotation(node)

gskContainerNodeNew(children, n_children)

gskContainerNodeGetChild(node, idx)

gskContainerNodeGetNChildren(node)

gskCrossFadeNodeNew(start, end, progress)

gskCrossFadeNodeGetEndChild(node)

gskCrossFadeNodeGetProgress(node)

gskCrossFadeNodeGetStartChild(node)

gskDebugNodeNew(child, message)

gskDebugNodeGetChild(node)

gskDebugNodeGetMessage(node)

gskFillNodeNew(child, path, fill_rule)

gskFillNodeGetChild(node)

gskFillNodeGetFillRule(node)

gskFillNodeGetPath(node)

gskGlRendererNew()

gskGlShaderNewFromBytes(sourcecode)

gskGlShaderNewFromResource(resource_path)

gskGlShaderCompile(shader, renderer)

gskGlShaderFindUniformByName(shader, name)

gskGlShaderGetArgBool(shader, args, idx)

gskGlShaderGetArgFloat(shader, args, idx)

gskGlShaderGetArgInt(shader, args, idx)

gskGlShaderGetArgUint(shader, args, idx)

gskGlShaderGetArgVec2(shader, args, idx, out_value)

gskGlShaderGetArgVec3(shader, args, idx, out_value)

gskGlShaderGetArgVec4(shader, args, idx, out_value)

gskGlShaderGetArgsSize(shader)

gskGlShaderGetNTextures(shader)

gskGlShaderGetNUniforms(shader)

gskGlShaderGetResource(shader)

gskGlShaderGetSource(shader)

gskGlShaderGetUniformName(shader, idx)

gskGlShaderGetUniformOffset(shader, idx)

gskGlShaderGetUniformType(shader, idx)

gskGlShaderNodeNew(shader, bounds, args, children, n_children)

gskGlShaderNodeGetArgs(node)

gskGlShaderNodeGetChild(node, idx)

gskGlShaderNodeGetNChildren(node)

gskGlShaderNodeGetShader(node)

gskInsetShadowNodeNew(outline, color, dx, dy, spread, blur_radius)

gskInsetShadowNodeGetBlurRadius(node)

gskInsetShadowNodeGetColor(node)

gskInsetShadowNodeGetDx(node)

gskInsetShadowNodeGetDy(node)

gskInsetShadowNodeGetOutline(node)

gskInsetShadowNodeGetSpread(node)

gskLinearGradientNodeNew(bounds, start, end, color_stops, n_color_stops)

gskLinearGradientNodeGetColorStops(node)

gskLinearGradientNodeGetEnd(node)

gskLinearGradientNodeGetNColorStops(node)

gskLinearGradientNodeGetStart(node)

gskMaskNodeNew(source, mask, mask_mode)

gskMaskNodeGetMask(node)

gskMaskNodeGetMaskMode(node)

gskMaskNodeGetSource(node)

gskNglRendererNew()

gskOpacityNodeNew(child, opacity)

gskOpacityNodeGetChild(node)

gskOpacityNodeGetOpacity(node)

gskOutsetShadowNodeNew(outline, color, dx, dy, spread, blur_radius)

gskOutsetShadowNodeGetBlurRadius(node)

gskOutsetShadowNodeGetColor(node)

gskOutsetShadowNodeGetDx(node)

gskOutsetShadowNodeGetDy(node)

gskOutsetShadowNodeGetOutline(node)

gskOutsetShadowNodeGetSpread(node)

gskRadialGradientNodeNew(
  bounds,
  center,
  hradius,
  vradius,
  start,
  end,
  color_stops,
  n_color_stops
)

gskRadialGradientNodeGetCenter(node)

gskRadialGradientNodeGetColorStops(node)

gskRadialGradientNodeGetEnd(node)

gskRadialGradientNodeGetHradius(node)

gskRadialGradientNodeGetNColorStops(node)

gskRadialGradientNodeGetStart(node)

gskRadialGradientNodeGetVradius(node)

gskRenderNodeDeserialize(bytes, error_func)

gskRenderNodeDraw(node, cr)

gskRenderNodeGetBounds(node)

gskRenderNodeGetNodeType(node)

gskRenderNodeRef(node)

gskRenderNodeSerialize(node)

gskRenderNodeUnref(node)

gskRenderNodeWriteToFile(node, filename)

gskRendererNewForSurface(surface)

gskRendererGetSurface(renderer)

gskRendererIsRealized(renderer)

gskRendererRealize(renderer, surface)

gskRendererRealizeForDisplay(renderer, display)

gskRendererRender(renderer, root, region)

gskRendererRenderTexture(renderer, root, viewport)

gskRendererUnrealize(renderer)

gskRepeatNodeNew(bounds, child, child_bounds)

gskRepeatNodeGetChild(node)

gskRepeatNodeGetChildBounds(node)

gskRepeatingLinearGradientNodeNew(
  bounds,
  start,
  end,
  color_stops,
  n_color_stops
)

gskRepeatingRadialGradientNodeNew(
  bounds,
  center,
  hradius,
  vradius,
  start,
  end,
  color_stops,
  n_color_stops
)

gskRoundedClipNodeNew(child, clip)

gskRoundedClipNodeGetChild(node)

gskRoundedClipNodeGetClip(node)

gskRoundedRectContainsPoint(self, point)

gskRoundedRectContainsRect(self, rect)

gskRoundedRectInit(
  self,
  bounds,
  top_left,
  top_right,
  bottom_right,
  bottom_left
)

gskRoundedRectInitCopy(self, src)

gskRoundedRectInitFromRect(self, bounds, radius)

gskRoundedRectIntersectsRect(self, rect)

gskRoundedRectIsRectilinear(self)

gskRoundedRectNormalize(self)

gskRoundedRectOffset(self, dx, dy)

gskRoundedRectShrink(self, top, right, bottom, left)

gskSerializationErrorQuark()

gskShaderArgsBuilderNew(shader, initial_values)

gskShaderArgsBuilderRef(builder)

gskShaderArgsBuilderSetBool(builder, idx, value)

gskShaderArgsBuilderSetFloat(builder, idx, value)

gskShaderArgsBuilderSetInt(builder, idx, value)

gskShaderArgsBuilderSetUint(builder, idx, value)

gskShaderArgsBuilderSetVec2(builder, idx, value)

gskShaderArgsBuilderSetVec3(builder, idx, value)

gskShaderArgsBuilderSetVec4(builder, idx, value)

gskShaderArgsBuilderToArgs(builder)

gskShaderArgsBuilderUnref(builder)

gskShadowNodeNew(child, shadows, n_shadows)

gskShadowNodeGetChild(node)

gskShadowNodeGetNShadows(node)

gskShadowNodeGetShadow(node, i)

gskStrokeNew(line_width)

gskStrokeCopy(other)

gskStrokeFree(self)

gskStrokeGetDash(self)

gskStrokeGetDashOffset(self)

gskStrokeGetLineCap(self)

gskStrokeGetLineJoin(self)

gskStrokeGetLineWidth(self)

gskStrokeGetMiterLimit(self)

gskStrokeSetDash(self, dash, n_dash)

gskStrokeSetDashOffset(self, offset)

gskStrokeSetLineCap(self, line_cap)

gskStrokeSetLineJoin(self, line_join)

gskStrokeSetLineWidth(self, line_width)

gskStrokeSetMiterLimit(self, limit)

gskStrokeToCairo(self, cr)

gskStrokeEqual(stroke1, stroke2)

gskStrokeNodeNew(child, path, stroke)

gskStrokeNodeGetChild(node)

gskStrokeNodeGetPath(node)

gskStrokeNodeGetStroke(node)

gskSubsurfaceNodeGetChild(node)

gskTextNodeNew(font, glyphs, color, offset)

gskTextNodeGetColor(node)

gskTextNodeGetFont(node)

gskTextNodeGetGlyphs(node)

gskTextNodeGetNumGlyphs(node)

gskTextNodeGetOffset(node)

gskTextNodeHasColorGlyphs(node)

gskTextureNodeNew(texture, bounds)

gskTextureNodeGetTexture(node)

gskTextureScaleNodeNew(texture, bounds, filter)

gskTextureScaleNodeGetFilter(node)

gskTextureScaleNodeGetTexture(node)

gskTransformNew()

gskTransformEqual(first, second)

gskTransformGetCategory(self)

gskTransformInvert(self)

gskTransformMatrix(next_, matrix)

gskTransformPerspective(next_, depth)

gskTransformPrint(self, string)

gskTransformRef(self)

gskTransformRotate(next_, angle)

gskTransformRotate3d(next_, angle, axis)

gskTransformScale(next_, factor_x, factor_y)

gskTransformScale3d(next_, factor_x, factor_y, factor_z)

gskTransformSkew(next_, skew_x, skew_y)

gskTransformTo2d(self)

gskTransformTo2dComponents(self)

gskTransformToAffine(self)

gskTransformToMatrix(self)

gskTransformToString(self)

gskTransformToTranslate(self)

gskTransformTransform(next_, other)

gskTransformTransformBounds(self, rect)

gskTransformTransformPoint(self, point)

gskTransformTranslate(next_, point)

gskTransformTranslate3d(next_, point)

gskTransformUnref(self)

gskTransformParse(string)

gskTransformNodeNew(child, transform)

gskTransformNodeGetChild(node)

gskTransformNodeGetTransform(node)

gskVulkanRendererNew()

gskValueDupRenderNode(value)

gskValueGetRenderNode(value)

gskValueSetRenderNode(value, node)

gskValueTakeRenderNode(value, node)
}
\arguments{
\item{bottom}{gfloat}

\item{top}{gfloat}

\item{blend_mode}{BlendMode}

\item{node}{RenderNode}

\item{child}{RenderNode}

\item{radius}{gfloat}

\item{outline}{RoundedRect}

\item{border_width}{gfloat}

\item{border_color}{Gdk.RGBA}

\item{bounds}{Graphene.Rect}

\item{clip}{RoundedRect}

\item{color_matrix}{Graphene.Matrix}

\item{color_offset}{Graphene.Vec4}

\item{rgba}{Gdk.RGBA}

\item{center}{Graphene.Point}

\item{rotation}{gfloat}

\item{color_stops}{ColorStop}

\item{n_color_stops}{gsize}

\item{children}{RenderNode}

\item{n_children}{guint}

\item{idx}{gint}

\item{start}{gfloat}

\item{end}{gfloat}

\item{progress}{gfloat}

\item{message}{utf8}

\item{path}{Path}

\item{fill_rule}{FillRule}

\item{sourcecode}{GLib.Bytes}

\item{resource_path}{utf8}

\item{shader}{GLShader}

\item{renderer}{Renderer}

\item{name}{utf8}

\item{args}{GLib.Bytes}

\item{out_value}{Graphene.Vec4}

\item{color}{Gdk.RGBA}

\item{dx}{gfloat}

\item{dy}{gfloat}

\item{spread}{gfloat}

\item{blur_radius}{gfloat}

\item{source}{RenderNode}

\item{mask}{RenderNode}

\item{mask_mode}{MaskMode}

\item{opacity}{gfloat}

\item{hradius}{gfloat}

\item{vradius}{gfloat}

\item{bytes}{GLib.Bytes}

\item{error_func}{function — ParseErrorFunc function(start, end, error)}

\item{cr}{cairo.Context}

\item{filename}{filename}

\item{surface}{Gdk.Surface}

\item{display}{Gdk.Display}

\item{root}{RenderNode}

\item{region}{cairo.Region}

\item{viewport}{Graphene.Rect}

\item{child_bounds}{Graphene.Rect}

\item{self}{Transform}

\item{point}{Graphene.Point3D}

\item{rect}{Graphene.Rect}

\item{top_left}{Graphene.Size}

\item{top_right}{Graphene.Size}

\item{bottom_right}{Graphene.Size}

\item{bottom_left}{Graphene.Size}

\item{src}{RoundedRect}

\item{right}{gfloat}

\item{left}{gfloat}

\item{initial_values}{GLib.Bytes}

\item{builder}{ShaderArgsBuilder}

\item{value}{GObject.Value}

\item{shadows}{Shadow}

\item{n_shadows}{gsize}

\item{i}{gsize}

\item{line_width}{gfloat}

\item{other}{Transform}

\item{dash}{gfloat}

\item{n_dash}{gsize}

\item{offset}{Graphene.Point}

\item{line_cap}{LineCap}

\item{line_join}{LineJoin}

\item{limit}{gfloat}

\item{stroke1}{gpointer}

\item{stroke2}{gpointer}

\item{stroke}{Stroke}

\item{font}{Pango.Font}

\item{glyphs}{Pango.GlyphString}

\item{texture}{Gdk.Texture}

\item{filter}{ScalingFilter}

\item{first}{Transform}

\item{second}{Transform}

\item{next_}{Transform}

\item{matrix}{Graphene.Matrix}

\item{depth}{gfloat}

\item{string}{utf8}

\item{angle}{gfloat}

\item{axis}{Graphene.Vec3}

\item{factor_x}{gfloat}

\item{factor_y}{gfloat}

\item{factor_z}{gfloat}

\item{skew_x}{gfloat}

\item{skew_y}{gfloat}

\item{transform}{Transform}
}
\value{
BlendNode

BlendMode

RenderNode

RenderNode

BlurNode

RenderNode

gfloat

BorderNode

Gdk.RGBA

RoundedRect

gfloat

CairoNode

cairo.Context

cairo.Surface

Renderer

ClipNode

RenderNode

Graphene.Rect

ColorMatrixNode

RenderNode

Graphene.Matrix

Graphene.Vec4

ColorNode

Gdk.RGBA

ConicGradientNode

gfloat

Graphene.Point

ColorStop

gsize

gfloat

ContainerNode

RenderNode

guint

CrossFadeNode

RenderNode

gfloat

RenderNode

DebugNode

RenderNode

utf8

FillNode

RenderNode

FillRule

Path

Renderer

GLShader

GLShader

gboolean

gint

gboolean

gfloat

gint32

guint32

Return value from C function

Return value from C function

Return value from C function

gsize

gint

gint

utf8

GLib.Bytes

utf8

gint

GLUniformType

GLShaderNode

GLib.Bytes

RenderNode

guint

GLShader

InsetShadowNode

gfloat

Gdk.RGBA

gfloat

gfloat

RoundedRect

gfloat

LinearGradientNode

ColorStop

Graphene.Point

gsize

Graphene.Point

MaskNode

RenderNode

MaskMode

RenderNode

Renderer

OpacityNode

RenderNode

gfloat

OutsetShadowNode

gfloat

Gdk.RGBA

gfloat

gfloat

RoundedRect

gfloat

RadialGradientNode

Graphene.Point

ColorStop

gfloat

gfloat

gsize

gfloat

gfloat

RenderNode

Return value from C function

Return value from C function

RenderNodeType

RenderNode

GLib.Bytes

Return value from C function

gboolean

Renderer

Gdk.Surface

gboolean

gboolean

gboolean

Return value from C function

Gdk.Texture

Return value from C function

RepeatNode

RenderNode

Graphene.Rect

RepeatingLinearGradientNode

RepeatingRadialGradientNode

RoundedClipNode

RenderNode

RoundedRect

gboolean

gboolean

RoundedRect

RoundedRect

RoundedRect

gboolean

gboolean

RoundedRect

RoundedRect

RoundedRect

GLib.Quark

ShaderArgsBuilder

ShaderArgsBuilder

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

GLib.Bytes

Return value from C function

ShadowNode

RenderNode

gsize

Shadow

Stroke

Stroke

Return value from C function

gfloat

gfloat

LineCap

LineJoin

gfloat

gfloat

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

StrokeNode

RenderNode

Path

Stroke

RenderNode

TextNode

Gdk.RGBA

Pango.Font

Pango.GlyphInfo

guint

Graphene.Point

gboolean

TextureNode

Gdk.Texture

TextureScaleNode

ScalingFilter

Gdk.Texture

Transform

gboolean

TransformCategory

Transform

Transform

Transform

Return value from C function

Transform

Transform

Transform

Transform

Transform

Transform

Return value from C function

Return value from C function

Return value from C function

Return value from C function

utf8

Return value from C function

Transform

Return value from C function

Return value from C function

Transform

Transform

Return value from C function

gboolean

TransformNode

RenderNode

Transform

Renderer

RenderNode

RenderNode

Return value from C function

Return value from C function
}
\description{
gskBlendNodeNew

gskBlendNodeGetBlendMode

gskBlendNodeGetBottomChild

gskBlendNodeGetTopChild

gskBlurNodeNew

gskBlurNodeGetChild

gskBlurNodeGetRadius

gskBorderNodeNew

gskBorderNodeGetColors

gskBorderNodeGetOutline

gskBorderNodeGetWidths

gskCairoNodeNew

gskCairoNodeGetDrawContext

gskCairoNodeGetSurface

gskCairoRendererNew

gskClipNodeNew

gskClipNodeGetChild

gskClipNodeGetClip

gskColorMatrixNodeNew

gskColorMatrixNodeGetChild

gskColorMatrixNodeGetColorMatrix

gskColorMatrixNodeGetColorOffset

gskColorNodeNew

gskColorNodeGetColor

gskConicGradientNodeNew

gskConicGradientNodeGetAngle

gskConicGradientNodeGetCenter

gskConicGradientNodeGetColorStops

gskConicGradientNodeGetNColorStops

gskConicGradientNodeGetRotation

gskContainerNodeNew

gskContainerNodeGetChild

gskContainerNodeGetNChildren

gskCrossFadeNodeNew

gskCrossFadeNodeGetEndChild

gskCrossFadeNodeGetProgress

gskCrossFadeNodeGetStartChild

gskDebugNodeNew

gskDebugNodeGetChild

gskDebugNodeGetMessage

gskFillNodeNew

gskFillNodeGetChild

gskFillNodeGetFillRule

gskFillNodeGetPath

gskGlRendererNew

gskGlShaderNewFromBytes

gskGlShaderNewFromResource

gskGlShaderCompile

gskGlShaderFindUniformByName

gskGlShaderGetArgBool

gskGlShaderGetArgFloat

gskGlShaderGetArgInt

gskGlShaderGetArgUint

gskGlShaderGetArgVec2

gskGlShaderGetArgVec3

gskGlShaderGetArgVec4

gskGlShaderGetArgsSize

gskGlShaderGetNTextures

gskGlShaderGetNUniforms

gskGlShaderGetResource

gskGlShaderGetSource

gskGlShaderGetUniformName

gskGlShaderGetUniformOffset

gskGlShaderGetUniformType

gskGlShaderNodeNew

gskGlShaderNodeGetArgs

gskGlShaderNodeGetChild

gskGlShaderNodeGetNChildren

gskGlShaderNodeGetShader

gskInsetShadowNodeNew

gskInsetShadowNodeGetBlurRadius

gskInsetShadowNodeGetColor

gskInsetShadowNodeGetDx

gskInsetShadowNodeGetDy

gskInsetShadowNodeGetOutline

gskInsetShadowNodeGetSpread

gskLinearGradientNodeNew

gskLinearGradientNodeGetColorStops

gskLinearGradientNodeGetEnd

gskLinearGradientNodeGetNColorStops

gskLinearGradientNodeGetStart

gskMaskNodeNew

gskMaskNodeGetMask

gskMaskNodeGetMaskMode

gskMaskNodeGetSource

gskNglRendererNew

gskOpacityNodeNew

gskOpacityNodeGetChild

gskOpacityNodeGetOpacity

gskOutsetShadowNodeNew

gskOutsetShadowNodeGetBlurRadius

gskOutsetShadowNodeGetColor

gskOutsetShadowNodeGetDx

gskOutsetShadowNodeGetDy

gskOutsetShadowNodeGetOutline

gskOutsetShadowNodeGetSpread

gskRadialGradientNodeNew

gskRadialGradientNodeGetCenter

gskRadialGradientNodeGetColorStops

gskRadialGradientNodeGetEnd

gskRadialGradientNodeGetHradius

gskRadialGradientNodeGetNColorStops

gskRadialGradientNodeGetStart

gskRadialGradientNodeGetVradius

gskRenderNodeDeserialize

gskRenderNodeDraw

gskRenderNodeGetBounds

gskRenderNodeGetNodeType

gskRenderNodeRef

gskRenderNodeSerialize

gskRenderNodeUnref

gskRenderNodeWriteToFile

gskRendererNewForSurface

gskRendererGetSurface

gskRendererIsRealized

gskRendererRealize

gskRendererRealizeForDisplay

gskRendererRender

gskRendererRenderTexture

gskRendererUnrealize

gskRepeatNodeNew

gskRepeatNodeGetChild

gskRepeatNodeGetChildBounds

gskRepeatingLinearGradientNodeNew

gskRepeatingRadialGradientNodeNew

gskRoundedClipNodeNew

gskRoundedClipNodeGetChild

gskRoundedClipNodeGetClip

gskRoundedRectContainsPoint

gskRoundedRectContainsRect

gskRoundedRectInit

gskRoundedRectInitCopy

gskRoundedRectInitFromRect

gskRoundedRectIntersectsRect

gskRoundedRectIsRectilinear

gskRoundedRectNormalize

gskRoundedRectOffset

gskRoundedRectShrink

gskSerializationErrorQuark

gskShaderArgsBuilderNew

gskShaderArgsBuilderRef

gskShaderArgsBuilderSetBool

gskShaderArgsBuilderSetFloat

gskShaderArgsBuilderSetInt

gskShaderArgsBuilderSetUint

gskShaderArgsBuilderSetVec2

gskShaderArgsBuilderSetVec3

gskShaderArgsBuilderSetVec4

gskShaderArgsBuilderToArgs

gskShaderArgsBuilderUnref

gskShadowNodeNew

gskShadowNodeGetChild

gskShadowNodeGetNShadows

gskShadowNodeGetShadow

gskStrokeNew

gskStrokeCopy

gskStrokeFree

gskStrokeGetDash

gskStrokeGetDashOffset

gskStrokeGetLineCap

gskStrokeGetLineJoin

gskStrokeGetLineWidth

gskStrokeGetMiterLimit

gskStrokeSetDash

gskStrokeSetDashOffset

gskStrokeSetLineCap

gskStrokeSetLineJoin

gskStrokeSetLineWidth

gskStrokeSetMiterLimit

gskStrokeToCairo

gskStrokeEqual

gskStrokeNodeNew

gskStrokeNodeGetChild

gskStrokeNodeGetPath

gskStrokeNodeGetStroke

gskSubsurfaceNodeGetChild

gskTextNodeNew

gskTextNodeGetColor

gskTextNodeGetFont

gskTextNodeGetGlyphs

gskTextNodeGetNumGlyphs

gskTextNodeGetOffset

gskTextNodeHasColorGlyphs

gskTextureNodeNew

gskTextureNodeGetTexture

gskTextureScaleNodeNew

gskTextureScaleNodeGetFilter

gskTextureScaleNodeGetTexture

gskTransformNew

gskTransformEqual

gskTransformGetCategory

gskTransformInvert

gskTransformMatrix

gskTransformPerspective

gskTransformPrint

gskTransformRef

gskTransformRotate

gskTransformRotate3d

gskTransformScale

gskTransformScale3d

gskTransformSkew

gskTransformTo2d

gskTransformTo2dComponents

gskTransformToAffine

gskTransformToMatrix

gskTransformToString

gskTransformToTranslate

gskTransformTransform

gskTransformTransformBounds

gskTransformTransformPoint

gskTransformTranslate

gskTransformTranslate3d

gskTransformUnref

gskTransformParse

gskTransformNodeNew

gskTransformNodeGetChild

gskTransformNodeGetTransform

gskVulkanRendererNew

gskValueDupRenderNode

gskValueGetRenderNode

gskValueSetRenderNode

gskValueTakeRenderNode
}
