% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtk_autogen.R
\name{gtkApplicationNew}
\alias{gtkApplicationNew}
\alias{gtkApplicationAddWindow}
\alias{gtkApplicationGetAccelsForAction}
\alias{gtkApplicationGetActionsForAccel}
\alias{gtkApplicationGetActiveWindow}
\alias{gtkApplicationGetMenuById}
\alias{gtkApplicationGetMenubar}
\alias{gtkApplicationGetWindowById}
\alias{gtkApplicationGetWindows}
\alias{gtkApplicationInhibit}
\alias{gtkApplicationListActionDescriptions}
\alias{gtkApplicationRemoveWindow}
\alias{gtkApplicationSetAccelsForAction}
\alias{gtkApplicationSetMenubar}
\alias{gtkApplicationUninhibit}
\alias{gtkApplicationWindowNew}
\alias{gtkApplicationWindowGetHelpOverlay}
\alias{gtkApplicationWindowGetId}
\alias{gtkApplicationWindowGetShowMenubar}
\alias{gtkApplicationWindowSetHelpOverlay}
\alias{gtkApplicationWindowSetShowMenubar}
\title{gtkApplicationNew}
\usage{
gtkApplicationNew(application_id, flags)

gtkApplicationAddWindow(application, window)

gtkApplicationGetAccelsForAction(application, detailed_action_name)

gtkApplicationGetActionsForAccel(application, accel)

gtkApplicationGetActiveWindow(application)

gtkApplicationGetMenuById(application, id)

gtkApplicationGetMenubar(application)

gtkApplicationGetWindowById(application, id)

gtkApplicationGetWindows(application)

gtkApplicationInhibit(application, window, flags, reason)

gtkApplicationListActionDescriptions(application)

gtkApplicationRemoveWindow(application, window)

gtkApplicationSetAccelsForAction(application, detailed_action_name, accels)

gtkApplicationSetMenubar(application, menubar)

gtkApplicationUninhibit(application, cookie)

gtkApplicationWindowNew(application)

gtkApplicationWindowGetHelpOverlay(window)

gtkApplicationWindowGetId(window)

gtkApplicationWindowGetShowMenubar(window)

gtkApplicationWindowSetHelpOverlay(window, help_overlay)

gtkApplicationWindowSetShowMenubar(window, show_menubar)
}
\arguments{
\item{application_id}{utf8}

\item{flags}{ApplicationInhibitFlags}

\item{application}{Application}

\item{window}{ApplicationWindow}

\item{detailed_action_name}{utf8}

\item{accel}{utf8}

\item{id}{guint}

\item{reason}{utf8}

\item{accels}{utf8}

\item{menubar}{Gio.MenuModel}

\item{cookie}{guint}

\item{help_overlay}{ShortcutsWindow}

\item{show_menubar}{gboolean}
}
\value{
Application

Return value from C function

utf8

utf8

Window

Gio.Menu

Gio.MenuModel

Window

GLib.List

guint

utf8

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

ShortcutsWindow

guint

gboolean

Return value from C function

Return value from C function
}
\description{
gtkApplicationNew

gtkApplicationAddWindow

gtkApplicationGetAccelsForAction

gtkApplicationGetActionsForAccel

gtkApplicationGetActiveWindow

gtkApplicationGetMenuById

gtkApplicationGetMenubar

gtkApplicationGetWindowById

gtkApplicationGetWindows

gtkApplicationInhibit

gtkApplicationListActionDescriptions

gtkApplicationRemoveWindow

gtkApplicationSetAccelsForAction

gtkApplicationSetMenubar

gtkApplicationUninhibit

gtkApplicationWindowNew

gtkApplicationWindowGetHelpOverlay

gtkApplicationWindowGetId

gtkApplicationWindowGetShowMenubar

gtkApplicationWindowSetHelpOverlay

gtkApplicationWindowSetShowMenubar
}
