% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtk_autogen.R
\name{gtkButtonNew}
\alias{gtkButtonNew}
\alias{gtkButtonNewFromIconName}
\alias{gtkButtonNewWithLabel}
\alias{gtkButtonNewWithMnemonic}
\alias{gtkButtonGetCanShrink}
\alias{gtkButtonGetChild}
\alias{gtkButtonGetHasFrame}
\alias{gtkButtonGetIconName}
\alias{gtkButtonGetLabel}
\alias{gtkButtonGetUseUnderline}
\alias{gtkButtonSetCanShrink}
\alias{gtkButtonSetChild}
\alias{gtkButtonSetHasFrame}
\alias{gtkButtonSetIconName}
\alias{gtkButtonSetLabel}
\alias{gtkButtonSetUseUnderline}
\alias{gtkCheckButtonNew}
\alias{gtkCheckButtonNewWithLabel}
\alias{gtkCheckButtonNewWithMnemonic}
\alias{gtkCheckButtonGetActive}
\alias{gtkCheckButtonGetChild}
\alias{gtkCheckButtonGetInconsistent}
\alias{gtkCheckButtonGetLabel}
\alias{gtkCheckButtonGetUseUnderline}
\alias{gtkCheckButtonSetActive}
\alias{gtkCheckButtonSetChild}
\alias{gtkCheckButtonSetGroup}
\alias{gtkCheckButtonSetInconsistent}
\alias{gtkCheckButtonSetLabel}
\alias{gtkCheckButtonSetUseUnderline}
\alias{gtkToggleButtonNew}
\alias{gtkToggleButtonNewWithLabel}
\alias{gtkToggleButtonNewWithMnemonic}
\alias{gtkToggleButtonGetActive}
\alias{gtkToggleButtonSetActive}
\alias{gtkToggleButtonSetGroup}
\alias{gtkToggleButtonToggled}
\title{gtkButtonNew}
\usage{
gtkButtonNew()

gtkButtonNewFromIconName(icon_name)

gtkButtonNewWithLabel(label)

gtkButtonNewWithMnemonic(label)

gtkButtonGetCanShrink(button)

gtkButtonGetChild(button)

gtkButtonGetHasFrame(button)

gtkButtonGetIconName(button)

gtkButtonGetLabel(button)

gtkButtonGetUseUnderline(button)

gtkButtonSetCanShrink(button, can_shrink)

gtkButtonSetChild(button, child)

gtkButtonSetHasFrame(button, has_frame)

gtkButtonSetIconName(button, icon_name)

gtkButtonSetLabel(button, label)

gtkButtonSetUseUnderline(button, use_underline)

gtkCheckButtonNew()

gtkCheckButtonNewWithLabel(label)

gtkCheckButtonNewWithMnemonic(label)

gtkCheckButtonGetActive(self)

gtkCheckButtonGetChild(button)

gtkCheckButtonGetInconsistent(check_button)

gtkCheckButtonGetLabel(self)

gtkCheckButtonGetUseUnderline(self)

gtkCheckButtonSetActive(self, setting)

gtkCheckButtonSetChild(button, child)

gtkCheckButtonSetGroup(self, group)

gtkCheckButtonSetInconsistent(check_button, inconsistent)

gtkCheckButtonSetLabel(self, label)

gtkCheckButtonSetUseUnderline(self, setting)

gtkToggleButtonNew()

gtkToggleButtonNewWithLabel(label)

gtkToggleButtonNewWithMnemonic(label)

gtkToggleButtonGetActive(toggle_button)

gtkToggleButtonSetActive(toggle_button, is_active)

gtkToggleButtonSetGroup(toggle_button, group)

gtkToggleButtonToggled(toggle_button)
}
\arguments{
\item{icon_name}{utf8}

\item{label}{utf8}

\item{button}{CheckButton}

\item{can_shrink}{gboolean}

\item{child}{Widget}

\item{has_frame}{gboolean}

\item{use_underline}{gboolean}

\item{self}{CheckButton}

\item{check_button}{CheckButton}

\item{setting}{gboolean}

\item{group}{ToggleButton}

\item{inconsistent}{gboolean}

\item{toggle_button}{ToggleButton}

\item{is_active}{gboolean}
}
\value{
Widget

Widget

Widget

Widget

gboolean

Widget

gboolean

utf8

utf8

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

Widget

Widget

gboolean

Widget

gboolean

utf8

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

Widget

Widget

gboolean

Return value from C function

Return value from C function

Return value from C function
}
\description{
gtkButtonNew

gtkButtonNewFromIconName

gtkButtonNewWithLabel

gtkButtonNewWithMnemonic

gtkButtonGetCanShrink

gtkButtonGetChild

gtkButtonGetHasFrame

gtkButtonGetIconName

gtkButtonGetLabel

gtkButtonGetUseUnderline

gtkButtonSetCanShrink

gtkButtonSetChild

gtkButtonSetHasFrame

gtkButtonSetIconName

gtkButtonSetLabel

gtkButtonSetUseUnderline

gtkCheckButtonNew

gtkCheckButtonNewWithLabel

gtkCheckButtonNewWithMnemonic

gtkCheckButtonGetActive

gtkCheckButtonGetChild

gtkCheckButtonGetInconsistent

gtkCheckButtonGetLabel

gtkCheckButtonGetUseUnderline

gtkCheckButtonSetActive

gtkCheckButtonSetChild

gtkCheckButtonSetGroup

gtkCheckButtonSetInconsistent

gtkCheckButtonSetLabel

gtkCheckButtonSetUseUnderline

gtkToggleButtonNew

gtkToggleButtonNewWithLabel

gtkToggleButtonNewWithMnemonic

gtkToggleButtonGetActive

gtkToggleButtonSetActive

gtkToggleButtonSetGroup

gtkToggleButtonToggled
}
