% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtk_autogen.R
\name{gtkAboutDialogNew}
\alias{gtkAboutDialogNew}
\alias{gtkAboutDialogAddCreditSection}
\alias{gtkAboutDialogGetArtists}
\alias{gtkAboutDialogGetAuthors}
\alias{gtkAboutDialogGetComments}
\alias{gtkAboutDialogGetCopyright}
\alias{gtkAboutDialogGetDocumenters}
\alias{gtkAboutDialogGetLicense}
\alias{gtkAboutDialogGetLicenseType}
\alias{gtkAboutDialogGetLogo}
\alias{gtkAboutDialogGetLogoIconName}
\alias{gtkAboutDialogGetProgramName}
\alias{gtkAboutDialogGetSystemInformation}
\alias{gtkAboutDialogGetTranslatorCredits}
\alias{gtkAboutDialogGetVersion}
\alias{gtkAboutDialogGetWebsite}
\alias{gtkAboutDialogGetWebsiteLabel}
\alias{gtkAboutDialogGetWrapLicense}
\alias{gtkAboutDialogSetArtists}
\alias{gtkAboutDialogSetAuthors}
\alias{gtkAboutDialogSetComments}
\alias{gtkAboutDialogSetCopyright}
\alias{gtkAboutDialogSetDocumenters}
\alias{gtkAboutDialogSetLicense}
\alias{gtkAboutDialogSetLicenseType}
\alias{gtkAboutDialogSetLogo}
\alias{gtkAboutDialogSetLogoIconName}
\alias{gtkAboutDialogSetProgramName}
\alias{gtkAboutDialogSetSystemInformation}
\alias{gtkAboutDialogSetTranslatorCredits}
\alias{gtkAboutDialogSetVersion}
\alias{gtkAboutDialogSetWebsite}
\alias{gtkAboutDialogSetWebsiteLabel}
\alias{gtkAboutDialogSetWrapLicense}
\alias{gtkDialogNew}
\alias{gtkDialogAddActionWidget}
\alias{gtkDialogAddButton}
\alias{gtkDialogGetContentArea}
\alias{gtkDialogGetHeaderBar}
\alias{gtkDialogGetResponseForWidget}
\alias{gtkDialogGetWidgetForResponse}
\alias{gtkDialogResponse}
\alias{gtkDialogSetDefaultResponse}
\alias{gtkDialogSetResponseSensitive}
\alias{gtkDialogErrorQuark}
\alias{gtkMessageDialogGetMessageArea}
\alias{gtkMessageDialogSetMarkup}
\title{gtkAboutDialogNew}
\usage{
gtkAboutDialogNew()

gtkAboutDialogAddCreditSection(about, section_name, people)

gtkAboutDialogGetArtists(about)

gtkAboutDialogGetAuthors(about)

gtkAboutDialogGetComments(about)

gtkAboutDialogGetCopyright(about)

gtkAboutDialogGetDocumenters(about)

gtkAboutDialogGetLicense(about)

gtkAboutDialogGetLicenseType(about)

gtkAboutDialogGetLogo(about)

gtkAboutDialogGetLogoIconName(about)

gtkAboutDialogGetProgramName(about)

gtkAboutDialogGetSystemInformation(about)

gtkAboutDialogGetTranslatorCredits(about)

gtkAboutDialogGetVersion(about)

gtkAboutDialogGetWebsite(about)

gtkAboutDialogGetWebsiteLabel(about)

gtkAboutDialogGetWrapLicense(about)

gtkAboutDialogSetArtists(about, artists)

gtkAboutDialogSetAuthors(about, authors)

gtkAboutDialogSetComments(about, comments)

gtkAboutDialogSetCopyright(about, copyright)

gtkAboutDialogSetDocumenters(about, documenters)

gtkAboutDialogSetLicense(about, license)

gtkAboutDialogSetLicenseType(about, license_type)

gtkAboutDialogSetLogo(about, logo)

gtkAboutDialogSetLogoIconName(about, icon_name)

gtkAboutDialogSetProgramName(about, name)

gtkAboutDialogSetSystemInformation(about, system_information)

gtkAboutDialogSetTranslatorCredits(about, translator_credits)

gtkAboutDialogSetVersion(about, version)

gtkAboutDialogSetWebsite(about, website)

gtkAboutDialogSetWebsiteLabel(about, website_label)

gtkAboutDialogSetWrapLicense(about, wrap_license)

gtkDialogNew()

gtkDialogAddActionWidget(dialog, child, response_id)

gtkDialogAddButton(dialog, button_text, response_id)

gtkDialogGetContentArea(dialog)

gtkDialogGetHeaderBar(dialog)

gtkDialogGetResponseForWidget(dialog, widget)

gtkDialogGetWidgetForResponse(dialog, response_id)

gtkDialogResponse(dialog, response_id)

gtkDialogSetDefaultResponse(dialog, response_id)

gtkDialogSetResponseSensitive(dialog, response_id, setting)

gtkDialogErrorQuark()

gtkMessageDialogGetMessageArea(message_dialog)

gtkMessageDialogSetMarkup(message_dialog, str)
}
\arguments{
\item{about}{AboutDialog}

\item{section_name}{utf8}

\item{people}{utf8}

\item{artists}{utf8}

\item{authors}{utf8}

\item{comments}{utf8}

\item{copyright}{utf8}

\item{documenters}{utf8}

\item{license}{utf8}

\item{license_type}{License}

\item{logo}{Gdk.Paintable}

\item{icon_name}{utf8}

\item{name}{utf8}

\item{system_information}{utf8}

\item{translator_credits}{utf8}

\item{version}{utf8}

\item{website}{utf8}

\item{website_label}{utf8}

\item{wrap_license}{gboolean}

\item{dialog}{Dialog}

\item{child}{Widget}

\item{response_id}{gint}

\item{button_text}{utf8}

\item{widget}{Widget}

\item{setting}{gboolean}

\item{message_dialog}{MessageDialog}

\item{str}{utf8}
}
\value{
Widget

Return value from C function

utf8

utf8

utf8

utf8

utf8

utf8

License

Gdk.Paintable

utf8

utf8

utf8

utf8

utf8

utf8

utf8

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

Return value from C function

Widget

Box

HeaderBar

gint

Widget

Return value from C function

Return value from C function

Return value from C function

GLib.Quark

Widget

Return value from C function
}
\description{
gtkAboutDialogNew

gtkAboutDialogAddCreditSection

gtkAboutDialogGetArtists

gtkAboutDialogGetAuthors

gtkAboutDialogGetComments

gtkAboutDialogGetCopyright

gtkAboutDialogGetDocumenters

gtkAboutDialogGetLicense

gtkAboutDialogGetLicenseType

gtkAboutDialogGetLogo

gtkAboutDialogGetLogoIconName

gtkAboutDialogGetProgramName

gtkAboutDialogGetSystemInformation

gtkAboutDialogGetTranslatorCredits

gtkAboutDialogGetVersion

gtkAboutDialogGetWebsite

gtkAboutDialogGetWebsiteLabel

gtkAboutDialogGetWrapLicense

gtkAboutDialogSetArtists

gtkAboutDialogSetAuthors

gtkAboutDialogSetComments

gtkAboutDialogSetCopyright

gtkAboutDialogSetDocumenters

gtkAboutDialogSetLicense

gtkAboutDialogSetLicenseType

gtkAboutDialogSetLogo

gtkAboutDialogSetLogoIconName

gtkAboutDialogSetProgramName

gtkAboutDialogSetSystemInformation

gtkAboutDialogSetTranslatorCredits

gtkAboutDialogSetVersion

gtkAboutDialogSetWebsite

gtkAboutDialogSetWebsiteLabel

gtkAboutDialogSetWrapLicense

gtkDialogNew

gtkDialogAddActionWidget

gtkDialogAddButton

gtkDialogGetContentArea

gtkDialogGetHeaderBar

gtkDialogGetResponseForWidget

gtkDialogGetWidgetForResponse

gtkDialogResponse

gtkDialogSetDefaultResponse

gtkDialogSetResponseSensitive

gtkDialogErrorQuark

gtkMessageDialogGetMessageArea

gtkMessageDialogSetMarkup
}
