% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtk_autogen.R
\name{gtkFileChooserAddChoice}
\alias{gtkFileChooserAddChoice}
\alias{gtkFileChooserAddFilter}
\alias{gtkFileChooserAddShortcutFolder}
\alias{gtkFileChooserGetAction}
\alias{gtkFileChooserGetChoice}
\alias{gtkFileChooserGetCreateFolders}
\alias{gtkFileChooserGetCurrentFolder}
\alias{gtkFileChooserGetCurrentName}
\alias{gtkFileChooserGetFile}
\alias{gtkFileChooserGetFiles}
\alias{gtkFileChooserGetFilter}
\alias{gtkFileChooserGetFilters}
\alias{gtkFileChooserGetSelectMultiple}
\alias{gtkFileChooserGetShortcutFolders}
\alias{gtkFileChooserRemoveChoice}
\alias{gtkFileChooserRemoveFilter}
\alias{gtkFileChooserRemoveShortcutFolder}
\alias{gtkFileChooserSetAction}
\alias{gtkFileChooserSetChoice}
\alias{gtkFileChooserSetCreateFolders}
\alias{gtkFileChooserSetCurrentFolder}
\alias{gtkFileChooserSetCurrentName}
\alias{gtkFileChooserSetFile}
\alias{gtkFileChooserSetFilter}
\alias{gtkFileChooserSetSelectMultiple}
\alias{gtkFileChooserErrorQuark}
\alias{gtkFileChooserNativeNew}
\alias{gtkFileChooserNativeGetAcceptLabel}
\alias{gtkFileChooserNativeGetCancelLabel}
\alias{gtkFileChooserNativeSetAcceptLabel}
\alias{gtkFileChooserNativeSetCancelLabel}
\alias{gtkFileChooserWidgetNew}
\alias{gtkFileFilterNew}
\alias{gtkFileFilterNewFromGvariant}
\alias{gtkFileFilterAddMimeType}
\alias{gtkFileFilterAddPattern}
\alias{gtkFileFilterAddPixbufFormats}
\alias{gtkFileFilterAddSuffix}
\alias{gtkFileFilterGetAttributes}
\alias{gtkFileFilterGetName}
\alias{gtkFileFilterSetName}
\alias{gtkFileFilterToGvariant}
\title{gtkFileChooserAddChoice}
\usage{
gtkFileChooserAddChoice(chooser, id, label, options, option_labels)

gtkFileChooserAddFilter(chooser, filter)

gtkFileChooserAddShortcutFolder(chooser, folder)

gtkFileChooserGetAction(chooser)

gtkFileChooserGetChoice(chooser, id)

gtkFileChooserGetCreateFolders(chooser)

gtkFileChooserGetCurrentFolder(chooser)

gtkFileChooserGetCurrentName(chooser)

gtkFileChooserGetFile(chooser)

gtkFileChooserGetFiles(chooser)

gtkFileChooserGetFilter(chooser)

gtkFileChooserGetFilters(chooser)

gtkFileChooserGetSelectMultiple(chooser)

gtkFileChooserGetShortcutFolders(chooser)

gtkFileChooserRemoveChoice(chooser, id)

gtkFileChooserRemoveFilter(chooser, filter)

gtkFileChooserRemoveShortcutFolder(chooser, folder)

gtkFileChooserSetAction(chooser, action)

gtkFileChooserSetChoice(chooser, id, option)

gtkFileChooserSetCreateFolders(chooser, create_folders)

gtkFileChooserSetCurrentFolder(chooser, file)

gtkFileChooserSetCurrentName(chooser, name)

gtkFileChooserSetFile(chooser, file)

gtkFileChooserSetFilter(chooser, filter)

gtkFileChooserSetSelectMultiple(chooser, select_multiple)

gtkFileChooserErrorQuark()

gtkFileChooserNativeNew(title, parent, action, accept_label, cancel_label)

gtkFileChooserNativeGetAcceptLabel(self)

gtkFileChooserNativeGetCancelLabel(self)

gtkFileChooserNativeSetAcceptLabel(self, accept_label)

gtkFileChooserNativeSetCancelLabel(self, cancel_label)

gtkFileChooserWidgetNew(action)

gtkFileFilterNew()

gtkFileFilterNewFromGvariant(variant)

gtkFileFilterAddMimeType(filter, mime_type)

gtkFileFilterAddPattern(filter, pattern)

gtkFileFilterAddPixbufFormats(filter)

gtkFileFilterAddSuffix(filter, suffix)

gtkFileFilterGetAttributes(filter)

gtkFileFilterGetName(filter)

gtkFileFilterSetName(filter, name)

gtkFileFilterToGvariant(filter)
}
\arguments{
\item{chooser}{FileChooser}

\item{id}{utf8}

\item{label}{utf8}

\item{options}{utf8}

\item{option_labels}{utf8}

\item{filter}{FileFilter}

\item{folder}{Gio.File}

\item{action}{FileChooserAction}

\item{option}{utf8}

\item{create_folders}{gboolean}

\item{file}{Gio.File}

\item{name}{utf8}

\item{select_multiple}{gboolean}

\item{title}{utf8}

\item{parent}{Window}

\item{accept_label}{utf8}

\item{cancel_label}{utf8}

\item{self}{FileChooserNative}

\item{variant}{GLib.Variant}

\item{mime_type}{utf8}

\item{pattern}{utf8}

\item{suffix}{utf8}
}
\value{
Return value from C function

Return value from C function

gboolean

FileChooserAction

utf8

gboolean

Gio.File

utf8

Gio.File

Gio.ListModel

FileFilter

Gio.ListModel

gboolean

Gio.ListModel

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

gboolean

Return value from C function

Return value from C function

GLib.Quark

FileChooserNative

utf8

utf8

Return value from C function

Return value from C function

Widget

FileFilter

FileFilter

Return value from C function

Return value from C function

Return value from C function

Return value from C function

utf8

utf8

Return value from C function

GLib.Variant
}
\description{
gtkFileChooserAddChoice

gtkFileChooserAddFilter

gtkFileChooserAddShortcutFolder

gtkFileChooserGetAction

gtkFileChooserGetChoice

gtkFileChooserGetCreateFolders

gtkFileChooserGetCurrentFolder

gtkFileChooserGetCurrentName

gtkFileChooserGetFile

gtkFileChooserGetFiles

gtkFileChooserGetFilter

gtkFileChooserGetFilters

gtkFileChooserGetSelectMultiple

gtkFileChooserGetShortcutFolders

gtkFileChooserRemoveChoice

gtkFileChooserRemoveFilter

gtkFileChooserRemoveShortcutFolder

gtkFileChooserSetAction

gtkFileChooserSetChoice

gtkFileChooserSetCreateFolders

gtkFileChooserSetCurrentFolder

gtkFileChooserSetCurrentName

gtkFileChooserSetFile

gtkFileChooserSetFilter

gtkFileChooserSetSelectMultiple

gtkFileChooserErrorQuark

gtkFileChooserNativeNew

gtkFileChooserNativeGetAcceptLabel

gtkFileChooserNativeGetCancelLabel

gtkFileChooserNativeSetAcceptLabel

gtkFileChooserNativeSetCancelLabel

gtkFileChooserWidgetNew

gtkFileFilterNew

gtkFileFilterNewFromGvariant

gtkFileFilterAddMimeType

gtkFileFilterAddPattern

gtkFileFilterAddPixbufFormats

gtkFileFilterAddSuffix

gtkFileFilterGetAttributes

gtkFileFilterGetName

gtkFileFilterSetName

gtkFileFilterToGvariant
}
