% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtk_autogen.R
\name{gtkIconPaintableNewForFile}
\alias{gtkIconPaintableNewForFile}
\alias{gtkIconPaintableGetFile}
\alias{gtkIconPaintableGetIconName}
\alias{gtkIconPaintableIsSymbolic}
\alias{gtkIconThemeNew}
\alias{gtkIconThemeGetForDisplay}
\alias{gtkIconThemeAddResourcePath}
\alias{gtkIconThemeAddSearchPath}
\alias{gtkIconThemeGetDisplay}
\alias{gtkIconThemeGetIconNames}
\alias{gtkIconThemeGetIconSizes}
\alias{gtkIconThemeGetResourcePath}
\alias{gtkIconThemeGetSearchPath}
\alias{gtkIconThemeGetThemeName}
\alias{gtkIconThemeHasGicon}
\alias{gtkIconThemeHasIcon}
\alias{gtkIconThemeLookupByGicon}
\alias{gtkIconThemeLookupIcon}
\alias{gtkIconThemeSetResourcePath}
\alias{gtkIconThemeSetSearchPath}
\alias{gtkIconThemeSetThemeName}
\alias{gtkIconThemeErrorQuark}
\alias{gtkIconViewNew}
\alias{gtkIconViewNewWithArea}
\alias{gtkIconViewNewWithModel}
\alias{gtkIconViewCreateDragIcon}
\alias{gtkIconViewEnableModelDragDest}
\alias{gtkIconViewEnableModelDragSource}
\alias{gtkIconViewGetActivateOnSingleClick}
\alias{gtkIconViewGetCellRect}
\alias{gtkIconViewGetColumnSpacing}
\alias{gtkIconViewGetColumns}
\alias{gtkIconViewGetCursor}
\alias{gtkIconViewGetDestItemAtPos}
\alias{gtkIconViewGetDragDestItem}
\alias{gtkIconViewGetItemAtPos}
\alias{gtkIconViewGetItemColumn}
\alias{gtkIconViewGetItemOrientation}
\alias{gtkIconViewGetItemPadding}
\alias{gtkIconViewGetItemRow}
\alias{gtkIconViewGetItemWidth}
\alias{gtkIconViewGetMargin}
\alias{gtkIconViewGetMarkupColumn}
\alias{gtkIconViewGetModel}
\alias{gtkIconViewGetPathAtPos}
\alias{gtkIconViewGetPixbufColumn}
\alias{gtkIconViewGetReorderable}
\alias{gtkIconViewGetRowSpacing}
\alias{gtkIconViewGetSelectedItems}
\alias{gtkIconViewGetSelectionMode}
\alias{gtkIconViewGetSpacing}
\alias{gtkIconViewGetTextColumn}
\alias{gtkIconViewGetTooltipColumn}
\alias{gtkIconViewGetTooltipContext}
\alias{gtkIconViewGetVisibleRange}
\alias{gtkIconViewItemActivated}
\alias{gtkIconViewPathIsSelected}
\alias{gtkIconViewScrollToPath}
\alias{gtkIconViewSelectAll}
\alias{gtkIconViewSelectPath}
\alias{gtkIconViewSelectedForeach}
\alias{gtkIconViewSetActivateOnSingleClick}
\alias{gtkIconViewSetColumnSpacing}
\alias{gtkIconViewSetColumns}
\alias{gtkIconViewSetCursor}
\alias{gtkIconViewSetDragDestItem}
\alias{gtkIconViewSetItemOrientation}
\alias{gtkIconViewSetItemPadding}
\alias{gtkIconViewSetItemWidth}
\alias{gtkIconViewSetMargin}
\alias{gtkIconViewSetMarkupColumn}
\alias{gtkIconViewSetModel}
\alias{gtkIconViewSetPixbufColumn}
\alias{gtkIconViewSetReorderable}
\alias{gtkIconViewSetRowSpacing}
\alias{gtkIconViewSetSelectionMode}
\alias{gtkIconViewSetSpacing}
\alias{gtkIconViewSetTextColumn}
\alias{gtkIconViewSetTooltipCell}
\alias{gtkIconViewSetTooltipColumn}
\alias{gtkIconViewSetTooltipItem}
\alias{gtkIconViewUnselectAll}
\alias{gtkIconViewUnselectPath}
\alias{gtkIconViewUnsetModelDragDest}
\alias{gtkIconViewUnsetModelDragSource}
\alias{gtkImageNew}
\alias{gtkImageNewFromFile}
\alias{gtkImageNewFromGicon}
\alias{gtkImageNewFromIconName}
\alias{gtkImageNewFromPaintable}
\alias{gtkImageNewFromPixbuf}
\alias{gtkImageNewFromResource}
\alias{gtkImageClear}
\alias{gtkImageGetGicon}
\alias{gtkImageGetIconName}
\alias{gtkImageGetIconSize}
\alias{gtkImageGetPaintable}
\alias{gtkImageGetPixelSize}
\alias{gtkImageGetStorageType}
\alias{gtkImageSetFromFile}
\alias{gtkImageSetFromGicon}
\alias{gtkImageSetFromIconName}
\alias{gtkImageSetFromPaintable}
\alias{gtkImageSetFromPixbuf}
\alias{gtkImageSetFromResource}
\alias{gtkImageSetIconSize}
\alias{gtkImageSetPixelSize}
\alias{gtkPictureNew}
\alias{gtkPictureNewForFile}
\alias{gtkPictureNewForFilename}
\alias{gtkPictureNewForPaintable}
\alias{gtkPictureNewForPixbuf}
\alias{gtkPictureNewForResource}
\alias{gtkPictureGetAlternativeText}
\alias{gtkPictureGetCanShrink}
\alias{gtkPictureGetContentFit}
\alias{gtkPictureGetFile}
\alias{gtkPictureGetKeepAspectRatio}
\alias{gtkPictureGetPaintable}
\alias{gtkPictureSetAlternativeText}
\alias{gtkPictureSetCanShrink}
\alias{gtkPictureSetContentFit}
\alias{gtkPictureSetFile}
\alias{gtkPictureSetFilename}
\alias{gtkPictureSetKeepAspectRatio}
\alias{gtkPictureSetPaintable}
\alias{gtkPictureSetPixbuf}
\alias{gtkPictureSetResource}
\title{gtkIconPaintableNewForFile}
\usage{
gtkIconPaintableNewForFile(file, size, scale)

gtkIconPaintableGetFile(self)

gtkIconPaintableGetIconName(self)

gtkIconPaintableIsSymbolic(self)

gtkIconThemeNew()

gtkIconThemeGetForDisplay(display)

gtkIconThemeAddResourcePath(self, path)

gtkIconThemeAddSearchPath(self, path)

gtkIconThemeGetDisplay(self)

gtkIconThemeGetIconNames(self)

gtkIconThemeGetIconSizes(self, icon_name)

gtkIconThemeGetResourcePath(self)

gtkIconThemeGetSearchPath(self)

gtkIconThemeGetThemeName(self)

gtkIconThemeHasGicon(self, gicon)

gtkIconThemeHasIcon(self, icon_name)

gtkIconThemeLookupByGicon(self, icon, size, scale, direction, flags)

gtkIconThemeLookupIcon(
  self,
  icon_name,
  fallbacks,
  size,
  scale,
  direction,
  flags
)

gtkIconThemeSetResourcePath(self, path)

gtkIconThemeSetSearchPath(self, path)

gtkIconThemeSetThemeName(self, theme_name)

gtkIconThemeErrorQuark()

gtkIconViewNew()

gtkIconViewNewWithArea(area)

gtkIconViewNewWithModel(model)

gtkIconViewCreateDragIcon(icon_view, path)

gtkIconViewEnableModelDragDest(icon_view, formats, actions)

gtkIconViewEnableModelDragSource(
  icon_view,
  start_button_mask,
  formats,
  actions
)

gtkIconViewGetActivateOnSingleClick(icon_view)

gtkIconViewGetCellRect(icon_view, path, cell)

gtkIconViewGetColumnSpacing(icon_view)

gtkIconViewGetColumns(icon_view)

gtkIconViewGetCursor(icon_view)

gtkIconViewGetDestItemAtPos(icon_view, drag_x, drag_y)

gtkIconViewGetDragDestItem(icon_view)

gtkIconViewGetItemAtPos(icon_view, x, y)

gtkIconViewGetItemColumn(icon_view, path)

gtkIconViewGetItemOrientation(icon_view)

gtkIconViewGetItemPadding(icon_view)

gtkIconViewGetItemRow(icon_view, path)

gtkIconViewGetItemWidth(icon_view)

gtkIconViewGetMargin(icon_view)

gtkIconViewGetMarkupColumn(icon_view)

gtkIconViewGetModel(icon_view)

gtkIconViewGetPathAtPos(icon_view, x, y)

gtkIconViewGetPixbufColumn(icon_view)

gtkIconViewGetReorderable(icon_view)

gtkIconViewGetRowSpacing(icon_view)

gtkIconViewGetSelectedItems(icon_view)

gtkIconViewGetSelectionMode(icon_view)

gtkIconViewGetSpacing(icon_view)

gtkIconViewGetTextColumn(icon_view)

gtkIconViewGetTooltipColumn(icon_view)

gtkIconViewGetTooltipContext(icon_view, x, y, keyboard_tip)

gtkIconViewGetVisibleRange(icon_view)

gtkIconViewItemActivated(icon_view, path)

gtkIconViewPathIsSelected(icon_view, path)

gtkIconViewScrollToPath(icon_view, path, use_align, row_align, col_align)

gtkIconViewSelectAll(icon_view)

gtkIconViewSelectPath(icon_view, path)

gtkIconViewSelectedForeach(icon_view, func)

gtkIconViewSetActivateOnSingleClick(icon_view, single)

gtkIconViewSetColumnSpacing(icon_view, column_spacing)

gtkIconViewSetColumns(icon_view, columns)

gtkIconViewSetCursor(icon_view, path, cell, start_editing)

gtkIconViewSetDragDestItem(icon_view, path, pos)

gtkIconViewSetItemOrientation(icon_view, orientation)

gtkIconViewSetItemPadding(icon_view, item_padding)

gtkIconViewSetItemWidth(icon_view, item_width)

gtkIconViewSetMargin(icon_view, margin)

gtkIconViewSetMarkupColumn(icon_view, column)

gtkIconViewSetModel(icon_view, model)

gtkIconViewSetPixbufColumn(icon_view, column)

gtkIconViewSetReorderable(icon_view, reorderable)

gtkIconViewSetRowSpacing(icon_view, row_spacing)

gtkIconViewSetSelectionMode(icon_view, mode)

gtkIconViewSetSpacing(icon_view, spacing)

gtkIconViewSetTextColumn(icon_view, column)

gtkIconViewSetTooltipCell(icon_view, tooltip, path, cell)

gtkIconViewSetTooltipColumn(icon_view, column)

gtkIconViewSetTooltipItem(icon_view, tooltip, path)

gtkIconViewUnselectAll(icon_view)

gtkIconViewUnselectPath(icon_view, path)

gtkIconViewUnsetModelDragDest(icon_view)

gtkIconViewUnsetModelDragSource(icon_view)

gtkImageNew()

gtkImageNewFromFile(filename)

gtkImageNewFromGicon(icon)

gtkImageNewFromIconName(icon_name)

gtkImageNewFromPaintable(paintable)

gtkImageNewFromPixbuf(pixbuf)

gtkImageNewFromResource(resource_path)

gtkImageClear(image)

gtkImageGetGicon(image)

gtkImageGetIconName(image)

gtkImageGetIconSize(image)

gtkImageGetPaintable(image)

gtkImageGetPixelSize(image)

gtkImageGetStorageType(image)

gtkImageSetFromFile(image, filename)

gtkImageSetFromGicon(image, icon)

gtkImageSetFromIconName(image, icon_name)

gtkImageSetFromPaintable(image, paintable)

gtkImageSetFromPixbuf(image, pixbuf)

gtkImageSetFromResource(image, resource_path)

gtkImageSetIconSize(image, icon_size)

gtkImageSetPixelSize(image, pixel_size)

gtkPictureNew()

gtkPictureNewForFile(file)

gtkPictureNewForFilename(filename)

gtkPictureNewForPaintable(paintable)

gtkPictureNewForPixbuf(pixbuf)

gtkPictureNewForResource(resource_path)

gtkPictureGetAlternativeText(self)

gtkPictureGetCanShrink(self)

gtkPictureGetContentFit(self)

gtkPictureGetFile(self)

gtkPictureGetKeepAspectRatio(self)

gtkPictureGetPaintable(self)

gtkPictureSetAlternativeText(self, alternative_text)

gtkPictureSetCanShrink(self, can_shrink)

gtkPictureSetContentFit(self, content_fit)

gtkPictureSetFile(self, file)

gtkPictureSetFilename(self, filename)

gtkPictureSetKeepAspectRatio(self, keep_aspect_ratio)

gtkPictureSetPaintable(self, paintable)

gtkPictureSetPixbuf(self, pixbuf)

gtkPictureSetResource(self, resource_path)
}
\arguments{
\item{file}{Gio.File}

\item{size}{gint}

\item{scale}{gint}

\item{self}{Picture}

\item{display}{Gdk.Display}

\item{path}{TreePath}

\item{icon_name}{utf8}

\item{gicon}{Gio.Icon}

\item{icon}{Gio.Icon}

\item{direction}{TextDirection}

\item{flags}{IconLookupFlags}

\item{fallbacks}{utf8}

\item{theme_name}{utf8}

\item{area}{CellArea}

\item{model}{TreeModel}

\item{icon_view}{IconView}

\item{formats}{Gdk.ContentFormats}

\item{actions}{Gdk.DragAction}

\item{start_button_mask}{Gdk.ModifierType}

\item{cell}{CellRenderer}

\item{drag_x}{gint}

\item{drag_y}{gint}

\item{x}{gint}

\item{y}{gint}

\item{keyboard_tip}{gboolean}

\item{use_align}{gboolean}

\item{row_align}{gfloat}

\item{col_align}{gfloat}

\item{func}{function — IconViewForeachFunc callback}

\item{single}{gboolean}

\item{column_spacing}{gint}

\item{columns}{gint}

\item{start_editing}{gboolean}

\item{pos}{IconViewDropPosition}

\item{orientation}{Orientation}

\item{item_padding}{gint}

\item{item_width}{gint}

\item{margin}{gint}

\item{column}{gint}

\item{reorderable}{gboolean}

\item{row_spacing}{gint}

\item{mode}{SelectionMode}

\item{spacing}{gint}

\item{tooltip}{Tooltip}

\item{filename}{filename}

\item{paintable}{Gdk.Paintable}

\item{pixbuf}{GdkPixbuf.Pixbuf}

\item{resource_path}{utf8}

\item{image}{Image}

\item{icon_size}{IconSize}

\item{pixel_size}{gint}

\item{alternative_text}{utf8}

\item{can_shrink}{gboolean}

\item{content_fit}{ContentFit}

\item{keep_aspect_ratio}{gboolean}
}
\value{
IconPaintable

Gio.File

filename

gboolean

IconTheme

IconTheme

Return value from C function

Return value from C function

Gdk.Display

utf8

gint

utf8

filename

utf8

gboolean

gboolean

IconPaintable

IconPaintable

Return value from C function

Return value from C function

Return value from C function

GLib.Quark

Widget

Widget

Widget

Gdk.Paintable

Return value from C function

Return value from C function

gboolean

gboolean

gint

gint

gboolean

gboolean

Return value from C function

gboolean

gint

Orientation

gint

gint

gint

gint

gint

TreeModel

TreePath

gint

gboolean

gint

GLib.List

SelectionMode

gint

gint

gint

gboolean

gboolean

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

Widget

Widget

Widget

Widget

Widget

Widget

Return value from C function

Gio.Icon

utf8

IconSize

Gdk.Paintable

gint

ImageType

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

Widget

Widget

Widget

Widget

Widget

utf8

gboolean

ContentFit

Gio.File

gboolean

Gdk.Paintable

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function
}
\description{
gtkIconPaintableNewForFile

gtkIconPaintableGetFile

gtkIconPaintableGetIconName

gtkIconPaintableIsSymbolic

gtkIconThemeNew

gtkIconThemeGetForDisplay

gtkIconThemeAddResourcePath

gtkIconThemeAddSearchPath

gtkIconThemeGetDisplay

gtkIconThemeGetIconNames

gtkIconThemeGetIconSizes

gtkIconThemeGetResourcePath

gtkIconThemeGetSearchPath

gtkIconThemeGetThemeName

gtkIconThemeHasGicon

gtkIconThemeHasIcon

gtkIconThemeLookupByGicon

gtkIconThemeLookupIcon

gtkIconThemeSetResourcePath

gtkIconThemeSetSearchPath

gtkIconThemeSetThemeName

gtkIconThemeErrorQuark

gtkIconViewNew

gtkIconViewNewWithArea

gtkIconViewNewWithModel

gtkIconViewCreateDragIcon

gtkIconViewEnableModelDragDest

gtkIconViewEnableModelDragSource

gtkIconViewGetActivateOnSingleClick

gtkIconViewGetCellRect

gtkIconViewGetColumnSpacing

gtkIconViewGetColumns

gtkIconViewGetCursor

gtkIconViewGetDestItemAtPos

gtkIconViewGetDragDestItem

gtkIconViewGetItemAtPos

gtkIconViewGetItemColumn

gtkIconViewGetItemOrientation

gtkIconViewGetItemPadding

gtkIconViewGetItemRow

gtkIconViewGetItemWidth

gtkIconViewGetMargin

gtkIconViewGetMarkupColumn

gtkIconViewGetModel

gtkIconViewGetPathAtPos

gtkIconViewGetPixbufColumn

gtkIconViewGetReorderable

gtkIconViewGetRowSpacing

gtkIconViewGetSelectedItems

gtkIconViewGetSelectionMode

gtkIconViewGetSpacing

gtkIconViewGetTextColumn

gtkIconViewGetTooltipColumn

gtkIconViewGetTooltipContext

gtkIconViewGetVisibleRange

gtkIconViewItemActivated

gtkIconViewPathIsSelected

gtkIconViewScrollToPath

gtkIconViewSelectAll

gtkIconViewSelectPath

gtkIconViewSelectedForeach

gtkIconViewSetActivateOnSingleClick

gtkIconViewSetColumnSpacing

gtkIconViewSetColumns

gtkIconViewSetCursor

gtkIconViewSetDragDestItem

gtkIconViewSetItemOrientation

gtkIconViewSetItemPadding

gtkIconViewSetItemWidth

gtkIconViewSetMargin

gtkIconViewSetMarkupColumn

gtkIconViewSetModel

gtkIconViewSetPixbufColumn

gtkIconViewSetReorderable

gtkIconViewSetRowSpacing

gtkIconViewSetSelectionMode

gtkIconViewSetSpacing

gtkIconViewSetTextColumn

gtkIconViewSetTooltipCell

gtkIconViewSetTooltipColumn

gtkIconViewSetTooltipItem

gtkIconViewUnselectAll

gtkIconViewUnselectPath

gtkIconViewUnsetModelDragDest

gtkIconViewUnsetModelDragSource

gtkImageNew

gtkImageNewFromFile

gtkImageNewFromGicon

gtkImageNewFromIconName

gtkImageNewFromPaintable

gtkImageNewFromPixbuf

gtkImageNewFromResource

gtkImageClear

gtkImageGetGicon

gtkImageGetIconName

gtkImageGetIconSize

gtkImageGetPaintable

gtkImageGetPixelSize

gtkImageGetStorageType

gtkImageSetFromFile

gtkImageSetFromGicon

gtkImageSetFromIconName

gtkImageSetFromPaintable

gtkImageSetFromPixbuf

gtkImageSetFromResource

gtkImageSetIconSize

gtkImageSetPixelSize

gtkPictureNew

gtkPictureNewForFile

gtkPictureNewForFilename

gtkPictureNewForPaintable

gtkPictureNewForPixbuf

gtkPictureNewForResource

gtkPictureGetAlternativeText

gtkPictureGetCanShrink

gtkPictureGetContentFit

gtkPictureGetFile

gtkPictureGetKeepAspectRatio

gtkPictureGetPaintable

gtkPictureSetAlternativeText

gtkPictureSetCanShrink

gtkPictureSetContentFit

gtkPictureSetFile

gtkPictureSetFilename

gtkPictureSetKeepAspectRatio

gtkPictureSetPaintable

gtkPictureSetPixbuf

gtkPictureSetResource
}
