% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtk_autogen.R
\name{gtkBoxNew}
\alias{gtkBoxNew}
\alias{gtkBoxAppend}
\alias{gtkBoxGetBaselineChild}
\alias{gtkBoxGetBaselinePosition}
\alias{gtkBoxGetHomogeneous}
\alias{gtkBoxGetSpacing}
\alias{gtkBoxInsertChildAfter}
\alias{gtkBoxPrepend}
\alias{gtkBoxRemove}
\alias{gtkBoxReorderChildAfter}
\alias{gtkBoxSetBaselineChild}
\alias{gtkBoxSetBaselinePosition}
\alias{gtkBoxSetHomogeneous}
\alias{gtkBoxSetSpacing}
\alias{gtkBoxLayoutNew}
\alias{gtkBoxLayoutGetBaselineChild}
\alias{gtkBoxLayoutGetBaselinePosition}
\alias{gtkBoxLayoutGetHomogeneous}
\alias{gtkBoxLayoutGetSpacing}
\alias{gtkBoxLayoutSetBaselineChild}
\alias{gtkBoxLayoutSetBaselinePosition}
\alias{gtkBoxLayoutSetHomogeneous}
\alias{gtkBoxLayoutSetSpacing}
\alias{gtkGridNew}
\alias{gtkGridAttach}
\alias{gtkGridAttachNextTo}
\alias{gtkGridGetBaselineRow}
\alias{gtkGridGetChildAt}
\alias{gtkGridGetColumnHomogeneous}
\alias{gtkGridGetColumnSpacing}
\alias{gtkGridGetRowBaselinePosition}
\alias{gtkGridGetRowHomogeneous}
\alias{gtkGridGetRowSpacing}
\alias{gtkGridInsertColumn}
\alias{gtkGridInsertNextTo}
\alias{gtkGridInsertRow}
\alias{gtkGridQueryChild}
\alias{gtkGridRemove}
\alias{gtkGridRemoveColumn}
\alias{gtkGridRemoveRow}
\alias{gtkGridSetBaselineRow}
\alias{gtkGridSetColumnHomogeneous}
\alias{gtkGridSetColumnSpacing}
\alias{gtkGridSetRowBaselinePosition}
\alias{gtkGridSetRowHomogeneous}
\alias{gtkGridSetRowSpacing}
\alias{gtkGridLayoutNew}
\alias{gtkGridLayoutGetBaselineRow}
\alias{gtkGridLayoutGetColumnHomogeneous}
\alias{gtkGridLayoutGetColumnSpacing}
\alias{gtkGridLayoutGetRowBaselinePosition}
\alias{gtkGridLayoutGetRowHomogeneous}
\alias{gtkGridLayoutGetRowSpacing}
\alias{gtkGridLayoutSetBaselineRow}
\alias{gtkGridLayoutSetColumnHomogeneous}
\alias{gtkGridLayoutSetColumnSpacing}
\alias{gtkGridLayoutSetRowBaselinePosition}
\alias{gtkGridLayoutSetRowHomogeneous}
\alias{gtkGridLayoutSetRowSpacing}
\alias{gtkGridLayoutChildGetColumn}
\alias{gtkGridLayoutChildGetColumnSpan}
\alias{gtkGridLayoutChildGetRow}
\alias{gtkGridLayoutChildGetRowSpan}
\alias{gtkGridLayoutChildSetColumn}
\alias{gtkGridLayoutChildSetColumnSpan}
\alias{gtkGridLayoutChildSetRow}
\alias{gtkGridLayoutChildSetRowSpan}
\alias{gtkGridViewNew}
\alias{gtkGridViewGetEnableRubberband}
\alias{gtkGridViewGetFactory}
\alias{gtkGridViewGetMaxColumns}
\alias{gtkGridViewGetMinColumns}
\alias{gtkGridViewGetModel}
\alias{gtkGridViewGetSingleClickActivate}
\alias{gtkGridViewGetTabBehavior}
\alias{gtkGridViewScrollTo}
\alias{gtkGridViewSetEnableRubberband}
\alias{gtkGridViewSetFactory}
\alias{gtkGridViewSetMaxColumns}
\alias{gtkGridViewSetMinColumns}
\alias{gtkGridViewSetModel}
\alias{gtkGridViewSetSingleClickActivate}
\alias{gtkGridViewSetTabBehavior}
\alias{gtkPanedNew}
\alias{gtkPanedGetEndChild}
\alias{gtkPanedGetPosition}
\alias{gtkPanedGetResizeEndChild}
\alias{gtkPanedGetResizeStartChild}
\alias{gtkPanedGetShrinkEndChild}
\alias{gtkPanedGetShrinkStartChild}
\alias{gtkPanedGetStartChild}
\alias{gtkPanedGetWideHandle}
\alias{gtkPanedSetEndChild}
\alias{gtkPanedSetPosition}
\alias{gtkPanedSetResizeEndChild}
\alias{gtkPanedSetResizeStartChild}
\alias{gtkPanedSetShrinkEndChild}
\alias{gtkPanedSetShrinkStartChild}
\alias{gtkPanedSetStartChild}
\alias{gtkPanedSetWideHandle}
\title{gtkBoxNew}
\usage{
gtkBoxNew(orientation, spacing)

gtkBoxAppend(box, child)

gtkBoxGetBaselineChild(box)

gtkBoxGetBaselinePosition(box)

gtkBoxGetHomogeneous(box)

gtkBoxGetSpacing(box)

gtkBoxInsertChildAfter(box, child, sibling)

gtkBoxPrepend(box, child)

gtkBoxRemove(box, child)

gtkBoxReorderChildAfter(box, child, sibling)

gtkBoxSetBaselineChild(box, child)

gtkBoxSetBaselinePosition(box, position)

gtkBoxSetHomogeneous(box, homogeneous)

gtkBoxSetSpacing(box, spacing)

gtkBoxLayoutNew(orientation)

gtkBoxLayoutGetBaselineChild(box_layout)

gtkBoxLayoutGetBaselinePosition(box_layout)

gtkBoxLayoutGetHomogeneous(box_layout)

gtkBoxLayoutGetSpacing(box_layout)

gtkBoxLayoutSetBaselineChild(box_layout, child)

gtkBoxLayoutSetBaselinePosition(box_layout, position)

gtkBoxLayoutSetHomogeneous(box_layout, homogeneous)

gtkBoxLayoutSetSpacing(box_layout, spacing)

gtkGridNew()

gtkGridAttach(grid, child, column, row, width, height)

gtkGridAttachNextTo(grid, child, sibling, side, width, height)

gtkGridGetBaselineRow(grid)

gtkGridGetChildAt(grid, column, row)

gtkGridGetColumnHomogeneous(grid)

gtkGridGetColumnSpacing(grid)

gtkGridGetRowBaselinePosition(grid, row)

gtkGridGetRowHomogeneous(grid)

gtkGridGetRowSpacing(grid)

gtkGridInsertColumn(grid, position)

gtkGridInsertNextTo(grid, sibling, side)

gtkGridInsertRow(grid, position)

gtkGridQueryChild(grid, child)

gtkGridRemove(grid, child)

gtkGridRemoveColumn(grid, position)

gtkGridRemoveRow(grid, position)

gtkGridSetBaselineRow(grid, row)

gtkGridSetColumnHomogeneous(grid, homogeneous)

gtkGridSetColumnSpacing(grid, spacing)

gtkGridSetRowBaselinePosition(grid, row, pos)

gtkGridSetRowHomogeneous(grid, homogeneous)

gtkGridSetRowSpacing(grid, spacing)

gtkGridLayoutNew()

gtkGridLayoutGetBaselineRow(grid)

gtkGridLayoutGetColumnHomogeneous(grid)

gtkGridLayoutGetColumnSpacing(grid)

gtkGridLayoutGetRowBaselinePosition(grid, row)

gtkGridLayoutGetRowHomogeneous(grid)

gtkGridLayoutGetRowSpacing(grid)

gtkGridLayoutSetBaselineRow(grid, row)

gtkGridLayoutSetColumnHomogeneous(grid, homogeneous)

gtkGridLayoutSetColumnSpacing(grid, spacing)

gtkGridLayoutSetRowBaselinePosition(grid, row, pos)

gtkGridLayoutSetRowHomogeneous(grid, homogeneous)

gtkGridLayoutSetRowSpacing(grid, spacing)

gtkGridLayoutChildGetColumn(child)

gtkGridLayoutChildGetColumnSpan(child)

gtkGridLayoutChildGetRow(child)

gtkGridLayoutChildGetRowSpan(child)

gtkGridLayoutChildSetColumn(child, column)

gtkGridLayoutChildSetColumnSpan(child, span)

gtkGridLayoutChildSetRow(child, row)

gtkGridLayoutChildSetRowSpan(child, span)

gtkGridViewNew(model, factory)

gtkGridViewGetEnableRubberband(self)

gtkGridViewGetFactory(self)

gtkGridViewGetMaxColumns(self)

gtkGridViewGetMinColumns(self)

gtkGridViewGetModel(self)

gtkGridViewGetSingleClickActivate(self)

gtkGridViewGetTabBehavior(self)

gtkGridViewScrollTo(self, pos, flags, scroll)

gtkGridViewSetEnableRubberband(self, enable_rubberband)

gtkGridViewSetFactory(self, factory)

gtkGridViewSetMaxColumns(self, max_columns)

gtkGridViewSetMinColumns(self, min_columns)

gtkGridViewSetModel(self, model)

gtkGridViewSetSingleClickActivate(self, single_click_activate)

gtkGridViewSetTabBehavior(self, tab_behavior)

gtkPanedNew(orientation)

gtkPanedGetEndChild(paned)

gtkPanedGetPosition(paned)

gtkPanedGetResizeEndChild(paned)

gtkPanedGetResizeStartChild(paned)

gtkPanedGetShrinkEndChild(paned)

gtkPanedGetShrinkStartChild(paned)

gtkPanedGetStartChild(paned)

gtkPanedGetWideHandle(paned)

gtkPanedSetEndChild(paned, child)

gtkPanedSetPosition(paned, position)

gtkPanedSetResizeEndChild(paned, resize)

gtkPanedSetResizeStartChild(paned, resize)

gtkPanedSetShrinkEndChild(paned, resize)

gtkPanedSetShrinkStartChild(paned, resize)

gtkPanedSetStartChild(paned, child)

gtkPanedSetWideHandle(paned, wide)
}
\arguments{
\item{orientation}{Orientation}

\item{spacing}{guint}

\item{box}{Box}

\item{child}{Widget}

\item{sibling}{Widget}

\item{position}{gint}

\item{homogeneous}{gboolean}

\item{box_layout}{BoxLayout}

\item{grid}{GridLayout}

\item{column}{gint}

\item{row}{gint}

\item{width}{gint}

\item{height}{gint}

\item{side}{PositionType}

\item{pos}{guint}

\item{span}{gint}

\item{model}{SelectionModel}

\item{factory}{ListItemFactory}

\item{self}{GridView}

\item{flags}{ListScrollFlags}

\item{scroll}{ScrollInfo}

\item{enable_rubberband}{gboolean}

\item{max_columns}{guint}

\item{min_columns}{guint}

\item{single_click_activate}{gboolean}

\item{tab_behavior}{ListTabBehavior}

\item{paned}{Paned}

\item{resize}{gboolean}

\item{wide}{gboolean}
}
\value{
Widget

Return value from C function

gint

BaselinePosition

gboolean

gint

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

LayoutManager

gint

BaselinePosition

gboolean

guint

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

Return value from C function

Return value from C function

gint

Widget

gboolean

guint

BaselinePosition

gboolean

guint

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

LayoutManager

gint

gboolean

guint

BaselinePosition

gboolean

guint

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gint

gint

gint

gint

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

gboolean

ListItemFactory

guint

guint

SelectionModel

gboolean

ListTabBehavior

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

Widget

gint

gboolean

gboolean

gboolean

gboolean

Widget

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function
}
\description{
gtkBoxNew

gtkBoxAppend

gtkBoxGetBaselineChild

gtkBoxGetBaselinePosition

gtkBoxGetHomogeneous

gtkBoxGetSpacing

gtkBoxInsertChildAfter

gtkBoxPrepend

gtkBoxRemove

gtkBoxReorderChildAfter

gtkBoxSetBaselineChild

gtkBoxSetBaselinePosition

gtkBoxSetHomogeneous

gtkBoxSetSpacing

gtkBoxLayoutNew

gtkBoxLayoutGetBaselineChild

gtkBoxLayoutGetBaselinePosition

gtkBoxLayoutGetHomogeneous

gtkBoxLayoutGetSpacing

gtkBoxLayoutSetBaselineChild

gtkBoxLayoutSetBaselinePosition

gtkBoxLayoutSetHomogeneous

gtkBoxLayoutSetSpacing

gtkGridNew

gtkGridAttach

gtkGridAttachNextTo

gtkGridGetBaselineRow

gtkGridGetChildAt

gtkGridGetColumnHomogeneous

gtkGridGetColumnSpacing

gtkGridGetRowBaselinePosition

gtkGridGetRowHomogeneous

gtkGridGetRowSpacing

gtkGridInsertColumn

gtkGridInsertNextTo

gtkGridInsertRow

gtkGridQueryChild

gtkGridRemove

gtkGridRemoveColumn

gtkGridRemoveRow

gtkGridSetBaselineRow

gtkGridSetColumnHomogeneous

gtkGridSetColumnSpacing

gtkGridSetRowBaselinePosition

gtkGridSetRowHomogeneous

gtkGridSetRowSpacing

gtkGridLayoutNew

gtkGridLayoutGetBaselineRow

gtkGridLayoutGetColumnHomogeneous

gtkGridLayoutGetColumnSpacing

gtkGridLayoutGetRowBaselinePosition

gtkGridLayoutGetRowHomogeneous

gtkGridLayoutGetRowSpacing

gtkGridLayoutSetBaselineRow

gtkGridLayoutSetColumnHomogeneous

gtkGridLayoutSetColumnSpacing

gtkGridLayoutSetRowBaselinePosition

gtkGridLayoutSetRowHomogeneous

gtkGridLayoutSetRowSpacing

gtkGridLayoutChildGetColumn

gtkGridLayoutChildGetColumnSpan

gtkGridLayoutChildGetRow

gtkGridLayoutChildGetRowSpan

gtkGridLayoutChildSetColumn

gtkGridLayoutChildSetColumnSpan

gtkGridLayoutChildSetRow

gtkGridLayoutChildSetRowSpan

gtkGridViewNew

gtkGridViewGetEnableRubberband

gtkGridViewGetFactory

gtkGridViewGetMaxColumns

gtkGridViewGetMinColumns

gtkGridViewGetModel

gtkGridViewGetSingleClickActivate

gtkGridViewGetTabBehavior

gtkGridViewScrollTo

gtkGridViewSetEnableRubberband

gtkGridViewSetFactory

gtkGridViewSetMaxColumns

gtkGridViewSetMinColumns

gtkGridViewSetModel

gtkGridViewSetSingleClickActivate

gtkGridViewSetTabBehavior

gtkPanedNew

gtkPanedGetEndChild

gtkPanedGetPosition

gtkPanedGetResizeEndChild

gtkPanedGetResizeStartChild

gtkPanedGetShrinkEndChild

gtkPanedGetShrinkStartChild

gtkPanedGetStartChild

gtkPanedGetWideHandle

gtkPanedSetEndChild

gtkPanedSetPosition

gtkPanedSetResizeEndChild

gtkPanedSetResizeStartChild

gtkPanedSetShrinkEndChild

gtkPanedSetShrinkStartChild

gtkPanedSetStartChild

gtkPanedSetWideHandle
}
