% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtk_autogen.R
\name{gtkMenuButtonNew}
\alias{gtkMenuButtonNew}
\alias{gtkMenuButtonGetActive}
\alias{gtkMenuButtonGetAlwaysShowArrow}
\alias{gtkMenuButtonGetCanShrink}
\alias{gtkMenuButtonGetChild}
\alias{gtkMenuButtonGetDirection}
\alias{gtkMenuButtonGetHasFrame}
\alias{gtkMenuButtonGetIconName}
\alias{gtkMenuButtonGetLabel}
\alias{gtkMenuButtonGetMenuModel}
\alias{gtkMenuButtonGetPopover}
\alias{gtkMenuButtonGetPrimary}
\alias{gtkMenuButtonGetUseUnderline}
\alias{gtkMenuButtonPopdown}
\alias{gtkMenuButtonPopup}
\alias{gtkMenuButtonSetActive}
\alias{gtkMenuButtonSetAlwaysShowArrow}
\alias{gtkMenuButtonSetCanShrink}
\alias{gtkMenuButtonSetChild}
\alias{gtkMenuButtonSetCreatePopupFunc}
\alias{gtkMenuButtonSetDirection}
\alias{gtkMenuButtonSetHasFrame}
\alias{gtkMenuButtonSetIconName}
\alias{gtkMenuButtonSetLabel}
\alias{gtkMenuButtonSetMenuModel}
\alias{gtkMenuButtonSetPopover}
\alias{gtkMenuButtonSetPrimary}
\alias{gtkMenuButtonSetUseUnderline}
\alias{gtkPopoverNew}
\alias{gtkPopoverGetAutohide}
\alias{gtkPopoverGetCascadePopdown}
\alias{gtkPopoverGetChild}
\alias{gtkPopoverGetHasArrow}
\alias{gtkPopoverGetMnemonicsVisible}
\alias{gtkPopoverGetOffset}
\alias{gtkPopoverGetPointingTo}
\alias{gtkPopoverGetPosition}
\alias{gtkPopoverPopdown}
\alias{gtkPopoverPopup}
\alias{gtkPopoverPresent}
\alias{gtkPopoverSetAutohide}
\alias{gtkPopoverSetCascadePopdown}
\alias{gtkPopoverSetChild}
\alias{gtkPopoverSetDefaultWidget}
\alias{gtkPopoverSetHasArrow}
\alias{gtkPopoverSetMnemonicsVisible}
\alias{gtkPopoverSetOffset}
\alias{gtkPopoverSetPointingTo}
\alias{gtkPopoverSetPosition}
\alias{gtkPopoverMenuNewFromModel}
\alias{gtkPopoverMenuNewFromModelFull}
\alias{gtkPopoverMenuAddChild}
\alias{gtkPopoverMenuGetMenuModel}
\alias{gtkPopoverMenuRemoveChild}
\alias{gtkPopoverMenuSetMenuModel}
\alias{gtkPopoverMenuBarNewFromModel}
\alias{gtkPopoverMenuBarAddChild}
\alias{gtkPopoverMenuBarGetMenuModel}
\alias{gtkPopoverMenuBarRemoveChild}
\alias{gtkPopoverMenuBarSetMenuModel}
\title{gtkMenuButtonNew}
\usage{
gtkMenuButtonNew()

gtkMenuButtonGetActive(menu_button)

gtkMenuButtonGetAlwaysShowArrow(menu_button)

gtkMenuButtonGetCanShrink(menu_button)

gtkMenuButtonGetChild(menu_button)

gtkMenuButtonGetDirection(menu_button)

gtkMenuButtonGetHasFrame(menu_button)

gtkMenuButtonGetIconName(menu_button)

gtkMenuButtonGetLabel(menu_button)

gtkMenuButtonGetMenuModel(menu_button)

gtkMenuButtonGetPopover(menu_button)

gtkMenuButtonGetPrimary(menu_button)

gtkMenuButtonGetUseUnderline(menu_button)

gtkMenuButtonPopdown(menu_button)

gtkMenuButtonPopup(menu_button)

gtkMenuButtonSetActive(menu_button, active)

gtkMenuButtonSetAlwaysShowArrow(menu_button, always_show_arrow)

gtkMenuButtonSetCanShrink(menu_button, can_shrink)

gtkMenuButtonSetChild(menu_button, child)

gtkMenuButtonSetCreatePopupFunc(menu_button, func)

gtkMenuButtonSetDirection(menu_button, direction)

gtkMenuButtonSetHasFrame(menu_button, has_frame)

gtkMenuButtonSetIconName(menu_button, icon_name)

gtkMenuButtonSetLabel(menu_button, label)

gtkMenuButtonSetMenuModel(menu_button, menu_model)

gtkMenuButtonSetPopover(menu_button, popover)

gtkMenuButtonSetPrimary(menu_button, primary)

gtkMenuButtonSetUseUnderline(menu_button, use_underline)

gtkPopoverNew()

gtkPopoverGetAutohide(popover)

gtkPopoverGetCascadePopdown(popover)

gtkPopoverGetChild(popover)

gtkPopoverGetHasArrow(popover)

gtkPopoverGetMnemonicsVisible(popover)

gtkPopoverGetOffset(popover)

gtkPopoverGetPointingTo(popover)

gtkPopoverGetPosition(popover)

gtkPopoverPopdown(popover)

gtkPopoverPopup(popover)

gtkPopoverPresent(popover)

gtkPopoverSetAutohide(popover, autohide)

gtkPopoverSetCascadePopdown(popover, cascade_popdown)

gtkPopoverSetChild(popover, child)

gtkPopoverSetDefaultWidget(popover, widget)

gtkPopoverSetHasArrow(popover, has_arrow)

gtkPopoverSetMnemonicsVisible(popover, mnemonics_visible)

gtkPopoverSetOffset(popover, x_offset, y_offset)

gtkPopoverSetPointingTo(popover, rect)

gtkPopoverSetPosition(popover, position)

gtkPopoverMenuNewFromModel(model)

gtkPopoverMenuNewFromModelFull(model, flags)

gtkPopoverMenuAddChild(popover, child, id)

gtkPopoverMenuGetMenuModel(popover)

gtkPopoverMenuRemoveChild(popover, child)

gtkPopoverMenuSetMenuModel(popover, model)

gtkPopoverMenuBarNewFromModel(model)

gtkPopoverMenuBarAddChild(bar, child, id)

gtkPopoverMenuBarGetMenuModel(bar)

gtkPopoverMenuBarRemoveChild(bar, child)

gtkPopoverMenuBarSetMenuModel(bar, model)
}
\arguments{
\item{menu_button}{MenuButton}

\item{active}{gboolean}

\item{always_show_arrow}{gboolean}

\item{can_shrink}{gboolean}

\item{child}{Widget}

\item{func}{function — MenuButtonCreatePopupFunc callback}

\item{direction}{ArrowType}

\item{has_frame}{gboolean}

\item{icon_name}{utf8}

\item{label}{utf8}

\item{menu_model}{Gio.MenuModel}

\item{popover}{PopoverMenu}

\item{primary}{gboolean}

\item{use_underline}{gboolean}

\item{autohide}{gboolean}

\item{cascade_popdown}{gboolean}

\item{widget}{Widget}

\item{has_arrow}{gboolean}

\item{mnemonics_visible}{gboolean}

\item{x_offset}{gint}

\item{y_offset}{gint}

\item{rect}{Gdk.Rectangle}

\item{position}{PositionType}

\item{model}{Gio.MenuModel}

\item{flags}{PopoverMenuFlags}

\item{id}{utf8}

\item{bar}{PopoverMenuBar}
}
\value{
Widget

gboolean

gboolean

gboolean

Widget

ArrowType

gboolean

utf8

utf8

Gio.MenuModel

Popover

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

gboolean

gboolean

Widget

gboolean

gboolean

Return value from C function

gboolean

PositionType

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

Widget

gboolean

Gio.MenuModel

gboolean

Return value from C function

Widget

gboolean

Gio.MenuModel

gboolean

Return value from C function
}
\description{
gtkMenuButtonNew

gtkMenuButtonGetActive

gtkMenuButtonGetAlwaysShowArrow

gtkMenuButtonGetCanShrink

gtkMenuButtonGetChild

gtkMenuButtonGetDirection

gtkMenuButtonGetHasFrame

gtkMenuButtonGetIconName

gtkMenuButtonGetLabel

gtkMenuButtonGetMenuModel

gtkMenuButtonGetPopover

gtkMenuButtonGetPrimary

gtkMenuButtonGetUseUnderline

gtkMenuButtonPopdown

gtkMenuButtonPopup

gtkMenuButtonSetActive

gtkMenuButtonSetAlwaysShowArrow

gtkMenuButtonSetCanShrink

gtkMenuButtonSetChild

gtkMenuButtonSetCreatePopupFunc

gtkMenuButtonSetDirection

gtkMenuButtonSetHasFrame

gtkMenuButtonSetIconName

gtkMenuButtonSetLabel

gtkMenuButtonSetMenuModel

gtkMenuButtonSetPopover

gtkMenuButtonSetPrimary

gtkMenuButtonSetUseUnderline

gtkPopoverNew

gtkPopoverGetAutohide

gtkPopoverGetCascadePopdown

gtkPopoverGetChild

gtkPopoverGetHasArrow

gtkPopoverGetMnemonicsVisible

gtkPopoverGetOffset

gtkPopoverGetPointingTo

gtkPopoverGetPosition

gtkPopoverPopdown

gtkPopoverPopup

gtkPopoverPresent

gtkPopoverSetAutohide

gtkPopoverSetCascadePopdown

gtkPopoverSetChild

gtkPopoverSetDefaultWidget

gtkPopoverSetHasArrow

gtkPopoverSetMnemonicsVisible

gtkPopoverSetOffset

gtkPopoverSetPointingTo

gtkPopoverSetPosition

gtkPopoverMenuNewFromModel

gtkPopoverMenuNewFromModelFull

gtkPopoverMenuAddChild

gtkPopoverMenuGetMenuModel

gtkPopoverMenuRemoveChild

gtkPopoverMenuSetMenuModel

gtkPopoverMenuBarNewFromModel

gtkPopoverMenuBarAddChild

gtkPopoverMenuBarGetMenuModel

gtkPopoverMenuBarRemoveChild

gtkPopoverMenuBarSetMenuModel
}
