% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtk_autogen.R
\name{gtkEntryNew}
\alias{gtkEntryNew}
\alias{gtkEntryNewWithBuffer}
\alias{gtkEntryGetActivatesDefault}
\alias{gtkEntryGetAlignment}
\alias{gtkEntryGetAttributes}
\alias{gtkEntryGetBuffer}
\alias{gtkEntryGetCompletion}
\alias{gtkEntryGetCurrentIconDragSource}
\alias{gtkEntryGetExtraMenu}
\alias{gtkEntryGetHasFrame}
\alias{gtkEntryGetIconActivatable}
\alias{gtkEntryGetIconArea}
\alias{gtkEntryGetIconAtPos}
\alias{gtkEntryGetIconGicon}
\alias{gtkEntryGetIconName}
\alias{gtkEntryGetIconPaintable}
\alias{gtkEntryGetIconSensitive}
\alias{gtkEntryGetIconStorageType}
\alias{gtkEntryGetIconTooltipMarkup}
\alias{gtkEntryGetIconTooltipText}
\alias{gtkEntryGetInputHints}
\alias{gtkEntryGetInputPurpose}
\alias{gtkEntryGetInvisibleChar}
\alias{gtkEntryGetMaxLength}
\alias{gtkEntryGetMenuEntryIconText}
\alias{gtkEntryGetOverwriteMode}
\alias{gtkEntryGetPlaceholderText}
\alias{gtkEntryGetProgressFraction}
\alias{gtkEntryGetProgressPulseStep}
\alias{gtkEntryGetTabs}
\alias{gtkEntryGetTextLength}
\alias{gtkEntryGetVisibility}
\alias{gtkEntryGrabFocusWithoutSelecting}
\alias{gtkEntryProgressPulse}
\alias{gtkEntryResetImContext}
\alias{gtkEntrySetActivatesDefault}
\alias{gtkEntrySetAlignment}
\alias{gtkEntrySetAttributes}
\alias{gtkEntrySetBuffer}
\alias{gtkEntrySetCompletion}
\alias{gtkEntrySetExtraMenu}
\alias{gtkEntrySetHasFrame}
\alias{gtkEntrySetIconActivatable}
\alias{gtkEntrySetIconDragSource}
\alias{gtkEntrySetIconFromGicon}
\alias{gtkEntrySetIconFromIconName}
\alias{gtkEntrySetIconFromPaintable}
\alias{gtkEntrySetIconSensitive}
\alias{gtkEntrySetIconTooltipMarkup}
\alias{gtkEntrySetIconTooltipText}
\alias{gtkEntrySetInputHints}
\alias{gtkEntrySetInputPurpose}
\alias{gtkEntrySetInvisibleChar}
\alias{gtkEntrySetMaxLength}
\alias{gtkEntrySetMenuEntryIconText}
\alias{gtkEntrySetOverwriteMode}
\alias{gtkEntrySetPlaceholderText}
\alias{gtkEntrySetProgressFraction}
\alias{gtkEntrySetProgressPulseStep}
\alias{gtkEntrySetTabs}
\alias{gtkEntrySetVisibility}
\alias{gtkEntryUnsetInvisibleChar}
\alias{gtkEntryBufferNew}
\alias{gtkEntryBufferDeleteText}
\alias{gtkEntryBufferEmitDeletedText}
\alias{gtkEntryBufferEmitInsertedText}
\alias{gtkEntryBufferGetBytes}
\alias{gtkEntryBufferGetLength}
\alias{gtkEntryBufferGetMaxLength}
\alias{gtkEntryBufferGetText}
\alias{gtkEntryBufferInsertText}
\alias{gtkEntryBufferSetMaxLength}
\alias{gtkEntryBufferSetText}
\alias{gtkEntryCompletionNew}
\alias{gtkEntryCompletionNewWithArea}
\alias{gtkEntryCompletionComplete}
\alias{gtkEntryCompletionComputePrefix}
\alias{gtkEntryCompletionGetCompletionPrefix}
\alias{gtkEntryCompletionGetEntry}
\alias{gtkEntryCompletionGetInlineCompletion}
\alias{gtkEntryCompletionGetInlineSelection}
\alias{gtkEntryCompletionGetMinimumKeyLength}
\alias{gtkEntryCompletionGetModel}
\alias{gtkEntryCompletionGetPopupCompletion}
\alias{gtkEntryCompletionGetPopupSetWidth}
\alias{gtkEntryCompletionGetPopupSingleMatch}
\alias{gtkEntryCompletionGetTextColumn}
\alias{gtkEntryCompletionInsertPrefix}
\alias{gtkEntryCompletionSetInlineCompletion}
\alias{gtkEntryCompletionSetInlineSelection}
\alias{gtkEntryCompletionSetMinimumKeyLength}
\alias{gtkEntryCompletionSetModel}
\alias{gtkEntryCompletionSetPopupCompletion}
\alias{gtkEntryCompletionSetPopupSetWidth}
\alias{gtkEntryCompletionSetPopupSingleMatch}
\alias{gtkEntryCompletionSetTextColumn}
\alias{gtkLabelNew}
\alias{gtkLabelNewWithMnemonic}
\alias{gtkLabelGetAttributes}
\alias{gtkLabelGetCurrentUri}
\alias{gtkLabelGetEllipsize}
\alias{gtkLabelGetExtraMenu}
\alias{gtkLabelGetJustify}
\alias{gtkLabelGetLabel}
\alias{gtkLabelGetLayout}
\alias{gtkLabelGetLayoutOffsets}
\alias{gtkLabelGetLines}
\alias{gtkLabelGetMaxWidthChars}
\alias{gtkLabelGetMnemonicKeyval}
\alias{gtkLabelGetMnemonicWidget}
\alias{gtkLabelGetNaturalWrapMode}
\alias{gtkLabelGetSelectable}
\alias{gtkLabelGetSelectionBounds}
\alias{gtkLabelGetSingleLineMode}
\alias{gtkLabelGetTabs}
\alias{gtkLabelGetText}
\alias{gtkLabelGetUseMarkup}
\alias{gtkLabelGetUseUnderline}
\alias{gtkLabelGetWidthChars}
\alias{gtkLabelGetWrap}
\alias{gtkLabelGetWrapMode}
\alias{gtkLabelGetXalign}
\alias{gtkLabelGetYalign}
\alias{gtkLabelSelectRegion}
\alias{gtkLabelSetAttributes}
\alias{gtkLabelSetEllipsize}
\alias{gtkLabelSetExtraMenu}
\alias{gtkLabelSetJustify}
\alias{gtkLabelSetLabel}
\alias{gtkLabelSetLines}
\alias{gtkLabelSetMarkup}
\alias{gtkLabelSetMarkupWithMnemonic}
\alias{gtkLabelSetMaxWidthChars}
\alias{gtkLabelSetMnemonicWidget}
\alias{gtkLabelSetNaturalWrapMode}
\alias{gtkLabelSetSelectable}
\alias{gtkLabelSetSingleLineMode}
\alias{gtkLabelSetTabs}
\alias{gtkLabelSetText}
\alias{gtkLabelSetTextWithMnemonic}
\alias{gtkLabelSetUseMarkup}
\alias{gtkLabelSetUseUnderline}
\alias{gtkLabelSetWidthChars}
\alias{gtkLabelSetWrap}
\alias{gtkLabelSetWrapMode}
\alias{gtkLabelSetXalign}
\alias{gtkLabelSetYalign}
\alias{gtkTextNew}
\alias{gtkTextNewWithBuffer}
\alias{gtkTextComputeCursorExtents}
\alias{gtkTextGetActivatesDefault}
\alias{gtkTextGetAttributes}
\alias{gtkTextGetBuffer}
\alias{gtkTextGetEnableEmojiCompletion}
\alias{gtkTextGetExtraMenu}
\alias{gtkTextGetInputHints}
\alias{gtkTextGetInputPurpose}
\alias{gtkTextGetInvisibleChar}
\alias{gtkTextGetMaxLength}
\alias{gtkTextGetOverwriteMode}
\alias{gtkTextGetPlaceholderText}
\alias{gtkTextGetPropagateTextWidth}
\alias{gtkTextGetTabs}
\alias{gtkTextGetTextLength}
\alias{gtkTextGetTruncateMultiline}
\alias{gtkTextGetVisibility}
\alias{gtkTextGrabFocusWithoutSelecting}
\alias{gtkTextSetActivatesDefault}
\alias{gtkTextSetAttributes}
\alias{gtkTextSetBuffer}
\alias{gtkTextSetEnableEmojiCompletion}
\alias{gtkTextSetExtraMenu}
\alias{gtkTextSetInputHints}
\alias{gtkTextSetInputPurpose}
\alias{gtkTextSetInvisibleChar}
\alias{gtkTextSetMaxLength}
\alias{gtkTextSetOverwriteMode}
\alias{gtkTextSetPlaceholderText}
\alias{gtkTextSetPropagateTextWidth}
\alias{gtkTextSetTabs}
\alias{gtkTextSetTruncateMultiline}
\alias{gtkTextSetVisibility}
\alias{gtkTextUnsetInvisibleChar}
\alias{gtkTextBufferNew}
\alias{gtkTextBufferAddMark}
\alias{gtkTextBufferAddSelectionClipboard}
\alias{gtkTextBufferApplyTag}
\alias{gtkTextBufferApplyTagByName}
\alias{gtkTextBufferBackspace}
\alias{gtkTextBufferBeginIrreversibleAction}
\alias{gtkTextBufferBeginUserAction}
\alias{gtkTextBufferCopyClipboard}
\alias{gtkTextBufferCreateChildAnchor}
\alias{gtkTextBufferCreateMark}
\alias{gtkTextBufferCutClipboard}
\alias{gtkTextBufferDelete}
\alias{gtkTextBufferDeleteInteractive}
\alias{gtkTextBufferDeleteMark}
\alias{gtkTextBufferDeleteMarkByName}
\alias{gtkTextBufferDeleteSelection}
\alias{gtkTextBufferEndIrreversibleAction}
\alias{gtkTextBufferEndUserAction}
\alias{gtkTextBufferGetBounds}
\alias{gtkTextBufferGetCanRedo}
\alias{gtkTextBufferGetCanUndo}
\alias{gtkTextBufferGetCharCount}
\alias{gtkTextBufferGetEnableUndo}
\alias{gtkTextBufferGetEndIter}
\alias{gtkTextBufferGetHasSelection}
\alias{gtkTextBufferGetInsert}
\alias{gtkTextBufferGetIterAtChildAnchor}
\alias{gtkTextBufferGetIterAtLine}
\alias{gtkTextBufferGetIterAtLineIndex}
\alias{gtkTextBufferGetIterAtLineOffset}
\alias{gtkTextBufferGetIterAtMark}
\alias{gtkTextBufferGetIterAtOffset}
\alias{gtkTextBufferGetLineCount}
\alias{gtkTextBufferGetMark}
\alias{gtkTextBufferGetMaxUndoLevels}
\alias{gtkTextBufferGetModified}
\alias{gtkTextBufferGetSelectionBound}
\alias{gtkTextBufferGetSelectionBounds}
\alias{gtkTextBufferGetSelectionContent}
\alias{gtkTextBufferGetSlice}
\alias{gtkTextBufferGetStartIter}
\alias{gtkTextBufferGetTagTable}
\alias{gtkTextBufferGetText}
\alias{gtkTextBufferInsert}
\alias{gtkTextBufferInsertAtCursor}
\alias{gtkTextBufferInsertChildAnchor}
\alias{gtkTextBufferInsertInteractive}
\alias{gtkTextBufferInsertInteractiveAtCursor}
\alias{gtkTextBufferInsertMarkup}
\alias{gtkTextBufferInsertPaintable}
\alias{gtkTextBufferInsertRange}
\alias{gtkTextBufferInsertRangeInteractive}
\alias{gtkTextBufferMoveMark}
\alias{gtkTextBufferMoveMarkByName}
\alias{gtkTextBufferPasteClipboard}
\alias{gtkTextBufferPlaceCursor}
\alias{gtkTextBufferRedo}
\alias{gtkTextBufferRemoveAllTags}
\alias{gtkTextBufferRemoveCommitNotify}
\alias{gtkTextBufferRemoveSelectionClipboard}
\alias{gtkTextBufferRemoveTag}
\alias{gtkTextBufferRemoveTagByName}
\alias{gtkTextBufferSelectRange}
\alias{gtkTextBufferSetEnableUndo}
\alias{gtkTextBufferSetMaxUndoLevels}
\alias{gtkTextBufferSetModified}
\alias{gtkTextBufferSetText}
\alias{gtkTextBufferUndo}
\alias{gtkTextChildAnchorNew}
\alias{gtkTextChildAnchorNewWithReplacement}
\alias{gtkTextChildAnchorGetDeleted}
\alias{gtkTextChildAnchorGetWidgets}
\alias{gtkTextIterAssign}
\alias{gtkTextIterBackwardChar}
\alias{gtkTextIterBackwardChars}
\alias{gtkTextIterBackwardCursorPosition}
\alias{gtkTextIterBackwardCursorPositions}
\alias{gtkTextIterBackwardFindChar}
\alias{gtkTextIterBackwardLine}
\alias{gtkTextIterBackwardLines}
\alias{gtkTextIterBackwardSearch}
\alias{gtkTextIterBackwardSentenceStart}
\alias{gtkTextIterBackwardSentenceStarts}
\alias{gtkTextIterBackwardToTagToggle}
\alias{gtkTextIterBackwardVisibleCursorPosition}
\alias{gtkTextIterBackwardVisibleCursorPositions}
\alias{gtkTextIterBackwardVisibleLine}
\alias{gtkTextIterBackwardVisibleLines}
\alias{gtkTextIterBackwardVisibleWordStart}
\alias{gtkTextIterBackwardVisibleWordStarts}
\alias{gtkTextIterBackwardWordStart}
\alias{gtkTextIterBackwardWordStarts}
\alias{gtkTextIterCanInsert}
\alias{gtkTextIterCompare}
\alias{gtkTextIterCopy}
\alias{gtkTextIterEditable}
\alias{gtkTextIterEndsLine}
\alias{gtkTextIterEndsSentence}
\alias{gtkTextIterEndsTag}
\alias{gtkTextIterEndsWord}
\alias{gtkTextIterEqual}
\alias{gtkTextIterForwardChar}
\alias{gtkTextIterForwardChars}
\alias{gtkTextIterForwardCursorPosition}
\alias{gtkTextIterForwardCursorPositions}
\alias{gtkTextIterForwardFindChar}
\alias{gtkTextIterForwardLine}
\alias{gtkTextIterForwardLines}
\alias{gtkTextIterForwardSearch}
\alias{gtkTextIterForwardSentenceEnd}
\alias{gtkTextIterForwardSentenceEnds}
\alias{gtkTextIterForwardToEnd}
\alias{gtkTextIterForwardToLineEnd}
\alias{gtkTextIterForwardToTagToggle}
\alias{gtkTextIterForwardVisibleCursorPosition}
\alias{gtkTextIterForwardVisibleCursorPositions}
\alias{gtkTextIterForwardVisibleLine}
\alias{gtkTextIterForwardVisibleLines}
\alias{gtkTextIterForwardVisibleWordEnd}
\alias{gtkTextIterForwardVisibleWordEnds}
\alias{gtkTextIterForwardWordEnd}
\alias{gtkTextIterForwardWordEnds}
\alias{gtkTextIterFree}
\alias{gtkTextIterGetBuffer}
\alias{gtkTextIterGetBytesInLine}
\alias{gtkTextIterGetChar}
\alias{gtkTextIterGetCharsInLine}
\alias{gtkTextIterGetChildAnchor}
\alias{gtkTextIterGetLanguage}
\alias{gtkTextIterGetLine}
\alias{gtkTextIterGetLineIndex}
\alias{gtkTextIterGetLineOffset}
\alias{gtkTextIterGetMarks}
\alias{gtkTextIterGetOffset}
\alias{gtkTextIterGetPaintable}
\alias{gtkTextIterGetSlice}
\alias{gtkTextIterGetTags}
\alias{gtkTextIterGetText}
\alias{gtkTextIterGetToggledTags}
\alias{gtkTextIterGetVisibleLineIndex}
\alias{gtkTextIterGetVisibleLineOffset}
\alias{gtkTextIterGetVisibleSlice}
\alias{gtkTextIterGetVisibleText}
\alias{gtkTextIterHasTag}
\alias{gtkTextIterInRange}
\alias{gtkTextIterInsideSentence}
\alias{gtkTextIterInsideWord}
\alias{gtkTextIterIsCursorPosition}
\alias{gtkTextIterIsEnd}
\alias{gtkTextIterIsStart}
\alias{gtkTextIterOrder}
\alias{gtkTextIterSetLine}
\alias{gtkTextIterSetLineIndex}
\alias{gtkTextIterSetLineOffset}
\alias{gtkTextIterSetOffset}
\alias{gtkTextIterSetVisibleLineIndex}
\alias{gtkTextIterSetVisibleLineOffset}
\alias{gtkTextIterStartsLine}
\alias{gtkTextIterStartsSentence}
\alias{gtkTextIterStartsTag}
\alias{gtkTextIterStartsWord}
\alias{gtkTextIterTogglesTag}
\alias{gtkTextMarkNew}
\alias{gtkTextMarkGetBuffer}
\alias{gtkTextMarkGetDeleted}
\alias{gtkTextMarkGetLeftGravity}
\alias{gtkTextMarkGetName}
\alias{gtkTextMarkGetVisible}
\alias{gtkTextMarkSetVisible}
\alias{gtkTextTagNew}
\alias{gtkTextTagChanged}
\alias{gtkTextTagGetPriority}
\alias{gtkTextTagSetPriority}
\alias{gtkTextTagTableNew}
\alias{gtkTextTagTableAdd}
\alias{gtkTextTagTableForeach}
\alias{gtkTextTagTableGetSize}
\alias{gtkTextTagTableLookup}
\alias{gtkTextTagTableRemove}
\alias{gtkTextViewNew}
\alias{gtkTextViewNewWithBuffer}
\alias{gtkTextViewAddChildAtAnchor}
\alias{gtkTextViewAddOverlay}
\alias{gtkTextViewBackwardDisplayLine}
\alias{gtkTextViewBackwardDisplayLineStart}
\alias{gtkTextViewBufferToWindowCoords}
\alias{gtkTextViewForwardDisplayLine}
\alias{gtkTextViewForwardDisplayLineEnd}
\alias{gtkTextViewGetAcceptsTab}
\alias{gtkTextViewGetBottomMargin}
\alias{gtkTextViewGetBuffer}
\alias{gtkTextViewGetCursorLocations}
\alias{gtkTextViewGetCursorVisible}
\alias{gtkTextViewGetEditable}
\alias{gtkTextViewGetExtraMenu}
\alias{gtkTextViewGetGutter}
\alias{gtkTextViewGetIndent}
\alias{gtkTextViewGetInputHints}
\alias{gtkTextViewGetInputPurpose}
\alias{gtkTextViewGetIterAtLocation}
\alias{gtkTextViewGetIterAtPosition}
\alias{gtkTextViewGetIterLocation}
\alias{gtkTextViewGetJustification}
\alias{gtkTextViewGetLeftMargin}
\alias{gtkTextViewGetLineAtY}
\alias{gtkTextViewGetLineYrange}
\alias{gtkTextViewGetLtrContext}
\alias{gtkTextViewGetMonospace}
\alias{gtkTextViewGetOverwrite}
\alias{gtkTextViewGetPixelsAboveLines}
\alias{gtkTextViewGetPixelsBelowLines}
\alias{gtkTextViewGetPixelsInsideWrap}
\alias{gtkTextViewGetRightMargin}
\alias{gtkTextViewGetRtlContext}
\alias{gtkTextViewGetTabs}
\alias{gtkTextViewGetTopMargin}
\alias{gtkTextViewGetVisibleOffset}
\alias{gtkTextViewGetVisibleRect}
\alias{gtkTextViewGetWrapMode}
\alias{gtkTextViewImContextFilterKeypress}
\alias{gtkTextViewMoveMarkOnscreen}
\alias{gtkTextViewMoveOverlay}
\alias{gtkTextViewMoveVisually}
\alias{gtkTextViewPlaceCursorOnscreen}
\alias{gtkTextViewRemove}
\alias{gtkTextViewResetCursorBlink}
\alias{gtkTextViewResetImContext}
\alias{gtkTextViewScrollMarkOnscreen}
\alias{gtkTextViewScrollToIter}
\alias{gtkTextViewScrollToMark}
\alias{gtkTextViewSetAcceptsTab}
\alias{gtkTextViewSetBottomMargin}
\alias{gtkTextViewSetBuffer}
\alias{gtkTextViewSetCursorVisible}
\alias{gtkTextViewSetEditable}
\alias{gtkTextViewSetExtraMenu}
\alias{gtkTextViewSetGutter}
\alias{gtkTextViewSetIndent}
\alias{gtkTextViewSetInputHints}
\alias{gtkTextViewSetInputPurpose}
\alias{gtkTextViewSetJustification}
\alias{gtkTextViewSetLeftMargin}
\alias{gtkTextViewSetMonospace}
\alias{gtkTextViewSetOverwrite}
\alias{gtkTextViewSetPixelsAboveLines}
\alias{gtkTextViewSetPixelsBelowLines}
\alias{gtkTextViewSetPixelsInsideWrap}
\alias{gtkTextViewSetRightMargin}
\alias{gtkTextViewSetTabs}
\alias{gtkTextViewSetTopMargin}
\alias{gtkTextViewSetWrapMode}
\alias{gtkTextViewStartsDisplayLine}
\alias{gtkTextViewWindowToBufferCoords}
\title{gtkEntryNew}
\usage{
gtkEntryNew()

gtkEntryNewWithBuffer(buffer)

gtkEntryGetActivatesDefault(entry)

gtkEntryGetAlignment(entry)

gtkEntryGetAttributes(entry)

gtkEntryGetBuffer(entry)

gtkEntryGetCompletion(entry)

gtkEntryGetCurrentIconDragSource(entry)

gtkEntryGetExtraMenu(entry)

gtkEntryGetHasFrame(entry)

gtkEntryGetIconActivatable(entry, icon_pos)

gtkEntryGetIconArea(entry, icon_pos)

gtkEntryGetIconAtPos(entry, x, y)

gtkEntryGetIconGicon(entry, icon_pos)

gtkEntryGetIconName(entry, icon_pos)

gtkEntryGetIconPaintable(entry, icon_pos)

gtkEntryGetIconSensitive(entry, icon_pos)

gtkEntryGetIconStorageType(entry, icon_pos)

gtkEntryGetIconTooltipMarkup(entry, icon_pos)

gtkEntryGetIconTooltipText(entry, icon_pos)

gtkEntryGetInputHints(entry)

gtkEntryGetInputPurpose(entry)

gtkEntryGetInvisibleChar(entry)

gtkEntryGetMaxLength(entry)

gtkEntryGetMenuEntryIconText(entry, icon_pos)

gtkEntryGetOverwriteMode(entry)

gtkEntryGetPlaceholderText(entry)

gtkEntryGetProgressFraction(entry)

gtkEntryGetProgressPulseStep(entry)

gtkEntryGetTabs(entry)

gtkEntryGetTextLength(entry)

gtkEntryGetVisibility(entry)

gtkEntryGrabFocusWithoutSelecting(entry)

gtkEntryProgressPulse(entry)

gtkEntryResetImContext(entry)

gtkEntrySetActivatesDefault(entry, setting)

gtkEntrySetAlignment(entry, xalign)

gtkEntrySetAttributes(entry, attrs)

gtkEntrySetBuffer(entry, buffer)

gtkEntrySetCompletion(entry, completion)

gtkEntrySetExtraMenu(entry, model)

gtkEntrySetHasFrame(entry, setting)

gtkEntrySetIconActivatable(entry, icon_pos, activatable)

gtkEntrySetIconDragSource(entry, icon_pos, provider, actions)

gtkEntrySetIconFromGicon(entry, icon_pos, icon)

gtkEntrySetIconFromIconName(entry, icon_pos, icon_name)

gtkEntrySetIconFromPaintable(entry, icon_pos, paintable)

gtkEntrySetIconSensitive(entry, icon_pos, sensitive)

gtkEntrySetIconTooltipMarkup(entry, icon_pos, tooltip)

gtkEntrySetIconTooltipText(entry, icon_pos, tooltip)

gtkEntrySetInputHints(entry, hints)

gtkEntrySetInputPurpose(entry, purpose)

gtkEntrySetInvisibleChar(entry, ch)

gtkEntrySetMaxLength(entry, max)

gtkEntrySetMenuEntryIconText(entry, icon_pos, text)

gtkEntrySetOverwriteMode(entry, overwrite)

gtkEntrySetPlaceholderText(entry, text)

gtkEntrySetProgressFraction(entry, fraction)

gtkEntrySetProgressPulseStep(entry, fraction)

gtkEntrySetTabs(entry, tabs)

gtkEntrySetVisibility(entry, visible)

gtkEntryUnsetInvisibleChar(entry)

gtkEntryBufferNew(initial_chars, n_initial_chars)

gtkEntryBufferDeleteText(buffer, position, n_chars)

gtkEntryBufferEmitDeletedText(buffer, position, n_chars)

gtkEntryBufferEmitInsertedText(buffer, position, chars, n_chars)

gtkEntryBufferGetBytes(buffer)

gtkEntryBufferGetLength(buffer)

gtkEntryBufferGetMaxLength(buffer)

gtkEntryBufferGetText(buffer)

gtkEntryBufferInsertText(buffer, position, chars, n_chars)

gtkEntryBufferSetMaxLength(buffer, max_length)

gtkEntryBufferSetText(buffer, chars, n_chars)

gtkEntryCompletionNew()

gtkEntryCompletionNewWithArea(area)

gtkEntryCompletionComplete(completion)

gtkEntryCompletionComputePrefix(completion, key)

gtkEntryCompletionGetCompletionPrefix(completion)

gtkEntryCompletionGetEntry(completion)

gtkEntryCompletionGetInlineCompletion(completion)

gtkEntryCompletionGetInlineSelection(completion)

gtkEntryCompletionGetMinimumKeyLength(completion)

gtkEntryCompletionGetModel(completion)

gtkEntryCompletionGetPopupCompletion(completion)

gtkEntryCompletionGetPopupSetWidth(completion)

gtkEntryCompletionGetPopupSingleMatch(completion)

gtkEntryCompletionGetTextColumn(completion)

gtkEntryCompletionInsertPrefix(completion)

gtkEntryCompletionSetInlineCompletion(completion, inline_completion)

gtkEntryCompletionSetInlineSelection(completion, inline_selection)

gtkEntryCompletionSetMinimumKeyLength(completion, length)

gtkEntryCompletionSetModel(completion, model)

gtkEntryCompletionSetPopupCompletion(completion, popup_completion)

gtkEntryCompletionSetPopupSetWidth(completion, popup_set_width)

gtkEntryCompletionSetPopupSingleMatch(completion, popup_single_match)

gtkEntryCompletionSetTextColumn(completion, column)

gtkLabelNew(str)

gtkLabelNewWithMnemonic(str)

gtkLabelGetAttributes(self)

gtkLabelGetCurrentUri(self)

gtkLabelGetEllipsize(self)

gtkLabelGetExtraMenu(self)

gtkLabelGetJustify(self)

gtkLabelGetLabel(self)

gtkLabelGetLayout(self)

gtkLabelGetLayoutOffsets(self)

gtkLabelGetLines(self)

gtkLabelGetMaxWidthChars(self)

gtkLabelGetMnemonicKeyval(self)

gtkLabelGetMnemonicWidget(self)

gtkLabelGetNaturalWrapMode(self)

gtkLabelGetSelectable(self)

gtkLabelGetSelectionBounds(self)

gtkLabelGetSingleLineMode(self)

gtkLabelGetTabs(self)

gtkLabelGetText(self)

gtkLabelGetUseMarkup(self)

gtkLabelGetUseUnderline(self)

gtkLabelGetWidthChars(self)

gtkLabelGetWrap(self)

gtkLabelGetWrapMode(self)

gtkLabelGetXalign(self)

gtkLabelGetYalign(self)

gtkLabelSelectRegion(self, start_offset, end_offset)

gtkLabelSetAttributes(self, attrs)

gtkLabelSetEllipsize(self, mode)

gtkLabelSetExtraMenu(self, model)

gtkLabelSetJustify(self, jtype)

gtkLabelSetLabel(self, str)

gtkLabelSetLines(self, lines)

gtkLabelSetMarkup(self, str)

gtkLabelSetMarkupWithMnemonic(self, str)

gtkLabelSetMaxWidthChars(self, n_chars)

gtkLabelSetMnemonicWidget(self, widget)

gtkLabelSetNaturalWrapMode(self, wrap_mode)

gtkLabelSetSelectable(self, setting)

gtkLabelSetSingleLineMode(self, single_line_mode)

gtkLabelSetTabs(self, tabs)

gtkLabelSetText(self, str)

gtkLabelSetTextWithMnemonic(self, str)

gtkLabelSetUseMarkup(self, setting)

gtkLabelSetUseUnderline(self, setting)

gtkLabelSetWidthChars(self, n_chars)

gtkLabelSetWrap(self, wrap)

gtkLabelSetWrapMode(self, wrap_mode)

gtkLabelSetXalign(self, xalign)

gtkLabelSetYalign(self, yalign)

gtkTextNew()

gtkTextNewWithBuffer(buffer)

gtkTextComputeCursorExtents(self, position)

gtkTextGetActivatesDefault(self)

gtkTextGetAttributes(self)

gtkTextGetBuffer(self)

gtkTextGetEnableEmojiCompletion(self)

gtkTextGetExtraMenu(self)

gtkTextGetInputHints(self)

gtkTextGetInputPurpose(self)

gtkTextGetInvisibleChar(self)

gtkTextGetMaxLength(self)

gtkTextGetOverwriteMode(self)

gtkTextGetPlaceholderText(self)

gtkTextGetPropagateTextWidth(self)

gtkTextGetTabs(self)

gtkTextGetTextLength(self)

gtkTextGetTruncateMultiline(self)

gtkTextGetVisibility(self)

gtkTextGrabFocusWithoutSelecting(self)

gtkTextSetActivatesDefault(self, activates)

gtkTextSetAttributes(self, attrs)

gtkTextSetBuffer(self, buffer)

gtkTextSetEnableEmojiCompletion(self, enable_emoji_completion)

gtkTextSetExtraMenu(self, model)

gtkTextSetInputHints(self, hints)

gtkTextSetInputPurpose(self, purpose)

gtkTextSetInvisibleChar(self, ch)

gtkTextSetMaxLength(self, length)

gtkTextSetOverwriteMode(self, overwrite)

gtkTextSetPlaceholderText(self, text)

gtkTextSetPropagateTextWidth(self, propagate_text_width)

gtkTextSetTabs(self, tabs)

gtkTextSetTruncateMultiline(self, truncate_multiline)

gtkTextSetVisibility(self, visible)

gtkTextUnsetInvisibleChar(self)

gtkTextBufferNew(table)

gtkTextBufferAddMark(buffer, mark, where)

gtkTextBufferAddSelectionClipboard(buffer, clipboard)

gtkTextBufferApplyTag(buffer, tag, start, end)

gtkTextBufferApplyTagByName(buffer, name, start, end)

gtkTextBufferBackspace(buffer, iter, interactive, default_editable)

gtkTextBufferBeginIrreversibleAction(buffer)

gtkTextBufferBeginUserAction(buffer)

gtkTextBufferCopyClipboard(buffer, clipboard)

gtkTextBufferCreateChildAnchor(buffer, iter)

gtkTextBufferCreateMark(buffer, mark_name, where, left_gravity)

gtkTextBufferCutClipboard(buffer, clipboard, default_editable)

gtkTextBufferDelete(buffer, start, end)

gtkTextBufferDeleteInteractive(buffer, start_iter, end_iter, default_editable)

gtkTextBufferDeleteMark(buffer, mark)

gtkTextBufferDeleteMarkByName(buffer, name)

gtkTextBufferDeleteSelection(buffer, interactive, default_editable)

gtkTextBufferEndIrreversibleAction(buffer)

gtkTextBufferEndUserAction(buffer)

gtkTextBufferGetBounds(buffer)

gtkTextBufferGetCanRedo(buffer)

gtkTextBufferGetCanUndo(buffer)

gtkTextBufferGetCharCount(buffer)

gtkTextBufferGetEnableUndo(buffer)

gtkTextBufferGetEndIter(buffer)

gtkTextBufferGetHasSelection(buffer)

gtkTextBufferGetInsert(buffer)

gtkTextBufferGetIterAtChildAnchor(buffer, anchor)

gtkTextBufferGetIterAtLine(buffer, line_number)

gtkTextBufferGetIterAtLineIndex(buffer, line_number, byte_index)

gtkTextBufferGetIterAtLineOffset(buffer, line_number, char_offset)

gtkTextBufferGetIterAtMark(buffer, mark)

gtkTextBufferGetIterAtOffset(buffer, char_offset)

gtkTextBufferGetLineCount(buffer)

gtkTextBufferGetMark(buffer, name)

gtkTextBufferGetMaxUndoLevels(buffer)

gtkTextBufferGetModified(buffer)

gtkTextBufferGetSelectionBound(buffer)

gtkTextBufferGetSelectionBounds(buffer)

gtkTextBufferGetSelectionContent(buffer)

gtkTextBufferGetSlice(buffer, start, end, include_hidden_chars)

gtkTextBufferGetStartIter(buffer)

gtkTextBufferGetTagTable(buffer)

gtkTextBufferGetText(buffer, start, end, include_hidden_chars)

gtkTextBufferInsert(buffer, iter, text, len)

gtkTextBufferInsertAtCursor(buffer, text, len)

gtkTextBufferInsertChildAnchor(buffer, iter, anchor)

gtkTextBufferInsertInteractive(buffer, iter, text, len, default_editable)

gtkTextBufferInsertInteractiveAtCursor(buffer, text, len, default_editable)

gtkTextBufferInsertMarkup(buffer, iter, markup, len)

gtkTextBufferInsertPaintable(buffer, iter, paintable)

gtkTextBufferInsertRange(buffer, iter, start, end)

gtkTextBufferInsertRangeInteractive(buffer, iter, start, end, default_editable)

gtkTextBufferMoveMark(buffer, mark, where)

gtkTextBufferMoveMarkByName(buffer, name, where)

gtkTextBufferPasteClipboard(
  buffer,
  clipboard,
  override_location,
  default_editable
)

gtkTextBufferPlaceCursor(buffer, where)

gtkTextBufferRedo(buffer)

gtkTextBufferRemoveAllTags(buffer, start, end)

gtkTextBufferRemoveCommitNotify(buffer, commit_notify_handler)

gtkTextBufferRemoveSelectionClipboard(buffer, clipboard)

gtkTextBufferRemoveTag(buffer, tag, start, end)

gtkTextBufferRemoveTagByName(buffer, name, start, end)

gtkTextBufferSelectRange(buffer, ins, bound)

gtkTextBufferSetEnableUndo(buffer, enable_undo)

gtkTextBufferSetMaxUndoLevels(buffer, max_undo_levels)

gtkTextBufferSetModified(buffer, setting)

gtkTextBufferSetText(buffer, text, len)

gtkTextBufferUndo(buffer)

gtkTextChildAnchorNew()

gtkTextChildAnchorNewWithReplacement(character)

gtkTextChildAnchorGetDeleted(anchor)

gtkTextChildAnchorGetWidgets(anchor)

gtkTextIterAssign(iter, other)

gtkTextIterBackwardChar(iter)

gtkTextIterBackwardChars(iter, count)

gtkTextIterBackwardCursorPosition(iter)

gtkTextIterBackwardCursorPositions(iter, count)

gtkTextIterBackwardFindChar(iter, pred, user_data, limit)

gtkTextIterBackwardLine(iter)

gtkTextIterBackwardLines(iter, count)

gtkTextIterBackwardSearch(iter, str, flags, limit)

gtkTextIterBackwardSentenceStart(iter)

gtkTextIterBackwardSentenceStarts(iter, count)

gtkTextIterBackwardToTagToggle(iter, tag)

gtkTextIterBackwardVisibleCursorPosition(iter)

gtkTextIterBackwardVisibleCursorPositions(iter, count)

gtkTextIterBackwardVisibleLine(iter)

gtkTextIterBackwardVisibleLines(iter, count)

gtkTextIterBackwardVisibleWordStart(iter)

gtkTextIterBackwardVisibleWordStarts(iter, count)

gtkTextIterBackwardWordStart(iter)

gtkTextIterBackwardWordStarts(iter, count)

gtkTextIterCanInsert(iter, default_editability)

gtkTextIterCompare(lhs, rhs)

gtkTextIterCopy(iter)

gtkTextIterEditable(iter, default_setting)

gtkTextIterEndsLine(iter)

gtkTextIterEndsSentence(iter)

gtkTextIterEndsTag(iter, tag)

gtkTextIterEndsWord(iter)

gtkTextIterEqual(lhs, rhs)

gtkTextIterForwardChar(iter)

gtkTextIterForwardChars(iter, count)

gtkTextIterForwardCursorPosition(iter)

gtkTextIterForwardCursorPositions(iter, count)

gtkTextIterForwardFindChar(iter, pred, user_data, limit)

gtkTextIterForwardLine(iter)

gtkTextIterForwardLines(iter, count)

gtkTextIterForwardSearch(iter, str, flags, limit)

gtkTextIterForwardSentenceEnd(iter)

gtkTextIterForwardSentenceEnds(iter, count)

gtkTextIterForwardToEnd(iter)

gtkTextIterForwardToLineEnd(iter)

gtkTextIterForwardToTagToggle(iter, tag)

gtkTextIterForwardVisibleCursorPosition(iter)

gtkTextIterForwardVisibleCursorPositions(iter, count)

gtkTextIterForwardVisibleLine(iter)

gtkTextIterForwardVisibleLines(iter, count)

gtkTextIterForwardVisibleWordEnd(iter)

gtkTextIterForwardVisibleWordEnds(iter, count)

gtkTextIterForwardWordEnd(iter)

gtkTextIterForwardWordEnds(iter, count)

gtkTextIterFree(iter)

gtkTextIterGetBuffer(iter)

gtkTextIterGetBytesInLine(iter)

gtkTextIterGetChar(iter)

gtkTextIterGetCharsInLine(iter)

gtkTextIterGetChildAnchor(iter)

gtkTextIterGetLanguage(iter)

gtkTextIterGetLine(iter)

gtkTextIterGetLineIndex(iter)

gtkTextIterGetLineOffset(iter)

gtkTextIterGetMarks(iter)

gtkTextIterGetOffset(iter)

gtkTextIterGetPaintable(iter)

gtkTextIterGetSlice(start, end)

gtkTextIterGetTags(iter)

gtkTextIterGetText(start, end)

gtkTextIterGetToggledTags(iter, toggled_on)

gtkTextIterGetVisibleLineIndex(iter)

gtkTextIterGetVisibleLineOffset(iter)

gtkTextIterGetVisibleSlice(start, end)

gtkTextIterGetVisibleText(start, end)

gtkTextIterHasTag(iter, tag)

gtkTextIterInRange(iter, start, end)

gtkTextIterInsideSentence(iter)

gtkTextIterInsideWord(iter)

gtkTextIterIsCursorPosition(iter)

gtkTextIterIsEnd(iter)

gtkTextIterIsStart(iter)

gtkTextIterOrder(first, second)

gtkTextIterSetLine(iter, line_number)

gtkTextIterSetLineIndex(iter, byte_on_line)

gtkTextIterSetLineOffset(iter, char_on_line)

gtkTextIterSetOffset(iter, char_offset)

gtkTextIterSetVisibleLineIndex(iter, byte_on_line)

gtkTextIterSetVisibleLineOffset(iter, char_on_line)

gtkTextIterStartsLine(iter)

gtkTextIterStartsSentence(iter)

gtkTextIterStartsTag(iter, tag)

gtkTextIterStartsWord(iter)

gtkTextIterTogglesTag(iter, tag)

gtkTextMarkNew(name, left_gravity)

gtkTextMarkGetBuffer(mark)

gtkTextMarkGetDeleted(mark)

gtkTextMarkGetLeftGravity(mark)

gtkTextMarkGetName(mark)

gtkTextMarkGetVisible(mark)

gtkTextMarkSetVisible(mark, setting)

gtkTextTagNew(name)

gtkTextTagChanged(tag, size_changed)

gtkTextTagGetPriority(tag)

gtkTextTagSetPriority(tag, priority)

gtkTextTagTableNew()

gtkTextTagTableAdd(table, tag)

gtkTextTagTableForeach(table, func, data)

gtkTextTagTableGetSize(table)

gtkTextTagTableLookup(table, name)

gtkTextTagTableRemove(table, tag)

gtkTextViewNew()

gtkTextViewNewWithBuffer(buffer)

gtkTextViewAddChildAtAnchor(text_view, child, anchor)

gtkTextViewAddOverlay(text_view, child, xpos, ypos)

gtkTextViewBackwardDisplayLine(text_view, iter)

gtkTextViewBackwardDisplayLineStart(text_view, iter)

gtkTextViewBufferToWindowCoords(text_view, win, buffer_x, buffer_y)

gtkTextViewForwardDisplayLine(text_view, iter)

gtkTextViewForwardDisplayLineEnd(text_view, iter)

gtkTextViewGetAcceptsTab(text_view)

gtkTextViewGetBottomMargin(text_view)

gtkTextViewGetBuffer(text_view)

gtkTextViewGetCursorLocations(text_view, iter)

gtkTextViewGetCursorVisible(text_view)

gtkTextViewGetEditable(text_view)

gtkTextViewGetExtraMenu(text_view)

gtkTextViewGetGutter(text_view, win)

gtkTextViewGetIndent(text_view)

gtkTextViewGetInputHints(text_view)

gtkTextViewGetInputPurpose(text_view)

gtkTextViewGetIterAtLocation(text_view, x, y)

gtkTextViewGetIterAtPosition(text_view, x, y)

gtkTextViewGetIterLocation(text_view, iter)

gtkTextViewGetJustification(text_view)

gtkTextViewGetLeftMargin(text_view)

gtkTextViewGetLineAtY(text_view, y)

gtkTextViewGetLineYrange(text_view, iter)

gtkTextViewGetLtrContext(text_view)

gtkTextViewGetMonospace(text_view)

gtkTextViewGetOverwrite(text_view)

gtkTextViewGetPixelsAboveLines(text_view)

gtkTextViewGetPixelsBelowLines(text_view)

gtkTextViewGetPixelsInsideWrap(text_view)

gtkTextViewGetRightMargin(text_view)

gtkTextViewGetRtlContext(text_view)

gtkTextViewGetTabs(text_view)

gtkTextViewGetTopMargin(text_view)

gtkTextViewGetVisibleOffset(text_view)

gtkTextViewGetVisibleRect(text_view)

gtkTextViewGetWrapMode(text_view)

gtkTextViewImContextFilterKeypress(text_view, event)

gtkTextViewMoveMarkOnscreen(text_view, mark)

gtkTextViewMoveOverlay(text_view, child, xpos, ypos)

gtkTextViewMoveVisually(text_view, iter, count)

gtkTextViewPlaceCursorOnscreen(text_view)

gtkTextViewRemove(text_view, child)

gtkTextViewResetCursorBlink(text_view)

gtkTextViewResetImContext(text_view)

gtkTextViewScrollMarkOnscreen(text_view, mark)

gtkTextViewScrollToIter(
  text_view,
  iter,
  within_margin,
  use_align,
  xalign,
  yalign
)

gtkTextViewScrollToMark(
  text_view,
  mark,
  within_margin,
  use_align,
  xalign,
  yalign
)

gtkTextViewSetAcceptsTab(text_view, accepts_tab)

gtkTextViewSetBottomMargin(text_view, bottom_margin)

gtkTextViewSetBuffer(text_view, buffer)

gtkTextViewSetCursorVisible(text_view, setting)

gtkTextViewSetEditable(text_view, setting)

gtkTextViewSetExtraMenu(text_view, model)

gtkTextViewSetGutter(text_view, win, widget)

gtkTextViewSetIndent(text_view, indent)

gtkTextViewSetInputHints(text_view, hints)

gtkTextViewSetInputPurpose(text_view, purpose)

gtkTextViewSetJustification(text_view, justification)

gtkTextViewSetLeftMargin(text_view, left_margin)

gtkTextViewSetMonospace(text_view, monospace)

gtkTextViewSetOverwrite(text_view, overwrite)

gtkTextViewSetPixelsAboveLines(text_view, pixels_above_lines)

gtkTextViewSetPixelsBelowLines(text_view, pixels_below_lines)

gtkTextViewSetPixelsInsideWrap(text_view, pixels_inside_wrap)

gtkTextViewSetRightMargin(text_view, right_margin)

gtkTextViewSetTabs(text_view, tabs)

gtkTextViewSetTopMargin(text_view, top_margin)

gtkTextViewSetWrapMode(text_view, wrap_mode)

gtkTextViewStartsDisplayLine(text_view, iter)

gtkTextViewWindowToBufferCoords(text_view, win, window_x, window_y)
}
\arguments{
\item{buffer}{TextBuffer}

\item{entry}{Entry}

\item{icon_pos}{EntryIconPosition}

\item{x}{gint}

\item{y}{gint}

\item{setting}{gboolean}

\item{xalign}{gdouble}

\item{attrs}{Pango.AttrList}

\item{completion}{EntryCompletion}

\item{model}{Gio.MenuModel}

\item{activatable}{gboolean}

\item{provider}{Gdk.ContentProvider}

\item{actions}{Gdk.DragAction}

\item{icon}{Gio.Icon}

\item{icon_name}{utf8}

\item{paintable}{Gdk.Paintable}

\item{sensitive}{gboolean}

\item{tooltip}{utf8}

\item{hints}{InputHints}

\item{purpose}{InputPurpose}

\item{ch}{gunichar}

\item{max}{gint}

\item{text}{utf8}

\item{overwrite}{gboolean}

\item{fraction}{gdouble}

\item{tabs}{Pango.TabArray}

\item{visible}{gboolean}

\item{initial_chars}{utf8}

\item{n_initial_chars}{gint}

\item{position}{gsize}

\item{n_chars}{gint}

\item{chars}{utf8}

\item{max_length}{gint}

\item{area}{CellArea}

\item{key}{utf8}

\item{inline_completion}{gboolean}

\item{inline_selection}{gboolean}

\item{length}{gint}

\item{popup_completion}{gboolean}

\item{popup_set_width}{gboolean}

\item{popup_single_match}{gboolean}

\item{column}{gint}

\item{str}{utf8}

\item{self}{Text}

\item{start_offset}{gint}

\item{end_offset}{gint}

\item{mode}{Pango.EllipsizeMode}

\item{jtype}{Justification}

\item{lines}{gint}

\item{widget}{Widget}

\item{wrap_mode}{WrapMode}

\item{single_line_mode}{gboolean}

\item{wrap}{gboolean}

\item{yalign}{gdouble}

\item{activates}{gboolean}

\item{enable_emoji_completion}{gboolean}

\item{propagate_text_width}{gboolean}

\item{truncate_multiline}{gboolean}

\item{table}{TextTagTable}

\item{mark}{TextMark}

\item{where}{TextIter}

\item{clipboard}{Gdk.Clipboard}

\item{tag}{TextTag}

\item{start}{TextIter}

\item{end}{TextIter}

\item{name}{utf8}

\item{iter}{TextIter}

\item{interactive}{gboolean}

\item{default_editable}{gboolean}

\item{mark_name}{utf8}

\item{left_gravity}{gboolean}

\item{start_iter}{TextIter}

\item{end_iter}{TextIter}

\item{anchor}{TextChildAnchor}

\item{line_number}{gint}

\item{byte_index}{gint}

\item{char_offset}{gint}

\item{include_hidden_chars}{gboolean}

\item{len}{gint}

\item{markup}{utf8}

\item{override_location}{TextIter}

\item{commit_notify_handler}{guint}

\item{ins}{TextIter}

\item{bound}{TextIter}

\item{enable_undo}{gboolean}

\item{max_undo_levels}{guint}

\item{character}{utf8}

\item{other}{TextIter}

\item{count}{gint}

\item{pred}{TextCharPredicate}

\item{user_data}{gpointer}

\item{limit}{TextIter}

\item{flags}{TextSearchFlags}

\item{default_editability}{gboolean}

\item{lhs}{TextIter}

\item{rhs}{TextIter}

\item{default_setting}{gboolean}

\item{toggled_on}{gboolean}

\item{first}{TextIter}

\item{second}{TextIter}

\item{byte_on_line}{gint}

\item{char_on_line}{gint}

\item{size_changed}{gboolean}

\item{priority}{gint}

\item{func}{TextTagTableForeach}

\item{data}{gpointer}

\item{text_view}{TextView}

\item{child}{Widget}

\item{xpos}{gint}

\item{ypos}{gint}

\item{win}{TextWindowType}

\item{buffer_x}{gint}

\item{buffer_y}{gint}

\item{event}{Gdk.Event}

\item{within_margin}{gdouble}

\item{use_align}{gboolean}

\item{accepts_tab}{gboolean}

\item{bottom_margin}{gint}

\item{indent}{gint}

\item{justification}{Justification}

\item{left_margin}{gint}

\item{monospace}{gboolean}

\item{pixels_above_lines}{gint}

\item{pixels_below_lines}{gint}

\item{pixels_inside_wrap}{gint}

\item{right_margin}{gint}

\item{top_margin}{gint}

\item{window_x}{gint}

\item{window_y}{gint}
}
\value{
Widget

Widget

gboolean

gfloat

Pango.AttrList

EntryBuffer

EntryCompletion

gint

Gio.MenuModel

gboolean

gboolean

Return value from C function

gint

Gio.Icon

utf8

Gdk.Paintable

gboolean

ImageType

utf8

utf8

InputHints

InputPurpose

gunichar

gint

utf8

gboolean

utf8

gdouble

gdouble

Pango.TabArray

guint16

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

EntryBuffer

guint

Return value from C function

Return value from C function

gsize

guint

gint

utf8

guint

Return value from C function

Return value from C function

EntryCompletion

EntryCompletion

Return value from C function

utf8

utf8

Widget

gboolean

gboolean

gint

TreeModel

gboolean

gboolean

gboolean

gint

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

Widget

Pango.AttrList

utf8

Pango.EllipsizeMode

Gio.MenuModel

Justification

utf8

Pango.Layout

Return value from C function

gint

gint

guint

Widget

NaturalWrapMode

gboolean

gboolean

gboolean

Pango.TabArray

utf8

gboolean

gboolean

gint

gboolean

Pango.WrapMode

gfloat

gfloat

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

Widget

Return value from C function

gboolean

Pango.AttrList

EntryBuffer

gboolean

Gio.MenuModel

InputHints

InputPurpose

gunichar

gint

gboolean

utf8

gboolean

Pango.TabArray

guint16

gboolean

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

TextBuffer

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

TextChildAnchor

TextMark

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

gboolean

gboolean

gint

gboolean

Return value from C function

gboolean

TextMark

Return value from C function

gboolean

gboolean

gboolean

Return value from C function

Return value from C function

gint

TextMark

guint

gboolean

TextMark

gboolean

Gdk.ContentProvider

utf8

Return value from C function

TextTagTable

utf8

Return value from C function

Return value from C function

Return value from C function

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

TextChildAnchor

TextChildAnchor

gboolean

Widget

Return value from C function

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gint

TextIter

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

Return value from C function

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

Return value from C function

TextBuffer

gint

gunichar

gint

TextChildAnchor

Pango.Language

gint

gint

gint

GLib.SList

gint

Gdk.Paintable

utf8

GLib.SList

utf8

GLib.SList

gint

gint

utf8

utf8

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

gboolean

gboolean

gboolean

gboolean

TextMark

TextBuffer

gboolean

gboolean

utf8

gboolean

Return value from C function

TextTag

Return value from C function

gint

Return value from C function

TextTagTable

gboolean

Return value from C function

gint

TextTag

Return value from C function

Widget

Widget

Return value from C function

Return value from C function

gboolean

gboolean

Return value from C function

gboolean

gboolean

gboolean

gint

TextBuffer

Return value from C function

gboolean

gboolean

Gio.MenuModel

Widget

gint

InputHints

InputPurpose

gboolean

gboolean

Return value from C function

Justification

gint

Return value from C function

Return value from C function

Pango.Context

gboolean

gboolean

gint

gint

gint

gint

Pango.Context

Pango.TabArray

gint

Return value from C function

Return value from C function

WrapMode

gboolean

gboolean

Return value from C function

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function
}
\description{
gtkEntryNew

gtkEntryNewWithBuffer

gtkEntryGetActivatesDefault

gtkEntryGetAlignment

gtkEntryGetAttributes

gtkEntryGetBuffer

gtkEntryGetCompletion

gtkEntryGetCurrentIconDragSource

gtkEntryGetExtraMenu

gtkEntryGetHasFrame

gtkEntryGetIconActivatable

gtkEntryGetIconArea

gtkEntryGetIconAtPos

gtkEntryGetIconGicon

gtkEntryGetIconName

gtkEntryGetIconPaintable

gtkEntryGetIconSensitive

gtkEntryGetIconStorageType

gtkEntryGetIconTooltipMarkup

gtkEntryGetIconTooltipText

gtkEntryGetInputHints

gtkEntryGetInputPurpose

gtkEntryGetInvisibleChar

gtkEntryGetMaxLength

gtkEntryGetMenuEntryIconText

gtkEntryGetOverwriteMode

gtkEntryGetPlaceholderText

gtkEntryGetProgressFraction

gtkEntryGetProgressPulseStep

gtkEntryGetTabs

gtkEntryGetTextLength

gtkEntryGetVisibility

gtkEntryGrabFocusWithoutSelecting

gtkEntryProgressPulse

gtkEntryResetImContext

gtkEntrySetActivatesDefault

gtkEntrySetAlignment

gtkEntrySetAttributes

gtkEntrySetBuffer

gtkEntrySetCompletion

gtkEntrySetExtraMenu

gtkEntrySetHasFrame

gtkEntrySetIconActivatable

gtkEntrySetIconDragSource

gtkEntrySetIconFromGicon

gtkEntrySetIconFromIconName

gtkEntrySetIconFromPaintable

gtkEntrySetIconSensitive

gtkEntrySetIconTooltipMarkup

gtkEntrySetIconTooltipText

gtkEntrySetInputHints

gtkEntrySetInputPurpose

gtkEntrySetInvisibleChar

gtkEntrySetMaxLength

gtkEntrySetMenuEntryIconText

gtkEntrySetOverwriteMode

gtkEntrySetPlaceholderText

gtkEntrySetProgressFraction

gtkEntrySetProgressPulseStep

gtkEntrySetTabs

gtkEntrySetVisibility

gtkEntryUnsetInvisibleChar

gtkEntryBufferNew

gtkEntryBufferDeleteText

gtkEntryBufferEmitDeletedText

gtkEntryBufferEmitInsertedText

gtkEntryBufferGetBytes

gtkEntryBufferGetLength

gtkEntryBufferGetMaxLength

gtkEntryBufferGetText

gtkEntryBufferInsertText

gtkEntryBufferSetMaxLength

gtkEntryBufferSetText

gtkEntryCompletionNew

gtkEntryCompletionNewWithArea

gtkEntryCompletionComplete

gtkEntryCompletionComputePrefix

gtkEntryCompletionGetCompletionPrefix

gtkEntryCompletionGetEntry

gtkEntryCompletionGetInlineCompletion

gtkEntryCompletionGetInlineSelection

gtkEntryCompletionGetMinimumKeyLength

gtkEntryCompletionGetModel

gtkEntryCompletionGetPopupCompletion

gtkEntryCompletionGetPopupSetWidth

gtkEntryCompletionGetPopupSingleMatch

gtkEntryCompletionGetTextColumn

gtkEntryCompletionInsertPrefix

gtkEntryCompletionSetInlineCompletion

gtkEntryCompletionSetInlineSelection

gtkEntryCompletionSetMinimumKeyLength

gtkEntryCompletionSetModel

gtkEntryCompletionSetPopupCompletion

gtkEntryCompletionSetPopupSetWidth

gtkEntryCompletionSetPopupSingleMatch

gtkEntryCompletionSetTextColumn

gtkLabelNew

gtkLabelNewWithMnemonic

gtkLabelGetAttributes

gtkLabelGetCurrentUri

gtkLabelGetEllipsize

gtkLabelGetExtraMenu

gtkLabelGetJustify

gtkLabelGetLabel

gtkLabelGetLayout

gtkLabelGetLayoutOffsets

gtkLabelGetLines

gtkLabelGetMaxWidthChars

gtkLabelGetMnemonicKeyval

gtkLabelGetMnemonicWidget

gtkLabelGetNaturalWrapMode

gtkLabelGetSelectable

gtkLabelGetSelectionBounds

gtkLabelGetSingleLineMode

gtkLabelGetTabs

gtkLabelGetText

gtkLabelGetUseMarkup

gtkLabelGetUseUnderline

gtkLabelGetWidthChars

gtkLabelGetWrap

gtkLabelGetWrapMode

gtkLabelGetXalign

gtkLabelGetYalign

gtkLabelSelectRegion

gtkLabelSetAttributes

gtkLabelSetEllipsize

gtkLabelSetExtraMenu

gtkLabelSetJustify

gtkLabelSetLabel

gtkLabelSetLines

gtkLabelSetMarkup

gtkLabelSetMarkupWithMnemonic

gtkLabelSetMaxWidthChars

gtkLabelSetMnemonicWidget

gtkLabelSetNaturalWrapMode

gtkLabelSetSelectable

gtkLabelSetSingleLineMode

gtkLabelSetTabs

gtkLabelSetText

gtkLabelSetTextWithMnemonic

gtkLabelSetUseMarkup

gtkLabelSetUseUnderline

gtkLabelSetWidthChars

gtkLabelSetWrap

gtkLabelSetWrapMode

gtkLabelSetXalign

gtkLabelSetYalign

gtkTextNew

gtkTextNewWithBuffer

gtkTextComputeCursorExtents

gtkTextGetActivatesDefault

gtkTextGetAttributes

gtkTextGetBuffer

gtkTextGetEnableEmojiCompletion

gtkTextGetExtraMenu

gtkTextGetInputHints

gtkTextGetInputPurpose

gtkTextGetInvisibleChar

gtkTextGetMaxLength

gtkTextGetOverwriteMode

gtkTextGetPlaceholderText

gtkTextGetPropagateTextWidth

gtkTextGetTabs

gtkTextGetTextLength

gtkTextGetTruncateMultiline

gtkTextGetVisibility

gtkTextGrabFocusWithoutSelecting

gtkTextSetActivatesDefault

gtkTextSetAttributes

gtkTextSetBuffer

gtkTextSetEnableEmojiCompletion

gtkTextSetExtraMenu

gtkTextSetInputHints

gtkTextSetInputPurpose

gtkTextSetInvisibleChar

gtkTextSetMaxLength

gtkTextSetOverwriteMode

gtkTextSetPlaceholderText

gtkTextSetPropagateTextWidth

gtkTextSetTabs

gtkTextSetTruncateMultiline

gtkTextSetVisibility

gtkTextUnsetInvisibleChar

gtkTextBufferNew

gtkTextBufferAddMark

gtkTextBufferAddSelectionClipboard

gtkTextBufferApplyTag

gtkTextBufferApplyTagByName

gtkTextBufferBackspace

gtkTextBufferBeginIrreversibleAction

gtkTextBufferBeginUserAction

gtkTextBufferCopyClipboard

gtkTextBufferCreateChildAnchor

gtkTextBufferCreateMark

gtkTextBufferCutClipboard

gtkTextBufferDelete

gtkTextBufferDeleteInteractive

gtkTextBufferDeleteMark

gtkTextBufferDeleteMarkByName

gtkTextBufferDeleteSelection

gtkTextBufferEndIrreversibleAction

gtkTextBufferEndUserAction

gtkTextBufferGetBounds

gtkTextBufferGetCanRedo

gtkTextBufferGetCanUndo

gtkTextBufferGetCharCount

gtkTextBufferGetEnableUndo

gtkTextBufferGetEndIter

gtkTextBufferGetHasSelection

gtkTextBufferGetInsert

gtkTextBufferGetIterAtChildAnchor

gtkTextBufferGetIterAtLine

gtkTextBufferGetIterAtLineIndex

gtkTextBufferGetIterAtLineOffset

gtkTextBufferGetIterAtMark

gtkTextBufferGetIterAtOffset

gtkTextBufferGetLineCount

gtkTextBufferGetMark

gtkTextBufferGetMaxUndoLevels

gtkTextBufferGetModified

gtkTextBufferGetSelectionBound

gtkTextBufferGetSelectionBounds

gtkTextBufferGetSelectionContent

gtkTextBufferGetSlice

gtkTextBufferGetStartIter

gtkTextBufferGetTagTable

gtkTextBufferGetText

gtkTextBufferInsert

gtkTextBufferInsertAtCursor

gtkTextBufferInsertChildAnchor

gtkTextBufferInsertInteractive

gtkTextBufferInsertInteractiveAtCursor

gtkTextBufferInsertMarkup

gtkTextBufferInsertPaintable

gtkTextBufferInsertRange

gtkTextBufferInsertRangeInteractive

gtkTextBufferMoveMark

gtkTextBufferMoveMarkByName

gtkTextBufferPasteClipboard

gtkTextBufferPlaceCursor

gtkTextBufferRedo

gtkTextBufferRemoveAllTags

gtkTextBufferRemoveCommitNotify

gtkTextBufferRemoveSelectionClipboard

gtkTextBufferRemoveTag

gtkTextBufferRemoveTagByName

gtkTextBufferSelectRange

gtkTextBufferSetEnableUndo

gtkTextBufferSetMaxUndoLevels

gtkTextBufferSetModified

gtkTextBufferSetText

gtkTextBufferUndo

gtkTextChildAnchorNew

gtkTextChildAnchorNewWithReplacement

gtkTextChildAnchorGetDeleted

gtkTextChildAnchorGetWidgets

gtkTextIterAssign

gtkTextIterBackwardChar

gtkTextIterBackwardChars

gtkTextIterBackwardCursorPosition

gtkTextIterBackwardCursorPositions

gtkTextIterBackwardFindChar

gtkTextIterBackwardLine

gtkTextIterBackwardLines

gtkTextIterBackwardSearch

gtkTextIterBackwardSentenceStart

gtkTextIterBackwardSentenceStarts

gtkTextIterBackwardToTagToggle

gtkTextIterBackwardVisibleCursorPosition

gtkTextIterBackwardVisibleCursorPositions

gtkTextIterBackwardVisibleLine

gtkTextIterBackwardVisibleLines

gtkTextIterBackwardVisibleWordStart

gtkTextIterBackwardVisibleWordStarts

gtkTextIterBackwardWordStart

gtkTextIterBackwardWordStarts

gtkTextIterCanInsert

gtkTextIterCompare

gtkTextIterCopy

gtkTextIterEditable

gtkTextIterEndsLine

gtkTextIterEndsSentence

gtkTextIterEndsTag

gtkTextIterEndsWord

gtkTextIterEqual

gtkTextIterForwardChar

gtkTextIterForwardChars

gtkTextIterForwardCursorPosition

gtkTextIterForwardCursorPositions

gtkTextIterForwardFindChar

gtkTextIterForwardLine

gtkTextIterForwardLines

gtkTextIterForwardSearch

gtkTextIterForwardSentenceEnd

gtkTextIterForwardSentenceEnds

gtkTextIterForwardToEnd

gtkTextIterForwardToLineEnd

gtkTextIterForwardToTagToggle

gtkTextIterForwardVisibleCursorPosition

gtkTextIterForwardVisibleCursorPositions

gtkTextIterForwardVisibleLine

gtkTextIterForwardVisibleLines

gtkTextIterForwardVisibleWordEnd

gtkTextIterForwardVisibleWordEnds

gtkTextIterForwardWordEnd

gtkTextIterForwardWordEnds

gtkTextIterFree

gtkTextIterGetBuffer

gtkTextIterGetBytesInLine

gtkTextIterGetChar

gtkTextIterGetCharsInLine

gtkTextIterGetChildAnchor

gtkTextIterGetLanguage

gtkTextIterGetLine

gtkTextIterGetLineIndex

gtkTextIterGetLineOffset

gtkTextIterGetMarks

gtkTextIterGetOffset

gtkTextIterGetPaintable

gtkTextIterGetSlice

gtkTextIterGetTags

gtkTextIterGetText

gtkTextIterGetToggledTags

gtkTextIterGetVisibleLineIndex

gtkTextIterGetVisibleLineOffset

gtkTextIterGetVisibleSlice

gtkTextIterGetVisibleText

gtkTextIterHasTag

gtkTextIterInRange

gtkTextIterInsideSentence

gtkTextIterInsideWord

gtkTextIterIsCursorPosition

gtkTextIterIsEnd

gtkTextIterIsStart

gtkTextIterOrder

gtkTextIterSetLine

gtkTextIterSetLineIndex

gtkTextIterSetLineOffset

gtkTextIterSetOffset

gtkTextIterSetVisibleLineIndex

gtkTextIterSetVisibleLineOffset

gtkTextIterStartsLine

gtkTextIterStartsSentence

gtkTextIterStartsTag

gtkTextIterStartsWord

gtkTextIterTogglesTag

gtkTextMarkNew

gtkTextMarkGetBuffer

gtkTextMarkGetDeleted

gtkTextMarkGetLeftGravity

gtkTextMarkGetName

gtkTextMarkGetVisible

gtkTextMarkSetVisible

gtkTextTagNew

gtkTextTagChanged

gtkTextTagGetPriority

gtkTextTagSetPriority

gtkTextTagTableNew

gtkTextTagTableAdd

gtkTextTagTableForeach

gtkTextTagTableGetSize

gtkTextTagTableLookup

gtkTextTagTableRemove

gtkTextViewNew

gtkTextViewNewWithBuffer

gtkTextViewAddChildAtAnchor

gtkTextViewAddOverlay

gtkTextViewBackwardDisplayLine

gtkTextViewBackwardDisplayLineStart

gtkTextViewBufferToWindowCoords

gtkTextViewForwardDisplayLine

gtkTextViewForwardDisplayLineEnd

gtkTextViewGetAcceptsTab

gtkTextViewGetBottomMargin

gtkTextViewGetBuffer

gtkTextViewGetCursorLocations

gtkTextViewGetCursorVisible

gtkTextViewGetEditable

gtkTextViewGetExtraMenu

gtkTextViewGetGutter

gtkTextViewGetIndent

gtkTextViewGetInputHints

gtkTextViewGetInputPurpose

gtkTextViewGetIterAtLocation

gtkTextViewGetIterAtPosition

gtkTextViewGetIterLocation

gtkTextViewGetJustification

gtkTextViewGetLeftMargin

gtkTextViewGetLineAtY

gtkTextViewGetLineYrange

gtkTextViewGetLtrContext

gtkTextViewGetMonospace

gtkTextViewGetOverwrite

gtkTextViewGetPixelsAboveLines

gtkTextViewGetPixelsBelowLines

gtkTextViewGetPixelsInsideWrap

gtkTextViewGetRightMargin

gtkTextViewGetRtlContext

gtkTextViewGetTabs

gtkTextViewGetTopMargin

gtkTextViewGetVisibleOffset

gtkTextViewGetVisibleRect

gtkTextViewGetWrapMode

gtkTextViewImContextFilterKeypress

gtkTextViewMoveMarkOnscreen

gtkTextViewMoveOverlay

gtkTextViewMoveVisually

gtkTextViewPlaceCursorOnscreen

gtkTextViewRemove

gtkTextViewResetCursorBlink

gtkTextViewResetImContext

gtkTextViewScrollMarkOnscreen

gtkTextViewScrollToIter

gtkTextViewScrollToMark

gtkTextViewSetAcceptsTab

gtkTextViewSetBottomMargin

gtkTextViewSetBuffer

gtkTextViewSetCursorVisible

gtkTextViewSetEditable

gtkTextViewSetExtraMenu

gtkTextViewSetGutter

gtkTextViewSetIndent

gtkTextViewSetInputHints

gtkTextViewSetInputPurpose

gtkTextViewSetJustification

gtkTextViewSetLeftMargin

gtkTextViewSetMonospace

gtkTextViewSetOverwrite

gtkTextViewSetPixelsAboveLines

gtkTextViewSetPixelsBelowLines

gtkTextViewSetPixelsInsideWrap

gtkTextViewSetRightMargin

gtkTextViewSetTabs

gtkTextViewSetTopMargin

gtkTextViewSetWrapMode

gtkTextViewStartsDisplayLine

gtkTextViewWindowToBufferCoords
}
