% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtk_autogen.R
\name{gtkListStoreNewv}
\alias{gtkListStoreNewv}
\alias{gtkListStoreAppend}
\alias{gtkListStoreClear}
\alias{gtkListStoreInsert}
\alias{gtkListStoreInsertAfter}
\alias{gtkListStoreInsertBefore}
\alias{gtkListStoreInsertWithValuesv}
\alias{gtkListStoreIterIsValid}
\alias{gtkListStoreMoveAfter}
\alias{gtkListStoreMoveBefore}
\alias{gtkListStorePrepend}
\alias{gtkListStoreRemove}
\alias{gtkListStoreReorder}
\alias{gtkListStoreSetColumnTypes}
\alias{gtkListStoreSetValue}
\alias{gtkListStoreSetValuesv}
\alias{gtkListStoreSwap}
\alias{gtkTreeModelFilterNew}
\alias{gtkTreeModelGetColumnType}
\alias{gtkTreeModelGetFlags}
\alias{gtkTreeModelGetIter}
\alias{gtkTreeModelGetIterFirst}
\alias{gtkTreeModelGetIterFromString}
\alias{gtkTreeModelGetNColumns}
\alias{gtkTreeModelGetPath}
\alias{gtkTreeModelGetStringFromIter}
\alias{gtkTreeModelGetValue}
\alias{gtkTreeModelIterChildren}
\alias{gtkTreeModelIterHasChild}
\alias{gtkTreeModelIterNChildren}
\alias{gtkTreeModelIterNext}
\alias{gtkTreeModelIterNthChild}
\alias{gtkTreeModelIterParent}
\alias{gtkTreeModelIterPrevious}
\alias{gtkTreeModelRefNode}
\alias{gtkTreeModelRowChanged}
\alias{gtkTreeModelRowDeleted}
\alias{gtkTreeModelRowHasChildToggled}
\alias{gtkTreeModelRowInserted}
\alias{gtkTreeModelRowsReorderedWithLength}
\alias{gtkTreeModelUnrefNode}
\alias{gtkTreeModelFilterClearCache}
\alias{gtkTreeModelFilterConvertChildIterToIter}
\alias{gtkTreeModelFilterConvertChildPathToPath}
\alias{gtkTreeModelFilterConvertIterToChildIter}
\alias{gtkTreeModelFilterConvertPathToChildPath}
\alias{gtkTreeModelFilterGetModel}
\alias{gtkTreeModelFilterRefilter}
\alias{gtkTreeModelFilterSetVisibleColumn}
\alias{gtkTreeModelSortNewWithModel}
\alias{gtkTreeModelSortClearCache}
\alias{gtkTreeModelSortConvertChildIterToIter}
\alias{gtkTreeModelSortConvertChildPathToPath}
\alias{gtkTreeModelSortConvertIterToChildIter}
\alias{gtkTreeModelSortConvertPathToChildPath}
\alias{gtkTreeModelSortGetModel}
\alias{gtkTreeModelSortIterIsValid}
\alias{gtkTreeModelSortResetDefaultSortFunc}
\alias{gtkTreeSelectionCountSelectedRows}
\alias{gtkTreeSelectionGetMode}
\alias{gtkTreeSelectionGetSelected}
\alias{gtkTreeSelectionGetSelectedRows}
\alias{gtkTreeSelectionGetTreeView}
\alias{gtkTreeSelectionIterIsSelected}
\alias{gtkTreeSelectionPathIsSelected}
\alias{gtkTreeSelectionSelectAll}
\alias{gtkTreeSelectionSelectIter}
\alias{gtkTreeSelectionSelectPath}
\alias{gtkTreeSelectionSelectRange}
\alias{gtkTreeSelectionSetMode}
\alias{gtkTreeSelectionUnselectAll}
\alias{gtkTreeSelectionUnselectIter}
\alias{gtkTreeSelectionUnselectPath}
\alias{gtkTreeSelectionUnselectRange}
\alias{gtkTreeStoreNewv}
\alias{gtkTreeStoreAppend}
\alias{gtkTreeStoreClear}
\alias{gtkTreeStoreInsert}
\alias{gtkTreeStoreInsertAfter}
\alias{gtkTreeStoreInsertBefore}
\alias{gtkTreeStoreInsertWithValuesv}
\alias{gtkTreeStoreIsAncestor}
\alias{gtkTreeStoreIterDepth}
\alias{gtkTreeStoreIterIsValid}
\alias{gtkTreeStoreMoveAfter}
\alias{gtkTreeStoreMoveBefore}
\alias{gtkTreeStorePrepend}
\alias{gtkTreeStoreRemove}
\alias{gtkTreeStoreSetColumnTypes}
\alias{gtkTreeStoreSetValue}
\alias{gtkTreeStoreSetValuesv}
\alias{gtkTreeStoreSwap}
\alias{gtkTreeViewNew}
\alias{gtkTreeViewNewWithModel}
\alias{gtkTreeViewAppendColumn}
\alias{gtkTreeViewCollapseAll}
\alias{gtkTreeViewCollapseRow}
\alias{gtkTreeViewColumnsAutosize}
\alias{gtkTreeViewConvertBinWindowToTreeCoords}
\alias{gtkTreeViewConvertBinWindowToWidgetCoords}
\alias{gtkTreeViewConvertTreeToBinWindowCoords}
\alias{gtkTreeViewConvertTreeToWidgetCoords}
\alias{gtkTreeViewConvertWidgetToBinWindowCoords}
\alias{gtkTreeViewConvertWidgetToTreeCoords}
\alias{gtkTreeViewCreateRowDragIcon}
\alias{gtkTreeViewEnableModelDragDest}
\alias{gtkTreeViewEnableModelDragSource}
\alias{gtkTreeViewExpandAll}
\alias{gtkTreeViewExpandRow}
\alias{gtkTreeViewExpandToPath}
\alias{gtkTreeViewGetActivateOnSingleClick}
\alias{gtkTreeViewGetBackgroundArea}
\alias{gtkTreeViewGetCellArea}
\alias{gtkTreeViewGetColumn}
\alias{gtkTreeViewGetColumns}
\alias{gtkTreeViewGetCursor}
\alias{gtkTreeViewGetDestRowAtPos}
\alias{gtkTreeViewGetDragDestRow}
\alias{gtkTreeViewGetEnableSearch}
\alias{gtkTreeViewGetEnableTreeLines}
\alias{gtkTreeViewGetExpanderColumn}
\alias{gtkTreeViewGetFixedHeightMode}
\alias{gtkTreeViewGetGridLines}
\alias{gtkTreeViewGetHeadersClickable}
\alias{gtkTreeViewGetHeadersVisible}
\alias{gtkTreeViewGetHoverExpand}
\alias{gtkTreeViewGetHoverSelection}
\alias{gtkTreeViewGetLevelIndentation}
\alias{gtkTreeViewGetModel}
\alias{gtkTreeViewGetNColumns}
\alias{gtkTreeViewGetPathAtPos}
\alias{gtkTreeViewGetReorderable}
\alias{gtkTreeViewGetRubberBanding}
\alias{gtkTreeViewGetSearchColumn}
\alias{gtkTreeViewGetSearchEntry}
\alias{gtkTreeViewGetSelection}
\alias{gtkTreeViewGetShowExpanders}
\alias{gtkTreeViewGetTooltipColumn}
\alias{gtkTreeViewGetTooltipContext}
\alias{gtkTreeViewGetVisibleRange}
\alias{gtkTreeViewGetVisibleRect}
\alias{gtkTreeViewInsertColumn}
\alias{gtkTreeViewIsBlankAtPos}
\alias{gtkTreeViewIsRubberBandingActive}
\alias{gtkTreeViewMoveColumnAfter}
\alias{gtkTreeViewRemoveColumn}
\alias{gtkTreeViewRowActivated}
\alias{gtkTreeViewRowExpanded}
\alias{gtkTreeViewScrollToCell}
\alias{gtkTreeViewScrollToPoint}
\alias{gtkTreeViewSetActivateOnSingleClick}
\alias{gtkTreeViewSetCursor}
\alias{gtkTreeViewSetCursorOnCell}
\alias{gtkTreeViewSetDragDestRow}
\alias{gtkTreeViewSetEnableSearch}
\alias{gtkTreeViewSetEnableTreeLines}
\alias{gtkTreeViewSetExpanderColumn}
\alias{gtkTreeViewSetFixedHeightMode}
\alias{gtkTreeViewSetGridLines}
\alias{gtkTreeViewSetHeadersClickable}
\alias{gtkTreeViewSetHeadersVisible}
\alias{gtkTreeViewSetHoverExpand}
\alias{gtkTreeViewSetHoverSelection}
\alias{gtkTreeViewSetLevelIndentation}
\alias{gtkTreeViewSetModel}
\alias{gtkTreeViewSetReorderable}
\alias{gtkTreeViewSetRubberBanding}
\alias{gtkTreeViewSetSearchColumn}
\alias{gtkTreeViewSetSearchEntry}
\alias{gtkTreeViewSetShowExpanders}
\alias{gtkTreeViewSetTooltipCell}
\alias{gtkTreeViewSetTooltipColumn}
\alias{gtkTreeViewSetTooltipRow}
\alias{gtkTreeViewUnsetRowsDragDest}
\alias{gtkTreeViewUnsetRowsDragSource}
\alias{gtkTreeViewColumnNew}
\alias{gtkTreeViewColumnNewWithArea}
\alias{gtkTreeViewColumnAddAttribute}
\alias{gtkTreeViewColumnCellGetPosition}
\alias{gtkTreeViewColumnCellGetSize}
\alias{gtkTreeViewColumnCellIsVisible}
\alias{gtkTreeViewColumnCellSetCellData}
\alias{gtkTreeViewColumnClear}
\alias{gtkTreeViewColumnClearAttributes}
\alias{gtkTreeViewColumnClicked}
\alias{gtkTreeViewColumnFocusCell}
\alias{gtkTreeViewColumnGetAlignment}
\alias{gtkTreeViewColumnGetButton}
\alias{gtkTreeViewColumnGetClickable}
\alias{gtkTreeViewColumnGetExpand}
\alias{gtkTreeViewColumnGetFixedWidth}
\alias{gtkTreeViewColumnGetMaxWidth}
\alias{gtkTreeViewColumnGetMinWidth}
\alias{gtkTreeViewColumnGetReorderable}
\alias{gtkTreeViewColumnGetResizable}
\alias{gtkTreeViewColumnGetSizing}
\alias{gtkTreeViewColumnGetSortColumnId}
\alias{gtkTreeViewColumnGetSortIndicator}
\alias{gtkTreeViewColumnGetSortOrder}
\alias{gtkTreeViewColumnGetSpacing}
\alias{gtkTreeViewColumnGetTitle}
\alias{gtkTreeViewColumnGetTreeView}
\alias{gtkTreeViewColumnGetVisible}
\alias{gtkTreeViewColumnGetWidget}
\alias{gtkTreeViewColumnGetWidth}
\alias{gtkTreeViewColumnGetXOffset}
\alias{gtkTreeViewColumnPackEnd}
\alias{gtkTreeViewColumnPackStart}
\alias{gtkTreeViewColumnQueueResize}
\alias{gtkTreeViewColumnSetAlignment}
\alias{gtkTreeViewColumnSetClickable}
\alias{gtkTreeViewColumnSetExpand}
\alias{gtkTreeViewColumnSetFixedWidth}
\alias{gtkTreeViewColumnSetMaxWidth}
\alias{gtkTreeViewColumnSetMinWidth}
\alias{gtkTreeViewColumnSetReorderable}
\alias{gtkTreeViewColumnSetResizable}
\alias{gtkTreeViewColumnSetSizing}
\alias{gtkTreeViewColumnSetSortColumnId}
\alias{gtkTreeViewColumnSetSortIndicator}
\alias{gtkTreeViewColumnSetSortOrder}
\alias{gtkTreeViewColumnSetSpacing}
\alias{gtkTreeViewColumnSetTitle}
\alias{gtkTreeViewColumnSetVisible}
\alias{gtkTreeViewColumnSetWidget}
\title{gtkListStoreNewv}
\usage{
gtkListStoreNewv(n_columns, types)

gtkListStoreAppend(list_store)

gtkListStoreClear(list_store)

gtkListStoreInsert(list_store, position)

gtkListStoreInsertAfter(list_store, sibling)

gtkListStoreInsertBefore(list_store, sibling)

gtkListStoreInsertWithValuesv(list_store, position, columns, values, n_values)

gtkListStoreIterIsValid(list_store, iter)

gtkListStoreMoveAfter(store, iter, position)

gtkListStoreMoveBefore(store, iter, position)

gtkListStorePrepend(list_store)

gtkListStoreRemove(list_store, iter)

gtkListStoreReorder(store, new_order)

gtkListStoreSetColumnTypes(list_store, n_columns, types)

gtkListStoreSetValue(list_store, iter, column, value)

gtkListStoreSetValuesv(list_store, iter, columns, values, n_values)

gtkListStoreSwap(store, a, b)

gtkTreeModelFilterNew(child_model, root)

gtkTreeModelGetColumnType(tree_model, index_)

gtkTreeModelGetFlags(tree_model)

gtkTreeModelGetIter(tree_model, path)

gtkTreeModelGetIterFirst(tree_model)

gtkTreeModelGetIterFromString(tree_model, path_string)

gtkTreeModelGetNColumns(tree_model)

gtkTreeModelGetPath(tree_model, iter)

gtkTreeModelGetStringFromIter(tree_model, iter)

gtkTreeModelGetValue(tree_model, iter, column)

gtkTreeModelIterChildren(tree_model, parent)

gtkTreeModelIterHasChild(tree_model, iter)

gtkTreeModelIterNChildren(tree_model, iter)

gtkTreeModelIterNext(tree_model, iter)

gtkTreeModelIterNthChild(tree_model, parent, n)

gtkTreeModelIterParent(tree_model, child)

gtkTreeModelIterPrevious(tree_model, iter)

gtkTreeModelRefNode(tree_model, iter)

gtkTreeModelRowChanged(tree_model, path, iter)

gtkTreeModelRowDeleted(tree_model, path)

gtkTreeModelRowHasChildToggled(tree_model, path, iter)

gtkTreeModelRowInserted(tree_model, path, iter)

gtkTreeModelRowsReorderedWithLength(tree_model, path, iter, new_order, length)

gtkTreeModelUnrefNode(tree_model, iter)

gtkTreeModelFilterClearCache(filter)

gtkTreeModelFilterConvertChildIterToIter(filter, child_iter)

gtkTreeModelFilterConvertChildPathToPath(filter, child_path)

gtkTreeModelFilterConvertIterToChildIter(filter, filter_iter)

gtkTreeModelFilterConvertPathToChildPath(filter, filter_path)

gtkTreeModelFilterGetModel(filter)

gtkTreeModelFilterRefilter(filter)

gtkTreeModelFilterSetVisibleColumn(filter, column)

gtkTreeModelSortNewWithModel(child_model)

gtkTreeModelSortClearCache(tree_model_sort)

gtkTreeModelSortConvertChildIterToIter(tree_model_sort, child_iter)

gtkTreeModelSortConvertChildPathToPath(tree_model_sort, child_path)

gtkTreeModelSortConvertIterToChildIter(tree_model_sort, sorted_iter)

gtkTreeModelSortConvertPathToChildPath(tree_model_sort, sorted_path)

gtkTreeModelSortGetModel(tree_model)

gtkTreeModelSortIterIsValid(tree_model_sort, iter)

gtkTreeModelSortResetDefaultSortFunc(tree_model_sort)

gtkTreeSelectionCountSelectedRows(selection)

gtkTreeSelectionGetMode(selection)

gtkTreeSelectionGetSelected(selection)

gtkTreeSelectionGetSelectedRows(selection)

gtkTreeSelectionGetTreeView(selection)

gtkTreeSelectionIterIsSelected(selection, iter)

gtkTreeSelectionPathIsSelected(selection, path)

gtkTreeSelectionSelectAll(selection)

gtkTreeSelectionSelectIter(selection, iter)

gtkTreeSelectionSelectPath(selection, path)

gtkTreeSelectionSelectRange(selection, start_path, end_path)

gtkTreeSelectionSetMode(selection, type)

gtkTreeSelectionUnselectAll(selection)

gtkTreeSelectionUnselectIter(selection, iter)

gtkTreeSelectionUnselectPath(selection, path)

gtkTreeSelectionUnselectRange(selection, start_path, end_path)

gtkTreeStoreNewv(n_columns, types)

gtkTreeStoreAppend(tree_store, parent)

gtkTreeStoreClear(tree_store)

gtkTreeStoreInsert(tree_store, parent, position)

gtkTreeStoreInsertAfter(tree_store, parent, sibling)

gtkTreeStoreInsertBefore(tree_store, parent, sibling)

gtkTreeStoreInsertWithValuesv(
  tree_store,
  parent,
  position,
  columns,
  values,
  n_values
)

gtkTreeStoreIsAncestor(tree_store, iter, descendant)

gtkTreeStoreIterDepth(tree_store, iter)

gtkTreeStoreIterIsValid(tree_store, iter)

gtkTreeStoreMoveAfter(tree_store, iter, position)

gtkTreeStoreMoveBefore(tree_store, iter, position)

gtkTreeStorePrepend(tree_store, parent)

gtkTreeStoreRemove(tree_store, iter)

gtkTreeStoreSetColumnTypes(tree_store, n_columns, types)

gtkTreeStoreSetValue(tree_store, iter, column, value)

gtkTreeStoreSetValuesv(tree_store, iter, columns, values, n_values)

gtkTreeStoreSwap(tree_store, a, b)

gtkTreeViewNew()

gtkTreeViewNewWithModel(model)

gtkTreeViewAppendColumn(tree_view, column)

gtkTreeViewCollapseAll(tree_view)

gtkTreeViewCollapseRow(tree_view, path)

gtkTreeViewColumnsAutosize(tree_view)

gtkTreeViewConvertBinWindowToTreeCoords(tree_view, bx, by)

gtkTreeViewConvertBinWindowToWidgetCoords(tree_view, bx, by)

gtkTreeViewConvertTreeToBinWindowCoords(tree_view, tx, ty)

gtkTreeViewConvertTreeToWidgetCoords(tree_view, tx, ty)

gtkTreeViewConvertWidgetToBinWindowCoords(tree_view, wx, wy)

gtkTreeViewConvertWidgetToTreeCoords(tree_view, wx, wy)

gtkTreeViewCreateRowDragIcon(tree_view, path)

gtkTreeViewEnableModelDragDest(tree_view, formats, actions)

gtkTreeViewEnableModelDragSource(
  tree_view,
  start_button_mask,
  formats,
  actions
)

gtkTreeViewExpandAll(tree_view)

gtkTreeViewExpandRow(tree_view, path, open_all)

gtkTreeViewExpandToPath(tree_view, path)

gtkTreeViewGetActivateOnSingleClick(tree_view)

gtkTreeViewGetBackgroundArea(tree_view, path, column)

gtkTreeViewGetCellArea(tree_view, path, column)

gtkTreeViewGetColumn(tree_view, n)

gtkTreeViewGetColumns(tree_view)

gtkTreeViewGetCursor(tree_view)

gtkTreeViewGetDestRowAtPos(tree_view, drag_x, drag_y)

gtkTreeViewGetDragDestRow(tree_view)

gtkTreeViewGetEnableSearch(tree_view)

gtkTreeViewGetEnableTreeLines(tree_view)

gtkTreeViewGetExpanderColumn(tree_view)

gtkTreeViewGetFixedHeightMode(tree_view)

gtkTreeViewGetGridLines(tree_view)

gtkTreeViewGetHeadersClickable(tree_view)

gtkTreeViewGetHeadersVisible(tree_view)

gtkTreeViewGetHoverExpand(tree_view)

gtkTreeViewGetHoverSelection(tree_view)

gtkTreeViewGetLevelIndentation(tree_view)

gtkTreeViewGetModel(tree_view)

gtkTreeViewGetNColumns(tree_view)

gtkTreeViewGetPathAtPos(tree_view, x, y)

gtkTreeViewGetReorderable(tree_view)

gtkTreeViewGetRubberBanding(tree_view)

gtkTreeViewGetSearchColumn(tree_view)

gtkTreeViewGetSearchEntry(tree_view)

gtkTreeViewGetSelection(tree_view)

gtkTreeViewGetShowExpanders(tree_view)

gtkTreeViewGetTooltipColumn(tree_view)

gtkTreeViewGetTooltipContext(tree_view, x, y, keyboard_tip)

gtkTreeViewGetVisibleRange(tree_view)

gtkTreeViewGetVisibleRect(tree_view)

gtkTreeViewInsertColumn(tree_view, column, position)

gtkTreeViewIsBlankAtPos(tree_view, x, y)

gtkTreeViewIsRubberBandingActive(tree_view)

gtkTreeViewMoveColumnAfter(tree_view, column, base_column)

gtkTreeViewRemoveColumn(tree_view, column)

gtkTreeViewRowActivated(tree_view, path, column)

gtkTreeViewRowExpanded(tree_view, path)

gtkTreeViewScrollToCell(
  tree_view,
  path,
  column,
  use_align,
  row_align,
  col_align
)

gtkTreeViewScrollToPoint(tree_view, tree_x, tree_y)

gtkTreeViewSetActivateOnSingleClick(tree_view, single)

gtkTreeViewSetCursor(tree_view, path, focus_column, start_editing)

gtkTreeViewSetCursorOnCell(
  tree_view,
  path,
  focus_column,
  focus_cell,
  start_editing
)

gtkTreeViewSetDragDestRow(tree_view, path, pos)

gtkTreeViewSetEnableSearch(tree_view, enable_search)

gtkTreeViewSetEnableTreeLines(tree_view, enabled)

gtkTreeViewSetExpanderColumn(tree_view, column)

gtkTreeViewSetFixedHeightMode(tree_view, enable)

gtkTreeViewSetGridLines(tree_view, grid_lines)

gtkTreeViewSetHeadersClickable(tree_view, setting)

gtkTreeViewSetHeadersVisible(tree_view, headers_visible)

gtkTreeViewSetHoverExpand(tree_view, expand)

gtkTreeViewSetHoverSelection(tree_view, hover)

gtkTreeViewSetLevelIndentation(tree_view, indentation)

gtkTreeViewSetModel(tree_view, model)

gtkTreeViewSetReorderable(tree_view, reorderable)

gtkTreeViewSetRubberBanding(tree_view, enable)

gtkTreeViewSetSearchColumn(tree_view, column)

gtkTreeViewSetSearchEntry(tree_view, entry)

gtkTreeViewSetShowExpanders(tree_view, enabled)

gtkTreeViewSetTooltipCell(tree_view, tooltip, path, column, cell)

gtkTreeViewSetTooltipColumn(tree_view, column)

gtkTreeViewSetTooltipRow(tree_view, tooltip, path)

gtkTreeViewUnsetRowsDragDest(tree_view)

gtkTreeViewUnsetRowsDragSource(tree_view)

gtkTreeViewColumnNew()

gtkTreeViewColumnNewWithArea(area)

gtkTreeViewColumnAddAttribute(tree_column, cell_renderer, attribute, column)

gtkTreeViewColumnCellGetPosition(tree_column, cell_renderer)

gtkTreeViewColumnCellGetSize(tree_column)

gtkTreeViewColumnCellIsVisible(tree_column)

gtkTreeViewColumnCellSetCellData(
  tree_column,
  tree_model,
  iter,
  is_expander,
  is_expanded
)

gtkTreeViewColumnClear(tree_column)

gtkTreeViewColumnClearAttributes(tree_column, cell_renderer)

gtkTreeViewColumnClicked(tree_column)

gtkTreeViewColumnFocusCell(tree_column, cell)

gtkTreeViewColumnGetAlignment(tree_column)

gtkTreeViewColumnGetButton(tree_column)

gtkTreeViewColumnGetClickable(tree_column)

gtkTreeViewColumnGetExpand(tree_column)

gtkTreeViewColumnGetFixedWidth(tree_column)

gtkTreeViewColumnGetMaxWidth(tree_column)

gtkTreeViewColumnGetMinWidth(tree_column)

gtkTreeViewColumnGetReorderable(tree_column)

gtkTreeViewColumnGetResizable(tree_column)

gtkTreeViewColumnGetSizing(tree_column)

gtkTreeViewColumnGetSortColumnId(tree_column)

gtkTreeViewColumnGetSortIndicator(tree_column)

gtkTreeViewColumnGetSortOrder(tree_column)

gtkTreeViewColumnGetSpacing(tree_column)

gtkTreeViewColumnGetTitle(tree_column)

gtkTreeViewColumnGetTreeView(tree_column)

gtkTreeViewColumnGetVisible(tree_column)

gtkTreeViewColumnGetWidget(tree_column)

gtkTreeViewColumnGetWidth(tree_column)

gtkTreeViewColumnGetXOffset(tree_column)

gtkTreeViewColumnPackEnd(tree_column, cell, expand)

gtkTreeViewColumnPackStart(tree_column, cell, expand)

gtkTreeViewColumnQueueResize(tree_column)

gtkTreeViewColumnSetAlignment(tree_column, xalign)

gtkTreeViewColumnSetClickable(tree_column, clickable)

gtkTreeViewColumnSetExpand(tree_column, expand)

gtkTreeViewColumnSetFixedWidth(tree_column, fixed_width)

gtkTreeViewColumnSetMaxWidth(tree_column, max_width)

gtkTreeViewColumnSetMinWidth(tree_column, min_width)

gtkTreeViewColumnSetReorderable(tree_column, reorderable)

gtkTreeViewColumnSetResizable(tree_column, resizable)

gtkTreeViewColumnSetSizing(tree_column, type)

gtkTreeViewColumnSetSortColumnId(tree_column, sort_column_id)

gtkTreeViewColumnSetSortIndicator(tree_column, setting)

gtkTreeViewColumnSetSortOrder(tree_column, order)

gtkTreeViewColumnSetSpacing(tree_column, spacing)

gtkTreeViewColumnSetTitle(tree_column, title)

gtkTreeViewColumnSetVisible(tree_column, visible)

gtkTreeViewColumnSetWidget(tree_column, widget)
}
\arguments{
\item{n_columns}{gint}

\item{types}{GType}

\item{list_store}{ListStore}

\item{position}{gint}

\item{sibling}{TreeIter}

\item{columns}{gint}

\item{values}{GObject.Value}

\item{n_values}{gint}

\item{iter}{TreeIter}

\item{store}{ListStore}

\item{new_order}{gint}

\item{column}{gint}

\item{value}{GObject.Value}

\item{a}{TreeIter}

\item{b}{TreeIter}

\item{child_model}{TreeModel}

\item{root}{TreePath}

\item{tree_model}{TreeModel}

\item{index_}{gint}

\item{path}{TreePath}

\item{path_string}{utf8}

\item{parent}{TreeIter}

\item{n}{gint}

\item{child}{TreeIter}

\item{length}{gint}

\item{filter}{TreeModelFilter}

\item{child_iter}{TreeIter}

\item{child_path}{TreePath}

\item{filter_iter}{TreeIter}

\item{filter_path}{TreePath}

\item{tree_model_sort}{TreeModelSort}

\item{sorted_iter}{TreeIter}

\item{sorted_path}{TreePath}

\item{selection}{TreeSelection}

\item{start_path}{TreePath}

\item{end_path}{TreePath}

\item{type}{TreeViewColumnSizing}

\item{tree_store}{TreeStore}

\item{descendant}{TreeIter}

\item{model}{TreeModel}

\item{tree_view}{TreeView}

\item{bx}{gint}

\item{by}{gint}

\item{tx}{gint}

\item{ty}{gint}

\item{wx}{gint}

\item{wy}{gint}

\item{formats}{Gdk.ContentFormats}

\item{actions}{Gdk.DragAction}

\item{start_button_mask}{Gdk.ModifierType}

\item{open_all}{gboolean}

\item{drag_x}{gint}

\item{drag_y}{gint}

\item{x}{gint}

\item{y}{gint}

\item{keyboard_tip}{gboolean}

\item{base_column}{TreeViewColumn}

\item{use_align}{gboolean}

\item{row_align}{gfloat}

\item{col_align}{gfloat}

\item{tree_x}{gint}

\item{tree_y}{gint}

\item{single}{gboolean}

\item{focus_column}{TreeViewColumn}

\item{start_editing}{gboolean}

\item{focus_cell}{CellRenderer}

\item{pos}{TreeViewDropPosition}

\item{enable_search}{gboolean}

\item{enabled}{gboolean}

\item{enable}{gboolean}

\item{grid_lines}{TreeViewGridLines}

\item{setting}{gboolean}

\item{headers_visible}{gboolean}

\item{expand}{gboolean}

\item{hover}{gboolean}

\item{indentation}{gint}

\item{reorderable}{gboolean}

\item{entry}{Editable}

\item{tooltip}{Tooltip}

\item{cell}{CellRenderer}

\item{area}{CellArea}

\item{tree_column}{TreeViewColumn}

\item{cell_renderer}{CellRenderer}

\item{attribute}{utf8}

\item{is_expander}{gboolean}

\item{is_expanded}{gboolean}

\item{xalign}{gfloat}

\item{clickable}{gboolean}

\item{fixed_width}{gint}

\item{max_width}{gint}

\item{min_width}{gint}

\item{resizable}{gboolean}

\item{sort_column_id}{gint}

\item{order}{SortType}

\item{spacing}{gint}

\item{title}{utf8}

\item{visible}{gboolean}

\item{widget}{Widget}
}
\value{
ListStore

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

TreeModel

GType

TreeModelFlags

gboolean

gboolean

gboolean

gint

TreePath

utf8

Return value from C function

gboolean

gboolean

gint

gboolean

gboolean

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

TreePath

Return value from C function

TreePath

TreeModel

Return value from C function

Return value from C function

TreeModelSort

Return value from C function

gboolean

TreePath

Return value from C function

TreePath

TreeModel

gboolean

Return value from C function

gint

SelectionMode

gboolean

GLib.List

TreeView

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

TreeStore

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

gint

gboolean

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

Widget

gint

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Gdk.Paintable

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

gboolean

Return value from C function

Return value from C function

TreeViewColumn

GLib.List

Return value from C function

gboolean

Return value from C function

gboolean

gboolean

TreeViewColumn

gboolean

TreeViewGridLines

gboolean

gboolean

gboolean

gboolean

gint

TreeModel

guint

gboolean

gboolean

gboolean

gint

Editable

TreeSelection

gboolean

gint

gboolean

gboolean

Return value from C function

gint

gboolean

gboolean

Return value from C function

gint

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

TreeViewColumn

TreeViewColumn

Return value from C function

gboolean

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gfloat

Widget

gboolean

gboolean

gint

gint

gint

gboolean

gboolean

TreeViewColumnSizing

gint

gboolean

SortType

gint

utf8

Widget

gboolean

Widget

gint

gint

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function
}
\description{
gtkListStoreNewv

gtkListStoreAppend

gtkListStoreClear

gtkListStoreInsert

gtkListStoreInsertAfter

gtkListStoreInsertBefore

gtkListStoreInsertWithValuesv

gtkListStoreIterIsValid

gtkListStoreMoveAfter

gtkListStoreMoveBefore

gtkListStorePrepend

gtkListStoreRemove

gtkListStoreReorder

gtkListStoreSetColumnTypes

gtkListStoreSetValue

gtkListStoreSetValuesv

gtkListStoreSwap

gtkTreeModelFilterNew

gtkTreeModelGetColumnType

gtkTreeModelGetFlags

gtkTreeModelGetIter

gtkTreeModelGetIterFirst

gtkTreeModelGetIterFromString

gtkTreeModelGetNColumns

gtkTreeModelGetPath

gtkTreeModelGetStringFromIter

gtkTreeModelGetValue

gtkTreeModelIterChildren

gtkTreeModelIterHasChild

gtkTreeModelIterNChildren

gtkTreeModelIterNext

gtkTreeModelIterNthChild

gtkTreeModelIterParent

gtkTreeModelIterPrevious

gtkTreeModelRefNode

gtkTreeModelRowChanged

gtkTreeModelRowDeleted

gtkTreeModelRowHasChildToggled

gtkTreeModelRowInserted

gtkTreeModelRowsReorderedWithLength

gtkTreeModelUnrefNode

gtkTreeModelFilterClearCache

gtkTreeModelFilterConvertChildIterToIter

gtkTreeModelFilterConvertChildPathToPath

gtkTreeModelFilterConvertIterToChildIter

gtkTreeModelFilterConvertPathToChildPath

gtkTreeModelFilterGetModel

gtkTreeModelFilterRefilter

gtkTreeModelFilterSetVisibleColumn

gtkTreeModelSortNewWithModel

gtkTreeModelSortClearCache

gtkTreeModelSortConvertChildIterToIter

gtkTreeModelSortConvertChildPathToPath

gtkTreeModelSortConvertIterToChildIter

gtkTreeModelSortConvertPathToChildPath

gtkTreeModelSortGetModel

gtkTreeModelSortIterIsValid

gtkTreeModelSortResetDefaultSortFunc

gtkTreeSelectionCountSelectedRows

gtkTreeSelectionGetMode

gtkTreeSelectionGetSelected

gtkTreeSelectionGetSelectedRows

gtkTreeSelectionGetTreeView

gtkTreeSelectionIterIsSelected

gtkTreeSelectionPathIsSelected

gtkTreeSelectionSelectAll

gtkTreeSelectionSelectIter

gtkTreeSelectionSelectPath

gtkTreeSelectionSelectRange

gtkTreeSelectionSetMode

gtkTreeSelectionUnselectAll

gtkTreeSelectionUnselectIter

gtkTreeSelectionUnselectPath

gtkTreeSelectionUnselectRange

gtkTreeStoreNewv

gtkTreeStoreAppend

gtkTreeStoreClear

gtkTreeStoreInsert

gtkTreeStoreInsertAfter

gtkTreeStoreInsertBefore

gtkTreeStoreInsertWithValuesv

gtkTreeStoreIsAncestor

gtkTreeStoreIterDepth

gtkTreeStoreIterIsValid

gtkTreeStoreMoveAfter

gtkTreeStoreMoveBefore

gtkTreeStorePrepend

gtkTreeStoreRemove

gtkTreeStoreSetColumnTypes

gtkTreeStoreSetValue

gtkTreeStoreSetValuesv

gtkTreeStoreSwap

gtkTreeViewNew

gtkTreeViewNewWithModel

gtkTreeViewAppendColumn

gtkTreeViewCollapseAll

gtkTreeViewCollapseRow

gtkTreeViewColumnsAutosize

gtkTreeViewConvertBinWindowToTreeCoords

gtkTreeViewConvertBinWindowToWidgetCoords

gtkTreeViewConvertTreeToBinWindowCoords

gtkTreeViewConvertTreeToWidgetCoords

gtkTreeViewConvertWidgetToBinWindowCoords

gtkTreeViewConvertWidgetToTreeCoords

gtkTreeViewCreateRowDragIcon

gtkTreeViewEnableModelDragDest

gtkTreeViewEnableModelDragSource

gtkTreeViewExpandAll

gtkTreeViewExpandRow

gtkTreeViewExpandToPath

gtkTreeViewGetActivateOnSingleClick

gtkTreeViewGetBackgroundArea

gtkTreeViewGetCellArea

gtkTreeViewGetColumn

gtkTreeViewGetColumns

gtkTreeViewGetCursor

gtkTreeViewGetDestRowAtPos

gtkTreeViewGetDragDestRow

gtkTreeViewGetEnableSearch

gtkTreeViewGetEnableTreeLines

gtkTreeViewGetExpanderColumn

gtkTreeViewGetFixedHeightMode

gtkTreeViewGetGridLines

gtkTreeViewGetHeadersClickable

gtkTreeViewGetHeadersVisible

gtkTreeViewGetHoverExpand

gtkTreeViewGetHoverSelection

gtkTreeViewGetLevelIndentation

gtkTreeViewGetModel

gtkTreeViewGetNColumns

gtkTreeViewGetPathAtPos

gtkTreeViewGetReorderable

gtkTreeViewGetRubberBanding

gtkTreeViewGetSearchColumn

gtkTreeViewGetSearchEntry

gtkTreeViewGetSelection

gtkTreeViewGetShowExpanders

gtkTreeViewGetTooltipColumn

gtkTreeViewGetTooltipContext

gtkTreeViewGetVisibleRange

gtkTreeViewGetVisibleRect

gtkTreeViewInsertColumn

gtkTreeViewIsBlankAtPos

gtkTreeViewIsRubberBandingActive

gtkTreeViewMoveColumnAfter

gtkTreeViewRemoveColumn

gtkTreeViewRowActivated

gtkTreeViewRowExpanded

gtkTreeViewScrollToCell

gtkTreeViewScrollToPoint

gtkTreeViewSetActivateOnSingleClick

gtkTreeViewSetCursor

gtkTreeViewSetCursorOnCell

gtkTreeViewSetDragDestRow

gtkTreeViewSetEnableSearch

gtkTreeViewSetEnableTreeLines

gtkTreeViewSetExpanderColumn

gtkTreeViewSetFixedHeightMode

gtkTreeViewSetGridLines

gtkTreeViewSetHeadersClickable

gtkTreeViewSetHeadersVisible

gtkTreeViewSetHoverExpand

gtkTreeViewSetHoverSelection

gtkTreeViewSetLevelIndentation

gtkTreeViewSetModel

gtkTreeViewSetReorderable

gtkTreeViewSetRubberBanding

gtkTreeViewSetSearchColumn

gtkTreeViewSetSearchEntry

gtkTreeViewSetShowExpanders

gtkTreeViewSetTooltipCell

gtkTreeViewSetTooltipColumn

gtkTreeViewSetTooltipRow

gtkTreeViewUnsetRowsDragDest

gtkTreeViewUnsetRowsDragSource

gtkTreeViewColumnNew

gtkTreeViewColumnNewWithArea

gtkTreeViewColumnAddAttribute

gtkTreeViewColumnCellGetPosition

gtkTreeViewColumnCellGetSize

gtkTreeViewColumnCellIsVisible

gtkTreeViewColumnCellSetCellData

gtkTreeViewColumnClear

gtkTreeViewColumnClearAttributes

gtkTreeViewColumnClicked

gtkTreeViewColumnFocusCell

gtkTreeViewColumnGetAlignment

gtkTreeViewColumnGetButton

gtkTreeViewColumnGetClickable

gtkTreeViewColumnGetExpand

gtkTreeViewColumnGetFixedWidth

gtkTreeViewColumnGetMaxWidth

gtkTreeViewColumnGetMinWidth

gtkTreeViewColumnGetReorderable

gtkTreeViewColumnGetResizable

gtkTreeViewColumnGetSizing

gtkTreeViewColumnGetSortColumnId

gtkTreeViewColumnGetSortIndicator

gtkTreeViewColumnGetSortOrder

gtkTreeViewColumnGetSpacing

gtkTreeViewColumnGetTitle

gtkTreeViewColumnGetTreeView

gtkTreeViewColumnGetVisible

gtkTreeViewColumnGetWidget

gtkTreeViewColumnGetWidth

gtkTreeViewColumnGetXOffset

gtkTreeViewColumnPackEnd

gtkTreeViewColumnPackStart

gtkTreeViewColumnQueueResize

gtkTreeViewColumnSetAlignment

gtkTreeViewColumnSetClickable

gtkTreeViewColumnSetExpand

gtkTreeViewColumnSetFixedWidth

gtkTreeViewColumnSetMaxWidth

gtkTreeViewColumnSetMinWidth

gtkTreeViewColumnSetReorderable

gtkTreeViewColumnSetResizable

gtkTreeViewColumnSetSizing

gtkTreeViewColumnSetSortColumnId

gtkTreeViewColumnSetSortIndicator

gtkTreeViewColumnSetSortOrder

gtkTreeViewColumnSetSpacing

gtkTreeViewColumnSetTitle

gtkTreeViewColumnSetVisible

gtkTreeViewColumnSetWidget
}
