% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtk_autogen.R
\name{gtkWidgetGetDefaultDirection}
\alias{gtkWidgetGetDefaultDirection}
\alias{gtkWidgetSetDefaultDirection}
\alias{gtkWidgetActionSetEnabled}
\alias{gtkWidgetActivate}
\alias{gtkWidgetActivateActionVariant}
\alias{gtkWidgetActivateDefault}
\alias{gtkWidgetAddController}
\alias{gtkWidgetAddCssClass}
\alias{gtkWidgetAddMnemonicLabel}
\alias{gtkWidgetAddTickCallback}
\alias{gtkWidgetAllocate}
\alias{gtkWidgetChildFocus}
\alias{gtkWidgetComputeBounds}
\alias{gtkWidgetComputeExpand}
\alias{gtkWidgetComputePoint}
\alias{gtkWidgetComputeTransform}
\alias{gtkWidgetContains}
\alias{gtkWidgetCreatePangoContext}
\alias{gtkWidgetCreatePangoLayout}
\alias{gtkWidgetDisposeTemplate}
\alias{gtkWidgetErrorBell}
\alias{gtkWidgetGetAllocatedBaseline}
\alias{gtkWidgetGetAllocatedHeight}
\alias{gtkWidgetGetAllocatedWidth}
\alias{gtkWidgetGetAllocation}
\alias{gtkWidgetGetAncestor}
\alias{gtkWidgetGetBaseline}
\alias{gtkWidgetGetCanFocus}
\alias{gtkWidgetGetCanTarget}
\alias{gtkWidgetGetChildVisible}
\alias{gtkWidgetGetClipboard}
\alias{gtkWidgetGetColor}
\alias{gtkWidgetGetCssClasses}
\alias{gtkWidgetGetCssName}
\alias{gtkWidgetGetCursor}
\alias{gtkWidgetGetDirection}
\alias{gtkWidgetGetDisplay}
\alias{gtkWidgetGetFirstChild}
\alias{gtkWidgetGetFocusChild}
\alias{gtkWidgetGetFocusOnClick}
\alias{gtkWidgetGetFocusable}
\alias{gtkWidgetGetFontMap}
\alias{gtkWidgetGetFontOptions}
\alias{gtkWidgetGetFrameClock}
\alias{gtkWidgetGetHalign}
\alias{gtkWidgetGetHasTooltip}
\alias{gtkWidgetGetHeight}
\alias{gtkWidgetGetHexpand}
\alias{gtkWidgetGetHexpandSet}
\alias{gtkWidgetGetLastChild}
\alias{gtkWidgetGetLayoutManager}
\alias{gtkWidgetGetMapped}
\alias{gtkWidgetGetMarginBottom}
\alias{gtkWidgetGetMarginEnd}
\alias{gtkWidgetGetMarginStart}
\alias{gtkWidgetGetMarginTop}
\alias{gtkWidgetGetName}
\alias{gtkWidgetGetNative}
\alias{gtkWidgetGetNextSibling}
\alias{gtkWidgetGetOpacity}
\alias{gtkWidgetGetOverflow}
\alias{gtkWidgetGetPangoContext}
\alias{gtkWidgetGetParent}
\alias{gtkWidgetGetPreferredSize}
\alias{gtkWidgetGetPrevSibling}
\alias{gtkWidgetGetPrimaryClipboard}
\alias{gtkWidgetGetRealized}
\alias{gtkWidgetGetReceivesDefault}
\alias{gtkWidgetGetRequestMode}
\alias{gtkWidgetGetRoot}
\alias{gtkWidgetGetScaleFactor}
\alias{gtkWidgetGetSensitive}
\alias{gtkWidgetGetSettings}
\alias{gtkWidgetGetSize}
\alias{gtkWidgetGetSizeRequest}
\alias{gtkWidgetGetStateFlags}
\alias{gtkWidgetGetStyleContext}
\alias{gtkWidgetGetTemplateChild}
\alias{gtkWidgetGetTooltipMarkup}
\alias{gtkWidgetGetTooltipText}
\alias{gtkWidgetGetValign}
\alias{gtkWidgetGetVexpand}
\alias{gtkWidgetGetVexpandSet}
\alias{gtkWidgetGetVisible}
\alias{gtkWidgetGetWidth}
\alias{gtkWidgetGrabFocus}
\alias{gtkWidgetHasCssClass}
\alias{gtkWidgetHasDefault}
\alias{gtkWidgetHasFocus}
\alias{gtkWidgetHasVisibleFocus}
\alias{gtkWidgetHide}
\alias{gtkWidgetInDestruction}
\alias{gtkWidgetInitTemplate}
\alias{gtkWidgetInsertActionGroup}
\alias{gtkWidgetInsertAfter}
\alias{gtkWidgetInsertBefore}
\alias{gtkWidgetIsAncestor}
\alias{gtkWidgetIsDrawable}
\alias{gtkWidgetIsFocus}
\alias{gtkWidgetIsSensitive}
\alias{gtkWidgetIsVisible}
\alias{gtkWidgetKeynavFailed}
\alias{gtkWidgetListMnemonicLabels}
\alias{gtkWidgetMap}
\alias{gtkWidgetMeasure}
\alias{gtkWidgetMnemonicActivate}
\alias{gtkWidgetObserveChildren}
\alias{gtkWidgetObserveControllers}
\alias{gtkWidgetPick}
\alias{gtkWidgetQueueAllocate}
\alias{gtkWidgetQueueDraw}
\alias{gtkWidgetQueueResize}
\alias{gtkWidgetRealize}
\alias{gtkWidgetRemoveController}
\alias{gtkWidgetRemoveCssClass}
\alias{gtkWidgetRemoveMnemonicLabel}
\alias{gtkWidgetRemoveTickCallback}
\alias{gtkWidgetSetCanFocus}
\alias{gtkWidgetSetCanTarget}
\alias{gtkWidgetSetChildVisible}
\alias{gtkWidgetSetCssClasses}
\alias{gtkWidgetSetCursor}
\alias{gtkWidgetSetCursorFromName}
\alias{gtkWidgetSetDirection}
\alias{gtkWidgetSetFocusChild}
\alias{gtkWidgetSetFocusOnClick}
\alias{gtkWidgetSetFocusable}
\alias{gtkWidgetSetFontMap}
\alias{gtkWidgetSetFontOptions}
\alias{gtkWidgetSetHalign}
\alias{gtkWidgetSetHasTooltip}
\alias{gtkWidgetSetHexpand}
\alias{gtkWidgetSetHexpandSet}
\alias{gtkWidgetSetLayoutManager}
\alias{gtkWidgetSetMarginBottom}
\alias{gtkWidgetSetMarginEnd}
\alias{gtkWidgetSetMarginStart}
\alias{gtkWidgetSetMarginTop}
\alias{gtkWidgetSetName}
\alias{gtkWidgetSetOpacity}
\alias{gtkWidgetSetOverflow}
\alias{gtkWidgetSetParent}
\alias{gtkWidgetSetReceivesDefault}
\alias{gtkWidgetSetSensitive}
\alias{gtkWidgetSetSizeRequest}
\alias{gtkWidgetSetStateFlags}
\alias{gtkWidgetSetTooltipMarkup}
\alias{gtkWidgetSetTooltipText}
\alias{gtkWidgetSetValign}
\alias{gtkWidgetSetVexpand}
\alias{gtkWidgetSetVexpandSet}
\alias{gtkWidgetSetVisible}
\alias{gtkWidgetShouldLayout}
\alias{gtkWidgetShow}
\alias{gtkWidgetSizeAllocate}
\alias{gtkWidgetSnapshotChild}
\alias{gtkWidgetTranslateCoordinates}
\alias{gtkWidgetTriggerTooltipQuery}
\alias{gtkWidgetUnmap}
\alias{gtkWidgetUnparent}
\alias{gtkWidgetUnrealize}
\alias{gtkWidgetUnsetStateFlags}
\alias{gtkWidgetClassAddShortcut}
\alias{gtkWidgetClassBindTemplateCallbackFull}
\alias{gtkWidgetClassBindTemplateChildFull}
\alias{gtkWidgetClassGetAccessibleRole}
\alias{gtkWidgetClassGetActivateSignal}
\alias{gtkWidgetClassGetCssName}
\alias{gtkWidgetClassGetLayoutManagerType}
\alias{gtkWidgetClassInstallAction}
\alias{gtkWidgetClassInstallPropertyAction}
\alias{gtkWidgetClassQueryAction}
\alias{gtkWidgetClassSetAccessibleRole}
\alias{gtkWidgetClassSetActivateSignal}
\alias{gtkWidgetClassSetActivateSignalFromName}
\alias{gtkWidgetClassSetCssName}
\alias{gtkWidgetClassSetLayoutManagerType}
\alias{gtkWidgetClassSetTemplate}
\alias{gtkWidgetClassSetTemplateFromResource}
\alias{gtkWidgetClassSetTemplateScope}
\alias{gtkWidgetPaintableNew}
\alias{gtkWidgetPaintableGetWidget}
\alias{gtkWidgetPaintableSetWidget}
\title{gtkWidgetGetDefaultDirection}
\usage{
gtkWidgetGetDefaultDirection()

gtkWidgetSetDefaultDirection(dir)

gtkWidgetActionSetEnabled(widget, action_name, enabled)

gtkWidgetActivate(widget)

gtkWidgetActivateActionVariant(widget, name, args)

gtkWidgetActivateDefault(widget)

gtkWidgetAddController(widget, controller)

gtkWidgetAddCssClass(widget, css_class)

gtkWidgetAddMnemonicLabel(widget, label)

gtkWidgetAddTickCallback(widget, callback)

gtkWidgetAllocate(widget, width, height, baseline, transform)

gtkWidgetChildFocus(widget, direction)

gtkWidgetComputeBounds(widget, target)

gtkWidgetComputeExpand(widget, orientation)

gtkWidgetComputePoint(widget, target, point)

gtkWidgetComputeTransform(widget, target)

gtkWidgetContains(widget, x, y)

gtkWidgetCreatePangoContext(widget)

gtkWidgetCreatePangoLayout(widget, text)

gtkWidgetDisposeTemplate(widget, widget_type)

gtkWidgetErrorBell(widget)

gtkWidgetGetAllocatedBaseline(widget)

gtkWidgetGetAllocatedHeight(widget)

gtkWidgetGetAllocatedWidth(widget)

gtkWidgetGetAllocation(widget)

gtkWidgetGetAncestor(widget, widget_type)

gtkWidgetGetBaseline(widget)

gtkWidgetGetCanFocus(widget)

gtkWidgetGetCanTarget(widget)

gtkWidgetGetChildVisible(widget)

gtkWidgetGetClipboard(widget)

gtkWidgetGetColor(widget)

gtkWidgetGetCssClasses(widget)

gtkWidgetGetCssName(self)

gtkWidgetGetCursor(widget)

gtkWidgetGetDirection(widget)

gtkWidgetGetDisplay(widget)

gtkWidgetGetFirstChild(widget)

gtkWidgetGetFocusChild(widget)

gtkWidgetGetFocusOnClick(widget)

gtkWidgetGetFocusable(widget)

gtkWidgetGetFontMap(widget)

gtkWidgetGetFontOptions(widget)

gtkWidgetGetFrameClock(widget)

gtkWidgetGetHalign(widget)

gtkWidgetGetHasTooltip(widget)

gtkWidgetGetHeight(widget)

gtkWidgetGetHexpand(widget)

gtkWidgetGetHexpandSet(widget)

gtkWidgetGetLastChild(widget)

gtkWidgetGetLayoutManager(widget)

gtkWidgetGetMapped(widget)

gtkWidgetGetMarginBottom(widget)

gtkWidgetGetMarginEnd(widget)

gtkWidgetGetMarginStart(widget)

gtkWidgetGetMarginTop(widget)

gtkWidgetGetName(widget)

gtkWidgetGetNative(widget)

gtkWidgetGetNextSibling(widget)

gtkWidgetGetOpacity(widget)

gtkWidgetGetOverflow(widget)

gtkWidgetGetPangoContext(widget)

gtkWidgetGetParent(widget)

gtkWidgetGetPreferredSize(widget)

gtkWidgetGetPrevSibling(widget)

gtkWidgetGetPrimaryClipboard(widget)

gtkWidgetGetRealized(widget)

gtkWidgetGetReceivesDefault(widget)

gtkWidgetGetRequestMode(widget)

gtkWidgetGetRoot(widget)

gtkWidgetGetScaleFactor(widget)

gtkWidgetGetSensitive(widget)

gtkWidgetGetSettings(widget)

gtkWidgetGetSize(widget, orientation)

gtkWidgetGetSizeRequest(widget)

gtkWidgetGetStateFlags(widget)

gtkWidgetGetStyleContext(widget)

gtkWidgetGetTemplateChild(widget, widget_type, name)

gtkWidgetGetTooltipMarkup(widget)

gtkWidgetGetTooltipText(widget)

gtkWidgetGetValign(widget)

gtkWidgetGetVexpand(widget)

gtkWidgetGetVexpandSet(widget)

gtkWidgetGetVisible(widget)

gtkWidgetGetWidth(widget)

gtkWidgetGrabFocus(widget)

gtkWidgetHasCssClass(widget, css_class)

gtkWidgetHasDefault(widget)

gtkWidgetHasFocus(widget)

gtkWidgetHasVisibleFocus(widget)

gtkWidgetHide(widget)

gtkWidgetInDestruction(widget)

gtkWidgetInitTemplate(widget)

gtkWidgetInsertActionGroup(widget, name, group)

gtkWidgetInsertAfter(widget, parent, previous_sibling)

gtkWidgetInsertBefore(widget, parent, next_sibling)

gtkWidgetIsAncestor(widget, ancestor)

gtkWidgetIsDrawable(widget)

gtkWidgetIsFocus(widget)

gtkWidgetIsSensitive(widget)

gtkWidgetIsVisible(widget)

gtkWidgetKeynavFailed(widget, direction)

gtkWidgetListMnemonicLabels(widget)

gtkWidgetMap(widget)

gtkWidgetMeasure(widget, orientation, for_size)

gtkWidgetMnemonicActivate(widget, group_cycling)

gtkWidgetObserveChildren(widget)

gtkWidgetObserveControllers(widget)

gtkWidgetPick(widget, x, y, flags)

gtkWidgetQueueAllocate(widget)

gtkWidgetQueueDraw(widget)

gtkWidgetQueueResize(widget)

gtkWidgetRealize(widget)

gtkWidgetRemoveController(widget, controller)

gtkWidgetRemoveCssClass(widget, css_class)

gtkWidgetRemoveMnemonicLabel(widget, label)

gtkWidgetRemoveTickCallback(widget, id)

gtkWidgetSetCanFocus(widget, can_focus)

gtkWidgetSetCanTarget(widget, can_target)

gtkWidgetSetChildVisible(widget, child_visible)

gtkWidgetSetCssClasses(widget, classes)

gtkWidgetSetCursor(widget, cursor)

gtkWidgetSetCursorFromName(widget, name)

gtkWidgetSetDirection(widget, dir)

gtkWidgetSetFocusChild(widget, child)

gtkWidgetSetFocusOnClick(widget, focus_on_click)

gtkWidgetSetFocusable(widget, focusable)

gtkWidgetSetFontMap(widget, font_map)

gtkWidgetSetFontOptions(widget, options)

gtkWidgetSetHalign(widget, align)

gtkWidgetSetHasTooltip(widget, has_tooltip)

gtkWidgetSetHexpand(widget, expand)

gtkWidgetSetHexpandSet(widget, set)

gtkWidgetSetLayoutManager(widget, layout_manager)

gtkWidgetSetMarginBottom(widget, margin)

gtkWidgetSetMarginEnd(widget, margin)

gtkWidgetSetMarginStart(widget, margin)

gtkWidgetSetMarginTop(widget, margin)

gtkWidgetSetName(widget, name)

gtkWidgetSetOpacity(widget, opacity)

gtkWidgetSetOverflow(widget, overflow)

gtkWidgetSetParent(widget, parent)

gtkWidgetSetReceivesDefault(widget, receives_default)

gtkWidgetSetSensitive(widget, sensitive)

gtkWidgetSetSizeRequest(widget, width, height)

gtkWidgetSetStateFlags(widget, flags, clear)

gtkWidgetSetTooltipMarkup(widget, markup)

gtkWidgetSetTooltipText(widget, text)

gtkWidgetSetValign(widget, align)

gtkWidgetSetVexpand(widget, expand)

gtkWidgetSetVexpandSet(widget, set)

gtkWidgetSetVisible(widget, visible)

gtkWidgetShouldLayout(widget)

gtkWidgetShow(widget)

gtkWidgetSizeAllocate(widget, allocation, baseline)

gtkWidgetSnapshotChild(widget, child, snapshot)

gtkWidgetTranslateCoordinates(src_widget, dest_widget, src_x, src_y)

gtkWidgetTriggerTooltipQuery(widget)

gtkWidgetUnmap(widget)

gtkWidgetUnparent(widget)

gtkWidgetUnrealize(widget)

gtkWidgetUnsetStateFlags(widget, flags)

gtkWidgetClassAddShortcut(widget_class, shortcut)

gtkWidgetClassBindTemplateCallbackFull(
  widget_class,
  callback_name,
  callback_symbol
)

gtkWidgetClassBindTemplateChildFull(
  widget_class,
  name,
  internal_child,
  struct_offset
)

gtkWidgetClassGetAccessibleRole(widget_class)

gtkWidgetClassGetActivateSignal(widget_class)

gtkWidgetClassGetCssName(widget_class)

gtkWidgetClassGetLayoutManagerType(widget_class)

gtkWidgetClassInstallAction(
  widget_class,
  action_name,
  parameter_type,
  activate
)

gtkWidgetClassInstallPropertyAction(widget_class, action_name, property_name)

gtkWidgetClassQueryAction(widget_class, index_)

gtkWidgetClassSetAccessibleRole(widget_class, accessible_role)

gtkWidgetClassSetActivateSignal(widget_class, signal_id)

gtkWidgetClassSetActivateSignalFromName(widget_class, signal_name)

gtkWidgetClassSetCssName(widget_class, name)

gtkWidgetClassSetLayoutManagerType(widget_class, type)

gtkWidgetClassSetTemplate(widget_class, template_bytes)

gtkWidgetClassSetTemplateFromResource(widget_class, resource_name)

gtkWidgetClassSetTemplateScope(widget_class, scope)

gtkWidgetPaintableNew(widget)

gtkWidgetPaintableGetWidget(self)

gtkWidgetPaintableSetWidget(self, widget)
}
\arguments{
\item{dir}{TextDirection}

\item{widget}{Widget}

\item{action_name}{utf8}

\item{enabled}{gboolean}

\item{name}{utf8}

\item{args}{GLib.Variant}

\item{controller}{EventController}

\item{css_class}{utf8}

\item{label}{Widget}

\item{callback}{function — TickCallback function(widget, frame_clock)}

\item{width}{gint}

\item{height}{gint}

\item{baseline}{gint}

\item{transform}{Gsk.Transform}

\item{direction}{DirectionType}

\item{target}{Widget}

\item{orientation}{Orientation}

\item{point}{Graphene.Point}

\item{x}{gdouble}

\item{y}{gdouble}

\item{text}{utf8}

\item{widget_type}{GType}

\item{self}{WidgetPaintable}

\item{group}{Gio.ActionGroup}

\item{parent}{Widget}

\item{previous_sibling}{Widget}

\item{next_sibling}{Widget}

\item{ancestor}{Widget}

\item{for_size}{gint}

\item{group_cycling}{gboolean}

\item{flags}{StateFlags}

\item{id}{guint}

\item{can_focus}{gboolean}

\item{can_target}{gboolean}

\item{child_visible}{gboolean}

\item{classes}{utf8}

\item{cursor}{Gdk.Cursor}

\item{child}{Widget}

\item{focus_on_click}{gboolean}

\item{focusable}{gboolean}

\item{font_map}{Pango.FontMap}

\item{options}{cairo.FontOptions}

\item{align}{Align}

\item{has_tooltip}{gboolean}

\item{expand}{gboolean}

\item{set}{gboolean}

\item{layout_manager}{LayoutManager}

\item{margin}{gint}

\item{opacity}{gdouble}

\item{overflow}{Overflow}

\item{receives_default}{gboolean}

\item{sensitive}{gboolean}

\item{clear}{gboolean}

\item{markup}{utf8}

\item{visible}{gboolean}

\item{allocation}{Allocation}

\item{snapshot}{Snapshot}

\item{src_widget}{Widget}

\item{dest_widget}{Widget}

\item{src_x}{gdouble}

\item{src_y}{gdouble}

\item{widget_class}{WidgetClass}

\item{shortcut}{Shortcut}

\item{callback_name}{utf8}

\item{callback_symbol}{function — GObject.Callback function()}

\item{internal_child}{gboolean}

\item{struct_offset}{gssize}

\item{parameter_type}{utf8}

\item{activate}{function — WidgetActionActivateFunc function(widget, action_name, parameter)}

\item{property_name}{utf8}

\item{index_}{guint}

\item{accessible_role}{AccessibleRole}

\item{signal_id}{guint}

\item{signal_name}{utf8}

\item{type}{GType}

\item{template_bytes}{GLib.Bytes}

\item{resource_name}{utf8}

\item{scope}{BuilderScope}
}
\value{
TextDirection

Return value from C function

Return value from C function

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

guint

Return value from C function

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

Pango.Context

Pango.Layout

Return value from C function

Return value from C function

gint

gint

gint

Return value from C function

Widget

gint

gboolean

gboolean

gboolean

Gdk.Clipboard

Return value from C function

utf8

utf8

Gdk.Cursor

TextDirection

Gdk.Display

Widget

Widget

gboolean

gboolean

Pango.FontMap

cairo.FontOptions

Gdk.FrameClock

Align

gboolean

gint

gboolean

gboolean

Widget

LayoutManager

gboolean

gint

gint

gint

gint

utf8

Native

Widget

gdouble

Overflow

Pango.Context

Widget

Return value from C function

Widget

Gdk.Clipboard

gboolean

gboolean

SizeRequestMode

Root

gint

gboolean

Settings

gint

Return value from C function

StateFlags

StyleContext

GObject.Object

utf8

utf8

Align

gboolean

gboolean

gboolean

gint

gboolean

gboolean

gboolean

gboolean

gboolean

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

GLib.List

Return value from C function

Return value from C function

gboolean

Gio.ListModel

Gio.ListModel

Widget

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

AccessibleRole

guint

utf8

GType

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

WidgetPaintable

Widget

Return value from C function
}
\description{
gtkWidgetGetDefaultDirection

gtkWidgetSetDefaultDirection

gtkWidgetActionSetEnabled

gtkWidgetActivate

gtkWidgetActivateActionVariant

gtkWidgetActivateDefault

gtkWidgetAddController

gtkWidgetAddCssClass

gtkWidgetAddMnemonicLabel

gtkWidgetAddTickCallback

gtkWidgetAllocate

gtkWidgetChildFocus

gtkWidgetComputeBounds

gtkWidgetComputeExpand

gtkWidgetComputePoint

gtkWidgetComputeTransform

gtkWidgetContains

gtkWidgetCreatePangoContext

gtkWidgetCreatePangoLayout

gtkWidgetDisposeTemplate

gtkWidgetErrorBell

gtkWidgetGetAllocatedBaseline

gtkWidgetGetAllocatedHeight

gtkWidgetGetAllocatedWidth

gtkWidgetGetAllocation

gtkWidgetGetAncestor

gtkWidgetGetBaseline

gtkWidgetGetCanFocus

gtkWidgetGetCanTarget

gtkWidgetGetChildVisible

gtkWidgetGetClipboard

gtkWidgetGetColor

gtkWidgetGetCssClasses

gtkWidgetGetCssName

gtkWidgetGetCursor

gtkWidgetGetDirection

gtkWidgetGetDisplay

gtkWidgetGetFirstChild

gtkWidgetGetFocusChild

gtkWidgetGetFocusOnClick

gtkWidgetGetFocusable

gtkWidgetGetFontMap

gtkWidgetGetFontOptions

gtkWidgetGetFrameClock

gtkWidgetGetHalign

gtkWidgetGetHasTooltip

gtkWidgetGetHeight

gtkWidgetGetHexpand

gtkWidgetGetHexpandSet

gtkWidgetGetLastChild

gtkWidgetGetLayoutManager

gtkWidgetGetMapped

gtkWidgetGetMarginBottom

gtkWidgetGetMarginEnd

gtkWidgetGetMarginStart

gtkWidgetGetMarginTop

gtkWidgetGetName

gtkWidgetGetNative

gtkWidgetGetNextSibling

gtkWidgetGetOpacity

gtkWidgetGetOverflow

gtkWidgetGetPangoContext

gtkWidgetGetParent

gtkWidgetGetPreferredSize

gtkWidgetGetPrevSibling

gtkWidgetGetPrimaryClipboard

gtkWidgetGetRealized

gtkWidgetGetReceivesDefault

gtkWidgetGetRequestMode

gtkWidgetGetRoot

gtkWidgetGetScaleFactor

gtkWidgetGetSensitive

gtkWidgetGetSettings

gtkWidgetGetSize

gtkWidgetGetSizeRequest

gtkWidgetGetStateFlags

gtkWidgetGetStyleContext

gtkWidgetGetTemplateChild

gtkWidgetGetTooltipMarkup

gtkWidgetGetTooltipText

gtkWidgetGetValign

gtkWidgetGetVexpand

gtkWidgetGetVexpandSet

gtkWidgetGetVisible

gtkWidgetGetWidth

gtkWidgetGrabFocus

gtkWidgetHasCssClass

gtkWidgetHasDefault

gtkWidgetHasFocus

gtkWidgetHasVisibleFocus

gtkWidgetHide

gtkWidgetInDestruction

gtkWidgetInitTemplate

gtkWidgetInsertActionGroup

gtkWidgetInsertAfter

gtkWidgetInsertBefore

gtkWidgetIsAncestor

gtkWidgetIsDrawable

gtkWidgetIsFocus

gtkWidgetIsSensitive

gtkWidgetIsVisible

gtkWidgetKeynavFailed

gtkWidgetListMnemonicLabels

gtkWidgetMap

gtkWidgetMeasure

gtkWidgetMnemonicActivate

gtkWidgetObserveChildren

gtkWidgetObserveControllers

gtkWidgetPick

gtkWidgetQueueAllocate

gtkWidgetQueueDraw

gtkWidgetQueueResize

gtkWidgetRealize

gtkWidgetRemoveController

gtkWidgetRemoveCssClass

gtkWidgetRemoveMnemonicLabel

gtkWidgetRemoveTickCallback

gtkWidgetSetCanFocus

gtkWidgetSetCanTarget

gtkWidgetSetChildVisible

gtkWidgetSetCssClasses

gtkWidgetSetCursor

gtkWidgetSetCursorFromName

gtkWidgetSetDirection

gtkWidgetSetFocusChild

gtkWidgetSetFocusOnClick

gtkWidgetSetFocusable

gtkWidgetSetFontMap

gtkWidgetSetFontOptions

gtkWidgetSetHalign

gtkWidgetSetHasTooltip

gtkWidgetSetHexpand

gtkWidgetSetHexpandSet

gtkWidgetSetLayoutManager

gtkWidgetSetMarginBottom

gtkWidgetSetMarginEnd

gtkWidgetSetMarginStart

gtkWidgetSetMarginTop

gtkWidgetSetName

gtkWidgetSetOpacity

gtkWidgetSetOverflow

gtkWidgetSetParent

gtkWidgetSetReceivesDefault

gtkWidgetSetSensitive

gtkWidgetSetSizeRequest

gtkWidgetSetStateFlags

gtkWidgetSetTooltipMarkup

gtkWidgetSetTooltipText

gtkWidgetSetValign

gtkWidgetSetVexpand

gtkWidgetSetVexpandSet

gtkWidgetSetVisible

gtkWidgetShouldLayout

gtkWidgetShow

gtkWidgetSizeAllocate

gtkWidgetSnapshotChild

gtkWidgetTranslateCoordinates

gtkWidgetTriggerTooltipQuery

gtkWidgetUnmap

gtkWidgetUnparent

gtkWidgetUnrealize

gtkWidgetUnsetStateFlags

gtkWidgetClassAddShortcut

gtkWidgetClassBindTemplateCallbackFull

gtkWidgetClassBindTemplateChildFull

gtkWidgetClassGetAccessibleRole

gtkWidgetClassGetActivateSignal

gtkWidgetClassGetCssName

gtkWidgetClassGetLayoutManagerType

gtkWidgetClassInstallAction

gtkWidgetClassInstallPropertyAction

gtkWidgetClassQueryAction

gtkWidgetClassSetAccessibleRole

gtkWidgetClassSetActivateSignal

gtkWidgetClassSetActivateSignalFromName

gtkWidgetClassSetCssName

gtkWidgetClassSetLayoutManagerType

gtkWidgetClassSetTemplate

gtkWidgetClassSetTemplateFromResource

gtkWidgetClassSetTemplateScope

gtkWidgetPaintableNew

gtkWidgetPaintableGetWidget

gtkWidgetPaintableSetWidget
}
