% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtk_autogen.R
\name{gtkWindowNew}
\alias{gtkWindowNew}
\alias{gtkWindowGetDefaultIconName}
\alias{gtkWindowGetToplevels}
\alias{gtkWindowListToplevels}
\alias{gtkWindowSetAutoStartupNotification}
\alias{gtkWindowSetDefaultIconName}
\alias{gtkWindowSetInteractiveDebugging}
\alias{gtkWindowClose}
\alias{gtkWindowDestroy}
\alias{gtkWindowFullscreen}
\alias{gtkWindowFullscreenOnMonitor}
\alias{gtkWindowGetApplication}
\alias{gtkWindowGetChild}
\alias{gtkWindowGetDecorated}
\alias{gtkWindowGetDefaultSize}
\alias{gtkWindowGetDefaultWidget}
\alias{gtkWindowGetDeletable}
\alias{gtkWindowGetDestroyWithParent}
\alias{gtkWindowGetFocus}
\alias{gtkWindowGetFocusVisible}
\alias{gtkWindowGetGravity}
\alias{gtkWindowGetGroup}
\alias{gtkWindowGetHandleMenubarAccel}
\alias{gtkWindowGetHideOnClose}
\alias{gtkWindowGetIconName}
\alias{gtkWindowGetMnemonicsVisible}
\alias{gtkWindowGetModal}
\alias{gtkWindowGetResizable}
\alias{gtkWindowGetTitle}
\alias{gtkWindowGetTitlebar}
\alias{gtkWindowGetTransientFor}
\alias{gtkWindowHasGroup}
\alias{gtkWindowIsActive}
\alias{gtkWindowIsFullscreen}
\alias{gtkWindowIsMaximized}
\alias{gtkWindowIsSuspended}
\alias{gtkWindowMaximize}
\alias{gtkWindowMinimize}
\alias{gtkWindowPresent}
\alias{gtkWindowPresentWithTime}
\alias{gtkWindowSetApplication}
\alias{gtkWindowSetChild}
\alias{gtkWindowSetDecorated}
\alias{gtkWindowSetDefaultSize}
\alias{gtkWindowSetDefaultWidget}
\alias{gtkWindowSetDeletable}
\alias{gtkWindowSetDestroyWithParent}
\alias{gtkWindowSetDisplay}
\alias{gtkWindowSetFocus}
\alias{gtkWindowSetFocusVisible}
\alias{gtkWindowSetGravity}
\alias{gtkWindowSetHandleMenubarAccel}
\alias{gtkWindowSetHideOnClose}
\alias{gtkWindowSetIconName}
\alias{gtkWindowSetMnemonicsVisible}
\alias{gtkWindowSetModal}
\alias{gtkWindowSetResizable}
\alias{gtkWindowSetStartupId}
\alias{gtkWindowSetTitle}
\alias{gtkWindowSetTitlebar}
\alias{gtkWindowSetTransientFor}
\alias{gtkWindowUnfullscreen}
\alias{gtkWindowUnmaximize}
\alias{gtkWindowUnminimize}
\alias{gtkWindowControlsNew}
\alias{gtkWindowControlsGetDecorationLayout}
\alias{gtkWindowControlsGetEmpty}
\alias{gtkWindowControlsGetSide}
\alias{gtkWindowControlsGetUseNativeControls}
\alias{gtkWindowControlsSetDecorationLayout}
\alias{gtkWindowControlsSetSide}
\alias{gtkWindowControlsSetUseNativeControls}
\alias{gtkWindowGroupNew}
\alias{gtkWindowGroupAddWindow}
\alias{gtkWindowGroupListWindows}
\alias{gtkWindowGroupRemoveWindow}
\alias{gtkWindowHandleNew}
\alias{gtkWindowHandleGetChild}
\alias{gtkWindowHandleSetChild}
\title{gtkWindowNew}
\usage{
gtkWindowNew()

gtkWindowGetDefaultIconName()

gtkWindowGetToplevels()

gtkWindowListToplevels()

gtkWindowSetAutoStartupNotification(setting)

gtkWindowSetDefaultIconName(name)

gtkWindowSetInteractiveDebugging(enable)

gtkWindowClose(window)

gtkWindowDestroy(window)

gtkWindowFullscreen(window)

gtkWindowFullscreenOnMonitor(window, monitor)

gtkWindowGetApplication(window)

gtkWindowGetChild(window)

gtkWindowGetDecorated(window)

gtkWindowGetDefaultSize(window)

gtkWindowGetDefaultWidget(window)

gtkWindowGetDeletable(window)

gtkWindowGetDestroyWithParent(window)

gtkWindowGetFocus(window)

gtkWindowGetFocusVisible(window)

gtkWindowGetGravity(window)

gtkWindowGetGroup(window)

gtkWindowGetHandleMenubarAccel(window)

gtkWindowGetHideOnClose(window)

gtkWindowGetIconName(window)

gtkWindowGetMnemonicsVisible(window)

gtkWindowGetModal(window)

gtkWindowGetResizable(window)

gtkWindowGetTitle(window)

gtkWindowGetTitlebar(window)

gtkWindowGetTransientFor(window)

gtkWindowHasGroup(window)

gtkWindowIsActive(window)

gtkWindowIsFullscreen(window)

gtkWindowIsMaximized(window)

gtkWindowIsSuspended(window)

gtkWindowMaximize(window)

gtkWindowMinimize(window)

gtkWindowPresent(window)

gtkWindowPresentWithTime(window, timestamp)

gtkWindowSetApplication(window, application)

gtkWindowSetChild(window, child)

gtkWindowSetDecorated(window, setting)

gtkWindowSetDefaultSize(window, width, height)

gtkWindowSetDefaultWidget(window, default_widget)

gtkWindowSetDeletable(window, setting)

gtkWindowSetDestroyWithParent(window, setting)

gtkWindowSetDisplay(window, display)

gtkWindowSetFocus(window, focus)

gtkWindowSetFocusVisible(window, setting)

gtkWindowSetGravity(window, gravity)

gtkWindowSetHandleMenubarAccel(window, handle_menubar_accel)

gtkWindowSetHideOnClose(window, setting)

gtkWindowSetIconName(window, name)

gtkWindowSetMnemonicsVisible(window, setting)

gtkWindowSetModal(window, modal)

gtkWindowSetResizable(window, resizable)

gtkWindowSetStartupId(window, startup_id)

gtkWindowSetTitle(window, title)

gtkWindowSetTitlebar(window, titlebar)

gtkWindowSetTransientFor(window, parent)

gtkWindowUnfullscreen(window)

gtkWindowUnmaximize(window)

gtkWindowUnminimize(window)

gtkWindowControlsNew(side)

gtkWindowControlsGetDecorationLayout(self)

gtkWindowControlsGetEmpty(self)

gtkWindowControlsGetSide(self)

gtkWindowControlsGetUseNativeControls(self)

gtkWindowControlsSetDecorationLayout(self, layout)

gtkWindowControlsSetSide(self, side)

gtkWindowControlsSetUseNativeControls(self, setting)

gtkWindowGroupNew()

gtkWindowGroupAddWindow(window_group, window)

gtkWindowGroupListWindows(window_group)

gtkWindowGroupRemoveWindow(window_group, window)

gtkWindowHandleNew()

gtkWindowHandleGetChild(self)

gtkWindowHandleSetChild(self, child)
}
\arguments{
\item{setting}{gboolean}

\item{name}{utf8}

\item{enable}{gboolean}

\item{window}{Window}

\item{monitor}{Gdk.Monitor}

\item{timestamp}{guint32}

\item{application}{Application}

\item{child}{Widget}

\item{width}{gint}

\item{height}{gint}

\item{default_widget}{Widget}

\item{display}{Gdk.Display}

\item{focus}{Widget}

\item{gravity}{WindowGravity}

\item{handle_menubar_accel}{gboolean}

\item{modal}{gboolean}

\item{resizable}{gboolean}

\item{startup_id}{utf8}

\item{title}{utf8}

\item{titlebar}{Widget}

\item{parent}{Window}

\item{side}{PackType}

\item{self}{WindowHandle}

\item{layout}{utf8}

\item{window_group}{WindowGroup}
}
\value{
Widget

utf8

Gio.ListModel

GLib.List

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Application

Widget

gboolean

Return value from C function

Widget

gboolean

gboolean

Widget

gboolean

WindowGravity

WindowGroup

gboolean

gboolean

utf8

gboolean

gboolean

gboolean

utf8

Widget

Window

gboolean

gboolean

gboolean

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Widget

utf8

gboolean

PackType

gboolean

Return value from C function

Return value from C function

Return value from C function

WindowGroup

Return value from C function

GLib.List

Return value from C function

Widget

Widget

Return value from C function
}
\description{
gtkWindowNew

gtkWindowGetDefaultIconName

gtkWindowGetToplevels

gtkWindowListToplevels

gtkWindowSetAutoStartupNotification

gtkWindowSetDefaultIconName

gtkWindowSetInteractiveDebugging

gtkWindowClose

gtkWindowDestroy

gtkWindowFullscreen

gtkWindowFullscreenOnMonitor

gtkWindowGetApplication

gtkWindowGetChild

gtkWindowGetDecorated

gtkWindowGetDefaultSize

gtkWindowGetDefaultWidget

gtkWindowGetDeletable

gtkWindowGetDestroyWithParent

gtkWindowGetFocus

gtkWindowGetFocusVisible

gtkWindowGetGravity

gtkWindowGetGroup

gtkWindowGetHandleMenubarAccel

gtkWindowGetHideOnClose

gtkWindowGetIconName

gtkWindowGetMnemonicsVisible

gtkWindowGetModal

gtkWindowGetResizable

gtkWindowGetTitle

gtkWindowGetTitlebar

gtkWindowGetTransientFor

gtkWindowHasGroup

gtkWindowIsActive

gtkWindowIsFullscreen

gtkWindowIsMaximized

gtkWindowIsSuspended

gtkWindowMaximize

gtkWindowMinimize

gtkWindowPresent

gtkWindowPresentWithTime

gtkWindowSetApplication

gtkWindowSetChild

gtkWindowSetDecorated

gtkWindowSetDefaultSize

gtkWindowSetDefaultWidget

gtkWindowSetDeletable

gtkWindowSetDestroyWithParent

gtkWindowSetDisplay

gtkWindowSetFocus

gtkWindowSetFocusVisible

gtkWindowSetGravity

gtkWindowSetHandleMenubarAccel

gtkWindowSetHideOnClose

gtkWindowSetIconName

gtkWindowSetMnemonicsVisible

gtkWindowSetModal

gtkWindowSetResizable

gtkWindowSetStartupId

gtkWindowSetTitle

gtkWindowSetTitlebar

gtkWindowSetTransientFor

gtkWindowUnfullscreen

gtkWindowUnmaximize

gtkWindowUnminimize

gtkWindowControlsNew

gtkWindowControlsGetDecorationLayout

gtkWindowControlsGetEmpty

gtkWindowControlsGetSide

gtkWindowControlsGetUseNativeControls

gtkWindowControlsSetDecorationLayout

gtkWindowControlsSetSide

gtkWindowControlsSetUseNativeControls

gtkWindowGroupNew

gtkWindowGroupAddWindow

gtkWindowGroupListWindows

gtkWindowGroupRemoveWindow

gtkWindowHandleNew

gtkWindowHandleGetChild

gtkWindowHandleSetChild
}
