% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gtkMainIteration}
\alias{gtkMainIteration}
\title{Process GTK Events}
\usage{
gtkMainIteration()
}
\value{
Logical indicating if events were processed
}
\description{
Manually process pending GTK events. Useful on Windows where automatic
event processing may not work as expected.
}
