% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gtkMainIterationDo}
\alias{gtkMainIterationDo}
\title{Process All GTK Events}
\usage{
gtkMainIterationDo(blocking = FALSE)
}
\arguments{
\item{blocking}{Whether to block until an event occurs}
}
\value{
Number of events processed
}
\description{
Process all pending GTK events in a loop.
}
