% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gtkMessageDialogNew}
\alias{gtkMessageDialogNew}
\title{Create a GtkMessageDialog}
\usage{
gtkMessageDialogNew(parent, flags, message_type, buttons_type, message)
}
\arguments{
\item{parent}{The parent GtkWindow (pointer)}

\item{flags}{Dialog flags (e.g., MODAL = 1L)}

\item{message_type}{Icon/Message type (e.g., QUESTION = 2L)}

\item{buttons_type}{Button set (e.g., YES_NO = 4L)}

\item{message}{The text message to display}
}
\description{
Create a GtkMessageDialog
}
