% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gtkSetAppIcon}
\alias{gtkSetAppIcon}
\title{Set application icon}
\usage{
gtkSetAppIcon(icon_path)
}
\arguments{
\item{icon_path}{Path to icon file (.png, .icns, or .ico)}
}
\description{
Sets a custom application icon. On macOS, sets the Dock icon.
On Linux, sets the default window icon. Windows support TBD.
}
