% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gtkStringListFromVector}
\alias{gtkStringListFromVector}
\title{Create GtkStringList from R character vector}
\usage{
gtkStringListFromVector(strings)
}
\arguments{
\item{strings}{Character vector}
}
\value{
External pointer to GtkStringList
}
\description{
Create GtkStringList from R character vector
}
