% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gtkWindowAddCloseShortcut}
\alias{gtkWindowAddCloseShortcut}
\title{Add keyboard shortcut to close window}
\usage{
gtkWindowAddCloseShortcut(window)
}
\arguments{
\item{window}{External pointer to GtkWindow}
}
\description{
Adds Cmd+W (macOS) or Ctrl+W (Windows/Linux) to close the window.
}
