% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pango_autogen.R
\name{pangoAttrFontDescNew}
\alias{pangoAttrFontDescNew}
\alias{pangoAttrFontFeaturesNew}
\alias{pangoAttrIteratorCopy}
\alias{pangoAttrIteratorDestroy}
\alias{pangoAttrIteratorGet}
\alias{pangoAttrIteratorGetAttrs}
\alias{pangoAttrIteratorGetFont}
\alias{pangoAttrIteratorNext}
\alias{pangoAttrIteratorRange}
\alias{pangoAttrLanguageNew}
\alias{pangoAttrListNew}
\alias{pangoAttrListChange}
\alias{pangoAttrListCopy}
\alias{pangoAttrListEqual}
\alias{pangoAttrListFilter}
\alias{pangoAttrListGetAttributes}
\alias{pangoAttrListGetIterator}
\alias{pangoAttrListInsert}
\alias{pangoAttrListInsertBefore}
\alias{pangoAttrListRef}
\alias{pangoAttrListSplice}
\alias{pangoAttrListToString}
\alias{pangoAttrListUnref}
\alias{pangoAttrListUpdate}
\alias{pangoAttrListFromString}
\alias{pangoAttrShapeNew}
\alias{pangoAttrShapeNewWithData}
\alias{pangoAttrSizeNew}
\alias{pangoAttrSizeNewAbsolute}
\alias{pangoAttrTypeGetName}
\alias{pangoAttrTypeRegister}
\alias{pangoAttributeAsColor}
\alias{pangoAttributeAsFloat}
\alias{pangoAttributeAsFontDesc}
\alias{pangoAttributeAsFontFeatures}
\alias{pangoAttributeAsInt}
\alias{pangoAttributeAsLanguage}
\alias{pangoAttributeAsShape}
\alias{pangoAttributeAsSize}
\alias{pangoAttributeAsString}
\alias{pangoAttributeCopy}
\alias{pangoAttributeDestroy}
\alias{pangoAttributeEqual}
\alias{pangoAttributeInit}
\alias{pangoBidiTypeForUnichar}
\alias{pangoColorCopy}
\alias{pangoColorFree}
\alias{pangoColorParse}
\alias{pangoColorParseWithAlpha}
\alias{pangoColorToString}
\alias{pangoContextNew}
\alias{pangoContextChanged}
\alias{pangoContextGetBaseDir}
\alias{pangoContextGetBaseGravity}
\alias{pangoContextGetFontDescription}
\alias{pangoContextGetFontMap}
\alias{pangoContextGetGravity}
\alias{pangoContextGetGravityHint}
\alias{pangoContextGetLanguage}
\alias{pangoContextGetMatrix}
\alias{pangoContextGetMetrics}
\alias{pangoContextGetRoundGlyphPositions}
\alias{pangoContextGetSerial}
\alias{pangoContextListFamilies}
\alias{pangoContextLoadFont}
\alias{pangoContextLoadFontset}
\alias{pangoContextSetBaseDir}
\alias{pangoContextSetBaseGravity}
\alias{pangoContextSetFontDescription}
\alias{pangoContextSetFontMap}
\alias{pangoContextSetGravityHint}
\alias{pangoContextSetLanguage}
\alias{pangoContextSetMatrix}
\alias{pangoContextSetRoundGlyphPositions}
\alias{pangoCoverageNew}
\alias{pangoCoverageFromBytes}
\alias{pangoCoverageCopy}
\alias{pangoCoverageGet}
\alias{pangoCoverageMax}
\alias{pangoCoverageRef}
\alias{pangoCoverageSet}
\alias{pangoCoverageToBytes}
\alias{pangoCoverageUnref}
\alias{pangoFontDescriptionsFree}
\alias{pangoFontDeserialize}
\alias{pangoFontDescribe}
\alias{pangoFontDescribeWithAbsoluteSize}
\alias{pangoFontGetCoverage}
\alias{pangoFontGetFace}
\alias{pangoFontGetFeatures}
\alias{pangoFontGetFontMap}
\alias{pangoFontGetGlyphExtents}
\alias{pangoFontGetLanguages}
\alias{pangoFontGetMetrics}
\alias{pangoFontHasChar}
\alias{pangoFontSerialize}
\alias{pangoFontDescriptionNew}
\alias{pangoFontDescriptionBetterMatch}
\alias{pangoFontDescriptionCopy}
\alias{pangoFontDescriptionCopyStatic}
\alias{pangoFontDescriptionEqual}
\alias{pangoFontDescriptionFree}
\alias{pangoFontDescriptionGetFamily}
\alias{pangoFontDescriptionGetGravity}
\alias{pangoFontDescriptionGetSetFields}
\alias{pangoFontDescriptionGetSize}
\alias{pangoFontDescriptionGetSizeIsAbsolute}
\alias{pangoFontDescriptionGetStretch}
\alias{pangoFontDescriptionGetStyle}
\alias{pangoFontDescriptionGetVariant}
\alias{pangoFontDescriptionGetVariations}
\alias{pangoFontDescriptionGetWeight}
\alias{pangoFontDescriptionHash}
\alias{pangoFontDescriptionMerge}
\alias{pangoFontDescriptionMergeStatic}
\alias{pangoFontDescriptionSetAbsoluteSize}
\alias{pangoFontDescriptionSetFamily}
\alias{pangoFontDescriptionSetFamilyStatic}
\alias{pangoFontDescriptionSetGravity}
\alias{pangoFontDescriptionSetSize}
\alias{pangoFontDescriptionSetStretch}
\alias{pangoFontDescriptionSetStyle}
\alias{pangoFontDescriptionSetVariant}
\alias{pangoFontDescriptionSetVariations}
\alias{pangoFontDescriptionSetVariationsStatic}
\alias{pangoFontDescriptionSetWeight}
\alias{pangoFontDescriptionToFilename}
\alias{pangoFontDescriptionToString}
\alias{pangoFontDescriptionUnsetFields}
\alias{pangoFontDescriptionFromString}
\alias{pangoFontFaceDescribe}
\alias{pangoFontFaceGetFaceName}
\alias{pangoFontFaceGetFamily}
\alias{pangoFontFaceIsSynthesized}
\alias{pangoFontFaceListSizes}
\alias{pangoFontFamilyGetFace}
\alias{pangoFontFamilyGetName}
\alias{pangoFontFamilyIsMonospace}
\alias{pangoFontFamilyIsVariable}
\alias{pangoFontFamilyListFaces}
\alias{pangoFontMapChanged}
\alias{pangoFontMapCreateContext}
\alias{pangoFontMapGetFamily}
\alias{pangoFontMapGetSerial}
\alias{pangoFontMapListFamilies}
\alias{pangoFontMapLoadFont}
\alias{pangoFontMapLoadFontset}
\alias{pangoFontMetricsGetApproximateCharWidth}
\alias{pangoFontMetricsGetApproximateDigitWidth}
\alias{pangoFontMetricsGetAscent}
\alias{pangoFontMetricsGetDescent}
\alias{pangoFontMetricsGetHeight}
\alias{pangoFontMetricsGetStrikethroughPosition}
\alias{pangoFontMetricsGetStrikethroughThickness}
\alias{pangoFontMetricsGetUnderlinePosition}
\alias{pangoFontMetricsGetUnderlineThickness}
\alias{pangoFontMetricsRef}
\alias{pangoFontMetricsUnref}
\alias{pangoFontsetForeach}
\alias{pangoFontsetGetFont}
\alias{pangoFontsetGetMetrics}
\alias{pangoFontsetSimpleNew}
\alias{pangoFontsetSimpleAppend}
\alias{pangoFontsetSimpleSize}
\alias{pangoGlyphItemApplyAttrs}
\alias{pangoGlyphItemCopy}
\alias{pangoGlyphItemFree}
\alias{pangoGlyphItemGetLogicalWidths}
\alias{pangoGlyphItemLetterSpace}
\alias{pangoGlyphItemSplit}
\alias{pangoGlyphItemIterCopy}
\alias{pangoGlyphItemIterFree}
\alias{pangoGlyphItemIterInitEnd}
\alias{pangoGlyphItemIterInitStart}
\alias{pangoGlyphItemIterNextCluster}
\alias{pangoGlyphItemIterPrevCluster}
\alias{pangoGlyphStringNew}
\alias{pangoGlyphStringCopy}
\alias{pangoGlyphStringExtents}
\alias{pangoGlyphStringExtentsRange}
\alias{pangoGlyphStringFree}
\alias{pangoGlyphStringGetLogicalWidths}
\alias{pangoGlyphStringGetWidth}
\alias{pangoGlyphStringIndexToX}
\alias{pangoGlyphStringIndexToXFull}
\alias{pangoGlyphStringSetSize}
\alias{pangoGlyphStringXToIndex}
\alias{pangoGravityGetForMatrix}
\alias{pangoGravityGetForScript}
\alias{pangoGravityGetForScriptAndWidth}
\alias{pangoGravityToRotation}
\alias{pangoItemNew}
\alias{pangoItemApplyAttrs}
\alias{pangoItemCopy}
\alias{pangoItemFree}
\alias{pangoItemSplit}
\alias{pangoLanguageGetSampleString}
\alias{pangoLanguageGetScripts}
\alias{pangoLanguageIncludesScript}
\alias{pangoLanguageMatches}
\alias{pangoLanguageToString}
\alias{pangoLanguageFromString}
\alias{pangoLanguageGetDefault}
\alias{pangoLanguageGetPreferred}
\alias{pangoLayoutNew}
\alias{pangoLayoutDeserialize}
\alias{pangoLayoutContextChanged}
\alias{pangoLayoutCopy}
\alias{pangoLayoutGetAlignment}
\alias{pangoLayoutGetAttributes}
\alias{pangoLayoutGetAutoDir}
\alias{pangoLayoutGetBaseline}
\alias{pangoLayoutGetCaretPos}
\alias{pangoLayoutGetCharacterCount}
\alias{pangoLayoutGetContext}
\alias{pangoLayoutGetCursorPos}
\alias{pangoLayoutGetDirection}
\alias{pangoLayoutGetEllipsize}
\alias{pangoLayoutGetExtents}
\alias{pangoLayoutGetFontDescription}
\alias{pangoLayoutGetHeight}
\alias{pangoLayoutGetIndent}
\alias{pangoLayoutGetIter}
\alias{pangoLayoutGetJustify}
\alias{pangoLayoutGetJustifyLastLine}
\alias{pangoLayoutGetLine}
\alias{pangoLayoutGetLineCount}
\alias{pangoLayoutGetLineReadonly}
\alias{pangoLayoutGetLineSpacing}
\alias{pangoLayoutGetLines}
\alias{pangoLayoutGetLinesReadonly}
\alias{pangoLayoutGetLogAttrs}
\alias{pangoLayoutGetLogAttrsReadonly}
\alias{pangoLayoutGetPixelExtents}
\alias{pangoLayoutGetPixelSize}
\alias{pangoLayoutGetSerial}
\alias{pangoLayoutGetSingleParagraphMode}
\alias{pangoLayoutGetSize}
\alias{pangoLayoutGetSpacing}
\alias{pangoLayoutGetTabs}
\alias{pangoLayoutGetText}
\alias{pangoLayoutGetUnknownGlyphsCount}
\alias{pangoLayoutGetWidth}
\alias{pangoLayoutGetWrap}
\alias{pangoLayoutIndexToLineX}
\alias{pangoLayoutIndexToPos}
\alias{pangoLayoutIsEllipsized}
\alias{pangoLayoutIsWrapped}
\alias{pangoLayoutMoveCursorVisually}
\alias{pangoLayoutSerialize}
\alias{pangoLayoutSetAlignment}
\alias{pangoLayoutSetAttributes}
\alias{pangoLayoutSetAutoDir}
\alias{pangoLayoutSetEllipsize}
\alias{pangoLayoutSetFontDescription}
\alias{pangoLayoutSetHeight}
\alias{pangoLayoutSetIndent}
\alias{pangoLayoutSetJustify}
\alias{pangoLayoutSetJustifyLastLine}
\alias{pangoLayoutSetLineSpacing}
\alias{pangoLayoutSetMarkup}
\alias{pangoLayoutSetMarkupWithAccel}
\alias{pangoLayoutSetSingleParagraphMode}
\alias{pangoLayoutSetSpacing}
\alias{pangoLayoutSetTabs}
\alias{pangoLayoutSetText}
\alias{pangoLayoutSetWidth}
\alias{pangoLayoutSetWrap}
\alias{pangoLayoutWriteToFile}
\alias{pangoLayoutXyToIndex}
\alias{pangoLayoutDeserializeErrorQuark}
\alias{pangoLayoutIterAtLastLine}
\alias{pangoLayoutIterCopy}
\alias{pangoLayoutIterFree}
\alias{pangoLayoutIterGetBaseline}
\alias{pangoLayoutIterGetCharExtents}
\alias{pangoLayoutIterGetClusterExtents}
\alias{pangoLayoutIterGetIndex}
\alias{pangoLayoutIterGetLayout}
\alias{pangoLayoutIterGetLayoutExtents}
\alias{pangoLayoutIterGetLine}
\alias{pangoLayoutIterGetLineExtents}
\alias{pangoLayoutIterGetLineReadonly}
\alias{pangoLayoutIterGetLineYrange}
\alias{pangoLayoutIterGetRun}
\alias{pangoLayoutIterGetRunBaseline}
\alias{pangoLayoutIterGetRunExtents}
\alias{pangoLayoutIterGetRunReadonly}
\alias{pangoLayoutIterNextChar}
\alias{pangoLayoutIterNextCluster}
\alias{pangoLayoutIterNextLine}
\alias{pangoLayoutIterNextRun}
\alias{pangoLayoutLineGetExtents}
\alias{pangoLayoutLineGetHeight}
\alias{pangoLayoutLineGetLength}
\alias{pangoLayoutLineGetPixelExtents}
\alias{pangoLayoutLineGetResolvedDirection}
\alias{pangoLayoutLineGetStartIndex}
\alias{pangoLayoutLineGetXRanges}
\alias{pangoLayoutLineIndexToX}
\alias{pangoLayoutLineIsParagraphStart}
\alias{pangoLayoutLineRef}
\alias{pangoLayoutLineUnref}
\alias{pangoLayoutLineXToIndex}
\alias{pangoMatrixConcat}
\alias{pangoMatrixCopy}
\alias{pangoMatrixFree}
\alias{pangoMatrixGetFontScaleFactor}
\alias{pangoMatrixGetFontScaleFactors}
\alias{pangoMatrixGetSlantRatio}
\alias{pangoMatrixRotate}
\alias{pangoMatrixScale}
\alias{pangoMatrixTransformDistance}
\alias{pangoMatrixTransformPixelRectangle}
\alias{pangoMatrixTransformPoint}
\alias{pangoMatrixTransformRectangle}
\alias{pangoMatrixTranslate}
\alias{pangoRendererActivate}
\alias{pangoRendererDeactivate}
\alias{pangoRendererDrawErrorUnderline}
\alias{pangoRendererDrawGlyph}
\alias{pangoRendererDrawGlyphItem}
\alias{pangoRendererDrawGlyphs}
\alias{pangoRendererDrawLayout}
\alias{pangoRendererDrawLayoutLine}
\alias{pangoRendererDrawRectangle}
\alias{pangoRendererDrawTrapezoid}
\alias{pangoRendererGetAlpha}
\alias{pangoRendererGetColor}
\alias{pangoRendererGetLayout}
\alias{pangoRendererGetLayoutLine}
\alias{pangoRendererGetMatrix}
\alias{pangoRendererPartChanged}
\alias{pangoRendererSetAlpha}
\alias{pangoRendererSetColor}
\alias{pangoRendererSetMatrix}
\alias{pangoScriptForUnichar}
\alias{pangoScriptGetSampleLanguage}
\alias{pangoScriptIterNew}
\alias{pangoScriptIterFree}
\alias{pangoScriptIterGetRange}
\alias{pangoScriptIterNext}
\alias{pangoTabArrayNew}
\alias{pangoTabArrayCopy}
\alias{pangoTabArrayFree}
\alias{pangoTabArrayGetDecimalPoint}
\alias{pangoTabArrayGetPositionsInPixels}
\alias{pangoTabArrayGetSize}
\alias{pangoTabArrayGetTab}
\alias{pangoTabArrayGetTabs}
\alias{pangoTabArrayResize}
\alias{pangoTabArraySetDecimalPoint}
\alias{pangoTabArraySetPositionsInPixels}
\alias{pangoTabArraySetTab}
\alias{pangoTabArraySort}
\alias{pangoTabArrayToString}
\alias{pangoTabArrayFromString}
\alias{pangoAttrAllowBreaksNew}
\alias{pangoAttrBackgroundAlphaNew}
\alias{pangoAttrBackgroundNew}
\alias{pangoAttrBaselineShiftNew}
\alias{pangoAttrBreak}
\alias{pangoAttrFallbackNew}
\alias{pangoAttrFamilyNew}
\alias{pangoAttrFontScaleNew}
\alias{pangoAttrForegroundAlphaNew}
\alias{pangoAttrForegroundNew}
\alias{pangoAttrGravityHintNew}
\alias{pangoAttrGravityNew}
\alias{pangoAttrInsertHyphensNew}
\alias{pangoAttrLetterSpacingNew}
\alias{pangoAttrLineHeightNew}
\alias{pangoAttrLineHeightNewAbsolute}
\alias{pangoAttrOverlineColorNew}
\alias{pangoAttrOverlineNew}
\alias{pangoAttrRiseNew}
\alias{pangoAttrScaleNew}
\alias{pangoAttrSentenceNew}
\alias{pangoAttrShowNew}
\alias{pangoAttrStretchNew}
\alias{pangoAttrStrikethroughColorNew}
\alias{pangoAttrStrikethroughNew}
\alias{pangoAttrStyleNew}
\alias{pangoAttrTextTransformNew}
\alias{pangoAttrUnderlineColorNew}
\alias{pangoAttrUnderlineNew}
\alias{pangoAttrVariantNew}
\alias{pangoAttrWeightNew}
\alias{pangoAttrWordNew}
\alias{pangoBreak}
\alias{pangoDefaultBreak}
\alias{pangoExtentsToPixels}
\alias{pangoFindBaseDir}
\alias{pangoFindParagraphBoundary}
\alias{pangoGetLogAttrs}
\alias{pangoGetMirrorChar}
\alias{pangoIsZeroWidth}
\alias{pangoItemize}
\alias{pangoItemizeWithBaseDir}
\alias{pangoLog2visGetEmbeddingLevels}
\alias{pangoMarkupParserFinish}
\alias{pangoMarkupParserNew}
\alias{pangoParseEnum}
\alias{pangoParseMarkup}
\alias{pangoParseStretch}
\alias{pangoParseStyle}
\alias{pangoParseVariant}
\alias{pangoParseWeight}
\alias{pangoQuantizeLineGeometry}
\alias{pangoReadLine}
\alias{pangoReorderItems}
\alias{pangoScanInt}
\alias{pangoScanString}
\alias{pangoScanWord}
\alias{pangoShape}
\alias{pangoShapeFull}
\alias{pangoShapeItem}
\alias{pangoShapeWithFlags}
\alias{pangoSkipSpace}
\alias{pangoSplitFileList}
\alias{pangoTailorBreak}
\alias{pangoTrimString}
\alias{pangoUnicharDirection}
\alias{pangoUnitsFromDouble}
\alias{pangoUnitsToDouble}
\alias{pangoVersion}
\alias{pangoVersionCheck}
\alias{pangoVersionString}
\title{pangoAttrFontDescNew}
\usage{
pangoAttrFontDescNew(desc)

pangoAttrFontFeaturesNew(features)

pangoAttrIteratorCopy(iterator)

pangoAttrIteratorDestroy(iterator)

pangoAttrIteratorGet(iterator, type)

pangoAttrIteratorGetAttrs(iterator)

pangoAttrIteratorGetFont(iterator, desc)

pangoAttrIteratorNext(iterator)

pangoAttrIteratorRange(iterator)

pangoAttrLanguageNew(language)

pangoAttrListNew()

pangoAttrListChange(list, attr)

pangoAttrListCopy(list)

pangoAttrListEqual(list, other_list)

pangoAttrListFilter(list, func)

pangoAttrListGetAttributes(list)

pangoAttrListGetIterator(list)

pangoAttrListInsert(list, attr)

pangoAttrListInsertBefore(list, attr)

pangoAttrListRef(list)

pangoAttrListSplice(list, other, pos, len)

pangoAttrListToString(list)

pangoAttrListUnref(list)

pangoAttrListUpdate(list, pos, remove, add)

pangoAttrListFromString(text)

pangoAttrShapeNew(ink_rect, logical_rect)

pangoAttrShapeNewWithData(ink_rect, logical_rect, data, copy_func)

pangoAttrSizeNew(size)

pangoAttrSizeNewAbsolute(size)

pangoAttrTypeGetName(type)

pangoAttrTypeRegister(name)

pangoAttributeAsColor(attr)

pangoAttributeAsFloat(attr)

pangoAttributeAsFontDesc(attr)

pangoAttributeAsFontFeatures(attr)

pangoAttributeAsInt(attr)

pangoAttributeAsLanguage(attr)

pangoAttributeAsShape(attr)

pangoAttributeAsSize(attr)

pangoAttributeAsString(attr)

pangoAttributeCopy(attr)

pangoAttributeDestroy(attr)

pangoAttributeEqual(attr1, attr2)

pangoAttributeInit(attr, klass)

pangoBidiTypeForUnichar(ch)

pangoColorCopy(src)

pangoColorFree(color)

pangoColorParse(color, spec)

pangoColorParseWithAlpha(color, spec)

pangoColorToString(color)

pangoContextNew()

pangoContextChanged(context)

pangoContextGetBaseDir(context)

pangoContextGetBaseGravity(context)

pangoContextGetFontDescription(context)

pangoContextGetFontMap(context)

pangoContextGetGravity(context)

pangoContextGetGravityHint(context)

pangoContextGetLanguage(context)

pangoContextGetMatrix(context)

pangoContextGetMetrics(context, desc, language)

pangoContextGetRoundGlyphPositions(context)

pangoContextGetSerial(context)

pangoContextListFamilies(context)

pangoContextLoadFont(context, desc)

pangoContextLoadFontset(context, desc, language)

pangoContextSetBaseDir(context, direction)

pangoContextSetBaseGravity(context, gravity)

pangoContextSetFontDescription(context, desc)

pangoContextSetFontMap(context, font_map)

pangoContextSetGravityHint(context, hint)

pangoContextSetLanguage(context, language)

pangoContextSetMatrix(context, matrix)

pangoContextSetRoundGlyphPositions(context, round_positions)

pangoCoverageNew()

pangoCoverageFromBytes(bytes, n_bytes)

pangoCoverageCopy(coverage)

pangoCoverageGet(coverage, index_)

pangoCoverageMax(coverage, other)

pangoCoverageRef(coverage)

pangoCoverageSet(coverage, index_, level)

pangoCoverageToBytes(coverage)

pangoCoverageUnref(coverage)

pangoFontDescriptionsFree(descs, n_descs)

pangoFontDeserialize(context, bytes)

pangoFontDescribe(font)

pangoFontDescribeWithAbsoluteSize(font)

pangoFontGetCoverage(font, language)

pangoFontGetFace(font)

pangoFontGetFeatures(font, len)

pangoFontGetFontMap(font)

pangoFontGetGlyphExtents(font, glyph)

pangoFontGetLanguages(font)

pangoFontGetMetrics(font, language)

pangoFontHasChar(font, wc)

pangoFontSerialize(font)

pangoFontDescriptionNew()

pangoFontDescriptionBetterMatch(desc, old_match, new_match)

pangoFontDescriptionCopy(desc)

pangoFontDescriptionCopyStatic(desc)

pangoFontDescriptionEqual(desc1, desc2)

pangoFontDescriptionFree(desc)

pangoFontDescriptionGetFamily(desc)

pangoFontDescriptionGetGravity(desc)

pangoFontDescriptionGetSetFields(desc)

pangoFontDescriptionGetSize(desc)

pangoFontDescriptionGetSizeIsAbsolute(desc)

pangoFontDescriptionGetStretch(desc)

pangoFontDescriptionGetStyle(desc)

pangoFontDescriptionGetVariant(desc)

pangoFontDescriptionGetVariations(desc)

pangoFontDescriptionGetWeight(desc)

pangoFontDescriptionHash(desc)

pangoFontDescriptionMerge(desc, desc_to_merge, replace_existing)

pangoFontDescriptionMergeStatic(desc, desc_to_merge, replace_existing)

pangoFontDescriptionSetAbsoluteSize(desc, size)

pangoFontDescriptionSetFamily(desc, family)

pangoFontDescriptionSetFamilyStatic(desc, family)

pangoFontDescriptionSetGravity(desc, gravity)

pangoFontDescriptionSetSize(desc, size)

pangoFontDescriptionSetStretch(desc, stretch)

pangoFontDescriptionSetStyle(desc, style)

pangoFontDescriptionSetVariant(desc, variant)

pangoFontDescriptionSetVariations(desc, variations)

pangoFontDescriptionSetVariationsStatic(desc, variations)

pangoFontDescriptionSetWeight(desc, weight)

pangoFontDescriptionToFilename(desc)

pangoFontDescriptionToString(desc)

pangoFontDescriptionUnsetFields(desc, to_unset)

pangoFontDescriptionFromString(str)

pangoFontFaceDescribe(face)

pangoFontFaceGetFaceName(face)

pangoFontFaceGetFamily(face)

pangoFontFaceIsSynthesized(face)

pangoFontFaceListSizes(face)

pangoFontFamilyGetFace(family, name)

pangoFontFamilyGetName(family)

pangoFontFamilyIsMonospace(family)

pangoFontFamilyIsVariable(family)

pangoFontFamilyListFaces(family)

pangoFontMapChanged(fontmap)

pangoFontMapCreateContext(fontmap)

pangoFontMapGetFamily(fontmap, name)

pangoFontMapGetSerial(fontmap)

pangoFontMapListFamilies(fontmap)

pangoFontMapLoadFont(fontmap, context, desc)

pangoFontMapLoadFontset(fontmap, context, desc, language)

pangoFontMetricsGetApproximateCharWidth(metrics)

pangoFontMetricsGetApproximateDigitWidth(metrics)

pangoFontMetricsGetAscent(metrics)

pangoFontMetricsGetDescent(metrics)

pangoFontMetricsGetHeight(metrics)

pangoFontMetricsGetStrikethroughPosition(metrics)

pangoFontMetricsGetStrikethroughThickness(metrics)

pangoFontMetricsGetUnderlinePosition(metrics)

pangoFontMetricsGetUnderlineThickness(metrics)

pangoFontMetricsRef(metrics)

pangoFontMetricsUnref(metrics)

pangoFontsetForeach(fontset, func)

pangoFontsetGetFont(fontset, wc)

pangoFontsetGetMetrics(fontset)

pangoFontsetSimpleNew(language)

pangoFontsetSimpleAppend(fontset, font)

pangoFontsetSimpleSize(fontset)

pangoGlyphItemApplyAttrs(glyph_item, text, list)

pangoGlyphItemCopy(orig)

pangoGlyphItemFree(glyph_item)

pangoGlyphItemGetLogicalWidths(glyph_item, text)

pangoGlyphItemLetterSpace(glyph_item, text, log_attrs, letter_spacing)

pangoGlyphItemSplit(orig, text, split_index)

pangoGlyphItemIterCopy(orig)

pangoGlyphItemIterFree(iter)

pangoGlyphItemIterInitEnd(iter, glyph_item, text)

pangoGlyphItemIterInitStart(iter, glyph_item, text)

pangoGlyphItemIterNextCluster(iter)

pangoGlyphItemIterPrevCluster(iter)

pangoGlyphStringNew()

pangoGlyphStringCopy(string)

pangoGlyphStringExtents(glyphs, font)

pangoGlyphStringExtentsRange(glyphs, start, end, font)

pangoGlyphStringFree(string)

pangoGlyphStringGetLogicalWidths(glyphs, text, length, embedding_level)

pangoGlyphStringGetWidth(glyphs)

pangoGlyphStringIndexToX(glyphs, text, length, analysis, index_, trailing)

pangoGlyphStringIndexToXFull(
  glyphs,
  text,
  length,
  analysis,
  attrs,
  index_,
  trailing
)

pangoGlyphStringSetSize(string, new_len)

pangoGlyphStringXToIndex(glyphs, text, length, analysis, x_pos)

pangoGravityGetForMatrix(matrix)

pangoGravityGetForScript(script, base_gravity, hint)

pangoGravityGetForScriptAndWidth(script, wide, base_gravity, hint)

pangoGravityToRotation(gravity)

pangoItemNew()

pangoItemApplyAttrs(item, iter)

pangoItemCopy(item)

pangoItemFree(item)

pangoItemSplit(orig, split_index, split_offset)

pangoLanguageGetSampleString(language)

pangoLanguageGetScripts(language)

pangoLanguageIncludesScript(language, script)

pangoLanguageMatches(language, range_list)

pangoLanguageToString(language)

pangoLanguageFromString(language)

pangoLanguageGetDefault()

pangoLanguageGetPreferred()

pangoLayoutNew(context)

pangoLayoutDeserialize(context, bytes, flags)

pangoLayoutContextChanged(layout)

pangoLayoutCopy(src)

pangoLayoutGetAlignment(layout)

pangoLayoutGetAttributes(layout)

pangoLayoutGetAutoDir(layout)

pangoLayoutGetBaseline(layout)

pangoLayoutGetCaretPos(layout, index_)

pangoLayoutGetCharacterCount(layout)

pangoLayoutGetContext(layout)

pangoLayoutGetCursorPos(layout, index_)

pangoLayoutGetDirection(layout, index)

pangoLayoutGetEllipsize(layout)

pangoLayoutGetExtents(layout)

pangoLayoutGetFontDescription(layout)

pangoLayoutGetHeight(layout)

pangoLayoutGetIndent(layout)

pangoLayoutGetIter(layout)

pangoLayoutGetJustify(layout)

pangoLayoutGetJustifyLastLine(layout)

pangoLayoutGetLine(layout, line)

pangoLayoutGetLineCount(layout)

pangoLayoutGetLineReadonly(layout, line)

pangoLayoutGetLineSpacing(layout)

pangoLayoutGetLines(layout)

pangoLayoutGetLinesReadonly(layout)

pangoLayoutGetLogAttrs(layout)

pangoLayoutGetLogAttrsReadonly(layout)

pangoLayoutGetPixelExtents(layout)

pangoLayoutGetPixelSize(layout)

pangoLayoutGetSerial(layout)

pangoLayoutGetSingleParagraphMode(layout)

pangoLayoutGetSize(layout)

pangoLayoutGetSpacing(layout)

pangoLayoutGetTabs(layout)

pangoLayoutGetText(layout)

pangoLayoutGetUnknownGlyphsCount(layout)

pangoLayoutGetWidth(layout)

pangoLayoutGetWrap(layout)

pangoLayoutIndexToLineX(layout, index_, trailing)

pangoLayoutIndexToPos(layout, index_)

pangoLayoutIsEllipsized(layout)

pangoLayoutIsWrapped(layout)

pangoLayoutMoveCursorVisually(
  layout,
  strong,
  old_index,
  old_trailing,
  direction
)

pangoLayoutSerialize(layout, flags)

pangoLayoutSetAlignment(layout, alignment)

pangoLayoutSetAttributes(layout, attrs)

pangoLayoutSetAutoDir(layout, auto_dir)

pangoLayoutSetEllipsize(layout, ellipsize)

pangoLayoutSetFontDescription(layout, desc)

pangoLayoutSetHeight(layout, height)

pangoLayoutSetIndent(layout, indent)

pangoLayoutSetJustify(layout, justify)

pangoLayoutSetJustifyLastLine(layout, justify)

pangoLayoutSetLineSpacing(layout, factor)

pangoLayoutSetMarkup(layout, markup, length)

pangoLayoutSetMarkupWithAccel(layout, markup, length, accel_marker)

pangoLayoutSetSingleParagraphMode(layout, setting)

pangoLayoutSetSpacing(layout, spacing)

pangoLayoutSetTabs(layout, tabs)

pangoLayoutSetText(layout, text, length)

pangoLayoutSetWidth(layout, width)

pangoLayoutSetWrap(layout, wrap)

pangoLayoutWriteToFile(layout, flags, filename)

pangoLayoutXyToIndex(layout, x, y)

pangoLayoutDeserializeErrorQuark()

pangoLayoutIterAtLastLine(iter)

pangoLayoutIterCopy(iter)

pangoLayoutIterFree(iter)

pangoLayoutIterGetBaseline(iter)

pangoLayoutIterGetCharExtents(iter)

pangoLayoutIterGetClusterExtents(iter)

pangoLayoutIterGetIndex(iter)

pangoLayoutIterGetLayout(iter)

pangoLayoutIterGetLayoutExtents(iter)

pangoLayoutIterGetLine(iter)

pangoLayoutIterGetLineExtents(iter)

pangoLayoutIterGetLineReadonly(iter)

pangoLayoutIterGetLineYrange(iter)

pangoLayoutIterGetRun(iter)

pangoLayoutIterGetRunBaseline(iter)

pangoLayoutIterGetRunExtents(iter)

pangoLayoutIterGetRunReadonly(iter)

pangoLayoutIterNextChar(iter)

pangoLayoutIterNextCluster(iter)

pangoLayoutIterNextLine(iter)

pangoLayoutIterNextRun(iter)

pangoLayoutLineGetExtents(line)

pangoLayoutLineGetHeight(line)

pangoLayoutLineGetLength(line)

pangoLayoutLineGetPixelExtents(layout_line)

pangoLayoutLineGetResolvedDirection(line)

pangoLayoutLineGetStartIndex(line)

pangoLayoutLineGetXRanges(line, start_index, end_index)

pangoLayoutLineIndexToX(line, index_, trailing)

pangoLayoutLineIsParagraphStart(line)

pangoLayoutLineRef(line)

pangoLayoutLineUnref(line)

pangoLayoutLineXToIndex(line, x_pos)

pangoMatrixConcat(matrix, new_matrix)

pangoMatrixCopy(matrix)

pangoMatrixFree(matrix)

pangoMatrixGetFontScaleFactor(matrix)

pangoMatrixGetFontScaleFactors(matrix)

pangoMatrixGetSlantRatio(matrix)

pangoMatrixRotate(matrix, degrees)

pangoMatrixScale(matrix, scale_x, scale_y)

pangoMatrixTransformDistance(matrix)

pangoMatrixTransformPixelRectangle(matrix)

pangoMatrixTransformPoint(matrix)

pangoMatrixTransformRectangle(matrix)

pangoMatrixTranslate(matrix, tx, ty)

pangoRendererActivate(renderer)

pangoRendererDeactivate(renderer)

pangoRendererDrawErrorUnderline(renderer, x, y, width, height)

pangoRendererDrawGlyph(renderer, font, glyph, x, y)

pangoRendererDrawGlyphItem(renderer, text, glyph_item, x, y)

pangoRendererDrawGlyphs(renderer, font, glyphs, x, y)

pangoRendererDrawLayout(renderer, layout, x, y)

pangoRendererDrawLayoutLine(renderer, line, x, y)

pangoRendererDrawRectangle(renderer, part, x, y, width, height)

pangoRendererDrawTrapezoid(renderer, part, y1_, x11, x21, y2, x12, x22)

pangoRendererGetAlpha(renderer, part)

pangoRendererGetColor(renderer, part)

pangoRendererGetLayout(renderer)

pangoRendererGetLayoutLine(renderer)

pangoRendererGetMatrix(renderer)

pangoRendererPartChanged(renderer, part)

pangoRendererSetAlpha(renderer, part, alpha)

pangoRendererSetColor(renderer, part, color)

pangoRendererSetMatrix(renderer, matrix)

pangoScriptForUnichar(ch)

pangoScriptGetSampleLanguage(script)

pangoScriptIterNew(text, length)

pangoScriptIterFree(iter)

pangoScriptIterGetRange(iter)

pangoScriptIterNext(iter)

pangoTabArrayNew(initial_size, positions_in_pixels)

pangoTabArrayCopy(src)

pangoTabArrayFree(tab_array)

pangoTabArrayGetDecimalPoint(tab_array, tab_index)

pangoTabArrayGetPositionsInPixels(tab_array)

pangoTabArrayGetSize(tab_array)

pangoTabArrayGetTab(tab_array, tab_index)

pangoTabArrayGetTabs(tab_array)

pangoTabArrayResize(tab_array, new_size)

pangoTabArraySetDecimalPoint(tab_array, tab_index, decimal_point)

pangoTabArraySetPositionsInPixels(tab_array, positions_in_pixels)

pangoTabArraySetTab(tab_array, tab_index, alignment, location)

pangoTabArraySort(tab_array)

pangoTabArrayToString(tab_array)

pangoTabArrayFromString(text)

pangoAttrAllowBreaksNew(allow_breaks)

pangoAttrBackgroundAlphaNew(alpha)

pangoAttrBackgroundNew(red, green, blue)

pangoAttrBaselineShiftNew(shift)

pangoAttrBreak(text, length, attr_list, offset, attrs_len)

pangoAttrFallbackNew(enable_fallback)

pangoAttrFamilyNew(family)

pangoAttrFontScaleNew(scale)

pangoAttrForegroundAlphaNew(alpha)

pangoAttrForegroundNew(red, green, blue)

pangoAttrGravityHintNew(hint)

pangoAttrGravityNew(gravity)

pangoAttrInsertHyphensNew(insert_hyphens)

pangoAttrLetterSpacingNew(letter_spacing)

pangoAttrLineHeightNew(factor)

pangoAttrLineHeightNewAbsolute(height)

pangoAttrOverlineColorNew(red, green, blue)

pangoAttrOverlineNew(overline)

pangoAttrRiseNew(rise)

pangoAttrScaleNew(scale_factor)

pangoAttrSentenceNew()

pangoAttrShowNew(flags)

pangoAttrStretchNew(stretch)

pangoAttrStrikethroughColorNew(red, green, blue)

pangoAttrStrikethroughNew(strikethrough)

pangoAttrStyleNew(style)

pangoAttrTextTransformNew(transform)

pangoAttrUnderlineColorNew(red, green, blue)

pangoAttrUnderlineNew(underline)

pangoAttrVariantNew(variant)

pangoAttrWeightNew(weight)

pangoAttrWordNew()

pangoBreak(text, length, analysis, attrs_len)

pangoDefaultBreak(text, length, analysis, attrs_len)

pangoExtentsToPixels()

pangoFindBaseDir(text, length)

pangoFindParagraphBoundary(text, length)

pangoGetLogAttrs(text, length, level, language, attrs_len)

pangoGetMirrorChar(ch)

pangoIsZeroWidth(ch)

pangoItemize(context, text, start_index, length, attrs, cached_iter)

pangoItemizeWithBaseDir(
  context,
  base_dir,
  text,
  start_index,
  length,
  attrs,
  cached_iter
)

pangoLog2visGetEmbeddingLevels(text, length)

pangoMarkupParserFinish(context)

pangoMarkupParserNew(accel_marker)

pangoParseEnum(type, str, warn)

pangoParseMarkup(markup_text, length, accel_marker)

pangoParseStretch(str, warn)

pangoParseStyle(str, warn)

pangoParseVariant(str, warn)

pangoParseWeight(str, warn)

pangoQuantizeLineGeometry()

pangoReadLine(stream, str)

pangoReorderItems(items)

pangoScanInt()

pangoScanString(out)

pangoScanWord(out)

pangoShape(text, length, analysis)

pangoShapeFull(
  item_text,
  item_length,
  paragraph_text,
  paragraph_length,
  analysis
)

pangoShapeItem(item, paragraph_text, paragraph_length, log_attrs, flags)

pangoShapeWithFlags(
  item_text,
  item_length,
  paragraph_text,
  paragraph_length,
  analysis,
  flags
)

pangoSkipSpace()

pangoSplitFileList(str)

pangoTailorBreak(text, length, analysis, offset, attrs_len)

pangoTrimString(str)

pangoUnicharDirection(ch)

pangoUnitsFromDouble(d)

pangoUnitsToDouble(i)

pangoVersion()

pangoVersionCheck(required_major, required_minor, required_micro)

pangoVersionString()
}
\arguments{
\item{desc}{FontDescription}

\item{features}{utf8}

\item{iterator}{AttrIterator}

\item{type}{GType}

\item{language}{Language}

\item{list}{AttrList}

\item{attr}{Attribute}

\item{other_list}{AttrList}

\item{func}{function — FontsetForeachFunc callback}

\item{other}{Coverage}

\item{pos}{gint}

\item{len}{guint}

\item{remove}{gint}

\item{add}{gint}

\item{text}{utf8}

\item{ink_rect}{Rectangle}

\item{logical_rect}{Rectangle}

\item{data}{gpointer}

\item{copy_func}{function — AttrDataCopyFunc callback}

\item{size}{gint}

\item{name}{utf8}

\item{attr1}{Attribute}

\item{attr2}{Attribute}

\item{klass}{AttrClass}

\item{ch}{gunichar}

\item{src}{TabArray}

\item{color}{Color}

\item{spec}{utf8}

\item{context}{GLib.MarkupParseContext}

\item{direction}{gint}

\item{gravity}{Gravity}

\item{font_map}{FontMap}

\item{hint}{GravityHint}

\item{matrix}{Matrix}

\item{round_positions}{gboolean}

\item{bytes}{GLib.Bytes}

\item{n_bytes}{gint}

\item{coverage}{Coverage}

\item{index_}{gint}

\item{level}{gint}

\item{descs}{FontDescription}

\item{n_descs}{gint}

\item{font}{Font}

\item{glyph}{Glyph}

\item{wc}{guint}

\item{old_match}{FontDescription}

\item{new_match}{FontDescription}

\item{desc1}{FontDescription}

\item{desc2}{FontDescription}

\item{desc_to_merge}{FontDescription}

\item{replace_existing}{gboolean}

\item{family}{utf8}

\item{stretch}{Stretch}

\item{style}{Style}

\item{variant}{Variant}

\item{variations}{utf8}

\item{weight}{Weight}

\item{to_unset}{FontMask}

\item{str}{utf8}

\item{face}{FontFace}

\item{fontmap}{FontMap}

\item{metrics}{FontMetrics}

\item{fontset}{FontsetSimple}

\item{glyph_item}{GlyphItem}

\item{orig}{Item}

\item{log_attrs}{LogAttr}

\item{letter_spacing}{gint}

\item{split_index}{gint}

\item{iter}{ScriptIter}

\item{string}{GlyphString}

\item{glyphs}{GlyphString}

\item{start}{gint}

\item{end}{gint}

\item{length}{gint}

\item{embedding_level}{gint}

\item{analysis}{Analysis}

\item{trailing}{gboolean}

\item{attrs}{AttrList}

\item{new_len}{gint}

\item{x_pos}{gint}

\item{script}{Script}

\item{base_gravity}{Gravity}

\item{wide}{gboolean}

\item{item}{Item}

\item{split_offset}{gint}

\item{range_list}{utf8}

\item{flags}{ShapeFlags}

\item{layout}{Layout}

\item{index}{gint}

\item{line}{LayoutLine}

\item{strong}{gboolean}

\item{old_index}{gint}

\item{old_trailing}{gint}

\item{alignment}{TabAlign}

\item{auto_dir}{gboolean}

\item{ellipsize}{EllipsizeMode}

\item{height}{gint}

\item{indent}{gint}

\item{justify}{gboolean}

\item{factor}{gdouble}

\item{markup}{utf8}

\item{accel_marker}{gunichar}

\item{setting}{gboolean}

\item{spacing}{gint}

\item{tabs}{TabArray}

\item{width}{gint}

\item{wrap}{WrapMode}

\item{filename}{filename}

\item{x}{gint}

\item{y}{gint}

\item{layout_line}{LayoutLine}

\item{start_index}{gint}

\item{end_index}{gint}

\item{new_matrix}{Matrix}

\item{degrees}{gdouble}

\item{scale_x}{gdouble}

\item{scale_y}{gdouble}

\item{tx}{gdouble}

\item{ty}{gdouble}

\item{renderer}{Renderer}

\item{part}{RenderPart}

\item{y1_}{gdouble}

\item{x11}{gdouble}

\item{x21}{gdouble}

\item{y2}{gdouble}

\item{x12}{gdouble}

\item{x22}{gdouble}

\item{alpha}{guint16}

\item{initial_size}{gint}

\item{positions_in_pixels}{gboolean}

\item{tab_array}{TabArray}

\item{tab_index}{gint}

\item{new_size}{gint}

\item{decimal_point}{gunichar}

\item{location}{gint}

\item{allow_breaks}{gboolean}

\item{red}{guint16}

\item{green}{guint16}

\item{blue}{guint16}

\item{shift}{gint}

\item{attr_list}{AttrList}

\item{offset}{gint}

\item{attrs_len}{gint}

\item{enable_fallback}{gboolean}

\item{scale}{FontScale}

\item{insert_hyphens}{gboolean}

\item{overline}{Overline}

\item{rise}{gint}

\item{scale_factor}{gdouble}

\item{strikethrough}{gboolean}

\item{transform}{TextTransform}

\item{underline}{Underline}

\item{cached_iter}{AttrIterator}

\item{base_dir}{Direction}

\item{warn}{gboolean}

\item{markup_text}{utf8}

\item{stream}{gpointer}

\item{items}{GLib.List}

\item{out}{GLib.String}

\item{item_text}{utf8}

\item{item_length}{gint}

\item{paragraph_text}{utf8}

\item{paragraph_length}{gint}

\item{d}{gdouble}

\item{i}{gint}

\item{required_major}{gint}

\item{required_minor}{gint}

\item{required_micro}{gint}
}
\value{
Attribute

Attribute

AttrIterator

Return value from C function

Attribute

GLib.SList

Return value from C function

gboolean

Return value from C function

Attribute

AttrList

Return value from C function

AttrList

gboolean

AttrList

GLib.SList

AttrIterator

Return value from C function

Return value from C function

AttrList

Return value from C function

utf8

Return value from C function

Return value from C function

AttrList

Attribute

Attribute

Attribute

Attribute

utf8

AttrType

AttrColor

AttrFloat

AttrFontDesc

AttrFontFeatures

AttrInt

AttrLanguage

AttrShape

AttrSize

AttrString

Attribute

Return value from C function

gboolean

Return value from C function

BidiType

Color

Return value from C function

gboolean

gboolean

utf8

Context

Return value from C function

Direction

Gravity

FontDescription

FontMap

Gravity

GravityHint

Language

Matrix

FontMetrics

gboolean

guint

Return value from C function

Font

Fontset

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Coverage

Coverage

Coverage

CoverageLevel

Return value from C function

Coverage

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Font

FontDescription

FontDescription

Coverage

FontFace

Return value from C function

FontMap

Return value from C function

Language

FontMetrics

gboolean

GLib.Bytes

FontDescription

gboolean

FontDescription

FontDescription

gboolean

Return value from C function

utf8

Gravity

FontMask

gint

gboolean

Stretch

Style

Variant

utf8

Weight

guint

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

utf8

utf8

Return value from C function

FontDescription

FontDescription

utf8

FontFamily

gboolean

Return value from C function

FontFace

utf8

gboolean

gboolean

Return value from C function

Return value from C function

Context

FontFamily

guint

Return value from C function

Font

Fontset

gint

gint

gint

gint

gint

gint

gint

gint

gint

FontMetrics

Return value from C function

Return value from C function

Font

FontMetrics

FontsetSimple

Return value from C function

gint

GLib.SList

GlyphItem

Return value from C function

Return value from C function

Return value from C function

GlyphItem

GlyphItemIter

Return value from C function

gboolean

gboolean

gboolean

gboolean

GlyphString

GlyphString

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gint

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Gravity

Gravity

Gravity

gdouble

Item

Return value from C function

Item

Return value from C function

Item

utf8

Script

gboolean

gboolean

utf8

Language

Language

Language

Layout

Layout

Return value from C function

Layout

Alignment

AttrList

gboolean

gint

Return value from C function

gint

Context

Return value from C function

Direction

EllipsizeMode

Return value from C function

FontDescription

gint

gint

LayoutIter

gboolean

gboolean

LayoutLine

gint

LayoutLine

gfloat

GLib.SList

GLib.SList

Return value from C function

LogAttr

Return value from C function

Return value from C function

guint

gboolean

Return value from C function

gint

TabArray

utf8

gint

gint

WrapMode

Return value from C function

Return value from C function

gboolean

gboolean

Return value from C function

GLib.Bytes

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

gboolean

GLib.Quark

gboolean

LayoutIter

Return value from C function

gint

Return value from C function

Return value from C function

gint

Layout

Return value from C function

LayoutLine

Return value from C function

LayoutLine

Return value from C function

LayoutRun

gint

Return value from C function

LayoutRun

gboolean

gboolean

gboolean

gboolean

Return value from C function

Return value from C function

gint

Return value from C function

Direction

gint

Return value from C function

Return value from C function

gboolean

LayoutLine

Return value from C function

gboolean

Return value from C function

Matrix

Return value from C function

gdouble

Return value from C function

gdouble

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

guint16

Color

Layout

LayoutLine

Matrix

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Script

Language

ScriptIter

Return value from C function

Return value from C function

gboolean

TabArray

TabArray

Return value from C function

gunichar

gboolean

gint

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

utf8

TabArray

Attribute

Attribute

Attribute

Attribute

Return value from C function

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Attribute

Return value from C function

Return value from C function

Return value from C function

Direction

Return value from C function

Return value from C function

gboolean

gboolean

GLib.List

GLib.List

guint8

gboolean

GLib.MarkupParseContext

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

Return value from C function

gint

GLib.List

gboolean

gboolean

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

utf8

Return value from C function

utf8

Direction

gint

gdouble

gint

utf8

utf8
}
\description{
pangoAttrFontDescNew

pangoAttrFontFeaturesNew

pangoAttrIteratorCopy

pangoAttrIteratorDestroy

pangoAttrIteratorGet

pangoAttrIteratorGetAttrs

pangoAttrIteratorGetFont

pangoAttrIteratorNext

pangoAttrIteratorRange

pangoAttrLanguageNew

pangoAttrListNew

pangoAttrListChange

pangoAttrListCopy

pangoAttrListEqual

pangoAttrListFilter

pangoAttrListGetAttributes

pangoAttrListGetIterator

pangoAttrListInsert

pangoAttrListInsertBefore

pangoAttrListRef

pangoAttrListSplice

pangoAttrListToString

pangoAttrListUnref

pangoAttrListUpdate

pangoAttrListFromString

pangoAttrShapeNew

pangoAttrShapeNewWithData

pangoAttrSizeNew

pangoAttrSizeNewAbsolute

pangoAttrTypeGetName

pangoAttrTypeRegister

pangoAttributeAsColor

pangoAttributeAsFloat

pangoAttributeAsFontDesc

pangoAttributeAsFontFeatures

pangoAttributeAsInt

pangoAttributeAsLanguage

pangoAttributeAsShape

pangoAttributeAsSize

pangoAttributeAsString

pangoAttributeCopy

pangoAttributeDestroy

pangoAttributeEqual

pangoAttributeInit

pangoBidiTypeForUnichar

pangoColorCopy

pangoColorFree

pangoColorParse

pangoColorParseWithAlpha

pangoColorToString

pangoContextNew

pangoContextChanged

pangoContextGetBaseDir

pangoContextGetBaseGravity

pangoContextGetFontDescription

pangoContextGetFontMap

pangoContextGetGravity

pangoContextGetGravityHint

pangoContextGetLanguage

pangoContextGetMatrix

pangoContextGetMetrics

pangoContextGetRoundGlyphPositions

pangoContextGetSerial

pangoContextListFamilies

pangoContextLoadFont

pangoContextLoadFontset

pangoContextSetBaseDir

pangoContextSetBaseGravity

pangoContextSetFontDescription

pangoContextSetFontMap

pangoContextSetGravityHint

pangoContextSetLanguage

pangoContextSetMatrix

pangoContextSetRoundGlyphPositions

pangoCoverageNew

pangoCoverageFromBytes

pangoCoverageCopy

pangoCoverageGet

pangoCoverageMax

pangoCoverageRef

pangoCoverageSet

pangoCoverageToBytes

pangoCoverageUnref

pangoFontDescriptionsFree

pangoFontDeserialize

pangoFontDescribe

pangoFontDescribeWithAbsoluteSize

pangoFontGetCoverage

pangoFontGetFace

pangoFontGetFeatures

pangoFontGetFontMap

pangoFontGetGlyphExtents

pangoFontGetLanguages

pangoFontGetMetrics

pangoFontHasChar

pangoFontSerialize

pangoFontDescriptionNew

pangoFontDescriptionBetterMatch

pangoFontDescriptionCopy

pangoFontDescriptionCopyStatic

pangoFontDescriptionEqual

pangoFontDescriptionFree

pangoFontDescriptionGetFamily

pangoFontDescriptionGetGravity

pangoFontDescriptionGetSetFields

pangoFontDescriptionGetSize

pangoFontDescriptionGetSizeIsAbsolute

pangoFontDescriptionGetStretch

pangoFontDescriptionGetStyle

pangoFontDescriptionGetVariant

pangoFontDescriptionGetVariations

pangoFontDescriptionGetWeight

pangoFontDescriptionHash

pangoFontDescriptionMerge

pangoFontDescriptionMergeStatic

pangoFontDescriptionSetAbsoluteSize

pangoFontDescriptionSetFamily

pangoFontDescriptionSetFamilyStatic

pangoFontDescriptionSetGravity

pangoFontDescriptionSetSize

pangoFontDescriptionSetStretch

pangoFontDescriptionSetStyle

pangoFontDescriptionSetVariant

pangoFontDescriptionSetVariations

pangoFontDescriptionSetVariationsStatic

pangoFontDescriptionSetWeight

pangoFontDescriptionToFilename

pangoFontDescriptionToString

pangoFontDescriptionUnsetFields

pangoFontDescriptionFromString

pangoFontFaceDescribe

pangoFontFaceGetFaceName

pangoFontFaceGetFamily

pangoFontFaceIsSynthesized

pangoFontFaceListSizes

pangoFontFamilyGetFace

pangoFontFamilyGetName

pangoFontFamilyIsMonospace

pangoFontFamilyIsVariable

pangoFontFamilyListFaces

pangoFontMapChanged

pangoFontMapCreateContext

pangoFontMapGetFamily

pangoFontMapGetSerial

pangoFontMapListFamilies

pangoFontMapLoadFont

pangoFontMapLoadFontset

pangoFontMetricsGetApproximateCharWidth

pangoFontMetricsGetApproximateDigitWidth

pangoFontMetricsGetAscent

pangoFontMetricsGetDescent

pangoFontMetricsGetHeight

pangoFontMetricsGetStrikethroughPosition

pangoFontMetricsGetStrikethroughThickness

pangoFontMetricsGetUnderlinePosition

pangoFontMetricsGetUnderlineThickness

pangoFontMetricsRef

pangoFontMetricsUnref

pangoFontsetForeach

pangoFontsetGetFont

pangoFontsetGetMetrics

pangoFontsetSimpleNew

pangoFontsetSimpleAppend

pangoFontsetSimpleSize

pangoGlyphItemApplyAttrs

pangoGlyphItemCopy

pangoGlyphItemFree

pangoGlyphItemGetLogicalWidths

pangoGlyphItemLetterSpace

pangoGlyphItemSplit

pangoGlyphItemIterCopy

pangoGlyphItemIterFree

pangoGlyphItemIterInitEnd

pangoGlyphItemIterInitStart

pangoGlyphItemIterNextCluster

pangoGlyphItemIterPrevCluster

pangoGlyphStringNew

pangoGlyphStringCopy

pangoGlyphStringExtents

pangoGlyphStringExtentsRange

pangoGlyphStringFree

pangoGlyphStringGetLogicalWidths

pangoGlyphStringGetWidth

pangoGlyphStringIndexToX

pangoGlyphStringIndexToXFull

pangoGlyphStringSetSize

pangoGlyphStringXToIndex

pangoGravityGetForMatrix

pangoGravityGetForScript

pangoGravityGetForScriptAndWidth

pangoGravityToRotation

pangoItemNew

pangoItemApplyAttrs

pangoItemCopy

pangoItemFree

pangoItemSplit

pangoLanguageGetSampleString

pangoLanguageGetScripts

pangoLanguageIncludesScript

pangoLanguageMatches

pangoLanguageToString

pangoLanguageFromString

pangoLanguageGetDefault

pangoLanguageGetPreferred

pangoLayoutNew

pangoLayoutDeserialize

pangoLayoutContextChanged

pangoLayoutCopy

pangoLayoutGetAlignment

pangoLayoutGetAttributes

pangoLayoutGetAutoDir

pangoLayoutGetBaseline

pangoLayoutGetCaretPos

pangoLayoutGetCharacterCount

pangoLayoutGetContext

pangoLayoutGetCursorPos

pangoLayoutGetDirection

pangoLayoutGetEllipsize

pangoLayoutGetExtents

pangoLayoutGetFontDescription

pangoLayoutGetHeight

pangoLayoutGetIndent

pangoLayoutGetIter

pangoLayoutGetJustify

pangoLayoutGetJustifyLastLine

pangoLayoutGetLine

pangoLayoutGetLineCount

pangoLayoutGetLineReadonly

pangoLayoutGetLineSpacing

pangoLayoutGetLines

pangoLayoutGetLinesReadonly

pangoLayoutGetLogAttrs

pangoLayoutGetLogAttrsReadonly

pangoLayoutGetPixelExtents

pangoLayoutGetPixelSize

pangoLayoutGetSerial

pangoLayoutGetSingleParagraphMode

pangoLayoutGetSize

pangoLayoutGetSpacing

pangoLayoutGetTabs

pangoLayoutGetText

pangoLayoutGetUnknownGlyphsCount

pangoLayoutGetWidth

pangoLayoutGetWrap

pangoLayoutIndexToLineX

pangoLayoutIndexToPos

pangoLayoutIsEllipsized

pangoLayoutIsWrapped

pangoLayoutMoveCursorVisually

pangoLayoutSerialize

pangoLayoutSetAlignment

pangoLayoutSetAttributes

pangoLayoutSetAutoDir

pangoLayoutSetEllipsize

pangoLayoutSetFontDescription

pangoLayoutSetHeight

pangoLayoutSetIndent

pangoLayoutSetJustify

pangoLayoutSetJustifyLastLine

pangoLayoutSetLineSpacing

pangoLayoutSetMarkup

pangoLayoutSetMarkupWithAccel

pangoLayoutSetSingleParagraphMode

pangoLayoutSetSpacing

pangoLayoutSetTabs

pangoLayoutSetText

pangoLayoutSetWidth

pangoLayoutSetWrap

pangoLayoutWriteToFile

pangoLayoutXyToIndex

pangoLayoutDeserializeErrorQuark

pangoLayoutIterAtLastLine

pangoLayoutIterCopy

pangoLayoutIterFree

pangoLayoutIterGetBaseline

pangoLayoutIterGetCharExtents

pangoLayoutIterGetClusterExtents

pangoLayoutIterGetIndex

pangoLayoutIterGetLayout

pangoLayoutIterGetLayoutExtents

pangoLayoutIterGetLine

pangoLayoutIterGetLineExtents

pangoLayoutIterGetLineReadonly

pangoLayoutIterGetLineYrange

pangoLayoutIterGetRun

pangoLayoutIterGetRunBaseline

pangoLayoutIterGetRunExtents

pangoLayoutIterGetRunReadonly

pangoLayoutIterNextChar

pangoLayoutIterNextCluster

pangoLayoutIterNextLine

pangoLayoutIterNextRun

pangoLayoutLineGetExtents

pangoLayoutLineGetHeight

pangoLayoutLineGetLength

pangoLayoutLineGetPixelExtents

pangoLayoutLineGetResolvedDirection

pangoLayoutLineGetStartIndex

pangoLayoutLineGetXRanges

pangoLayoutLineIndexToX

pangoLayoutLineIsParagraphStart

pangoLayoutLineRef

pangoLayoutLineUnref

pangoLayoutLineXToIndex

pangoMatrixConcat

pangoMatrixCopy

pangoMatrixFree

pangoMatrixGetFontScaleFactor

pangoMatrixGetFontScaleFactors

pangoMatrixGetSlantRatio

pangoMatrixRotate

pangoMatrixScale

pangoMatrixTransformDistance

pangoMatrixTransformPixelRectangle

pangoMatrixTransformPoint

pangoMatrixTransformRectangle

pangoMatrixTranslate

pangoRendererActivate

pangoRendererDeactivate

pangoRendererDrawErrorUnderline

pangoRendererDrawGlyph

pangoRendererDrawGlyphItem

pangoRendererDrawGlyphs

pangoRendererDrawLayout

pangoRendererDrawLayoutLine

pangoRendererDrawRectangle

pangoRendererDrawTrapezoid

pangoRendererGetAlpha

pangoRendererGetColor

pangoRendererGetLayout

pangoRendererGetLayoutLine

pangoRendererGetMatrix

pangoRendererPartChanged

pangoRendererSetAlpha

pangoRendererSetColor

pangoRendererSetMatrix

pangoScriptForUnichar

pangoScriptGetSampleLanguage

pangoScriptIterNew

pangoScriptIterFree

pangoScriptIterGetRange

pangoScriptIterNext

pangoTabArrayNew

pangoTabArrayCopy

pangoTabArrayFree

pangoTabArrayGetDecimalPoint

pangoTabArrayGetPositionsInPixels

pangoTabArrayGetSize

pangoTabArrayGetTab

pangoTabArrayGetTabs

pangoTabArrayResize

pangoTabArraySetDecimalPoint

pangoTabArraySetPositionsInPixels

pangoTabArraySetTab

pangoTabArraySort

pangoTabArrayToString

pangoTabArrayFromString

pangoAttrAllowBreaksNew

pangoAttrBackgroundAlphaNew

pangoAttrBackgroundNew

pangoAttrBaselineShiftNew

pangoAttrBreak

pangoAttrFallbackNew

pangoAttrFamilyNew

pangoAttrFontScaleNew

pangoAttrForegroundAlphaNew

pangoAttrForegroundNew

pangoAttrGravityHintNew

pangoAttrGravityNew

pangoAttrInsertHyphensNew

pangoAttrLetterSpacingNew

pangoAttrLineHeightNew

pangoAttrLineHeightNewAbsolute

pangoAttrOverlineColorNew

pangoAttrOverlineNew

pangoAttrRiseNew

pangoAttrScaleNew

pangoAttrSentenceNew

pangoAttrShowNew

pangoAttrStretchNew

pangoAttrStrikethroughColorNew

pangoAttrStrikethroughNew

pangoAttrStyleNew

pangoAttrTextTransformNew

pangoAttrUnderlineColorNew

pangoAttrUnderlineNew

pangoAttrVariantNew

pangoAttrWeightNew

pangoAttrWordNew

pangoBreak

pangoDefaultBreak

pangoExtentsToPixels

pangoFindBaseDir

pangoFindParagraphBoundary

pangoGetLogAttrs

pangoGetMirrorChar

pangoIsZeroWidth

pangoItemize

pangoItemizeWithBaseDir

pangoLog2visGetEmbeddingLevels

pangoMarkupParserFinish

pangoMarkupParserNew

pangoParseEnum

pangoParseMarkup

pangoParseStretch

pangoParseStyle

pangoParseVariant

pangoParseWeight

pangoQuantizeLineGeometry

pangoReadLine

pangoReorderItems

pangoScanInt

pangoScanString

pangoScanWord

pangoShape

pangoShapeFull

pangoShapeItem

pangoShapeWithFlags

pangoSkipSpace

pangoSplitFileList

pangoTailorBreak

pangoTrimString

pangoUnicharDirection

pangoUnitsFromDouble

pangoUnitsToDouble

pangoVersion

pangoVersionCheck

pangoVersionString
}
