% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print.GObject}
\alias{print.GObject}
\title{Print method for GObject external pointers}
\usage{
\method{print}{GObject}(x, ...)
}
\arguments{
\item{x}{External pointer with glib_type attribute}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for GObject external pointers
}
