% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print.RGtkObject}
\alias{print.RGtkObject}
\title{Pretty-print method for RGtk4 wrapped pointers.}
\usage{
\method{print}{RGtkObject}(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{additional arguments}
}
\description{
All extptrs returned from auto-generated bindings carry a trailing class
"RGtkObject" plus the GLib type name. This method prints them as
  <TypeName 0x...>
instead of the default base-R "<pointer: 0x...>" with a noisy attribute
dump.
}
