% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GdkPixbuf_autogen.R, R/Gdk_autogen.R
\name{gdkPixbufNew}
\alias{gdkPixbufNew}
\alias{gdkPixbufNewFromBytes}
\alias{gdkPixbufNewFromFile}
\alias{gdkPixbufNewFromFileAtScale}
\alias{gdkPixbufNewFromFileAtSize}
\alias{gdkPixbufNewFromInline}
\alias{gdkPixbufNewFromResource}
\alias{gdkPixbufNewFromResourceAtScale}
\alias{gdkPixbufNewFromStream}
\alias{gdkPixbufNewFromStreamAtScale}
\alias{gdkPixbufNewFromStreamFinish}
\alias{gdkPixbufNewFromXpmData}
\alias{gdkPixbufCalculateRowstride}
\alias{gdkPixbufGetFileInfo}
\alias{gdkPixbufGetFileInfoFinish}
\alias{gdkPixbufGetFormats}
\alias{gdkPixbufInitModules}
\alias{gdkPixbufSaveToStreamFinish}
\alias{gdkPixbufAddAlpha}
\alias{gdkPixbufApplyEmbeddedOrientation}
\alias{gdkPixbufComposite}
\alias{gdkPixbufCompositeColor}
\alias{gdkPixbufCompositeColorSimple}
\alias{gdkPixbufCopy}
\alias{gdkPixbufCopyArea}
\alias{gdkPixbufCopyOptions}
\alias{gdkPixbufFill}
\alias{gdkPixbufFlip}
\alias{gdkPixbufGetBitsPerSample}
\alias{gdkPixbufGetByteLength}
\alias{gdkPixbufGetColorspace}
\alias{gdkPixbufGetHasAlpha}
\alias{gdkPixbufGetHeight}
\alias{gdkPixbufGetNChannels}
\alias{gdkPixbufGetOption}
\alias{gdkPixbufGetOptions}
\alias{gdkPixbufGetPixels}
\alias{gdkPixbufGetPixelsWithLength}
\alias{gdkPixbufGetRowstride}
\alias{gdkPixbufGetWidth}
\alias{gdkPixbufNewSubpixbuf}
\alias{gdkPixbufReadPixelBytes}
\alias{gdkPixbufReadPixels}
\alias{gdkPixbufRemoveOption}
\alias{gdkPixbufRotateSimple}
\alias{gdkPixbufSaturateAndPixelate}
\alias{gdkPixbufSaveToBufferv}
\alias{gdkPixbufSaveToStreamv}
\alias{gdkPixbufSavev}
\alias{gdkPixbufScale}
\alias{gdkPixbufScaleSimple}
\alias{gdkPixbufSetOption}
\alias{gdkPixbufAnimationNewFromFile}
\alias{gdkPixbufAnimationNewFromResource}
\alias{gdkPixbufAnimationNewFromStream}
\alias{gdkPixbufAnimationNewFromStreamFinish}
\alias{gdkPixbufAnimationGetHeight}
\alias{gdkPixbufAnimationGetIter}
\alias{gdkPixbufAnimationGetStaticImage}
\alias{gdkPixbufAnimationGetWidth}
\alias{gdkPixbufAnimationIsStaticImage}
\alias{gdkPixbufAnimationIterAdvance}
\alias{gdkPixbufAnimationIterGetDelayTime}
\alias{gdkPixbufAnimationIterGetPixbuf}
\alias{gdkPixbufAnimationIterOnCurrentlyLoadingFrame}
\alias{gdkPixbufErrorQuark}
\alias{gdkPixbufFormatCopy}
\alias{gdkPixbufFormatFree}
\alias{gdkPixbufFormatGetDescription}
\alias{gdkPixbufFormatGetExtensions}
\alias{gdkPixbufFormatGetLicense}
\alias{gdkPixbufFormatGetMimeTypes}
\alias{gdkPixbufFormatGetName}
\alias{gdkPixbufFormatIsDisabled}
\alias{gdkPixbufFormatIsSaveOptionSupported}
\alias{gdkPixbufFormatIsScalable}
\alias{gdkPixbufFormatIsWritable}
\alias{gdkPixbufFormatSetDisabled}
\alias{gdkPixbufLoaderNew}
\alias{gdkPixbufLoaderNewWithMimeType}
\alias{gdkPixbufLoaderNewWithType}
\alias{gdkPixbufLoaderClose}
\alias{gdkPixbufLoaderGetAnimation}
\alias{gdkPixbufLoaderGetFormat}
\alias{gdkPixbufLoaderGetPixbuf}
\alias{gdkPixbufLoaderSetSize}
\alias{gdkPixbufLoaderWrite}
\alias{gdkPixbufLoaderWriteBytes}
\alias{gdkPixbufSimpleAnimNew}
\alias{gdkPixbufSimpleAnimAddFrame}
\alias{gdkPixbufSimpleAnimGetLoop}
\alias{gdkPixbufSimpleAnimSetLoop}
\alias{gdkTextureNewForPixbuf}
\alias{gdkTextureNewFromBytes}
\alias{gdkTextureNewFromFile}
\alias{gdkTextureNewFromFilename}
\alias{gdkTextureNewFromResource}
\alias{gdkTextureDownload}
\alias{gdkTextureGetFormat}
\alias{gdkTextureGetHeight}
\alias{gdkTextureGetWidth}
\alias{gdkTextureSaveToPng}
\alias{gdkTextureSaveToPngBytes}
\alias{gdkTextureSaveToTiff}
\alias{gdkTextureSaveToTiffBytes}
\alias{gdkTextureDownloaderNew}
\alias{gdkTextureDownloaderCopy}
\alias{gdkTextureDownloaderDownloadBytes}
\alias{gdkTextureDownloaderDownloadInto}
\alias{gdkTextureDownloaderFree}
\alias{gdkTextureDownloaderGetFormat}
\alias{gdkTextureDownloaderGetTexture}
\alias{gdkTextureDownloaderSetFormat}
\alias{gdkTextureDownloaderSetTexture}
\alias{gdkTextureErrorQuark}
\alias{gdkPixbufGetFromSurface}
\alias{gdkPixbufGetFromTexture}
\title{gdkPixbufNew}
\usage{
gdkPixbufNew(colorspace, has_alpha, bits_per_sample, width, height)

gdkPixbufNewFromBytes(
  data,
  colorspace,
  has_alpha,
  bits_per_sample,
  width,
  height,
  rowstride
)

gdkPixbufNewFromFile(filename)

gdkPixbufNewFromFileAtScale(filename, width, height, preserve_aspect_ratio)

gdkPixbufNewFromFileAtSize(filename, width, height)

gdkPixbufNewFromInline(data_length, data, copy_pixels)

gdkPixbufNewFromResource(resource_path)

gdkPixbufNewFromResourceAtScale(
  resource_path,
  width,
  height,
  preserve_aspect_ratio
)

gdkPixbufNewFromStream(stream, cancellable)

gdkPixbufNewFromStreamAtScale(
  stream,
  width,
  height,
  preserve_aspect_ratio,
  cancellable
)

gdkPixbufNewFromStreamFinish(async_result)

gdkPixbufNewFromXpmData(data)

gdkPixbufCalculateRowstride(
  colorspace,
  has_alpha,
  bits_per_sample,
  width,
  height
)

gdkPixbufGetFileInfo(filename)

gdkPixbufGetFileInfoFinish(async_result)

gdkPixbufGetFormats()

gdkPixbufInitModules(path)

gdkPixbufSaveToStreamFinish(async_result)

gdkPixbufAddAlpha(pixbuf, substitute_color, r, g, b)

gdkPixbufApplyEmbeddedOrientation(src)

gdkPixbufComposite(
  src,
  dest,
  dest_x,
  dest_y,
  dest_width,
  dest_height,
  offset_x,
  offset_y,
  scale_x,
  scale_y,
  interp_type,
  overall_alpha
)

gdkPixbufCompositeColor(
  src,
  dest,
  dest_x,
  dest_y,
  dest_width,
  dest_height,
  offset_x,
  offset_y,
  scale_x,
  scale_y,
  interp_type,
  overall_alpha,
  check_x,
  check_y,
  check_size,
  color1,
  color2
)

gdkPixbufCompositeColorSimple(
  src,
  dest_width,
  dest_height,
  interp_type,
  overall_alpha,
  check_size,
  color1,
  color2
)

gdkPixbufCopy(pixbuf)

gdkPixbufCopyArea(
  src_pixbuf,
  src_x,
  src_y,
  width,
  height,
  dest_pixbuf,
  dest_x,
  dest_y
)

gdkPixbufCopyOptions(src_pixbuf, dest_pixbuf)

gdkPixbufFill(pixbuf, pixel)

gdkPixbufFlip(src, horizontal)

gdkPixbufGetBitsPerSample(pixbuf)

gdkPixbufGetByteLength(pixbuf)

gdkPixbufGetColorspace(pixbuf)

gdkPixbufGetHasAlpha(pixbuf)

gdkPixbufGetHeight(pixbuf)

gdkPixbufGetNChannels(pixbuf)

gdkPixbufGetOption(pixbuf, key)

gdkPixbufGetOptions(pixbuf)

gdkPixbufGetPixels(pixbuf)

gdkPixbufGetPixelsWithLength(pixbuf)

gdkPixbufGetRowstride(pixbuf)

gdkPixbufGetWidth(pixbuf)

gdkPixbufNewSubpixbuf(src_pixbuf, src_x, src_y, width, height)

gdkPixbufReadPixelBytes(pixbuf)

gdkPixbufReadPixels(pixbuf)

gdkPixbufRemoveOption(pixbuf, key)

gdkPixbufRotateSimple(src, angle)

gdkPixbufSaturateAndPixelate(src, dest, saturation, pixelate)

gdkPixbufSaveToBufferv(pixbuf, type, option_keys, option_values)

gdkPixbufSaveToStreamv(
  pixbuf,
  stream,
  type,
  option_keys,
  option_values,
  cancellable
)

gdkPixbufSavev(pixbuf, filename, type, option_keys, option_values)

gdkPixbufScale(
  src,
  dest,
  dest_x,
  dest_y,
  dest_width,
  dest_height,
  offset_x,
  offset_y,
  scale_x,
  scale_y,
  interp_type
)

gdkPixbufScaleSimple(src, dest_width, dest_height, interp_type)

gdkPixbufSetOption(pixbuf, key, value)

gdkPixbufAnimationNewFromFile(filename)

gdkPixbufAnimationNewFromResource(resource_path)

gdkPixbufAnimationNewFromStream(stream, cancellable)

gdkPixbufAnimationNewFromStreamFinish(async_result)

gdkPixbufAnimationGetHeight(animation)

gdkPixbufAnimationGetIter(animation, start_time)

gdkPixbufAnimationGetStaticImage(animation)

gdkPixbufAnimationGetWidth(animation)

gdkPixbufAnimationIsStaticImage(animation)

gdkPixbufAnimationIterAdvance(iter, current_time)

gdkPixbufAnimationIterGetDelayTime(iter)

gdkPixbufAnimationIterGetPixbuf(iter)

gdkPixbufAnimationIterOnCurrentlyLoadingFrame(iter)

gdkPixbufErrorQuark()

gdkPixbufFormatCopy(format)

gdkPixbufFormatFree(format)

gdkPixbufFormatGetDescription(format)

gdkPixbufFormatGetExtensions(format)

gdkPixbufFormatGetLicense(format)

gdkPixbufFormatGetMimeTypes(format)

gdkPixbufFormatGetName(format)

gdkPixbufFormatIsDisabled(format)

gdkPixbufFormatIsSaveOptionSupported(format, option_key)

gdkPixbufFormatIsScalable(format)

gdkPixbufFormatIsWritable(format)

gdkPixbufFormatSetDisabled(format, disabled)

gdkPixbufLoaderNew()

gdkPixbufLoaderNewWithMimeType(mime_type)

gdkPixbufLoaderNewWithType(image_type)

gdkPixbufLoaderClose(loader)

gdkPixbufLoaderGetAnimation(loader)

gdkPixbufLoaderGetFormat(loader)

gdkPixbufLoaderGetPixbuf(loader)

gdkPixbufLoaderSetSize(loader, width, height)

gdkPixbufLoaderWrite(loader, buf, count)

gdkPixbufLoaderWriteBytes(loader, buffer)

gdkPixbufSimpleAnimNew(width, height, rate)

gdkPixbufSimpleAnimAddFrame(animation, pixbuf)

gdkPixbufSimpleAnimGetLoop(animation)

gdkPixbufSimpleAnimSetLoop(animation, loop)

gdkTextureNewForPixbuf(pixbuf)

gdkTextureNewFromBytes(bytes)

gdkTextureNewFromFile(file)

gdkTextureNewFromFilename(path)

gdkTextureNewFromResource(resource_path)

gdkTextureDownload(texture, data, stride)

gdkTextureGetFormat(self)

gdkTextureGetHeight(texture)

gdkTextureGetWidth(texture)

gdkTextureSaveToPng(texture, filename)

gdkTextureSaveToPngBytes(texture)

gdkTextureSaveToTiff(texture, filename)

gdkTextureSaveToTiffBytes(texture)

gdkTextureDownloaderNew(texture)

gdkTextureDownloaderCopy(self)

gdkTextureDownloaderDownloadBytes(self)

gdkTextureDownloaderDownloadInto(self, data, stride)

gdkTextureDownloaderFree(self)

gdkTextureDownloaderGetFormat(self)

gdkTextureDownloaderGetTexture(self)

gdkTextureDownloaderSetFormat(self, format)

gdkTextureDownloaderSetTexture(self, texture)

gdkTextureErrorQuark()

gdkPixbufGetFromSurface(surface, src_x, src_y, width, height)

gdkPixbufGetFromTexture(texture)
}
\arguments{
\item{colorspace}{Colorspace}

\item{has_alpha}{gboolean}

\item{bits_per_sample}{gint}

\item{width}{gint}

\item{height}{gint}

\item{data}{guint8}

\item{rowstride}{gint}

\item{filename}{filename}

\item{preserve_aspect_ratio}{gboolean}

\item{data_length}{gint}

\item{copy_pixels}{gboolean}

\item{resource_path}{utf8}

\item{stream}{Gio.InputStream}

\item{cancellable}{Gio.Cancellable}

\item{async_result}{Gio.AsyncResult}

\item{path}{filename}

\item{pixbuf}{GdkPixbuf.Pixbuf}

\item{substitute_color}{gboolean}

\item{r}{guint8}

\item{g}{guint8}

\item{b}{guint8}

\item{src}{Pixbuf}

\item{dest}{Pixbuf}

\item{dest_x}{gint}

\item{dest_y}{gint}

\item{dest_width}{gint}

\item{dest_height}{gint}

\item{offset_x}{gdouble}

\item{offset_y}{gdouble}

\item{scale_x}{gdouble}

\item{scale_y}{gdouble}

\item{interp_type}{InterpType}

\item{overall_alpha}{gint}

\item{check_x}{gint}

\item{check_y}{gint}

\item{check_size}{gint}

\item{color1}{guint32}

\item{color2}{guint32}

\item{src_pixbuf}{Pixbuf}

\item{src_x}{gint}

\item{src_y}{gint}

\item{dest_pixbuf}{Pixbuf}

\item{pixel}{guint32}

\item{horizontal}{gboolean}

\item{key}{utf8}

\item{angle}{PixbufRotation}

\item{saturation}{gfloat}

\item{pixelate}{gboolean}

\item{type}{utf8}

\item{option_keys}{utf8}

\item{option_values}{utf8}

\item{value}{utf8}

\item{animation}{PixbufSimpleAnim}

\item{start_time}{GLib.TimeVal}

\item{iter}{PixbufAnimationIter}

\item{current_time}{GLib.TimeVal}

\item{format}{MemoryFormat}

\item{option_key}{utf8}

\item{disabled}{gboolean}

\item{mime_type}{utf8}

\item{image_type}{utf8}

\item{loader}{PixbufLoader}

\item{buf}{guint8}

\item{count}{gsize}

\item{buffer}{GLib.Bytes}

\item{rate}{gfloat}

\item{loop}{gboolean}

\item{bytes}{GLib.Bytes}

\item{file}{Gio.File}

\item{texture}{Texture}

\item{stride}{gsize}

\item{self}{TextureDownloader}

\item{surface}{cairo.Surface}
}
\value{
Pixbuf

Pixbuf

Pixbuf

Pixbuf

Pixbuf

Pixbuf

Pixbuf

Pixbuf

Pixbuf

Pixbuf

Pixbuf

Pixbuf

gint

PixbufFormat

PixbufFormat

GLib.SList

gboolean

gboolean

Pixbuf

Pixbuf

Return value from C function

Return value from C function

Pixbuf

Pixbuf

Return value from C function

gboolean

Return value from C function

Pixbuf

gint

gsize

Colorspace

gboolean

gint

gint

utf8

GLib.HashTable

guint8

guint8

gint

gint

Pixbuf

GLib.Bytes

guint8

gboolean

Pixbuf

Return value from C function

gboolean

gboolean

gboolean

Return value from C function

Pixbuf

gboolean

PixbufAnimation

PixbufAnimation

PixbufAnimation

PixbufAnimation

gint

PixbufAnimationIter

Pixbuf

gint

gboolean

gboolean

gint

Pixbuf

gboolean

GLib.Quark

PixbufFormat

Return value from C function

utf8

utf8

utf8

utf8

utf8

gboolean

gboolean

gboolean

gboolean

Return value from C function

PixbufLoader

PixbufLoader

PixbufLoader

gboolean

PixbufAnimation

PixbufFormat

Pixbuf

Return value from C function

gboolean

gboolean

PixbufSimpleAnim

Return value from C function

gboolean

Return value from C function

Texture

Texture

Texture

Texture

Texture

Return value from C function

MemoryFormat

gint

gint

gboolean

GLib.Bytes

gboolean

GLib.Bytes

TextureDownloader

TextureDownloader

GLib.Bytes

Return value from C function

Return value from C function

MemoryFormat

Texture

Return value from C function

Return value from C function

GLib.Quark

GdkPixbuf.Pixbuf

GdkPixbuf.Pixbuf
}
\description{
gdkPixbufNew

gdkPixbufNewFromBytes

gdkPixbufNewFromFile

gdkPixbufNewFromFileAtScale

gdkPixbufNewFromFileAtSize

gdkPixbufNewFromInline

gdkPixbufNewFromResource

gdkPixbufNewFromResourceAtScale

gdkPixbufNewFromStream

gdkPixbufNewFromStreamAtScale

gdkPixbufNewFromStreamFinish

gdkPixbufNewFromXpmData

gdkPixbufCalculateRowstride

gdkPixbufGetFileInfo

gdkPixbufGetFileInfoFinish

gdkPixbufGetFormats

gdkPixbufInitModules

gdkPixbufSaveToStreamFinish

gdkPixbufAddAlpha

gdkPixbufApplyEmbeddedOrientation

gdkPixbufComposite

gdkPixbufCompositeColor

gdkPixbufCompositeColorSimple

gdkPixbufCopy

gdkPixbufCopyArea

gdkPixbufCopyOptions

gdkPixbufFill

gdkPixbufFlip

gdkPixbufGetBitsPerSample

gdkPixbufGetByteLength

gdkPixbufGetColorspace

gdkPixbufGetHasAlpha

gdkPixbufGetHeight

gdkPixbufGetNChannels

gdkPixbufGetOption

gdkPixbufGetOptions

gdkPixbufGetPixels

gdkPixbufGetPixelsWithLength

gdkPixbufGetRowstride

gdkPixbufGetWidth

gdkPixbufNewSubpixbuf

gdkPixbufReadPixelBytes

gdkPixbufReadPixels

gdkPixbufRemoveOption

gdkPixbufRotateSimple

gdkPixbufSaturateAndPixelate

gdkPixbufSaveToBufferv

gdkPixbufSaveToStreamv

gdkPixbufSavev

gdkPixbufScale

gdkPixbufScaleSimple

gdkPixbufSetOption

gdkPixbufAnimationNewFromFile

gdkPixbufAnimationNewFromResource

gdkPixbufAnimationNewFromStream

gdkPixbufAnimationNewFromStreamFinish

gdkPixbufAnimationGetHeight

gdkPixbufAnimationGetIter

gdkPixbufAnimationGetStaticImage

gdkPixbufAnimationGetWidth

gdkPixbufAnimationIsStaticImage

gdkPixbufAnimationIterAdvance

gdkPixbufAnimationIterGetDelayTime

gdkPixbufAnimationIterGetPixbuf

gdkPixbufAnimationIterOnCurrentlyLoadingFrame

gdkPixbufErrorQuark

gdkPixbufFormatCopy

gdkPixbufFormatFree

gdkPixbufFormatGetDescription

gdkPixbufFormatGetExtensions

gdkPixbufFormatGetLicense

gdkPixbufFormatGetMimeTypes

gdkPixbufFormatGetName

gdkPixbufFormatIsDisabled

gdkPixbufFormatIsSaveOptionSupported

gdkPixbufFormatIsScalable

gdkPixbufFormatIsWritable

gdkPixbufFormatSetDisabled

gdkPixbufLoaderNew

gdkPixbufLoaderNewWithMimeType

gdkPixbufLoaderNewWithType

gdkPixbufLoaderClose

gdkPixbufLoaderGetAnimation

gdkPixbufLoaderGetFormat

gdkPixbufLoaderGetPixbuf

gdkPixbufLoaderSetSize

gdkPixbufLoaderWrite

gdkPixbufLoaderWriteBytes

gdkPixbufSimpleAnimNew

gdkPixbufSimpleAnimAddFrame

gdkPixbufSimpleAnimGetLoop

gdkPixbufSimpleAnimSetLoop

gdkTextureNewForPixbuf

gdkTextureNewFromBytes

gdkTextureNewFromFile

gdkTextureNewFromFilename

gdkTextureNewFromResource

gdkTextureDownload

gdkTextureGetFormat

gdkTextureGetHeight

gdkTextureGetWidth

gdkTextureSaveToPng

gdkTextureSaveToPngBytes

gdkTextureSaveToTiff

gdkTextureSaveToTiffBytes

gdkTextureDownloaderNew

gdkTextureDownloaderCopy

gdkTextureDownloaderDownloadBytes

gdkTextureDownloaderDownloadInto

gdkTextureDownloaderFree

gdkTextureDownloaderGetFormat

gdkTextureDownloaderGetTexture

gdkTextureDownloaderSetFormat

gdkTextureDownloaderSetTexture

gdkTextureErrorQuark

gdkPixbufGetFromSurface

gdkPixbufGetFromTexture
}
