% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R, R/Gio_autogen.R
\name{gFileErrorFromErrno}
\alias{gFileErrorFromErrno}
\alias{gFileErrorQuark}
\alias{gFileGetContents}
\alias{gFileOpenTmp}
\alias{gFileReadLink}
\alias{gFileSetContents}
\alias{gFileTest}
\alias{gFileNewForCommandlineArg}
\alias{gFileNewForCommandlineArgAndCwd}
\alias{gFileNewForPath}
\alias{gFileNewForUri}
\alias{gFileNewTmp}
\alias{gFileParseName}
\alias{gFileAppendTo}
\alias{gFileAppendToAsync}
\alias{gFileAppendToFinish}
\alias{gFileCopy}
\alias{gFileCopyAsync}
\alias{gFileCopyAttributes}
\alias{gFileCopyFinish}
\alias{gFileCreate}
\alias{gFileCreateAsync}
\alias{gFileCreateFinish}
\alias{gFileCreateReadwrite}
\alias{gFileCreateReadwriteAsync}
\alias{gFileCreateReadwriteFinish}
\alias{gFileDelete}
\alias{gFileDeleteAsync}
\alias{gFileDeleteFinish}
\alias{gFileDup}
\alias{gFileEjectMountable}
\alias{gFileEjectMountableFinish}
\alias{gFileEjectMountableWithOperation}
\alias{gFileEjectMountableWithOperationFinish}
\alias{gFileEnumerateChildren}
\alias{gFileEnumerateChildrenAsync}
\alias{gFileEnumerateChildrenFinish}
\alias{gFileEqual}
\alias{gFileFindEnclosingMount}
\alias{gFileFindEnclosingMountAsync}
\alias{gFileFindEnclosingMountFinish}
\alias{gFileGetBasename}
\alias{gFileGetChild}
\alias{gFileGetChildForDisplayName}
\alias{gFileGetParent}
\alias{gFileGetParseName}
\alias{gFileGetPath}
\alias{gFileGetRelativePath}
\alias{gFileGetUri}
\alias{gFileGetUriScheme}
\alias{gFileHasParent}
\alias{gFileHasPrefix}
\alias{gFileHasUriScheme}
\alias{gFileHash}
\alias{gFileIsNative}
\alias{gFileLoadBytes}
\alias{gFileLoadBytesAsync}
\alias{gFileLoadBytesFinish}
\alias{gFileLoadContents}
\alias{gFileLoadContentsAsync}
\alias{gFileLoadContentsFinish}
\alias{gFileLoadPartialContentsFinish}
\alias{gFileMakeDirectory}
\alias{gFileMakeDirectoryAsync}
\alias{gFileMakeDirectoryFinish}
\alias{gFileMakeDirectoryWithParents}
\alias{gFileMakeSymbolicLink}
\alias{gFileMeasureDiskUsage}
\alias{gFileMeasureDiskUsageFinish}
\alias{gFileMonitor}
\alias{gFileMonitorDirectory}
\alias{gFileMonitorFile}
\alias{gFileMountEnclosingVolume}
\alias{gFileMountEnclosingVolumeFinish}
\alias{gFileMountMountable}
\alias{gFileMountMountableFinish}
\alias{gFileMove}
\alias{gFileOpenReadwrite}
\alias{gFileOpenReadwriteAsync}
\alias{gFileOpenReadwriteFinish}
\alias{gFilePeekPath}
\alias{gFilePollMountable}
\alias{gFilePollMountableFinish}
\alias{gFileQueryDefaultHandler}
\alias{gFileQueryExists}
\alias{gFileQueryFileType}
\alias{gFileQueryFilesystemInfo}
\alias{gFileQueryFilesystemInfoAsync}
\alias{gFileQueryFilesystemInfoFinish}
\alias{gFileQueryInfo}
\alias{gFileQueryInfoAsync}
\alias{gFileQueryInfoFinish}
\alias{gFileQuerySettableAttributes}
\alias{gFileQueryWritableNamespaces}
\alias{gFileRead}
\alias{gFileReadAsync}
\alias{gFileReadFinish}
\alias{gFileReplace}
\alias{gFileReplaceAsync}
\alias{gFileReplaceContents}
\alias{gFileReplaceContentsAsync}
\alias{gFileReplaceContentsBytesAsync}
\alias{gFileReplaceContentsFinish}
\alias{gFileReplaceFinish}
\alias{gFileReplaceReadwrite}
\alias{gFileReplaceReadwriteAsync}
\alias{gFileReplaceReadwriteFinish}
\alias{gFileResolveRelativePath}
\alias{gFileSetAttribute}
\alias{gFileSetAttributeByteString}
\alias{gFileSetAttributeInt32}
\alias{gFileSetAttributeInt64}
\alias{gFileSetAttributeString}
\alias{gFileSetAttributeUint32}
\alias{gFileSetAttributeUint64}
\alias{gFileSetAttributesAsync}
\alias{gFileSetAttributesFinish}
\alias{gFileSetAttributesFromInfo}
\alias{gFileSetDisplayName}
\alias{gFileSetDisplayNameAsync}
\alias{gFileSetDisplayNameFinish}
\alias{gFileStartMountable}
\alias{gFileStartMountableFinish}
\alias{gFileStopMountable}
\alias{gFileStopMountableFinish}
\alias{gFileSupportsThreadContexts}
\alias{gFileTrash}
\alias{gFileTrashAsync}
\alias{gFileTrashFinish}
\alias{gFileUnmountMountable}
\alias{gFileUnmountMountableFinish}
\alias{gFileUnmountMountableWithOperation}
\alias{gFileUnmountMountableWithOperationFinish}
\alias{gFileAttributeInfoListNew}
\alias{gFileAttributeInfoListAdd}
\alias{gFileAttributeInfoListDup}
\alias{gFileAttributeInfoListLookup}
\alias{gFileAttributeInfoListRef}
\alias{gFileAttributeInfoListUnref}
\alias{gFileAttributeMatcherNew}
\alias{gFileAttributeMatcherEnumerateNamespace}
\alias{gFileAttributeMatcherEnumerateNext}
\alias{gFileAttributeMatcherMatches}
\alias{gFileAttributeMatcherMatchesOnly}
\alias{gFileAttributeMatcherRef}
\alias{gFileAttributeMatcherSubtract}
\alias{gFileAttributeMatcherToString}
\alias{gFileAttributeMatcherUnref}
\alias{gFileEnumeratorClose}
\alias{gFileEnumeratorCloseAsync}
\alias{gFileEnumeratorCloseFinish}
\alias{gFileEnumeratorGetChild}
\alias{gFileEnumeratorGetContainer}
\alias{gFileEnumeratorHasPending}
\alias{gFileEnumeratorIsClosed}
\alias{gFileEnumeratorIterate}
\alias{gFileEnumeratorNextFile}
\alias{gFileEnumeratorNextFilesAsync}
\alias{gFileEnumeratorNextFilesFinish}
\alias{gFileEnumeratorSetPending}
\alias{gFileIoStreamGetEtag}
\alias{gFileIoStreamQueryInfo}
\alias{gFileIoStreamQueryInfoAsync}
\alias{gFileIoStreamQueryInfoFinish}
\alias{gFileIconNew}
\alias{gFileIconGetFile}
\alias{gFileInfoNew}
\alias{gFileInfoClearStatus}
\alias{gFileInfoCopyInto}
\alias{gFileInfoDup}
\alias{gFileInfoGetAttributeAsString}
\alias{gFileInfoGetAttributeBoolean}
\alias{gFileInfoGetAttributeByteString}
\alias{gFileInfoGetAttributeData}
\alias{gFileInfoGetAttributeInt32}
\alias{gFileInfoGetAttributeInt64}
\alias{gFileInfoGetAttributeObject}
\alias{gFileInfoGetAttributeStatus}
\alias{gFileInfoGetAttributeString}
\alias{gFileInfoGetAttributeStringv}
\alias{gFileInfoGetAttributeType}
\alias{gFileInfoGetAttributeUint32}
\alias{gFileInfoGetAttributeUint64}
\alias{gFileInfoGetContentType}
\alias{gFileInfoGetDeletionDate}
\alias{gFileInfoGetDisplayName}
\alias{gFileInfoGetEditName}
\alias{gFileInfoGetEtag}
\alias{gFileInfoGetFileType}
\alias{gFileInfoGetIcon}
\alias{gFileInfoGetIsBackup}
\alias{gFileInfoGetIsHidden}
\alias{gFileInfoGetIsSymlink}
\alias{gFileInfoGetModificationTime}
\alias{gFileInfoGetName}
\alias{gFileInfoGetSize}
\alias{gFileInfoGetSortOrder}
\alias{gFileInfoGetSymbolicIcon}
\alias{gFileInfoGetSymlinkTarget}
\alias{gFileInfoHasAttribute}
\alias{gFileInfoHasNamespace}
\alias{gFileInfoListAttributes}
\alias{gFileInfoRemoveAttribute}
\alias{gFileInfoSetAttribute}
\alias{gFileInfoSetAttributeBoolean}
\alias{gFileInfoSetAttributeByteString}
\alias{gFileInfoSetAttributeInt32}
\alias{gFileInfoSetAttributeInt64}
\alias{gFileInfoSetAttributeMask}
\alias{gFileInfoSetAttributeObject}
\alias{gFileInfoSetAttributeStatus}
\alias{gFileInfoSetAttributeString}
\alias{gFileInfoSetAttributeStringv}
\alias{gFileInfoSetAttributeUint32}
\alias{gFileInfoSetAttributeUint64}
\alias{gFileInfoSetContentType}
\alias{gFileInfoSetDisplayName}
\alias{gFileInfoSetEditName}
\alias{gFileInfoSetFileType}
\alias{gFileInfoSetIcon}
\alias{gFileInfoSetIsHidden}
\alias{gFileInfoSetIsSymlink}
\alias{gFileInfoSetModificationTime}
\alias{gFileInfoSetName}
\alias{gFileInfoSetSize}
\alias{gFileInfoSetSortOrder}
\alias{gFileInfoSetSymbolicIcon}
\alias{gFileInfoSetSymlinkTarget}
\alias{gFileInfoUnsetAttributeMask}
\alias{gFileInputStreamQueryInfo}
\alias{gFileInputStreamQueryInfoAsync}
\alias{gFileInputStreamQueryInfoFinish}
\alias{gFileMonitorCancel}
\alias{gFileMonitorEmitEvent}
\alias{gFileMonitorIsCancelled}
\alias{gFileMonitorSetRateLimit}
\alias{gFileOutputStreamGetEtag}
\alias{gFileOutputStreamQueryInfo}
\alias{gFileOutputStreamQueryInfoAsync}
\alias{gFileOutputStreamQueryInfoFinish}
\title{gFileErrorFromErrno}
\usage{
gFileErrorFromErrno(err_no)

gFileErrorQuark()

gFileGetContents(filename)

gFileOpenTmp(tmpl)

gFileReadLink(filename)

gFileSetContents(filename, contents, length)

gFileTest(filename, test)

gFileNewForCommandlineArg(arg)

gFileNewForCommandlineArgAndCwd(arg, cwd)

gFileNewForPath(path)

gFileNewForUri(uri)

gFileNewTmp(tmpl)

gFileParseName(parse_name)

gFileAppendTo(file, flags, cancellable)

gFileAppendToAsync(file, flags, io_priority, cancellable, callback)

gFileAppendToFinish(file, res)

gFileCopy(source, destination, flags, cancellable, progress_callback)

gFileCopyAsync(
  source,
  destination,
  flags,
  io_priority,
  cancellable,
  progress_callback,
  callback
)

gFileCopyAttributes(source, destination, flags, cancellable)

gFileCopyFinish(file, res)

gFileCreate(file, flags, cancellable)

gFileCreateAsync(file, flags, io_priority, cancellable, callback)

gFileCreateFinish(file, res)

gFileCreateReadwrite(file, flags, cancellable)

gFileCreateReadwriteAsync(file, flags, io_priority, cancellable, callback)

gFileCreateReadwriteFinish(file, res)

gFileDelete(file, cancellable)

gFileDeleteAsync(file, io_priority, cancellable, callback)

gFileDeleteFinish(file, result)

gFileDup(file)

gFileEjectMountable(file, flags, cancellable, callback)

gFileEjectMountableFinish(file, result)

gFileEjectMountableWithOperation(
  file,
  flags,
  mount_operation,
  cancellable,
  callback
)

gFileEjectMountableWithOperationFinish(file, result)

gFileEnumerateChildren(file, attributes, flags, cancellable)

gFileEnumerateChildrenAsync(
  file,
  attributes,
  flags,
  io_priority,
  cancellable,
  callback
)

gFileEnumerateChildrenFinish(file, res)

gFileEqual(file1, file2)

gFileFindEnclosingMount(file, cancellable)

gFileFindEnclosingMountAsync(file, io_priority, cancellable, callback)

gFileFindEnclosingMountFinish(file, res)

gFileGetBasename(file)

gFileGetChild(file, name)

gFileGetChildForDisplayName(file, display_name)

gFileGetParent(file)

gFileGetParseName(file)

gFileGetPath(file)

gFileGetRelativePath(parent, descendant)

gFileGetUri(file)

gFileGetUriScheme(file)

gFileHasParent(file, parent)

gFileHasPrefix(file, prefix)

gFileHasUriScheme(file, uri_scheme)

gFileHash(file)

gFileIsNative(file)

gFileLoadBytes(file, cancellable)

gFileLoadBytesAsync(file, cancellable, callback)

gFileLoadBytesFinish(file, result)

gFileLoadContents(file, cancellable)

gFileLoadContentsAsync(file, cancellable, callback)

gFileLoadContentsFinish(file, res)

gFileLoadPartialContentsFinish(file, res)

gFileMakeDirectory(file, cancellable)

gFileMakeDirectoryAsync(file, io_priority, cancellable, callback)

gFileMakeDirectoryFinish(file, result)

gFileMakeDirectoryWithParents(file, cancellable)

gFileMakeSymbolicLink(file, symlink_value, cancellable)

gFileMeasureDiskUsage(file, flags, cancellable, progress_callback)

gFileMeasureDiskUsageFinish(file, result)

gFileMonitor(file, flags, cancellable)

gFileMonitorDirectory(file, flags, cancellable)

gFileMonitorFile(file, flags, cancellable)

gFileMountEnclosingVolume(
  location,
  flags,
  mount_operation,
  cancellable,
  callback
)

gFileMountEnclosingVolumeFinish(location, result)

gFileMountMountable(file, flags, mount_operation, cancellable, callback)

gFileMountMountableFinish(file, result)

gFileMove(source, destination, flags, cancellable, progress_callback)

gFileOpenReadwrite(file, cancellable)

gFileOpenReadwriteAsync(file, io_priority, cancellable, callback)

gFileOpenReadwriteFinish(file, res)

gFilePeekPath(file)

gFilePollMountable(file, cancellable, callback)

gFilePollMountableFinish(file, result)

gFileQueryDefaultHandler(file, cancellable)

gFileQueryExists(file, cancellable)

gFileQueryFileType(file, flags, cancellable)

gFileQueryFilesystemInfo(file, attributes, cancellable)

gFileQueryFilesystemInfoAsync(
  file,
  attributes,
  io_priority,
  cancellable,
  callback
)

gFileQueryFilesystemInfoFinish(file, res)

gFileQueryInfo(file, attributes, flags, cancellable)

gFileQueryInfoAsync(
  file,
  attributes,
  flags,
  io_priority,
  cancellable,
  callback
)

gFileQueryInfoFinish(file, res)

gFileQuerySettableAttributes(file, cancellable)

gFileQueryWritableNamespaces(file, cancellable)

gFileRead(file, cancellable)

gFileReadAsync(file, io_priority, cancellable, callback)

gFileReadFinish(file, res)

gFileReplace(file, etag, make_backup, flags, cancellable)

gFileReplaceAsync(
  file,
  etag,
  make_backup,
  flags,
  io_priority,
  cancellable,
  callback
)

gFileReplaceContents(
  file,
  contents,
  length,
  etag,
  make_backup,
  flags,
  cancellable
)

gFileReplaceContentsAsync(
  file,
  contents,
  length,
  etag,
  make_backup,
  flags,
  cancellable,
  callback
)

gFileReplaceContentsBytesAsync(
  file,
  contents,
  etag,
  make_backup,
  flags,
  cancellable,
  callback
)

gFileReplaceContentsFinish(file, res)

gFileReplaceFinish(file, res)

gFileReplaceReadwrite(file, etag, make_backup, flags, cancellable)

gFileReplaceReadwriteAsync(
  file,
  etag,
  make_backup,
  flags,
  io_priority,
  cancellable,
  callback
)

gFileReplaceReadwriteFinish(file, res)

gFileResolveRelativePath(file, relative_path)

gFileSetAttribute(file, attribute, type, value_p, flags, cancellable)

gFileSetAttributeByteString(file, attribute, value, flags, cancellable)

gFileSetAttributeInt32(file, attribute, value, flags, cancellable)

gFileSetAttributeInt64(file, attribute, value, flags, cancellable)

gFileSetAttributeString(file, attribute, value, flags, cancellable)

gFileSetAttributeUint32(file, attribute, value, flags, cancellable)

gFileSetAttributeUint64(file, attribute, value, flags, cancellable)

gFileSetAttributesAsync(file, info, flags, io_priority, cancellable, callback)

gFileSetAttributesFinish(file, result)

gFileSetAttributesFromInfo(file, info, flags, cancellable)

gFileSetDisplayName(file, display_name, cancellable)

gFileSetDisplayNameAsync(
  file,
  display_name,
  io_priority,
  cancellable,
  callback
)

gFileSetDisplayNameFinish(file, res)

gFileStartMountable(file, flags, start_operation, cancellable, callback)

gFileStartMountableFinish(file, result)

gFileStopMountable(file, flags, mount_operation, cancellable, callback)

gFileStopMountableFinish(file, result)

gFileSupportsThreadContexts(file)

gFileTrash(file, cancellable)

gFileTrashAsync(file, io_priority, cancellable, callback)

gFileTrashFinish(file, result)

gFileUnmountMountable(file, flags, cancellable, callback)

gFileUnmountMountableFinish(file, result)

gFileUnmountMountableWithOperation(
  file,
  flags,
  mount_operation,
  cancellable,
  callback
)

gFileUnmountMountableWithOperationFinish(file, result)

gFileAttributeInfoListNew()

gFileAttributeInfoListAdd(list, name, type, flags)

gFileAttributeInfoListDup(list)

gFileAttributeInfoListLookup(list, name)

gFileAttributeInfoListRef(list)

gFileAttributeInfoListUnref(list)

gFileAttributeMatcherNew(attributes)

gFileAttributeMatcherEnumerateNamespace(matcher, ns)

gFileAttributeMatcherEnumerateNext(matcher)

gFileAttributeMatcherMatches(matcher, attribute)

gFileAttributeMatcherMatchesOnly(matcher, attribute)

gFileAttributeMatcherRef(matcher)

gFileAttributeMatcherSubtract(matcher, subtract)

gFileAttributeMatcherToString(matcher)

gFileAttributeMatcherUnref(matcher)

gFileEnumeratorClose(enumerator, cancellable)

gFileEnumeratorCloseAsync(enumerator, io_priority, cancellable, callback)

gFileEnumeratorCloseFinish(enumerator, result)

gFileEnumeratorGetChild(enumerator, info)

gFileEnumeratorGetContainer(enumerator)

gFileEnumeratorHasPending(enumerator)

gFileEnumeratorIsClosed(enumerator)

gFileEnumeratorIterate(direnum, cancellable)

gFileEnumeratorNextFile(enumerator, cancellable)

gFileEnumeratorNextFilesAsync(
  enumerator,
  num_files,
  io_priority,
  cancellable,
  callback
)

gFileEnumeratorNextFilesFinish(enumerator, result)

gFileEnumeratorSetPending(enumerator, pending)

gFileIoStreamGetEtag(stream)

gFileIoStreamQueryInfo(stream, attributes, cancellable)

gFileIoStreamQueryInfoAsync(
  stream,
  attributes,
  io_priority,
  cancellable,
  callback
)

gFileIoStreamQueryInfoFinish(stream, result)

gFileIconNew(file)

gFileIconGetFile(icon)

gFileInfoNew()

gFileInfoClearStatus(info)

gFileInfoCopyInto(src_info, dest_info)

gFileInfoDup(other)

gFileInfoGetAttributeAsString(info, attribute)

gFileInfoGetAttributeBoolean(info, attribute)

gFileInfoGetAttributeByteString(info, attribute)

gFileInfoGetAttributeData(info, attribute)

gFileInfoGetAttributeInt32(info, attribute)

gFileInfoGetAttributeInt64(info, attribute)

gFileInfoGetAttributeObject(info, attribute)

gFileInfoGetAttributeStatus(info, attribute)

gFileInfoGetAttributeString(info, attribute)

gFileInfoGetAttributeStringv(info, attribute)

gFileInfoGetAttributeType(info, attribute)

gFileInfoGetAttributeUint32(info, attribute)

gFileInfoGetAttributeUint64(info, attribute)

gFileInfoGetContentType(info)

gFileInfoGetDeletionDate(info)

gFileInfoGetDisplayName(info)

gFileInfoGetEditName(info)

gFileInfoGetEtag(info)

gFileInfoGetFileType(info)

gFileInfoGetIcon(info)

gFileInfoGetIsBackup(info)

gFileInfoGetIsHidden(info)

gFileInfoGetIsSymlink(info)

gFileInfoGetModificationTime(info)

gFileInfoGetName(info)

gFileInfoGetSize(info)

gFileInfoGetSortOrder(info)

gFileInfoGetSymbolicIcon(info)

gFileInfoGetSymlinkTarget(info)

gFileInfoHasAttribute(info, attribute)

gFileInfoHasNamespace(info, name_space)

gFileInfoListAttributes(info, name_space)

gFileInfoRemoveAttribute(info, attribute)

gFileInfoSetAttribute(info, attribute, type, value_p)

gFileInfoSetAttributeBoolean(info, attribute, attr_value)

gFileInfoSetAttributeByteString(info, attribute, attr_value)

gFileInfoSetAttributeInt32(info, attribute, attr_value)

gFileInfoSetAttributeInt64(info, attribute, attr_value)

gFileInfoSetAttributeMask(info, mask)

gFileInfoSetAttributeObject(info, attribute, attr_value)

gFileInfoSetAttributeStatus(info, attribute, status)

gFileInfoSetAttributeString(info, attribute, attr_value)

gFileInfoSetAttributeStringv(info, attribute, attr_value)

gFileInfoSetAttributeUint32(info, attribute, attr_value)

gFileInfoSetAttributeUint64(info, attribute, attr_value)

gFileInfoSetContentType(info, content_type)

gFileInfoSetDisplayName(info, display_name)

gFileInfoSetEditName(info, edit_name)

gFileInfoSetFileType(info, type)

gFileInfoSetIcon(info, icon)

gFileInfoSetIsHidden(info, is_hidden)

gFileInfoSetIsSymlink(info, is_symlink)

gFileInfoSetModificationTime(info, mtime)

gFileInfoSetName(info, name)

gFileInfoSetSize(info, size)

gFileInfoSetSortOrder(info, sort_order)

gFileInfoSetSymbolicIcon(info, icon)

gFileInfoSetSymlinkTarget(info, symlink_target)

gFileInfoUnsetAttributeMask(info)

gFileInputStreamQueryInfo(stream, attributes, cancellable)

gFileInputStreamQueryInfoAsync(
  stream,
  attributes,
  io_priority,
  cancellable,
  callback
)

gFileInputStreamQueryInfoFinish(stream, result)

gFileMonitorCancel(monitor)

gFileMonitorEmitEvent(monitor, child, other_file, event_type)

gFileMonitorIsCancelled(monitor)

gFileMonitorSetRateLimit(monitor, limit_msecs)

gFileOutputStreamGetEtag(stream)

gFileOutputStreamQueryInfo(stream, attributes, cancellable)

gFileOutputStreamQueryInfoAsync(
  stream,
  attributes,
  io_priority,
  cancellable,
  callback
)

gFileOutputStreamQueryInfoFinish(stream, result)
}
\arguments{
\item{err_no}{gint}

\item{filename}{filename}

\item{tmpl}{filename}

\item{contents}{GLib.Bytes}

\item{length}{gsize}

\item{test}{FileTest}

\item{arg}{filename}

\item{cwd}{filename}

\item{path}{filename}

\item{uri}{utf8}

\item{parse_name}{utf8}

\item{file}{File}

\item{flags}{FileAttributeInfoFlags}

\item{cancellable}{Cancellable}

\item{io_priority}{gint}

\item{callback}{function — AsyncReadyCallback function(source_object, res)}

\item{res}{AsyncResult}

\item{source}{File}

\item{destination}{File}

\item{progress_callback}{function — FileProgressCallback function(current_num_bytes, total_num_bytes)}

\item{result}{AsyncResult}

\item{mount_operation}{MountOperation}

\item{attributes}{utf8}

\item{file1}{File}

\item{file2}{File}

\item{name}{filename}

\item{display_name}{utf8}

\item{parent}{File}

\item{descendant}{File}

\item{prefix}{File}

\item{uri_scheme}{utf8}

\item{symlink_value}{filename}

\item{location}{File}

\item{etag}{utf8}

\item{make_backup}{gboolean}

\item{relative_path}{filename}

\item{attribute}{utf8}

\item{type}{FileType}

\item{value_p}{gpointer}

\item{value}{guint64}

\item{info}{FileInfo}

\item{start_operation}{MountOperation}

\item{list}{FileAttributeInfoList}

\item{matcher}{FileAttributeMatcher}

\item{ns}{utf8}

\item{subtract}{FileAttributeMatcher}

\item{enumerator}{FileEnumerator}

\item{direnum}{FileEnumerator}

\item{num_files}{gint}

\item{pending}{gboolean}

\item{stream}{FileOutputStream}

\item{icon}{Icon}

\item{src_info}{FileInfo}

\item{dest_info}{FileInfo}

\item{other}{FileInfo}

\item{name_space}{utf8}

\item{attr_value}{guint64}

\item{mask}{FileAttributeMatcher}

\item{status}{FileAttributeStatus}

\item{content_type}{utf8}

\item{edit_name}{utf8}

\item{is_hidden}{gboolean}

\item{is_symlink}{gboolean}

\item{mtime}{GLib.TimeVal}

\item{size}{gint64}

\item{sort_order}{gint32}

\item{symlink_target}{filename}

\item{monitor}{FileMonitor}

\item{child}{File}

\item{other_file}{File}

\item{event_type}{FileMonitorEvent}

\item{limit_msecs}{gint}
}
\value{
FileError

Quark

gboolean

gint

filename

gboolean

gboolean

File

File

File

File

File

File

FileOutputStream

Return value from C function

FileOutputStream

gboolean

Return value from C function

gboolean

gboolean

FileOutputStream

Return value from C function

FileOutputStream

FileIOStream

Return value from C function

FileIOStream

gboolean

Return value from C function

gboolean

File

Return value from C function

gboolean

Return value from C function

gboolean

FileEnumerator

Return value from C function

FileEnumerator

gboolean

Mount

Return value from C function

Mount

filename

File

File

File

utf8

filename

filename

utf8

utf8

gboolean

gboolean

gboolean

guint

gboolean

GLib.Bytes

Return value from C function

GLib.Bytes

gboolean

Return value from C function

gboolean

gboolean

gboolean

Return value from C function

gboolean

gboolean

gboolean

gboolean

gboolean

FileMonitor

FileMonitor

FileMonitor

Return value from C function

gboolean

Return value from C function

File

gboolean

FileIOStream

Return value from C function

FileIOStream

filename

Return value from C function

gboolean

AppInfo

gboolean

FileType

FileInfo

Return value from C function

FileInfo

FileInfo

Return value from C function

FileInfo

FileAttributeInfoList

FileAttributeInfoList

FileInputStream

Return value from C function

FileInputStream

FileOutputStream

Return value from C function

gboolean

Return value from C function

Return value from C function

gboolean

FileOutputStream

FileIOStream

Return value from C function

FileIOStream

File

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

Return value from C function

gboolean

gboolean

File

Return value from C function

File

Return value from C function

gboolean

Return value from C function

gboolean

gboolean

gboolean

Return value from C function

gboolean

Return value from C function

gboolean

Return value from C function

gboolean

FileAttributeInfoList

Return value from C function

FileAttributeInfoList

FileAttributeInfo

FileAttributeInfoList

Return value from C function

FileAttributeMatcher

gboolean

utf8

gboolean

gboolean

FileAttributeMatcher

FileAttributeMatcher

utf8

Return value from C function

gboolean

Return value from C function

gboolean

File

File

gboolean

gboolean

gboolean

FileInfo

Return value from C function

GLib.List

Return value from C function

utf8

FileInfo

Return value from C function

FileInfo

FileIcon

File

FileInfo

Return value from C function

Return value from C function

FileInfo

utf8

gboolean

utf8

gboolean

gint32

gint64

GObject.Object

FileAttributeStatus

utf8

utf8

FileAttributeType

guint32

guint64

utf8

GLib.DateTime

utf8

utf8

utf8

FileType

Icon

gboolean

gboolean

gboolean

Return value from C function

filename

gint64

gint32

Icon

filename

gboolean

gboolean

utf8

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

FileInfo

Return value from C function

FileInfo

gboolean

Return value from C function

gboolean

Return value from C function

utf8

FileInfo

Return value from C function

FileInfo
}
\description{
gFileErrorFromErrno

gFileErrorQuark

gFileGetContents

gFileOpenTmp

gFileReadLink

gFileSetContents

gFileTest

gFileNewForCommandlineArg

gFileNewForCommandlineArgAndCwd

gFileNewForPath

gFileNewForUri

gFileNewTmp

gFileParseName

gFileAppendTo

gFileAppendToAsync

gFileAppendToFinish

gFileCopy

gFileCopyAsync

gFileCopyAttributes

gFileCopyFinish

gFileCreate

gFileCreateAsync

gFileCreateFinish

gFileCreateReadwrite

gFileCreateReadwriteAsync

gFileCreateReadwriteFinish

gFileDelete

gFileDeleteAsync

gFileDeleteFinish

gFileDup

gFileEjectMountable

gFileEjectMountableFinish

gFileEjectMountableWithOperation

gFileEjectMountableWithOperationFinish

gFileEnumerateChildren

gFileEnumerateChildrenAsync

gFileEnumerateChildrenFinish

gFileEqual

gFileFindEnclosingMount

gFileFindEnclosingMountAsync

gFileFindEnclosingMountFinish

gFileGetBasename

gFileGetChild

gFileGetChildForDisplayName

gFileGetParent

gFileGetParseName

gFileGetPath

gFileGetRelativePath

gFileGetUri

gFileGetUriScheme

gFileHasParent

gFileHasPrefix

gFileHasUriScheme

gFileHash

gFileIsNative

gFileLoadBytes

gFileLoadBytesAsync

gFileLoadBytesFinish

gFileLoadContents

gFileLoadContentsAsync

gFileLoadContentsFinish

gFileLoadPartialContentsFinish

gFileMakeDirectory

gFileMakeDirectoryAsync

gFileMakeDirectoryFinish

gFileMakeDirectoryWithParents

gFileMakeSymbolicLink

gFileMeasureDiskUsage

gFileMeasureDiskUsageFinish

gFileMonitor

gFileMonitorDirectory

gFileMonitorFile

gFileMountEnclosingVolume

gFileMountEnclosingVolumeFinish

gFileMountMountable

gFileMountMountableFinish

gFileMove

gFileOpenReadwrite

gFileOpenReadwriteAsync

gFileOpenReadwriteFinish

gFilePeekPath

gFilePollMountable

gFilePollMountableFinish

gFileQueryDefaultHandler

gFileQueryExists

gFileQueryFileType

gFileQueryFilesystemInfo

gFileQueryFilesystemInfoAsync

gFileQueryFilesystemInfoFinish

gFileQueryInfo

gFileQueryInfoAsync

gFileQueryInfoFinish

gFileQuerySettableAttributes

gFileQueryWritableNamespaces

gFileRead

gFileReadAsync

gFileReadFinish

gFileReplace

gFileReplaceAsync

gFileReplaceContents

gFileReplaceContentsAsync

gFileReplaceContentsBytesAsync

gFileReplaceContentsFinish

gFileReplaceFinish

gFileReplaceReadwrite

gFileReplaceReadwriteAsync

gFileReplaceReadwriteFinish

gFileResolveRelativePath

gFileSetAttribute

gFileSetAttributeByteString

gFileSetAttributeInt32

gFileSetAttributeInt64

gFileSetAttributeString

gFileSetAttributeUint32

gFileSetAttributeUint64

gFileSetAttributesAsync

gFileSetAttributesFinish

gFileSetAttributesFromInfo

gFileSetDisplayName

gFileSetDisplayNameAsync

gFileSetDisplayNameFinish

gFileStartMountable

gFileStartMountableFinish

gFileStopMountable

gFileStopMountableFinish

gFileSupportsThreadContexts

gFileTrash

gFileTrashAsync

gFileTrashFinish

gFileUnmountMountable

gFileUnmountMountableFinish

gFileUnmountMountableWithOperation

gFileUnmountMountableWithOperationFinish

gFileAttributeInfoListNew

gFileAttributeInfoListAdd

gFileAttributeInfoListDup

gFileAttributeInfoListLookup

gFileAttributeInfoListRef

gFileAttributeInfoListUnref

gFileAttributeMatcherNew

gFileAttributeMatcherEnumerateNamespace

gFileAttributeMatcherEnumerateNext

gFileAttributeMatcherMatches

gFileAttributeMatcherMatchesOnly

gFileAttributeMatcherRef

gFileAttributeMatcherSubtract

gFileAttributeMatcherToString

gFileAttributeMatcherUnref

gFileEnumeratorClose

gFileEnumeratorCloseAsync

gFileEnumeratorCloseFinish

gFileEnumeratorGetChild

gFileEnumeratorGetContainer

gFileEnumeratorHasPending

gFileEnumeratorIsClosed

gFileEnumeratorIterate

gFileEnumeratorNextFile

gFileEnumeratorNextFilesAsync

gFileEnumeratorNextFilesFinish

gFileEnumeratorSetPending

gFileIoStreamGetEtag

gFileIoStreamQueryInfo

gFileIoStreamQueryInfoAsync

gFileIoStreamQueryInfoFinish

gFileIconNew

gFileIconGetFile

gFileInfoNew

gFileInfoClearStatus

gFileInfoCopyInto

gFileInfoDup

gFileInfoGetAttributeAsString

gFileInfoGetAttributeBoolean

gFileInfoGetAttributeByteString

gFileInfoGetAttributeData

gFileInfoGetAttributeInt32

gFileInfoGetAttributeInt64

gFileInfoGetAttributeObject

gFileInfoGetAttributeStatus

gFileInfoGetAttributeString

gFileInfoGetAttributeStringv

gFileInfoGetAttributeType

gFileInfoGetAttributeUint32

gFileInfoGetAttributeUint64

gFileInfoGetContentType

gFileInfoGetDeletionDate

gFileInfoGetDisplayName

gFileInfoGetEditName

gFileInfoGetEtag

gFileInfoGetFileType

gFileInfoGetIcon

gFileInfoGetIsBackup

gFileInfoGetIsHidden

gFileInfoGetIsSymlink

gFileInfoGetModificationTime

gFileInfoGetName

gFileInfoGetSize

gFileInfoGetSortOrder

gFileInfoGetSymbolicIcon

gFileInfoGetSymlinkTarget

gFileInfoHasAttribute

gFileInfoHasNamespace

gFileInfoListAttributes

gFileInfoRemoveAttribute

gFileInfoSetAttribute

gFileInfoSetAttributeBoolean

gFileInfoSetAttributeByteString

gFileInfoSetAttributeInt32

gFileInfoSetAttributeInt64

gFileInfoSetAttributeMask

gFileInfoSetAttributeObject

gFileInfoSetAttributeStatus

gFileInfoSetAttributeString

gFileInfoSetAttributeStringv

gFileInfoSetAttributeUint32

gFileInfoSetAttributeUint64

gFileInfoSetContentType

gFileInfoSetDisplayName

gFileInfoSetEditName

gFileInfoSetFileType

gFileInfoSetIcon

gFileInfoSetIsHidden

gFileInfoSetIsSymlink

gFileInfoSetModificationTime

gFileInfoSetName

gFileInfoSetSize

gFileInfoSetSortOrder

gFileInfoSetSymbolicIcon

gFileInfoSetSymlinkTarget

gFileInfoUnsetAttributeMask

gFileInputStreamQueryInfo

gFileInputStreamQueryInfoAsync

gFileInputStreamQueryInfoFinish

gFileMonitorCancel

gFileMonitorEmitEvent

gFileMonitorIsCancelled

gFileMonitorSetRateLimit

gFileOutputStreamGetEtag

gFileOutputStreamQueryInfo

gFileOutputStreamQueryInfoAsync

gFileOutputStreamQueryInfoFinish
}
