% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R, R/Gio_autogen.R
\name{gIoChannelNewFile}
\alias{gIoChannelNewFile}
\alias{gIoChannelUnixNew}
\alias{gIoChannelClose}
\alias{gIoChannelFlush}
\alias{gIoChannelGetBufferCondition}
\alias{gIoChannelGetBufferSize}
\alias{gIoChannelGetBuffered}
\alias{gIoChannelGetCloseOnUnref}
\alias{gIoChannelGetEncoding}
\alias{gIoChannelGetFlags}
\alias{gIoChannelGetLineTerm}
\alias{gIoChannelInit}
\alias{gIoChannelRead}
\alias{gIoChannelReadChars}
\alias{gIoChannelReadLine}
\alias{gIoChannelReadLineString}
\alias{gIoChannelReadToEnd}
\alias{gIoChannelReadUnichar}
\alias{gIoChannelRef}
\alias{gIoChannelSeek}
\alias{gIoChannelSeekPosition}
\alias{gIoChannelSetBufferSize}
\alias{gIoChannelSetBuffered}
\alias{gIoChannelSetCloseOnUnref}
\alias{gIoChannelSetEncoding}
\alias{gIoChannelSetFlags}
\alias{gIoChannelSetLineTerm}
\alias{gIoChannelShutdown}
\alias{gIoChannelUnixGetFd}
\alias{gIoChannelUnref}
\alias{gIoChannelWrite}
\alias{gIoChannelWriteChars}
\alias{gIoChannelWriteUnichar}
\alias{gIoChannelErrorFromErrno}
\alias{gIoChannelErrorQuark}
\alias{gIoCreateWatch}
\alias{gIoExtensionGetName}
\alias{gIoExtensionGetPriority}
\alias{gIoExtensionGetType}
\alias{gIoExtensionPointGetExtensionByName}
\alias{gIoExtensionPointGetExtensions}
\alias{gIoExtensionPointGetRequiredType}
\alias{gIoExtensionPointSetRequiredType}
\alias{gIoExtensionPointImplement}
\alias{gIoExtensionPointLookup}
\alias{gIoExtensionPointRegister}
\alias{gIoStreamSpliceFinish}
\alias{gIoStreamClearPending}
\alias{gIoStreamClose}
\alias{gIoStreamCloseFinish}
\alias{gIoStreamGetInputStream}
\alias{gIoStreamGetOutputStream}
\alias{gIoStreamHasPending}
\alias{gIoStreamIsClosed}
\alias{gIoStreamSetPending}
\alias{gInputStreamClearPending}
\alias{gInputStreamClose}
\alias{gInputStreamCloseFinish}
\alias{gInputStreamHasPending}
\alias{gInputStreamIsClosed}
\alias{gInputStreamRead}
\alias{gInputStreamReadAll}
\alias{gInputStreamReadAllFinish}
\alias{gInputStreamReadBytes}
\alias{gInputStreamReadBytesFinish}
\alias{gInputStreamReadFinish}
\alias{gInputStreamSetPending}
\alias{gInputStreamSkip}
\alias{gInputStreamSkipFinish}
\alias{gOutputStreamClearPending}
\alias{gOutputStreamClose}
\alias{gOutputStreamCloseFinish}
\alias{gOutputStreamFlush}
\alias{gOutputStreamFlushFinish}
\alias{gOutputStreamHasPending}
\alias{gOutputStreamIsClosed}
\alias{gOutputStreamIsClosing}
\alias{gOutputStreamSetPending}
\alias{gOutputStreamSplice}
\alias{gOutputStreamSpliceFinish}
\alias{gOutputStreamWrite}
\alias{gOutputStreamWriteAll}
\alias{gOutputStreamWriteAllFinish}
\alias{gOutputStreamWriteBytes}
\alias{gOutputStreamWriteBytesFinish}
\alias{gOutputStreamWriteFinish}
\alias{gIoErrorFromErrno}
\alias{gIoErrorQuark}
\alias{gIoModulesLoadAllInDirectory}
\alias{gIoModulesLoadAllInDirectoryWithScope}
\alias{gIoModulesScanAllInDirectory}
\alias{gIoModulesScanAllInDirectoryWithScope}
\alias{gIoSchedulerCancelAllJobs}
\title{gIoChannelNewFile}
\usage{
gIoChannelNewFile(filename, mode)

gIoChannelUnixNew(fd)

gIoChannelClose(channel)

gIoChannelFlush(channel)

gIoChannelGetBufferCondition(channel)

gIoChannelGetBufferSize(channel)

gIoChannelGetBuffered(channel)

gIoChannelGetCloseOnUnref(channel)

gIoChannelGetEncoding(channel)

gIoChannelGetFlags(channel)

gIoChannelGetLineTerm(channel)

gIoChannelInit(channel)

gIoChannelRead(channel, buf, count, bytes_read)

gIoChannelReadChars(channel, count)

gIoChannelReadLine(channel)

gIoChannelReadLineString(channel, buffer, terminator_pos)

gIoChannelReadToEnd(channel)

gIoChannelReadUnichar(channel)

gIoChannelRef(channel)

gIoChannelSeek(channel, offset, type)

gIoChannelSeekPosition(channel, offset, type)

gIoChannelSetBufferSize(channel, size)

gIoChannelSetBuffered(channel, buffered)

gIoChannelSetCloseOnUnref(channel, do_close)

gIoChannelSetEncoding(channel, encoding)

gIoChannelSetFlags(channel, flags)

gIoChannelSetLineTerm(channel, line_term, length)

gIoChannelShutdown(channel, flush)

gIoChannelUnixGetFd(channel)

gIoChannelUnref(channel)

gIoChannelWrite(channel, buf, count, bytes_written)

gIoChannelWriteChars(channel, buf, count)

gIoChannelWriteUnichar(channel, thechar)

gIoChannelErrorFromErrno(en)

gIoChannelErrorQuark()

gIoCreateWatch(channel, condition)

gIoExtensionGetName(extension)

gIoExtensionGetPriority(extension)

gIoExtensionGetType(extension)

gIoExtensionPointGetExtensionByName(extension_point, name)

gIoExtensionPointGetExtensions(extension_point)

gIoExtensionPointGetRequiredType(extension_point)

gIoExtensionPointSetRequiredType(extension_point, type)

gIoExtensionPointImplement(
  extension_point_name,
  type,
  extension_name,
  priority
)

gIoExtensionPointLookup(name)

gIoExtensionPointRegister(name)

gIoStreamSpliceFinish(result)

gIoStreamClearPending(stream)

gIoStreamClose(stream, cancellable)

gIoStreamCloseFinish(stream, result)

gIoStreamGetInputStream(stream)

gIoStreamGetOutputStream(stream)

gIoStreamHasPending(stream)

gIoStreamIsClosed(stream)

gIoStreamSetPending(stream)

gInputStreamClearPending(stream)

gInputStreamClose(stream, cancellable)

gInputStreamCloseFinish(stream, result)

gInputStreamHasPending(stream)

gInputStreamIsClosed(stream)

gInputStreamRead(stream, count, cancellable)

gInputStreamReadAll(stream, count, cancellable)

gInputStreamReadAllFinish(stream, result)

gInputStreamReadBytes(stream, count, cancellable)

gInputStreamReadBytesFinish(stream, result)

gInputStreamReadFinish(stream, result)

gInputStreamSetPending(stream)

gInputStreamSkip(stream, count, cancellable)

gInputStreamSkipFinish(stream, result)

gOutputStreamClearPending(stream)

gOutputStreamClose(stream, cancellable)

gOutputStreamCloseFinish(stream, result)

gOutputStreamFlush(stream, cancellable)

gOutputStreamFlushFinish(stream, result)

gOutputStreamHasPending(stream)

gOutputStreamIsClosed(stream)

gOutputStreamIsClosing(stream)

gOutputStreamSetPending(stream)

gOutputStreamSplice(stream, source, flags, cancellable)

gOutputStreamSpliceFinish(stream, result)

gOutputStreamWrite(stream, buffer, count, cancellable)

gOutputStreamWriteAll(stream, buffer, count, cancellable)

gOutputStreamWriteAllFinish(stream, result)

gOutputStreamWriteBytes(stream, bytes, cancellable)

gOutputStreamWriteBytesFinish(stream, result)

gOutputStreamWriteFinish(stream, result)

gIoErrorFromErrno(err_no)

gIoErrorQuark()

gIoModulesLoadAllInDirectory(dirname)

gIoModulesLoadAllInDirectoryWithScope(dirname, scope)

gIoModulesScanAllInDirectory(dirname)

gIoModulesScanAllInDirectoryWithScope(dirname, scope)

gIoSchedulerCancelAllJobs()
}
\arguments{
\item{filename}{filename}

\item{mode}{utf8}

\item{fd}{gint}

\item{channel}{IOChannel}

\item{buf}{guint8}

\item{count}{gsize}

\item{bytes_read}{gsize}

\item{buffer}{guint8}

\item{terminator_pos}{gsize}

\item{offset}{gint64}

\item{type}{GType}

\item{size}{gsize}

\item{buffered}{gboolean}

\item{do_close}{gboolean}

\item{encoding}{utf8}

\item{flags}{OutputStreamSpliceFlags}

\item{line_term}{utf8}

\item{length}{gint}

\item{flush}{gboolean}

\item{bytes_written}{gsize}

\item{thechar}{gunichar}

\item{en}{gint}

\item{condition}{IOCondition}

\item{extension}{IOExtension}

\item{extension_point}{IOExtensionPoint}

\item{name}{utf8}

\item{extension_point_name}{utf8}

\item{extension_name}{utf8}

\item{priority}{gint}

\item{result}{AsyncResult}

\item{stream}{OutputStream}

\item{cancellable}{Cancellable}

\item{source}{InputStream}

\item{bytes}{GLib.Bytes}

\item{err_no}{gint}

\item{dirname}{filename}

\item{scope}{IOModuleScope}
}
\value{
IOChannel

IOChannel

Return value from C function

IOStatus

IOCondition

gsize

gboolean

gboolean

utf8

IOFlags

utf8

Return value from C function

IOError

IOStatus

IOStatus

IOStatus

IOStatus

IOStatus

IOChannel

IOError

IOStatus

Return value from C function

Return value from C function

Return value from C function

IOStatus

IOStatus

Return value from C function

IOStatus

gint

Return value from C function

IOError

IOStatus

IOStatus

IOChannelError

Quark

Source

utf8

gint

GType

IOExtension

GLib.List

GType

Return value from C function

IOExtension

IOExtensionPoint

IOExtensionPoint

gboolean

Return value from C function

gboolean

gboolean

InputStream

OutputStream

gboolean

gboolean

gboolean

Return value from C function

gboolean

gboolean

gboolean

gboolean

gssize

gboolean

gboolean

GLib.Bytes

GLib.Bytes

gssize

gboolean

gssize

gssize

Return value from C function

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gssize

gssize

gssize

gboolean

gboolean

gssize

gssize

gssize

IOErrorEnum

GLib.Quark

GLib.List

GLib.List

Return value from C function

Return value from C function

Return value from C function
}
\description{
gIoChannelNewFile

gIoChannelUnixNew

gIoChannelClose

gIoChannelFlush

gIoChannelGetBufferCondition

gIoChannelGetBufferSize

gIoChannelGetBuffered

gIoChannelGetCloseOnUnref

gIoChannelGetEncoding

gIoChannelGetFlags

gIoChannelGetLineTerm

gIoChannelInit

gIoChannelRead

gIoChannelReadChars

gIoChannelReadLine

gIoChannelReadLineString

gIoChannelReadToEnd

gIoChannelReadUnichar

gIoChannelRef

gIoChannelSeek

gIoChannelSeekPosition

gIoChannelSetBufferSize

gIoChannelSetBuffered

gIoChannelSetCloseOnUnref

gIoChannelSetEncoding

gIoChannelSetFlags

gIoChannelSetLineTerm

gIoChannelShutdown

gIoChannelUnixGetFd

gIoChannelUnref

gIoChannelWrite

gIoChannelWriteChars

gIoChannelWriteUnichar

gIoChannelErrorFromErrno

gIoChannelErrorQuark

gIoCreateWatch

gIoExtensionGetName

gIoExtensionGetPriority

gIoExtensionGetType

gIoExtensionPointGetExtensionByName

gIoExtensionPointGetExtensions

gIoExtensionPointGetRequiredType

gIoExtensionPointSetRequiredType

gIoExtensionPointImplement

gIoExtensionPointLookup

gIoExtensionPointRegister

gIoStreamSpliceFinish

gIoStreamClearPending

gIoStreamClose

gIoStreamCloseFinish

gIoStreamGetInputStream

gIoStreamGetOutputStream

gIoStreamHasPending

gIoStreamIsClosed

gIoStreamSetPending

gInputStreamClearPending

gInputStreamClose

gInputStreamCloseFinish

gInputStreamHasPending

gInputStreamIsClosed

gInputStreamRead

gInputStreamReadAll

gInputStreamReadAllFinish

gInputStreamReadBytes

gInputStreamReadBytesFinish

gInputStreamReadFinish

gInputStreamSetPending

gInputStreamSkip

gInputStreamSkipFinish

gOutputStreamClearPending

gOutputStreamClose

gOutputStreamCloseFinish

gOutputStreamFlush

gOutputStreamFlushFinish

gOutputStreamHasPending

gOutputStreamIsClosed

gOutputStreamIsClosing

gOutputStreamSetPending

gOutputStreamSplice

gOutputStreamSpliceFinish

gOutputStreamWrite

gOutputStreamWriteAll

gOutputStreamWriteAllFinish

gOutputStreamWriteBytes

gOutputStreamWriteBytesFinish

gOutputStreamWriteFinish

gIoErrorFromErrno

gIoErrorQuark

gIoModulesLoadAllInDirectory

gIoModulesLoadAllInDirectoryWithScope

gIoModulesScanAllInDirectory

gIoModulesScanAllInDirectoryWithScope

gIoSchedulerCancelAllJobs
}
