% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R
\name{gErrorNewLiteral}
\alias{gErrorNewLiteral}
\alias{gErrorCopy}
\alias{gErrorFree}
\alias{gErrorMatches}
\alias{gLogDefaultHandler}
\alias{gLogGetAlwaysFatal}
\alias{gLogGetDebugEnabled}
\alias{gLogRemoveHandler}
\alias{gLogSetAlwaysFatal}
\alias{gLogSetDebugEnabled}
\alias{gLogSetFatalMask}
\alias{gLogStructuredArray}
\alias{gLogVariant}
\alias{gLogWriterDefault}
\alias{gLogWriterDefaultSetDebugDomains}
\alias{gLogWriterDefaultSetUseStderr}
\alias{gLogWriterDefaultWouldDrop}
\alias{gLogWriterFormatFields}
\alias{gLogWriterIsJournald}
\alias{gLogWriterJournald}
\alias{gLogWriterStandardStreams}
\alias{gLogWriterSupportsColor}
\alias{gLogWriterSyslog}
\title{gErrorNewLiteral}
\usage{
gErrorNewLiteral(domain, code, message)

gErrorCopy(error)

gErrorFree(error)

gErrorMatches(error, domain, code)

gLogDefaultHandler(log_domain, log_level, message, unused_data)

gLogGetAlwaysFatal()

gLogGetDebugEnabled()

gLogRemoveHandler(log_domain, handler_id)

gLogSetAlwaysFatal(fatal_mask)

gLogSetDebugEnabled(enabled)

gLogSetFatalMask(log_domain, fatal_mask)

gLogStructuredArray(log_level, fields, n_fields)

gLogVariant(log_domain, log_level, fields)

gLogWriterDefault(log_level, fields, n_fields, user_data)

gLogWriterDefaultSetDebugDomains(domains)

gLogWriterDefaultSetUseStderr(use_stderr)

gLogWriterDefaultWouldDrop(log_level, log_domain)

gLogWriterFormatFields(log_level, fields, n_fields, use_color)

gLogWriterIsJournald(output_fd)

gLogWriterJournald(log_level, fields, n_fields, user_data)

gLogWriterStandardStreams(log_level, fields, n_fields, user_data)

gLogWriterSupportsColor(output_fd)

gLogWriterSyslog(log_level, fields, n_fields, user_data)
}
\arguments{
\item{domain}{Quark}

\item{code}{gint}

\item{message}{utf8}

\item{error}{Error}

\item{log_domain}{utf8}

\item{log_level}{LogLevelFlags}

\item{unused_data}{gpointer}

\item{handler_id}{guint}

\item{fatal_mask}{LogLevelFlags}

\item{enabled}{gboolean}

\item{fields}{LogField}

\item{n_fields}{gsize}

\item{user_data}{gpointer}

\item{domains}{utf8}

\item{use_stderr}{gboolean}

\item{use_color}{gboolean}

\item{output_fd}{gint}
}
\value{
Error

Error

Return value from C function

gboolean

Return value from C function

LogLevelFlags

gboolean

Return value from C function

LogLevelFlags

Return value from C function

LogLevelFlags

Return value from C function

Return value from C function

LogWriterOutput

Return value from C function

Return value from C function

gboolean

utf8

gboolean

LogWriterOutput

LogWriterOutput

gboolean

LogWriterOutput
}
\description{
gErrorNewLiteral

gErrorCopy

gErrorFree

gErrorMatches

gLogDefaultHandler

gLogGetAlwaysFatal

gLogGetDebugEnabled

gLogRemoveHandler

gLogSetAlwaysFatal

gLogSetDebugEnabled

gLogSetFatalMask

gLogStructuredArray

gLogVariant

gLogWriterDefault

gLogWriterDefaultSetDebugDomains

gLogWriterDefaultSetUseStderr

gLogWriterDefaultWouldDrop

gLogWriterFormatFields

gLogWriterIsJournald

gLogWriterJournald

gLogWriterStandardStreams

gLogWriterSupportsColor

gLogWriterSyslog
}
