% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R
\name{gHashTableAdd}
\alias{gHashTableAdd}
\alias{gHashTableContains}
\alias{gHashTableDestroy}
\alias{gHashTableInsert}
\alias{gHashTableLookup}
\alias{gHashTableLookupExtended}
\alias{gHashTableNewSimilar}
\alias{gHashTableRef}
\alias{gHashTableRemove}
\alias{gHashTableRemoveAll}
\alias{gHashTableReplace}
\alias{gHashTableSize}
\alias{gHashTableSteal}
\alias{gHashTableStealAll}
\alias{gHashTableStealExtended}
\alias{gHashTableUnref}
\alias{gHashTableIterGetHashTable}
\alias{gHashTableIterInit}
\alias{gHashTableIterNext}
\alias{gHashTableIterRemove}
\alias{gHashTableIterReplace}
\alias{gHashTableIterSteal}
\title{gHashTableAdd}
\usage{
gHashTableAdd(hash_table, key)

gHashTableContains(hash_table, key)

gHashTableDestroy(hash_table)

gHashTableInsert(hash_table, key, value)

gHashTableLookup(hash_table, key)

gHashTableLookupExtended(hash_table, lookup_key)

gHashTableNewSimilar(other_hash_table)

gHashTableRef(hash_table)

gHashTableRemove(hash_table, key)

gHashTableRemoveAll(hash_table)

gHashTableReplace(hash_table, key, value)

gHashTableSize(hash_table)

gHashTableSteal(hash_table, key)

gHashTableStealAll(hash_table)

gHashTableStealExtended(hash_table, lookup_key)

gHashTableUnref(hash_table)

gHashTableIterGetHashTable(iter)

gHashTableIterInit(iter, hash_table)

gHashTableIterNext(iter)

gHashTableIterRemove(iter)

gHashTableIterReplace(iter, value)

gHashTableIterSteal(iter)
}
\arguments{
\item{hash_table}{GLib.HashTable}

\item{key}{gpointer}

\item{value}{gpointer}

\item{lookup_key}{gpointer}

\item{other_hash_table}{GLib.HashTable}

\item{iter}{HashTableIter}
}
\value{
gboolean

gboolean

Return value from C function

gboolean

gpointer

gboolean

GLib.HashTable

GLib.HashTable

gboolean

Return value from C function

gboolean

guint

gboolean

Return value from C function

gboolean

Return value from C function

GLib.HashTable

Return value from C function

gboolean

Return value from C function

Return value from C function

Return value from C function
}
\description{
gHashTableAdd

gHashTableContains

gHashTableDestroy

gHashTableInsert

gHashTableLookup

gHashTableLookupExtended

gHashTableNewSimilar

gHashTableRef

gHashTableRemove

gHashTableRemoveAll

gHashTableReplace

gHashTableSize

gHashTableSteal

gHashTableStealAll

gHashTableStealExtended

gHashTableUnref

gHashTableIterGetHashTable

gHashTableIterInit

gHashTableIterNext

gHashTableIterRemove

gHashTableIterReplace

gHashTableIterSteal
}
