% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R, R/Gio_autogen.R
\name{gListPopAllocator}
\alias{gListPopAllocator}
\alias{gListPushAllocator}
\alias{gQueueClear}
\alias{gQueueForeach}
\alias{gQueueFree}
\alias{gQueueFreeFull}
\alias{gQueueGetLength}
\alias{gQueueIndex}
\alias{gQueueInit}
\alias{gQueueInsertSorted}
\alias{gQueueIsEmpty}
\alias{gQueuePeekHead}
\alias{gQueuePeekNth}
\alias{gQueuePeekTail}
\alias{gQueuePopHead}
\alias{gQueuePopNth}
\alias{gQueuePopTail}
\alias{gQueuePushHead}
\alias{gQueuePushNth}
\alias{gQueuePushTail}
\alias{gQueueRemove}
\alias{gQueueRemoveAll}
\alias{gQueueReverse}
\alias{gQueueSort}
\alias{gSlistPopAllocator}
\alias{gSlistPushAllocator}
\alias{gListModelGetItemType}
\alias{gListModelGetNItems}
\alias{gListModelGetObject}
\alias{gListModelItemsChanged}
\alias{gListStoreNew}
\alias{gListStoreAppend}
\alias{gListStoreInsert}
\alias{gListStoreInsertSorted}
\alias{gListStoreRemove}
\alias{gListStoreRemoveAll}
\alias{gListStoreSort}
\alias{gListStoreSplice}
\title{gListPopAllocator}
\usage{
gListPopAllocator()

gListPushAllocator(allocator)

gQueueClear(queue)

gQueueForeach(queue, func)

gQueueFree(queue)

gQueueFreeFull(queue, free_func)

gQueueGetLength(queue)

gQueueIndex(queue, data)

gQueueInit(queue)

gQueueInsertSorted(queue, data, func)

gQueueIsEmpty(queue)

gQueuePeekHead(queue)

gQueuePeekNth(queue, n)

gQueuePeekTail(queue)

gQueuePopHead(queue)

gQueuePopNth(queue, n)

gQueuePopTail(queue)

gQueuePushHead(queue, data)

gQueuePushNth(queue, data, n)

gQueuePushTail(queue, data)

gQueueRemove(queue, data)

gQueueRemoveAll(queue, data)

gQueueReverse(queue)

gQueueSort(queue, compare_func)

gSlistPopAllocator()

gSlistPushAllocator(allocator)

gListModelGetItemType(list)

gListModelGetNItems(list)

gListModelGetObject(list, position)

gListModelItemsChanged(list, position, removed, added)

gListStoreNew(item_type)

gListStoreAppend(store, item)

gListStoreInsert(store, position, item)

gListStoreInsertSorted(store, item, compare_func)

gListStoreRemove(store, position)

gListStoreRemoveAll(store)

gListStoreSort(store, compare_func)

gListStoreSplice(store, position, n_removals, additions, n_additions)
}
\arguments{
\item{allocator}{Allocator}

\item{queue}{Queue}

\item{func}{function — CompareDataFunc callback}

\item{free_func}{DestroyNotify}

\item{data}{gpointer}

\item{n}{gint}

\item{compare_func}{function — GLib.CompareDataFunc callback}

\item{list}{ListModel}

\item{position}{guint}

\item{removed}{guint}

\item{added}{guint}

\item{item_type}{GType}

\item{store}{ListStore}

\item{item}{GObject.Object}

\item{n_removals}{guint}

\item{additions}{GObject.Object}

\item{n_additions}{guint}
}
\value{
Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

guint

gint

Return value from C function

Return value from C function

gboolean

gpointer

gpointer

gpointer

gpointer

gpointer

gpointer

Return value from C function

Return value from C function

Return value from C function

gboolean

guint

Return value from C function

Return value from C function

Return value from C function

Return value from C function

GType

guint

GObject.Object

Return value from C function

ListStore

Return value from C function

Return value from C function

guint

Return value from C function

Return value from C function

Return value from C function

Return value from C function
}
\description{
gListPopAllocator

gListPushAllocator

gQueueClear

gQueueForeach

gQueueFree

gQueueFreeFull

gQueueGetLength

gQueueIndex

gQueueInit

gQueueInsertSorted

gQueueIsEmpty

gQueuePeekHead

gQueuePeekNth

gQueuePeekTail

gQueuePopHead

gQueuePopNth

gQueuePopTail

gQueuePushHead

gQueuePushNth

gQueuePushTail

gQueueRemove

gQueueRemoveAll

gQueueReverse

gQueueSort

gSlistPopAllocator

gSlistPushAllocator

gListModelGetItemType

gListModelGetNItems

gListModelGetObject

gListModelItemsChanged

gListStoreNew

gListStoreAppend

gListStoreInsert

gListStoreInsertSorted

gListStoreRemove

gListStoreRemoveAll

gListStoreSort

gListStoreSplice
}
