% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLib_autogen.R, R/GObject_autogen.R
\name{gStringNew}
\alias{gStringNew}
\alias{gStringNewLen}
\alias{gStringSizedNew}
\alias{gStringAppend}
\alias{gStringAppendC}
\alias{gStringAppendLen}
\alias{gStringAppendUnichar}
\alias{gStringAppendUriEscaped}
\alias{gStringAsciiDown}
\alias{gStringAsciiUp}
\alias{gStringAssign}
\alias{gStringDown}
\alias{gStringEqual}
\alias{gStringErase}
\alias{gStringFree}
\alias{gStringFreeToBytes}
\alias{gStringHash}
\alias{gStringInsert}
\alias{gStringInsertC}
\alias{gStringInsertLen}
\alias{gStringInsertUnichar}
\alias{gStringOverwrite}
\alias{gStringOverwriteLen}
\alias{gStringPrepend}
\alias{gStringPrependC}
\alias{gStringPrependLen}
\alias{gStringPrependUnichar}
\alias{gStringSetSize}
\alias{gStringTruncate}
\alias{gStringUp}
\alias{gStringChunkClear}
\alias{gStringChunkFree}
\alias{gStringChunkInsert}
\alias{gStringChunkInsertConst}
\alias{gStringChunkInsertLen}
\alias{gStrvBuilderAdd}
\alias{gStrvBuilderAddv}
\alias{gStrvBuilderEnd}
\alias{gStrvBuilderTake}
\alias{gStrEqual}
\alias{gStrHasPrefix}
\alias{gStrHasSuffix}
\alias{gStrHash}
\alias{gStrIsAscii}
\alias{gStrMatchString}
\alias{gStrToAscii}
\alias{gStrTokenizeAndFold}
\alias{gStripContext}
\alias{gStrrstrLen}
\alias{gStrsplitSet}
\alias{gStrstrLen}
\alias{gStrvContains}
\alias{gStrvGetType}
\alias{gStrvLength}
\alias{gStrdupValueContents}
\title{gStringNew}
\usage{
gStringNew(init)

gStringNewLen(init, len)

gStringSizedNew(dfl_size)

gStringAppend(string, val)

gStringAppendC(string, c)

gStringAppendLen(string, val, len)

gStringAppendUnichar(string, wc)

gStringAppendUriEscaped(string, unescaped, reserved_chars_allowed, allow_utf8)

gStringAsciiDown(string)

gStringAsciiUp(string)

gStringAssign(string, rval)

gStringDown(string)

gStringEqual(v, v2)

gStringErase(string, pos, len)

gStringFree(string, free_segment)

gStringFreeToBytes(string)

gStringHash(str)

gStringInsert(string, pos, val)

gStringInsertC(string, pos, c)

gStringInsertLen(string, pos, val, len)

gStringInsertUnichar(string, pos, wc)

gStringOverwrite(string, pos, val)

gStringOverwriteLen(string, pos, val, len)

gStringPrepend(string, val)

gStringPrependC(string, c)

gStringPrependLen(string, val, len)

gStringPrependUnichar(string, wc)

gStringSetSize(string, len)

gStringTruncate(string, len)

gStringUp(string)

gStringChunkClear(chunk)

gStringChunkFree(chunk)

gStringChunkInsert(chunk, string)

gStringChunkInsertConst(chunk, string)

gStringChunkInsertLen(chunk, string, len)

gStrvBuilderAdd(builder, value)

gStrvBuilderAddv(builder, value)

gStrvBuilderEnd(builder)

gStrvBuilderTake(builder, value)

gStrEqual(v1, v2)

gStrHasPrefix(str, prefix)

gStrHasSuffix(str, suffix)

gStrHash(v)

gStrIsAscii(str)

gStrMatchString(search_term, potential_hit, accept_alternates)

gStrToAscii(str, from_locale)

gStrTokenizeAndFold(string, translit_locale)

gStripContext(msgid, msgval)

gStrrstrLen(haystack, haystack_len, needle)

gStrsplitSet(string, delimiters, max_tokens)

gStrstrLen(haystack, haystack_len, needle)

gStrvContains(strv, str)

gStrvGetType()

gStrvLength(str_array)

gStrdupValueContents(value)
}
\arguments{
\item{init}{utf8}

\item{len}{gssize}

\item{dfl_size}{gsize}

\item{string}{utf8}

\item{val}{utf8}

\item{c}{gchar}

\item{wc}{gunichar}

\item{unescaped}{utf8}

\item{reserved_chars_allowed}{utf8}

\item{allow_utf8}{gboolean}

\item{rval}{utf8}

\item{v}{gpointer}

\item{v2}{gpointer}

\item{pos}{gsize}

\item{free_segment}{gboolean}

\item{str}{utf8}

\item{chunk}{StringChunk}

\item{builder}{StrvBuilder}

\item{value}{Value}

\item{v1}{gpointer}

\item{prefix}{utf8}

\item{suffix}{utf8}

\item{search_term}{utf8}

\item{potential_hit}{utf8}

\item{accept_alternates}{gboolean}

\item{from_locale}{utf8}

\item{translit_locale}{utf8}

\item{msgid}{utf8}

\item{msgval}{utf8}

\item{haystack}{utf8}

\item{haystack_len}{gssize}

\item{needle}{utf8}

\item{delimiters}{guint8}

\item{max_tokens}{gint}

\item{strv}{utf8}

\item{str_array}{utf8}
}
\value{
String

String

String

String

String

String

String

String

String

String

String

String

gboolean

String

utf8

Bytes

guint

String

String

String

String

String

String

String

String

String

String

String

String

String

Return value from C function

Return value from C function

utf8

utf8

utf8

Return value from C function

Return value from C function

utf8

Return value from C function

gboolean

gboolean

gboolean

guint

gboolean

gboolean

utf8

utf8

utf8

utf8

utf8

utf8

gboolean

GType

guint

utf8
}
\description{
gStringNew

gStringNewLen

gStringSizedNew

gStringAppend

gStringAppendC

gStringAppendLen

gStringAppendUnichar

gStringAppendUriEscaped

gStringAsciiDown

gStringAsciiUp

gStringAssign

gStringDown

gStringEqual

gStringErase

gStringFree

gStringFreeToBytes

gStringHash

gStringInsert

gStringInsertC

gStringInsertLen

gStringInsertUnichar

gStringOverwrite

gStringOverwriteLen

gStringPrepend

gStringPrependC

gStringPrependLen

gStringPrependUnichar

gStringSetSize

gStringTruncate

gStringUp

gStringChunkClear

gStringChunkFree

gStringChunkInsert

gStringChunkInsertConst

gStringChunkInsertLen

gStrvBuilderAdd

gStrvBuilderAddv

gStrvBuilderEnd

gStrvBuilderTake

gStrEqual

gStrHasPrefix

gStrHasSuffix

gStrHash

gStrIsAscii

gStrMatchString

gStrToAscii

gStrTokenizeAndFold

gStripContext

gStrrstrLen

gStrsplitSet

gStrstrLen

gStrvContains

gStrvGetType

gStrvLength

gStrdupValueContents
}
