% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gsk_autogen.R
\name{gskBlendNodeNew}
\alias{gskBlendNodeNew}
\alias{gskBlendNodeGetBlendMode}
\alias{gskBlendNodeGetBottomChild}
\alias{gskBlendNodeGetTopChild}
\alias{gskBlurNodeNew}
\alias{gskBlurNodeGetChild}
\alias{gskBlurNodeGetRadius}
\alias{gskBorderNodeNew}
\alias{gskBorderNodeGetColors}
\alias{gskBorderNodeGetOutline}
\alias{gskBorderNodeGetWidths}
\alias{gskCairoNodeNew}
\alias{gskCairoNodeGetDrawContext}
\alias{gskCairoNodeGetSurface}
\alias{gskCairoRendererNew}
\alias{gskClipNodeNew}
\alias{gskClipNodeGetChild}
\alias{gskClipNodeGetClip}
\alias{gskColorMatrixNodeNew}
\alias{gskColorMatrixNodeGetChild}
\alias{gskColorMatrixNodeGetColorMatrix}
\alias{gskColorMatrixNodeGetColorOffset}
\alias{gskColorNodeNew}
\alias{gskColorNodeGetColor}
\alias{gskComponentTransferNewDiscrete}
\alias{gskComponentTransferNewGamma}
\alias{gskComponentTransferNewIdentity}
\alias{gskComponentTransferNewLevels}
\alias{gskComponentTransferNewLinear}
\alias{gskComponentTransferNewTable}
\alias{gskComponentTransferCopy}
\alias{gskComponentTransferFree}
\alias{gskComponentTransferEqual}
\alias{gskComponentTransferNodeNew}
\alias{gskComponentTransferNodeGetChild}
\alias{gskComponentTransferNodeGetTransfer}
\alias{gskCompositeNodeNew}
\alias{gskCompositeNodeGetChild}
\alias{gskCompositeNodeGetMask}
\alias{gskCompositeNodeGetOperator}
\alias{gskConicGradientNodeNew}
\alias{gskConicGradientNodeGetAngle}
\alias{gskConicGradientNodeGetCenter}
\alias{gskConicGradientNodeGetColorStops}
\alias{gskConicGradientNodeGetNColorStops}
\alias{gskConicGradientNodeGetRotation}
\alias{gskContainerNodeNew}
\alias{gskContainerNodeGetChild}
\alias{gskContainerNodeGetNChildren}
\alias{gskCopyNodeNew}
\alias{gskCopyNodeGetChild}
\alias{gskCrossFadeNodeNew}
\alias{gskCrossFadeNodeGetEndChild}
\alias{gskCrossFadeNodeGetProgress}
\alias{gskCrossFadeNodeGetStartChild}
\alias{gskDebugNodeNew}
\alias{gskDebugNodeGetChild}
\alias{gskDebugNodeGetMessage}
\alias{gskFillNodeNew}
\alias{gskFillNodeGetChild}
\alias{gskFillNodeGetFillRule}
\alias{gskFillNodeGetPath}
\alias{gskGlRendererNew}
\alias{gskGlShaderNewFromBytes}
\alias{gskGlShaderNewFromResource}
\alias{gskGlShaderCompile}
\alias{gskGlShaderFindUniformByName}
\alias{gskGlShaderGetArgBool}
\alias{gskGlShaderGetArgFloat}
\alias{gskGlShaderGetArgInt}
\alias{gskGlShaderGetArgUint}
\alias{gskGlShaderGetArgVec2}
\alias{gskGlShaderGetArgVec3}
\alias{gskGlShaderGetArgVec4}
\alias{gskGlShaderGetArgsSize}
\alias{gskGlShaderGetNTextures}
\alias{gskGlShaderGetNUniforms}
\alias{gskGlShaderGetResource}
\alias{gskGlShaderGetSource}
\alias{gskGlShaderGetUniformName}
\alias{gskGlShaderGetUniformOffset}
\alias{gskGlShaderGetUniformType}
\alias{gskGlShaderNodeNew}
\alias{gskGlShaderNodeGetArgs}
\alias{gskGlShaderNodeGetChild}
\alias{gskGlShaderNodeGetNChildren}
\alias{gskGlShaderNodeGetShader}
\alias{gskInsetShadowNodeNew}
\alias{gskInsetShadowNodeGetBlurRadius}
\alias{gskInsetShadowNodeGetColor}
\alias{gskInsetShadowNodeGetDx}
\alias{gskInsetShadowNodeGetDy}
\alias{gskInsetShadowNodeGetOutline}
\alias{gskInsetShadowNodeGetSpread}
\alias{gskIsolationNodeNew}
\alias{gskIsolationNodeGetChild}
\alias{gskIsolationNodeGetIsolations}
\alias{gskLinearGradientNodeNew}
\alias{gskLinearGradientNodeGetColorStops}
\alias{gskLinearGradientNodeGetEnd}
\alias{gskLinearGradientNodeGetNColorStops}
\alias{gskLinearGradientNodeGetStart}
\alias{gskMaskNodeNew}
\alias{gskMaskNodeGetMask}
\alias{gskMaskNodeGetMaskMode}
\alias{gskMaskNodeGetSource}
\alias{gskNglRendererNew}
\alias{gskOpacityNodeNew}
\alias{gskOpacityNodeGetChild}
\alias{gskOpacityNodeGetOpacity}
\alias{gskOutsetShadowNodeNew}
\alias{gskOutsetShadowNodeGetBlurRadius}
\alias{gskOutsetShadowNodeGetColor}
\alias{gskOutsetShadowNodeGetDx}
\alias{gskOutsetShadowNodeGetDy}
\alias{gskOutsetShadowNodeGetOutline}
\alias{gskOutsetShadowNodeGetSpread}
\alias{gskPasteNodeNew}
\alias{gskPasteNodeGetDepth}
\alias{gskPathEqual}
\alias{gskPathForeach}
\alias{gskPathForeachIntersection}
\alias{gskPathGetBounds}
\alias{gskPathGetClosestPoint}
\alias{gskPathGetEndPoint}
\alias{gskPathGetNext}
\alias{gskPathGetPrevious}
\alias{gskPathGetStartPoint}
\alias{gskPathGetStrokeBounds}
\alias{gskPathGetTightBounds}
\alias{gskPathInFill}
\alias{gskPathIsClosed}
\alias{gskPathIsEmpty}
\alias{gskPathPrint}
\alias{gskPathRef}
\alias{gskPathToCairo}
\alias{gskPathToString}
\alias{gskPathUnref}
\alias{gskPathParse}
\alias{gskPathBuilderNew}
\alias{gskPathBuilderAddCairoPath}
\alias{gskPathBuilderAddCircle}
\alias{gskPathBuilderAddLayout}
\alias{gskPathBuilderAddPath}
\alias{gskPathBuilderAddRect}
\alias{gskPathBuilderAddReversePath}
\alias{gskPathBuilderAddRoundedRect}
\alias{gskPathBuilderAddSegment}
\alias{gskPathBuilderArcTo}
\alias{gskPathBuilderClose}
\alias{gskPathBuilderConicTo}
\alias{gskPathBuilderCubicTo}
\alias{gskPathBuilderGetCurrentPoint}
\alias{gskPathBuilderHtmlArcTo}
\alias{gskPathBuilderLineTo}
\alias{gskPathBuilderMoveTo}
\alias{gskPathBuilderQuadTo}
\alias{gskPathBuilderRef}
\alias{gskPathBuilderRelArcTo}
\alias{gskPathBuilderRelConicTo}
\alias{gskPathBuilderRelCubicTo}
\alias{gskPathBuilderRelHtmlArcTo}
\alias{gskPathBuilderRelLineTo}
\alias{gskPathBuilderRelMoveTo}
\alias{gskPathBuilderRelQuadTo}
\alias{gskPathBuilderRelSvgArcTo}
\alias{gskPathBuilderSvgArcTo}
\alias{gskPathBuilderToPath}
\alias{gskPathBuilderUnref}
\alias{gskPathMeasureNew}
\alias{gskPathMeasureNewWithTolerance}
\alias{gskPathMeasureGetLength}
\alias{gskPathMeasureGetPath}
\alias{gskPathMeasureGetPoint}
\alias{gskPathMeasureGetTolerance}
\alias{gskPathMeasureRef}
\alias{gskPathMeasureUnref}
\alias{gskPathPointCompare}
\alias{gskPathPointCopy}
\alias{gskPathPointEqual}
\alias{gskPathPointFree}
\alias{gskPathPointGetCurvature}
\alias{gskPathPointGetDistance}
\alias{gskPathPointGetPosition}
\alias{gskPathPointGetRotation}
\alias{gskPathPointGetTangent}
\alias{gskRadialGradientNodeNew}
\alias{gskRadialGradientNodeGetCenter}
\alias{gskRadialGradientNodeGetColorStops}
\alias{gskRadialGradientNodeGetEnd}
\alias{gskRadialGradientNodeGetHradius}
\alias{gskRadialGradientNodeGetNColorStops}
\alias{gskRadialGradientNodeGetStart}
\alias{gskRadialGradientNodeGetVradius}
\alias{gskRenderNodeDeserialize}
\alias{gskRenderNodeDraw}
\alias{gskRenderNodeGetBounds}
\alias{gskRenderNodeGetChildren}
\alias{gskRenderNodeGetNodeType}
\alias{gskRenderNodeGetOpaqueRect}
\alias{gskRenderNodeRef}
\alias{gskRenderNodeSerialize}
\alias{gskRenderNodeUnref}
\alias{gskRenderNodeWriteToFile}
\alias{gskRenderReplayNew}
\alias{gskRenderReplayDefault}
\alias{gskRenderReplayFilterFont}
\alias{gskRenderReplayFilterNode}
\alias{gskRenderReplayFilterTexture}
\alias{gskRenderReplayFree}
\alias{gskRenderReplaySetFontFilter}
\alias{gskRenderReplaySetNodeFilter}
\alias{gskRenderReplaySetTextureFilter}
\alias{gskRendererNewForSurface}
\alias{gskRendererGetSurface}
\alias{gskRendererIsRealized}
\alias{gskRendererRealize}
\alias{gskRendererRealizeForDisplay}
\alias{gskRendererRender}
\alias{gskRendererRenderTexture}
\alias{gskRendererUnrealize}
\alias{gskRepeatNodeNew}
\alias{gskRepeatNodeGetChild}
\alias{gskRepeatNodeGetChildBounds}
\alias{gskRepeatingLinearGradientNodeNew}
\alias{gskRepeatingRadialGradientNodeNew}
\alias{gskRoundedClipNodeNew}
\alias{gskRoundedClipNodeGetChild}
\alias{gskRoundedClipNodeGetClip}
\alias{gskRoundedRectContainsPoint}
\alias{gskRoundedRectContainsRect}
\alias{gskRoundedRectInit}
\alias{gskRoundedRectInitCopy}
\alias{gskRoundedRectInitFromRect}
\alias{gskRoundedRectIntersectsRect}
\alias{gskRoundedRectIsRectilinear}
\alias{gskRoundedRectNormalize}
\alias{gskRoundedRectOffset}
\alias{gskRoundedRectShrink}
\alias{gskSerializationErrorQuark}
\alias{gskShaderArgsBuilderNew}
\alias{gskShaderArgsBuilderRef}
\alias{gskShaderArgsBuilderSetBool}
\alias{gskShaderArgsBuilderSetFloat}
\alias{gskShaderArgsBuilderSetInt}
\alias{gskShaderArgsBuilderSetUint}
\alias{gskShaderArgsBuilderSetVec2}
\alias{gskShaderArgsBuilderSetVec3}
\alias{gskShaderArgsBuilderSetVec4}
\alias{gskShaderArgsBuilderToArgs}
\alias{gskShaderArgsBuilderUnref}
\alias{gskShadowNodeNew}
\alias{gskShadowNodeGetChild}
\alias{gskShadowNodeGetNShadows}
\alias{gskShadowNodeGetShadow}
\alias{gskStrokeNew}
\alias{gskStrokeCopy}
\alias{gskStrokeFree}
\alias{gskStrokeGetDash}
\alias{gskStrokeGetDashOffset}
\alias{gskStrokeGetLineCap}
\alias{gskStrokeGetLineJoin}
\alias{gskStrokeGetLineWidth}
\alias{gskStrokeGetMiterLimit}
\alias{gskStrokeSetDash}
\alias{gskStrokeSetDashOffset}
\alias{gskStrokeSetLineCap}
\alias{gskStrokeSetLineJoin}
\alias{gskStrokeSetLineWidth}
\alias{gskStrokeSetMiterLimit}
\alias{gskStrokeToCairo}
\alias{gskStrokeEqual}
\alias{gskStrokeNodeNew}
\alias{gskStrokeNodeGetChild}
\alias{gskStrokeNodeGetPath}
\alias{gskStrokeNodeGetStroke}
\alias{gskSubsurfaceNodeGetChild}
\alias{gskTextNodeNew}
\alias{gskTextNodeGetColor}
\alias{gskTextNodeGetFont}
\alias{gskTextNodeGetGlyphs}
\alias{gskTextNodeGetNumGlyphs}
\alias{gskTextNodeGetOffset}
\alias{gskTextNodeHasColorGlyphs}
\alias{gskTextureNodeNew}
\alias{gskTextureNodeGetTexture}
\alias{gskTextureScaleNodeNew}
\alias{gskTextureScaleNodeGetFilter}
\alias{gskTextureScaleNodeGetTexture}
\alias{gskTransformNew}
\alias{gskTransformEqual}
\alias{gskTransformGetCategory}
\alias{gskTransformInvert}
\alias{gskTransformMatrix}
\alias{gskTransformMatrix2d}
\alias{gskTransformPerspective}
\alias{gskTransformPrint}
\alias{gskTransformRef}
\alias{gskTransformRotate}
\alias{gskTransformRotate3d}
\alias{gskTransformScale}
\alias{gskTransformScale3d}
\alias{gskTransformSkew}
\alias{gskTransformTo2d}
\alias{gskTransformTo2dComponents}
\alias{gskTransformToAffine}
\alias{gskTransformToMatrix}
\alias{gskTransformToString}
\alias{gskTransformToTranslate}
\alias{gskTransformTransform}
\alias{gskTransformTransformBounds}
\alias{gskTransformTransformPoint}
\alias{gskTransformTranslate}
\alias{gskTransformTranslate3d}
\alias{gskTransformUnref}
\alias{gskTransformParse}
\alias{gskTransformNodeNew}
\alias{gskTransformNodeGetChild}
\alias{gskTransformNodeGetTransform}
\alias{gskVulkanRendererNew}
\alias{gskValueDupRenderNode}
\alias{gskValueGetRenderNode}
\alias{gskValueSetRenderNode}
\alias{gskValueTakeRenderNode}
\title{gskBlendNodeNew}
\usage{
gskBlendNodeNew(bottom, top, blend_mode)

gskBlendNodeGetBlendMode(node)

gskBlendNodeGetBottomChild(node)

gskBlendNodeGetTopChild(node)

gskBlurNodeNew(child, radius)

gskBlurNodeGetChild(node)

gskBlurNodeGetRadius(node)

gskBorderNodeNew(outline, border_width, border_color)

gskBorderNodeGetColors(node)

gskBorderNodeGetOutline(node)

gskBorderNodeGetWidths(node)

gskCairoNodeNew(bounds)

gskCairoNodeGetDrawContext(node)

gskCairoNodeGetSurface(node)

gskCairoRendererNew()

gskClipNodeNew(child, clip)

gskClipNodeGetChild(node)

gskClipNodeGetClip(node)

gskColorMatrixNodeNew(child, color_matrix, color_offset)

gskColorMatrixNodeGetChild(node)

gskColorMatrixNodeGetColorMatrix(node)

gskColorMatrixNodeGetColorOffset(node)

gskColorNodeNew(rgba, bounds)

gskColorNodeGetColor(node)

gskComponentTransferNewDiscrete(n, values)

gskComponentTransferNewGamma(amp, exp, ofs)

gskComponentTransferNewIdentity()

gskComponentTransferNewLevels(n)

gskComponentTransferNewLinear(m, b)

gskComponentTransferNewTable(n, values)

gskComponentTransferCopy(other)

gskComponentTransferFree(self)

gskComponentTransferEqual(self, other)

gskComponentTransferNodeNew(child, r, g, b, a)

gskComponentTransferNodeGetChild(node)

gskComponentTransferNodeGetTransfer(node, component)

gskCompositeNodeNew(child, mask, op)

gskCompositeNodeGetChild(node)

gskCompositeNodeGetMask(node)

gskCompositeNodeGetOperator(node)

gskConicGradientNodeNew(bounds, center, rotation, color_stops, n_color_stops)

gskConicGradientNodeGetAngle(node)

gskConicGradientNodeGetCenter(node)

gskConicGradientNodeGetColorStops(node)

gskConicGradientNodeGetNColorStops(node)

gskConicGradientNodeGetRotation(node)

gskContainerNodeNew(children, n_children)

gskContainerNodeGetChild(node, idx)

gskContainerNodeGetNChildren(node)

gskCopyNodeNew(child)

gskCopyNodeGetChild(node)

gskCrossFadeNodeNew(start, end, progress)

gskCrossFadeNodeGetEndChild(node)

gskCrossFadeNodeGetProgress(node)

gskCrossFadeNodeGetStartChild(node)

gskDebugNodeNew(child, message)

gskDebugNodeGetChild(node)

gskDebugNodeGetMessage(node)

gskFillNodeNew(child, path, fill_rule)

gskFillNodeGetChild(node)

gskFillNodeGetFillRule(node)

gskFillNodeGetPath(node)

gskGlRendererNew()

gskGlShaderNewFromBytes(sourcecode)

gskGlShaderNewFromResource(resource_path)

gskGlShaderCompile(shader, renderer)

gskGlShaderFindUniformByName(shader, name)

gskGlShaderGetArgBool(shader, args, idx)

gskGlShaderGetArgFloat(shader, args, idx)

gskGlShaderGetArgInt(shader, args, idx)

gskGlShaderGetArgUint(shader, args, idx)

gskGlShaderGetArgVec2(shader, args, idx, out_value)

gskGlShaderGetArgVec3(shader, args, idx, out_value)

gskGlShaderGetArgVec4(shader, args, idx, out_value)

gskGlShaderGetArgsSize(shader)

gskGlShaderGetNTextures(shader)

gskGlShaderGetNUniforms(shader)

gskGlShaderGetResource(shader)

gskGlShaderGetSource(shader)

gskGlShaderGetUniformName(shader, idx)

gskGlShaderGetUniformOffset(shader, idx)

gskGlShaderGetUniformType(shader, idx)

gskGlShaderNodeNew(shader, bounds, args, children, n_children)

gskGlShaderNodeGetArgs(node)

gskGlShaderNodeGetChild(node, idx)

gskGlShaderNodeGetNChildren(node)

gskGlShaderNodeGetShader(node)

gskInsetShadowNodeNew(outline, color, dx, dy, spread, blur_radius)

gskInsetShadowNodeGetBlurRadius(node)

gskInsetShadowNodeGetColor(node)

gskInsetShadowNodeGetDx(node)

gskInsetShadowNodeGetDy(node)

gskInsetShadowNodeGetOutline(node)

gskInsetShadowNodeGetSpread(node)

gskIsolationNodeNew(child, isolations)

gskIsolationNodeGetChild(node)

gskIsolationNodeGetIsolations(node)

gskLinearGradientNodeNew(bounds, start, end, color_stops, n_color_stops)

gskLinearGradientNodeGetColorStops(node)

gskLinearGradientNodeGetEnd(node)

gskLinearGradientNodeGetNColorStops(node)

gskLinearGradientNodeGetStart(node)

gskMaskNodeNew(source, mask, mask_mode)

gskMaskNodeGetMask(node)

gskMaskNodeGetMaskMode(node)

gskMaskNodeGetSource(node)

gskNglRendererNew()

gskOpacityNodeNew(child, opacity)

gskOpacityNodeGetChild(node)

gskOpacityNodeGetOpacity(node)

gskOutsetShadowNodeNew(outline, color, dx, dy, spread, blur_radius)

gskOutsetShadowNodeGetBlurRadius(node)

gskOutsetShadowNodeGetColor(node)

gskOutsetShadowNodeGetDx(node)

gskOutsetShadowNodeGetDy(node)

gskOutsetShadowNodeGetOutline(node)

gskOutsetShadowNodeGetSpread(node)

gskPasteNodeNew(bounds, depth)

gskPasteNodeGetDepth(node)

gskPathEqual(path1, path2)

gskPathForeach(self, flags, func)

gskPathForeachIntersection(path1, path2, func)

gskPathGetBounds(self)

gskPathGetClosestPoint(self, point, threshold)

gskPathGetEndPoint(self)

gskPathGetNext(self)

gskPathGetPrevious(self)

gskPathGetStartPoint(self)

gskPathGetStrokeBounds(self, stroke)

gskPathGetTightBounds(self)

gskPathInFill(self, point, fill_rule)

gskPathIsClosed(self)

gskPathIsEmpty(self)

gskPathPrint(self, string)

gskPathRef(self)

gskPathToCairo(self, cr)

gskPathToString(self)

gskPathUnref(self)

gskPathParse(string)

gskPathBuilderNew()

gskPathBuilderAddCairoPath(self, path)

gskPathBuilderAddCircle(self, center, radius)

gskPathBuilderAddLayout(self, layout)

gskPathBuilderAddPath(self, path)

gskPathBuilderAddRect(self, rect)

gskPathBuilderAddReversePath(self, path)

gskPathBuilderAddRoundedRect(self, rect)

gskPathBuilderAddSegment(self, path, start, end)

gskPathBuilderArcTo(self, x1, y1, x2, y2)

gskPathBuilderClose(self)

gskPathBuilderConicTo(self, x1, y1, x2, y2, weight)

gskPathBuilderCubicTo(self, x1, y1, x2, y2, x3, y3)

gskPathBuilderGetCurrentPoint(self)

gskPathBuilderHtmlArcTo(self, x1, y1, x2, y2, radius)

gskPathBuilderLineTo(self, x, y)

gskPathBuilderMoveTo(self, x, y)

gskPathBuilderQuadTo(self, x1, y1, x2, y2)

gskPathBuilderRef(self)

gskPathBuilderRelArcTo(self, x1, y1, x2, y2)

gskPathBuilderRelConicTo(self, x1, y1, x2, y2, weight)

gskPathBuilderRelCubicTo(self, x1, y1, x2, y2, x3, y3)

gskPathBuilderRelHtmlArcTo(self, x1, y1, x2, y2, radius)

gskPathBuilderRelLineTo(self, x, y)

gskPathBuilderRelMoveTo(self, x, y)

gskPathBuilderRelQuadTo(self, x1, y1, x2, y2)

gskPathBuilderRelSvgArcTo(
  self,
  rx,
  ry,
  x_axis_rotation,
  large_arc,
  positive_sweep,
  x,
  y
)

gskPathBuilderSvgArcTo(
  self,
  rx,
  ry,
  x_axis_rotation,
  large_arc,
  positive_sweep,
  x,
  y
)

gskPathBuilderToPath(self)

gskPathBuilderUnref(self)

gskPathMeasureNew(path)

gskPathMeasureNewWithTolerance(path, tolerance)

gskPathMeasureGetLength(self)

gskPathMeasureGetPath(self)

gskPathMeasureGetPoint(self, distance)

gskPathMeasureGetTolerance(self)

gskPathMeasureRef(self)

gskPathMeasureUnref(self)

gskPathPointCompare(point1, point2)

gskPathPointCopy(point)

gskPathPointEqual(point1, point2)

gskPathPointFree(point)

gskPathPointGetCurvature(point, path, direction)

gskPathPointGetDistance(point, measure)

gskPathPointGetPosition(point, path)

gskPathPointGetRotation(point, path, direction)

gskPathPointGetTangent(point, path, direction)

gskRadialGradientNodeNew(
  bounds,
  center,
  hradius,
  vradius,
  start,
  end,
  color_stops,
  n_color_stops
)

gskRadialGradientNodeGetCenter(node)

gskRadialGradientNodeGetColorStops(node)

gskRadialGradientNodeGetEnd(node)

gskRadialGradientNodeGetHradius(node)

gskRadialGradientNodeGetNColorStops(node)

gskRadialGradientNodeGetStart(node)

gskRadialGradientNodeGetVradius(node)

gskRenderNodeDeserialize(bytes, error_func)

gskRenderNodeDraw(node, cr)

gskRenderNodeGetBounds(node)

gskRenderNodeGetChildren(self)

gskRenderNodeGetNodeType(node)

gskRenderNodeGetOpaqueRect(self)

gskRenderNodeRef(node)

gskRenderNodeSerialize(node)

gskRenderNodeUnref(node)

gskRenderNodeWriteToFile(node, filename)

gskRenderReplayNew()

gskRenderReplayDefault(self, node)

gskRenderReplayFilterFont(self, font)

gskRenderReplayFilterNode(self, node)

gskRenderReplayFilterTexture(self, texture)

gskRenderReplayFree(self)

gskRenderReplaySetFontFilter(self, filter)

gskRenderReplaySetNodeFilter(self, filter)

gskRenderReplaySetTextureFilter(self, filter)

gskRendererNewForSurface(surface)

gskRendererGetSurface(renderer)

gskRendererIsRealized(renderer)

gskRendererRealize(renderer, surface)

gskRendererRealizeForDisplay(renderer, display)

gskRendererRender(renderer, root, region)

gskRendererRenderTexture(renderer, root, viewport)

gskRendererUnrealize(renderer)

gskRepeatNodeNew(bounds, child, child_bounds)

gskRepeatNodeGetChild(node)

gskRepeatNodeGetChildBounds(node)

gskRepeatingLinearGradientNodeNew(
  bounds,
  start,
  end,
  color_stops,
  n_color_stops
)

gskRepeatingRadialGradientNodeNew(
  bounds,
  center,
  hradius,
  vradius,
  start,
  end,
  color_stops,
  n_color_stops
)

gskRoundedClipNodeNew(child, clip)

gskRoundedClipNodeGetChild(node)

gskRoundedClipNodeGetClip(node)

gskRoundedRectContainsPoint(self, point)

gskRoundedRectContainsRect(self, rect)

gskRoundedRectInit(
  self,
  bounds,
  top_left,
  top_right,
  bottom_right,
  bottom_left
)

gskRoundedRectInitCopy(self, src)

gskRoundedRectInitFromRect(self, bounds, radius)

gskRoundedRectIntersectsRect(self, rect)

gskRoundedRectIsRectilinear(self)

gskRoundedRectNormalize(self)

gskRoundedRectOffset(self, dx, dy)

gskRoundedRectShrink(self, top, right, bottom, left)

gskSerializationErrorQuark()

gskShaderArgsBuilderNew(shader, initial_values)

gskShaderArgsBuilderRef(builder)

gskShaderArgsBuilderSetBool(builder, idx, value)

gskShaderArgsBuilderSetFloat(builder, idx, value)

gskShaderArgsBuilderSetInt(builder, idx, value)

gskShaderArgsBuilderSetUint(builder, idx, value)

gskShaderArgsBuilderSetVec2(builder, idx, value)

gskShaderArgsBuilderSetVec3(builder, idx, value)

gskShaderArgsBuilderSetVec4(builder, idx, value)

gskShaderArgsBuilderToArgs(builder)

gskShaderArgsBuilderUnref(builder)

gskShadowNodeNew(child, shadows, n_shadows)

gskShadowNodeGetChild(node)

gskShadowNodeGetNShadows(node)

gskShadowNodeGetShadow(node, i)

gskStrokeNew(line_width)

gskStrokeCopy(other)

gskStrokeFree(self)

gskStrokeGetDash(self)

gskStrokeGetDashOffset(self)

gskStrokeGetLineCap(self)

gskStrokeGetLineJoin(self)

gskStrokeGetLineWidth(self)

gskStrokeGetMiterLimit(self)

gskStrokeSetDash(self, dash, n_dash)

gskStrokeSetDashOffset(self, offset)

gskStrokeSetLineCap(self, line_cap)

gskStrokeSetLineJoin(self, line_join)

gskStrokeSetLineWidth(self, line_width)

gskStrokeSetMiterLimit(self, limit)

gskStrokeToCairo(self, cr)

gskStrokeEqual(stroke1, stroke2)

gskStrokeNodeNew(child, path, stroke)

gskStrokeNodeGetChild(node)

gskStrokeNodeGetPath(node)

gskStrokeNodeGetStroke(node)

gskSubsurfaceNodeGetChild(node)

gskTextNodeNew(font, glyphs, color, offset)

gskTextNodeGetColor(node)

gskTextNodeGetFont(node)

gskTextNodeGetGlyphs(node)

gskTextNodeGetNumGlyphs(node)

gskTextNodeGetOffset(node)

gskTextNodeHasColorGlyphs(node)

gskTextureNodeNew(texture, bounds)

gskTextureNodeGetTexture(node)

gskTextureScaleNodeNew(texture, bounds, filter)

gskTextureScaleNodeGetFilter(node)

gskTextureScaleNodeGetTexture(node)

gskTransformNew()

gskTransformEqual(first, second)

gskTransformGetCategory(self)

gskTransformInvert(self)

gskTransformMatrix(next_, matrix)

gskTransformMatrix2d(next_, xx, yx, xy, yy, dx, dy)

gskTransformPerspective(next_, depth)

gskTransformPrint(self, string)

gskTransformRef(self)

gskTransformRotate(next_, angle)

gskTransformRotate3d(next_, angle, axis)

gskTransformScale(next_, factor_x, factor_y)

gskTransformScale3d(next_, factor_x, factor_y, factor_z)

gskTransformSkew(next_, skew_x, skew_y)

gskTransformTo2d(self)

gskTransformTo2dComponents(self)

gskTransformToAffine(self)

gskTransformToMatrix(self)

gskTransformToString(self)

gskTransformToTranslate(self)

gskTransformTransform(next_, other)

gskTransformTransformBounds(self, rect)

gskTransformTransformPoint(self, point)

gskTransformTranslate(next_, point)

gskTransformTranslate3d(next_, point)

gskTransformUnref(self)

gskTransformParse(string)

gskTransformNodeNew(child, transform)

gskTransformNodeGetChild(node)

gskTransformNodeGetTransform(node)

gskVulkanRendererNew()

gskValueDupRenderNode(value)

gskValueGetRenderNode(value)

gskValueSetRenderNode(value, node)

gskValueTakeRenderNode(value, node)
}
\arguments{
\item{bottom}{gfloat}

\item{top}{gfloat}

\item{blend_mode}{BlendMode}

\item{node}{RenderNode}

\item{child}{RenderNode}

\item{radius}{gfloat}

\item{outline}{RoundedRect}

\item{border_width}{gfloat}

\item{border_color}{Gdk.RGBA}

\item{bounds}{Graphene.Rect}

\item{clip}{RoundedRect}

\item{color_matrix}{Graphene.Matrix}

\item{color_offset}{Graphene.Vec4}

\item{rgba}{Gdk.RGBA}

\item{n}{guint}

\item{values}{gfloat}

\item{amp}{gfloat}

\item{exp}{gfloat}

\item{ofs}{gfloat}

\item{m}{gfloat}

\item{b}{ComponentTransfer}

\item{other}{Transform}

\item{self}{Transform}

\item{r}{ComponentTransfer}

\item{g}{ComponentTransfer}

\item{a}{ComponentTransfer}

\item{component}{Gdk.ColorChannel}

\item{mask}{RenderNode}

\item{op}{PorterDuff}

\item{center}{Graphene.Point}

\item{rotation}{gfloat}

\item{color_stops}{ColorStop}

\item{n_color_stops}{gsize}

\item{children}{RenderNode}

\item{n_children}{guint}

\item{idx}{gint}

\item{start}{gfloat}

\item{end}{gfloat}

\item{progress}{gfloat}

\item{message}{utf8}

\item{path}{Path}

\item{fill_rule}{FillRule}

\item{sourcecode}{GLib.Bytes}

\item{resource_path}{utf8}

\item{shader}{GLShader}

\item{renderer}{Renderer}

\item{name}{utf8}

\item{args}{GLib.Bytes}

\item{out_value}{Graphene.Vec4}

\item{color}{Gdk.RGBA}

\item{dx}{gfloat}

\item{dy}{gfloat}

\item{spread}{gfloat}

\item{blur_radius}{gfloat}

\item{isolations}{Isolation}

\item{source}{RenderNode}

\item{mask_mode}{MaskMode}

\item{opacity}{gfloat}

\item{depth}{gfloat}

\item{path1}{Path}

\item{path2}{Path}

\item{flags}{PathForeachFlags}

\item{func}{function — PathIntersectionFunc callback}

\item{point}{Graphene.Point3D}

\item{threshold}{gfloat}

\item{stroke}{Stroke}

\item{string}{utf8}

\item{cr}{cairo.Context}

\item{layout}{Pango.Layout}

\item{rect}{Graphene.Rect}

\item{x1}{gfloat}

\item{y1}{gfloat}

\item{x2}{gfloat}

\item{y2}{gfloat}

\item{weight}{gfloat}

\item{x3}{gfloat}

\item{y3}{gfloat}

\item{x}{gfloat}

\item{y}{gfloat}

\item{rx}{gfloat}

\item{ry}{gfloat}

\item{x_axis_rotation}{gfloat}

\item{large_arc}{gboolean}

\item{positive_sweep}{gboolean}

\item{tolerance}{gfloat}

\item{distance}{gfloat}

\item{point1}{PathPoint}

\item{point2}{PathPoint}

\item{direction}{PathDirection}

\item{measure}{PathMeasure}

\item{hradius}{gfloat}

\item{vradius}{gfloat}

\item{bytes}{GLib.Bytes}

\item{error_func}{function — ParseErrorFunc callback}

\item{filename}{filename}

\item{font}{Pango.Font}

\item{texture}{Gdk.Texture}

\item{filter}{ScalingFilter}

\item{surface}{Gdk.Surface}

\item{display}{Gdk.Display}

\item{root}{RenderNode}

\item{region}{cairo.Region}

\item{viewport}{Graphene.Rect}

\item{child_bounds}{Graphene.Rect}

\item{top_left}{Graphene.Size}

\item{top_right}{Graphene.Size}

\item{bottom_right}{Graphene.Size}

\item{bottom_left}{Graphene.Size}

\item{src}{RoundedRect}

\item{right}{gfloat}

\item{left}{gfloat}

\item{initial_values}{GLib.Bytes}

\item{builder}{ShaderArgsBuilder}

\item{value}{GObject.Value}

\item{shadows}{Shadow}

\item{n_shadows}{gsize}

\item{i}{gsize}

\item{line_width}{gfloat}

\item{dash}{gfloat}

\item{n_dash}{gsize}

\item{offset}{Graphene.Point}

\item{line_cap}{LineCap}

\item{line_join}{LineJoin}

\item{limit}{gfloat}

\item{stroke1}{gpointer}

\item{stroke2}{gpointer}

\item{glyphs}{Pango.GlyphString}

\item{first}{Transform}

\item{second}{Transform}

\item{next_}{Transform}

\item{matrix}{Graphene.Matrix}

\item{xx}{gfloat}

\item{yx}{gfloat}

\item{xy}{gfloat}

\item{yy}{gfloat}

\item{angle}{gfloat}

\item{axis}{Graphene.Vec3}

\item{factor_x}{gfloat}

\item{factor_y}{gfloat}

\item{factor_z}{gfloat}

\item{skew_x}{gfloat}

\item{skew_y}{gfloat}

\item{transform}{Transform}
}
\value{
BlendNode

BlendMode

RenderNode

RenderNode

BlurNode

RenderNode

gfloat

BorderNode

Gdk.RGBA

RoundedRect

gfloat

CairoNode

cairo.Context

cairo.Surface

Renderer

ClipNode

RenderNode

Graphene.Rect

ColorMatrixNode

RenderNode

Graphene.Matrix

Graphene.Vec4

ColorNode

Gdk.RGBA

ComponentTransfer

ComponentTransfer

ComponentTransfer

ComponentTransfer

ComponentTransfer

ComponentTransfer

ComponentTransfer

Return value from C function

gboolean

ComponentTransferNode

RenderNode

ComponentTransfer

CompositeNode

RenderNode

RenderNode

PorterDuff

ConicGradientNode

gfloat

Graphene.Point

ColorStop

gsize

gfloat

ContainerNode

RenderNode

guint

CopyNode

RenderNode

CrossFadeNode

RenderNode

gfloat

RenderNode

DebugNode

RenderNode

utf8

FillNode

RenderNode

FillRule

Path

Renderer

GLShader

GLShader

gboolean

gint

gboolean

gfloat

gint32

guint32

Return value from C function

Return value from C function

Return value from C function

gsize

gint

gint

utf8

GLib.Bytes

utf8

gint

GLUniformType

GLShaderNode

GLib.Bytes

RenderNode

guint

GLShader

InsetShadowNode

gfloat

Gdk.RGBA

gfloat

gfloat

RoundedRect

gfloat

IsolationNode

RenderNode

Isolation

LinearGradientNode

ColorStop

Graphene.Point

gsize

Graphene.Point

MaskNode

RenderNode

MaskMode

RenderNode

Renderer

OpacityNode

RenderNode

gfloat

OutsetShadowNode

gfloat

Gdk.RGBA

gfloat

gfloat

RoundedRect

gfloat

PasteNode

gsize

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

gboolean

Return value from C function

Path

Return value from C function

utf8

Return value from C function

Path

PathBuilder

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Graphene.Point

Return value from C function

Return value from C function

Return value from C function

Return value from C function

PathBuilder

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Path

Return value from C function

PathMeasure

PathMeasure

gfloat

Path

gboolean

gfloat

PathMeasure

Return value from C function

gint

PathPoint

gboolean

Return value from C function

gfloat

gfloat

Return value from C function

gfloat

Return value from C function

RadialGradientNode

Graphene.Point

ColorStop

gfloat

gfloat

gsize

gfloat

gfloat

RenderNode

Return value from C function

Return value from C function

RenderNode

RenderNodeType

gboolean

RenderNode

GLib.Bytes

Return value from C function

gboolean

RenderReplay

RenderNode

Pango.Font

RenderNode

Gdk.Texture

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Renderer

Gdk.Surface

gboolean

gboolean

gboolean

Return value from C function

Gdk.Texture

Return value from C function

RepeatNode

RenderNode

Graphene.Rect

RepeatingLinearGradientNode

RepeatingRadialGradientNode

RoundedClipNode

RenderNode

RoundedRect

gboolean

gboolean

RoundedRect

RoundedRect

RoundedRect

gboolean

gboolean

RoundedRect

RoundedRect

RoundedRect

GLib.Quark

ShaderArgsBuilder

ShaderArgsBuilder

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

GLib.Bytes

Return value from C function

ShadowNode

RenderNode

gsize

Shadow

Stroke

Stroke

Return value from C function

gfloat

gfloat

LineCap

LineJoin

gfloat

gfloat

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

Return value from C function

gboolean

StrokeNode

RenderNode

Path

Stroke

RenderNode

TextNode

Gdk.RGBA

Pango.Font

Pango.GlyphInfo

guint

Graphene.Point

gboolean

TextureNode

Gdk.Texture

TextureScaleNode

ScalingFilter

Gdk.Texture

Transform

gboolean

TransformCategory

Transform

Transform

Transform

Transform

Return value from C function

Transform

Transform

Transform

Transform

Transform

Transform

Return value from C function

Return value from C function

Return value from C function

Return value from C function

utf8

Return value from C function

Transform

Return value from C function

Return value from C function

Transform

Transform

Return value from C function

gboolean

TransformNode

RenderNode

Transform

Renderer

RenderNode

RenderNode

Return value from C function

Return value from C function
}
\description{
gskBlendNodeNew

gskBlendNodeGetBlendMode

gskBlendNodeGetBottomChild

gskBlendNodeGetTopChild

gskBlurNodeNew

gskBlurNodeGetChild

gskBlurNodeGetRadius

gskBorderNodeNew

gskBorderNodeGetColors

gskBorderNodeGetOutline

gskBorderNodeGetWidths

gskCairoNodeNew

gskCairoNodeGetDrawContext

gskCairoNodeGetSurface

gskCairoRendererNew

gskClipNodeNew

gskClipNodeGetChild

gskClipNodeGetClip

gskColorMatrixNodeNew

gskColorMatrixNodeGetChild

gskColorMatrixNodeGetColorMatrix

gskColorMatrixNodeGetColorOffset

gskColorNodeNew

gskColorNodeGetColor

gskComponentTransferNewDiscrete

gskComponentTransferNewGamma

gskComponentTransferNewIdentity

gskComponentTransferNewLevels

gskComponentTransferNewLinear

gskComponentTransferNewTable

gskComponentTransferCopy

gskComponentTransferFree

gskComponentTransferEqual

gskComponentTransferNodeNew

gskComponentTransferNodeGetChild

gskComponentTransferNodeGetTransfer

gskCompositeNodeNew

gskCompositeNodeGetChild

gskCompositeNodeGetMask

gskCompositeNodeGetOperator

gskConicGradientNodeNew

gskConicGradientNodeGetAngle

gskConicGradientNodeGetCenter

gskConicGradientNodeGetColorStops

gskConicGradientNodeGetNColorStops

gskConicGradientNodeGetRotation

gskContainerNodeNew

gskContainerNodeGetChild

gskContainerNodeGetNChildren

gskCopyNodeNew

gskCopyNodeGetChild

gskCrossFadeNodeNew

gskCrossFadeNodeGetEndChild

gskCrossFadeNodeGetProgress

gskCrossFadeNodeGetStartChild

gskDebugNodeNew

gskDebugNodeGetChild

gskDebugNodeGetMessage

gskFillNodeNew

gskFillNodeGetChild

gskFillNodeGetFillRule

gskFillNodeGetPath

gskGlRendererNew

gskGlShaderNewFromBytes

gskGlShaderNewFromResource

gskGlShaderCompile

gskGlShaderFindUniformByName

gskGlShaderGetArgBool

gskGlShaderGetArgFloat

gskGlShaderGetArgInt

gskGlShaderGetArgUint

gskGlShaderGetArgVec2

gskGlShaderGetArgVec3

gskGlShaderGetArgVec4

gskGlShaderGetArgsSize

gskGlShaderGetNTextures

gskGlShaderGetNUniforms

gskGlShaderGetResource

gskGlShaderGetSource

gskGlShaderGetUniformName

gskGlShaderGetUniformOffset

gskGlShaderGetUniformType

gskGlShaderNodeNew

gskGlShaderNodeGetArgs

gskGlShaderNodeGetChild

gskGlShaderNodeGetNChildren

gskGlShaderNodeGetShader

gskInsetShadowNodeNew

gskInsetShadowNodeGetBlurRadius

gskInsetShadowNodeGetColor

gskInsetShadowNodeGetDx

gskInsetShadowNodeGetDy

gskInsetShadowNodeGetOutline

gskInsetShadowNodeGetSpread

gskIsolationNodeNew

gskIsolationNodeGetChild

gskIsolationNodeGetIsolations

gskLinearGradientNodeNew

gskLinearGradientNodeGetColorStops

gskLinearGradientNodeGetEnd

gskLinearGradientNodeGetNColorStops

gskLinearGradientNodeGetStart

gskMaskNodeNew

gskMaskNodeGetMask

gskMaskNodeGetMaskMode

gskMaskNodeGetSource

gskNglRendererNew

gskOpacityNodeNew

gskOpacityNodeGetChild

gskOpacityNodeGetOpacity

gskOutsetShadowNodeNew

gskOutsetShadowNodeGetBlurRadius

gskOutsetShadowNodeGetColor

gskOutsetShadowNodeGetDx

gskOutsetShadowNodeGetDy

gskOutsetShadowNodeGetOutline

gskOutsetShadowNodeGetSpread

gskPasteNodeNew

gskPasteNodeGetDepth

gskPathEqual

gskPathForeach

gskPathForeachIntersection

gskPathGetBounds

gskPathGetClosestPoint

gskPathGetEndPoint

gskPathGetNext

gskPathGetPrevious

gskPathGetStartPoint

gskPathGetStrokeBounds

gskPathGetTightBounds

gskPathInFill

gskPathIsClosed

gskPathIsEmpty

gskPathPrint

gskPathRef

gskPathToCairo

gskPathToString

gskPathUnref

gskPathParse

gskPathBuilderNew

gskPathBuilderAddCairoPath

gskPathBuilderAddCircle

gskPathBuilderAddLayout

gskPathBuilderAddPath

gskPathBuilderAddRect

gskPathBuilderAddReversePath

gskPathBuilderAddRoundedRect

gskPathBuilderAddSegment

gskPathBuilderArcTo

gskPathBuilderClose

gskPathBuilderConicTo

gskPathBuilderCubicTo

gskPathBuilderGetCurrentPoint

gskPathBuilderHtmlArcTo

gskPathBuilderLineTo

gskPathBuilderMoveTo

gskPathBuilderQuadTo

gskPathBuilderRef

gskPathBuilderRelArcTo

gskPathBuilderRelConicTo

gskPathBuilderRelCubicTo

gskPathBuilderRelHtmlArcTo

gskPathBuilderRelLineTo

gskPathBuilderRelMoveTo

gskPathBuilderRelQuadTo

gskPathBuilderRelSvgArcTo

gskPathBuilderSvgArcTo

gskPathBuilderToPath

gskPathBuilderUnref

gskPathMeasureNew

gskPathMeasureNewWithTolerance

gskPathMeasureGetLength

gskPathMeasureGetPath

gskPathMeasureGetPoint

gskPathMeasureGetTolerance

gskPathMeasureRef

gskPathMeasureUnref

gskPathPointCompare

gskPathPointCopy

gskPathPointEqual

gskPathPointFree

gskPathPointGetCurvature

gskPathPointGetDistance

gskPathPointGetPosition

gskPathPointGetRotation

gskPathPointGetTangent

gskRadialGradientNodeNew

gskRadialGradientNodeGetCenter

gskRadialGradientNodeGetColorStops

gskRadialGradientNodeGetEnd

gskRadialGradientNodeGetHradius

gskRadialGradientNodeGetNColorStops

gskRadialGradientNodeGetStart

gskRadialGradientNodeGetVradius

gskRenderNodeDeserialize

gskRenderNodeDraw

gskRenderNodeGetBounds

gskRenderNodeGetChildren

gskRenderNodeGetNodeType

gskRenderNodeGetOpaqueRect

gskRenderNodeRef

gskRenderNodeSerialize

gskRenderNodeUnref

gskRenderNodeWriteToFile

gskRenderReplayNew

gskRenderReplayDefault

gskRenderReplayFilterFont

gskRenderReplayFilterNode

gskRenderReplayFilterTexture

gskRenderReplayFree

gskRenderReplaySetFontFilter

gskRenderReplaySetNodeFilter

gskRenderReplaySetTextureFilter

gskRendererNewForSurface

gskRendererGetSurface

gskRendererIsRealized

gskRendererRealize

gskRendererRealizeForDisplay

gskRendererRender

gskRendererRenderTexture

gskRendererUnrealize

gskRepeatNodeNew

gskRepeatNodeGetChild

gskRepeatNodeGetChildBounds

gskRepeatingLinearGradientNodeNew

gskRepeatingRadialGradientNodeNew

gskRoundedClipNodeNew

gskRoundedClipNodeGetChild

gskRoundedClipNodeGetClip

gskRoundedRectContainsPoint

gskRoundedRectContainsRect

gskRoundedRectInit

gskRoundedRectInitCopy

gskRoundedRectInitFromRect

gskRoundedRectIntersectsRect

gskRoundedRectIsRectilinear

gskRoundedRectNormalize

gskRoundedRectOffset

gskRoundedRectShrink

gskSerializationErrorQuark

gskShaderArgsBuilderNew

gskShaderArgsBuilderRef

gskShaderArgsBuilderSetBool

gskShaderArgsBuilderSetFloat

gskShaderArgsBuilderSetInt

gskShaderArgsBuilderSetUint

gskShaderArgsBuilderSetVec2

gskShaderArgsBuilderSetVec3

gskShaderArgsBuilderSetVec4

gskShaderArgsBuilderToArgs

gskShaderArgsBuilderUnref

gskShadowNodeNew

gskShadowNodeGetChild

gskShadowNodeGetNShadows

gskShadowNodeGetShadow

gskStrokeNew

gskStrokeCopy

gskStrokeFree

gskStrokeGetDash

gskStrokeGetDashOffset

gskStrokeGetLineCap

gskStrokeGetLineJoin

gskStrokeGetLineWidth

gskStrokeGetMiterLimit

gskStrokeSetDash

gskStrokeSetDashOffset

gskStrokeSetLineCap

gskStrokeSetLineJoin

gskStrokeSetLineWidth

gskStrokeSetMiterLimit

gskStrokeToCairo

gskStrokeEqual

gskStrokeNodeNew

gskStrokeNodeGetChild

gskStrokeNodeGetPath

gskStrokeNodeGetStroke

gskSubsurfaceNodeGetChild

gskTextNodeNew

gskTextNodeGetColor

gskTextNodeGetFont

gskTextNodeGetGlyphs

gskTextNodeGetNumGlyphs

gskTextNodeGetOffset

gskTextNodeHasColorGlyphs

gskTextureNodeNew

gskTextureNodeGetTexture

gskTextureScaleNodeNew

gskTextureScaleNodeGetFilter

gskTextureScaleNodeGetTexture

gskTransformNew

gskTransformEqual

gskTransformGetCategory

gskTransformInvert

gskTransformMatrix

gskTransformMatrix2d

gskTransformPerspective

gskTransformPrint

gskTransformRef

gskTransformRotate

gskTransformRotate3d

gskTransformScale

gskTransformScale3d

gskTransformSkew

gskTransformTo2d

gskTransformTo2dComponents

gskTransformToAffine

gskTransformToMatrix

gskTransformToString

gskTransformToTranslate

gskTransformTransform

gskTransformTransformBounds

gskTransformTransformPoint

gskTransformTranslate

gskTransformTranslate3d

gskTransformUnref

gskTransformParse

gskTransformNodeNew

gskTransformNodeGetChild

gskTransformNodeGetTransform

gskVulkanRendererNew

gskValueDupRenderNode

gskValueGetRenderNode

gskValueSetRenderNode

gskValueTakeRenderNode
}
